Updated check kernel to search from HASH_LENGTH instead of STATE_LENGTH Moved kernels into shaders folder.
