+++ src/shaders/headers.js
@@ -1,0 +1,7 @@
+export default 
+`#define HASH_LENGTH 243
+#define STATE_LENGTH 3 * HASH_LENGTH
+#define HALF_LENGTH 364
+#define HIGH_BITS 0xFFFFFFFF
+#define LOW_BITS 0x00000000
+`
