--- src/test.js
@@ -1,61 +1,0 @@
-export default [
-	[
-		[-1227133514,-1,-1,0,-1,0,0,-1,-1,-1,0,-1,0,0,-1,0,-1,-1,-1,-1,0,-1,-1,-1,0,-1,-1,-1,-1,-1,-1,0,0,0,-1,0,-1,0,0,0,-1,-1,-1,-1,-1,0,-1,0,0,0,0,-1,-1,0,0,0,-1,-1,-1,-1,-1,0,-1,0,-1,-1,0,-1,-1,-1,-1,0,0,0,0,0,-1,0,-1,-1,-1,-1,-1,0,-1,-1,0,-1,0,-1,-1,-1,-1,-1,-1,0,-1,-1,0,0,-1,0,-1,-1,-1,0,-1,-1,-1,0,-1,-1,-1,-1,-1,0,-1,0,-1,0,-1,-1,-1,-1,-1,-1,-1,0,0,-1,0,-1,0,-1,0,-1,-1,-1,-1,-1,0,-1,-1,-1,0,0,-1,-1,-1,-1,-1,-1,0,-1,0,0,-1,-1,0,-1,-1,-1,0,0,-1,-1,0,0,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,0,-1,-1,-1,-1,-1,-1,-1,0,-1,-1,-1,0,-1,-1,-1,0,-1,-1,-1,-1,-1,-1,0,0,-1,0,0,-1,0,-1,-1,-1,0,-1,0,0,0,-1,-1,-1,-1,-1,-1,-1,-1,0,-1,-1,-1,0,-1,-1,-1,0,-1,-1,0,-1,-1,0,0,0,0,0,0,-1,-1,-1,-1,-1,0,-1,-1,0,-1,0,-1,-1,-1,-1,0,-1,-1,0,0,-1,-1,0,-1,-1,-1,-1,-1,-1,-1,0,-1,-1,0,-1,0,0,-1,0,0,0,-1,-1,-1,-1,-1,0,-1,-1,-1,0,-1,-1,-1,-1,-1,0,-1,0,-1,-1,-1,-1,0,-1,0,-1,-1,-1,-1,0,-1,-1,-1,-1,-1,-1,-1,0,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,0,0,0,0,-1,0,0,0,0,-1,0,0,-1,-1,0,-1,-1,0,-1,-1,-1,0,-1,-1,-1,0,0,0,0,-1,-1,-1,-1,0,0,0,-1,-1,0,-1,0,-1,-1,-1,0,-1,-1,-1,-1,-1,-1,-1,0,0,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,0,-1,-1,0,-1,0,-1,-1,-1,-1,-1,0,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,0,-1,-1,-1,0,-1,-1,-1,-1,-1,-1,-1,-1,0,-1,0,0,-1,-1,-1,-1,-1,0,0,-1,0,-1,-1,0,0,-1,-1,-1,0,-1,-1,0,0,-1,-1,0,-1,-1,0,-1,-1,-1,-1,0,-1,0,-1,-1,-1,-1,0,-1,-1,-1,-1,0,0,-1,0,0,-1,-1,-1,-1,-1,0,-1,0,0,-1,-1,-1,0,-1,0,0,-1,-1,0,0,-1,-1,-1,0,-1,-1,0,-1,-1,-1,-1,0,-1,-1,-1,-1,0,-1,-1,-1,0,-1,-1,-1,-1,0,0,0,-1,0,-1,0,0,0,-1,0,0,0,-1,0,-1,0,-1,-1,-1,-1,-1,-1,0,-1,0,-1,-1,0,-1,-1,0,-1,-1,-1,-1,0,-1,-1,-1,-1,-1,-1,-1,0,-1,-1,-1,-1,0,-1,-1,-1,-1,-1,0,-1,-1,-1,-1,0,0,0,-1,-1,0,-1,-1,-1,-1,-1,-1,-1,-1,0,-1,0,-1,-1,-1,-1,0,-1,-1,-1,-1,-1,-1,-1,-1,-1,0,0,0,-1,-1,-1,-1,-1,0,0,-1,0,0,0,-1,0,-1,-1,0,0,-1,0,-1,-1,-1,-1,-1,-1,-1,0,-1,0,0,-1,0,-1,-1,-1,0,-1,-1,-1,0,0,0,0,-1,0,-1,-1,-1,-1,-1,-1,0,-1,-1,0,-1,-1,-1,-1,0,-1,-1,-1,-1,0,-1,0,-1,-1,0,0,-1,-1,0,-1,-1,0,-1,-1,-1,0,0,-1,0,-1,-1,0,0,0,-1,-1,-1,0,-1,-1,4194303,4194303,2147475471,-4186113,235340673,2118066063,-1],
-		[-613566757,0,-1227133514,2118066063,-235340674,-1,-1,-1,-2143289360,-1,-4186113,-1,-1,-4194304,-4194304,-1,0,-1,0,-1,-1,-1,-1,-1,0,0,-1,-1,-1,0,-1,0,-1,-1,0,0,0,-1,-1,-1,0,-1,-1,-1,-1,-1,0,-1,-1,-1,-1,-1,-1,0,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,0,-1,0,0,0,-1,-1,-1,0,0,-1,-1,-1,-1,-1,-1,0,0,0,0,-1,-1,-1,0,0,0,0,-1,0,0,0,-1,-1,-1,0,-1,-1,-1,-1,0,0,-1,0,0,0,0,-1,0,-1,0,-1,-1,0,-1,0,-1,-1,-1,-1,-1,-1,-1,0,-1,-1,-1,0,-1,-1,-1,0,-1,0,-1,0,-1,-1,-1,0,-1,-1,-1,0,-1,-1,0,0,-1,-1,0,-1,-1,-1,0,-1,-1,-1,0,-1,0,-1,-1,-1,-1,-1,-1,-1,-1,0,-1,-1,-1,0,-1,0,-1,-1,-1,0,-1,0,-1,-1,-1,0,0,-1,-1,0,-1,-1,0,-1,-1,0,-1,0,0,-1,0,-1,-1,0,-1,-1,-1,-1,0,-1,0,0,0,-1,0,0,0,-1,-1,-1,-1,-1,-1,0,0,0,-1,-1,-1,0,0,-1,-1,-1,-1,-1,-1,0,-1,-1,-1,-1,-1,-1,0,-1,0,-1,0,-1,0,0,0,0,-1,-1,0,0,-1,0,0,-1,0,-1,-1,-1,-1,0,-1,-1,0,0,-1,-1,-1,-1,-1,-1,-1,0,0,-1,0,-1,0,0,0,0,0,0,-1,-1,0,-1,-1,-1,-1,-1,0,-1,-1,-1,0,-1,-1,0,0,-1,-1,0,0,-1,-1,-1,0,-1,-1,-1,-1,-1,-1,-1,0,-1,-1,0,-1,0,0,-1,-1,0,-1,0,0,-1,-1,-1,-1,0,-1,-1,-1,0,-1,-1,-1,-1,-1,-1,0,-1,0,-1,0,-1,-1,-1,-1,0,-1,-1,0,0,-1,-1,-1,0,-1,-1,0,-1,-1,0,-1,-1,-1,-1,0,-1,-1,0,0,-1,-1,0,-1,-1,-1,-1,0,-1,0,-1,0,-1,-1,-1,-1,-1,0,-1,0,0,-1,0,-1,-1,-1,-1,-1,-1,0,0,-1,-1,-1,-1,0,-1,-1,-1,-1,-1,0,-1,-1,0,0,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,0,-1,-1,-1,0,0,-1,-1,-1,-1,0,-1,-1,-1,-1,-1,-1,0,-1,-1,0,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,0,-1,0,-1,-1,-1,0,-1,0,-1,-1,0,-1,0,-1,-1,-1,-1,-1,0,0,-1,-1,0,-1,-1,-1,-1,-1,0,0,-1,-1,0,0,0,-1,-1,-1,-1,0,-1,-1,-1,-1,-1,0,0,-1,-1,-1,-1,-1,-1,-1,0,-1,-1,-1,-1,0,0,-1,0,-1,-1,-1,-1,-1,0,-1,0,-1,-1,0,-1,0,-1,-1,-1,-1,-1,-1,-1,0,-1,-1,0,-1,-1,0,0,-1,0,-1,-1,-1,-1,0,0,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,0,-1,-1,-1,0,-1,-1,0,0,-1,-1,0,0,-1,-1,0,-1,0,0,-1,-1,0,-1,-1,0,0,0,-1,-1,-1,0,-1,0,-1,0,-1,-1,-1,0,-1,0,-1,-1,0,0,-1,0,-1,0,-1,-1,-1,0,-1,0,-1,0,-1,-1,-1,-1,-1,-1,-1,0,-1,-1,-1,0,-1,-1,-1,-1,0,-1,-1,0,0,-1,-1,-1,-1,-1,-1,0,0,0,0,-1,0,-1,-1,-1,-1,0,0,-1,-1,-1,0,-1,0,-1,0,-1,-1,-1,0,0,0,0,-1,-1,0,-1,-1,-1,-1,-1,-1,0,-1,-1,0,0,-1,-1,0,-1,-1,-1,0,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,0,1227133513],
-		[-1227133514,1227133513,-1,-1,-1,0,-1,0,0,-1,-1,0,0,-1,-1,0,-1,-1,0,0,0,-1,0,-1,0,0,0,-1,-1,0,0,-1,0,0,0,-1,-1,0,0,0,-1,0,-1,-1,-1,0,-1,-1,-1,-1,-1,-1,0,-1,0,-1,0,0,0,-1,-1,0,0,-1,-1,-1,0,-1,0,-1,-1,-1,-1,-1,-1,0,-1,-1,-1,-1,-1,0,0,0,-1,0,0,0,-1,-1,-1,0,-1,-1,0,0,-1,0,0,0,-1,-1,-1,-1,-1,-1,0,0,-1,-1,-1,-1,-1,-1,-1,0,-1,-1,-1,-1,-1,0,-1,-1,-1,-1,0,-1,0,0,0,0,0,0,0,0,-1,-1,-1,-1,0,-1,-1,-1,-1,-1,-1,-1,0,-1,-1,-1,-1,-1,-1,-1,0,-1,-1,-1,-1,-1,-1,-1,-1,0,-1,-1,0,0,-1,-1,0,0,0,0,0,-1,-1,-1,-1,-1,-1,0,0,0,-1,-1,0,-1,-1,0,-1,-1,-1,0,0,-1,-1,0,0,-1,-1,-1,-1,0,0,-1,-1,-1,-1,0,0,-1,-1,-1,0,-1,-1,-1,-1,-1,-1,0,-1,-1,-1,-1,0,-1,-1,-1,0,-1,-1,-1,0,-1,-1,-1,-1,0,-1,-1,0,0,-1,-1,-1,0,0,0,-1,0,-1,-1,-1,0,-1,-1,-1,-1,0,-1,-1,0,-1,0,-1,-1,0,-1,-1,-1,-1,-1,-1,-1,-1,-1,0,-1,-1,-1,0,0,-1,-1,0,-1,-1,-1,0,-1,-1,-1,0,-1,0,-1,0,-1,0,-1,0,-1,0,0,-1,-1,0,-1,-1,-1,-1,-1,-1,-1,0,-1,-1,-1,-1,0,-1,-1,0,-1,-1,0,-1,-1,0,-1,-1,0,-1,-1,0,0,0,0,-1,-1,0,-1,-1,0,-1,-1,-1,-1,-1,0,0,0,-1,-1,-1,-1,-1,0,-1,-1,-1,0,-1,0,-1,-1,-1,-1,-1,0,0,-1,0,-1,-1,-1,-1,-1,0,-1,-1,-1,0,0,-1,-1,-1,0,0,-1,-1,0,-1,-1,0,0,-1,-1,-1,-1,0,-1,-1,-1,0,0,-1,-1,-1,-1,0,0,-1,-1,-1,0,-1,-1,0,-1,-1,-1,0,-1,-1,-1,-1,0,-1,-1,0,-1,-1,-1,0,-1,-1,0,-1,-1,-1,0,-1,-1,0,-1,-1,-1,-1,-1,-1,-1,0,-1,0,-1,-1,0,0,-1,-1,-1,0,-1,-1,0,-1,0,-1,0,-1,-1,-1,0,-1,-1,0,-1,-1,-1,-1,0,0,-1,-1,0,0,-1,0,-1,-1,-1,-1,-1,-1,-1,0,-1,0,0,-1,-1,-1,0,-1,-1,-1,-1,0,-1,0,-1,0,-1,0,-1,0,-1,-1,-1,0,-1,0,-1,-1,0,-1,-1,-1,-1,-1,0,-1,0,-1,-1,-1,-1,0,-1,0,-1,0,-1,-1,-1,0,-1,-1,-1,-1,-1,0,-1,-1,-1,0,-1,-1,-1,-1,-1,-1,-1,0,-1,0,-1,-1,0,0,0,-1,-1,0,0,-1,-1,-1,-1,-1,0,-1,0,0,-1,0,-1,-1,-1,-1,-1,-1,0,-1,-1,0,0,-1,0,-1,-1,0,-1,-1,0,0,0,-1,0,-1,-1,0,0,-1,0,-1,0,-1,-1,-1,-1,-1,-1,-1,0,0,0,-1,0,0,0,0,-1,-1,0,-1,-1,0,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,0,-1,-1,0,-1,-1,-1,-1,0,0,0,-1,-1,-1,0,-1,-1,0,-1,-1,-1,0,0,-1,-1,-1,-1,-1,-1,0,-1,-1,-1,-1,-1,0,-1,-1,-1,-1,0,0,-1,-1,-1,4194303,-1,4194303,-1,-1,4194303,-1,-1,4186112,2147475471,-1,-2143289360,2143289359,2147475471,-1,-2147475472,-1,-235340674,-1,-235340674,235340673,-1882725391,-2118066064,-1882725391,-1,-1,-1,-1,-613566757],
-		[1840700269,-613566757,1840700269,-1,-1,-1,0,-1,1227133513,0,1227133513,-1882725391,-235340674,-2118066064,2118066063,1882725390,-235340674,235340673,-1,-1,235340673,-1,-235340674,-1,235340673,-1,-235340674,-2147475472,2147475471,0,-2147475472,-2147475472,-4186113,-2143289360,-1,-1,-1,-1,-2143289360,-2147475472,-4186113,-1,-4186113,0,-1,-1,-1,-1,-1,-1,-4194304,-1,-4194304,-1,-4194304,0,-4194304,4194303,-1,-1,-1,0,0,0,0,-1,-1,-1,-1,-1,-1,-1,-1,-1,0,-1,0,-1,-1,-1,-1,-1,0,-1,0,0,0,-1,0,-1,-1,-1,0,-1,0,-1,-1,0,-1,-1,-1,-1,0,-1,-1,-1,-1,0,0,-1,0,-1,0,-1,-1,-1,-1,-1,-1,-1,-1,-1,0,0,0,-1,-1,-1,-1,-1,-1,0,-1,-1,0,-1,0,-1,0,0,0,-1,-1,-1,0,-1,0,0,-1,-1,-1,-1,-1,-1,0,0,0,0,0,-1,0,-1,0,-1,-1,-1,-1,0,-1,-1,0,-1,0,-1,-1,0,-1,-1,-1,0,-1,-1,0,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,0,-1,-1,0,-1,-1,-1,0,-1,-1,-1,-1,-1,-1,0,-1,0,0,-1,-1,-1,-1,-1,0,-1,-1,-1,0,0,-1,-1,-1,-1,-1,0,-1,0,-1,-1,0,0,0,-1,-1,-1,-1,-1,-1,-1,-1,0,0,-1,-1,-1,-1,0,-1,-1,0,0,-1,-1,-1,-1,-1,0,0,0,0,-1,-1,0,0,-1,-1,-1,-1,0,-1,0,0,0,0,-1,-1,0,-1,-1,-1,-1,0,-1,0,-1,0,-1,0,-1,0,0,0,-1,-1,0,0,-1,0,-1,0,0,-1,-1,-1,-1,-1,-1,0,-1,0,0,0,-1,-1,-1,-1,-1,-1,-1,-1,0,-1,0,-1,-1,-1,0,-1,-1,0,-1,-1,0,-1,-1,-1,0,0,0,0,0,0,-1,-1,-1,-1,0,0,0,-1,-1,0,-1,0,-1,-1,0,-1,-1,-1,0,-1,-1,-1,0,-1,-1,0,0,0,-1,0,0,-1,-1,-1,-1,-1,0,-1,-1,-1,0,0,-1,0,0,-1,-1,-1,-1,-1,-1,0,-1,-1,-1,-1,-1,-1,0,0,-1,-1,-1,-1,-1,-1,-1,0,0,-1,0,-1,-1,0,-1,0,-1,-1,0,-1,-1,0,0,-1,-1,0,0,-1,-1,-1,-1,0,-1,0,-1,-1,0,0,-1,0,0,-1,0,-1,0,0,-1,0,-1,-1,0,0,-1,-1,-1,-1,0,-1,-1,-1,0,-1,0,-1,0,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,0,0,-1,0,0,-1,-1,-1,-1,0,0,0,0,-1,-1,0,0,-1,-1,-1,0,0,-1,-1,-1,-1,-1,-1,0,0,-1,-1,0,-1,-1,-1,-1,-1,0,-1,0,0,-1,-1,0,-1,-1,0,-1,0,-1,-1,-1,-1,-1,-1,-1,0,-1,0,-1,-1,0,0,0,-1,-1,0,0,0,0,-1,-1,0,-1,0,-1,0,-1,0,-1,-1,-1,0,-1,-1,-1,-1,-1,0,0,-1,0,-1,0,-1,-1,0,-1,0,0,-1,0,-1,0,0,-1,-1,0,-1,0,-1,-1,0,-1,-1,-1,0,-1,-1,-1,0,0,-1,-1,-1,-1,0,-1,0,-1,-1,-1,-1,-1,-1,-1,0,-1,-1,-1,0,-1,-1,-1,0,-1,-1,0,-1,0,0,-1,0,0,-1,0,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,0,-1,0,-1,-1,-1,-1,0,-1,0,0,-1,-1,0,-1,0,-1,-1,-1,-1,-1,-1,-1,-1,-1,0,0,-1,-1,-1,0,0,-1,0,-1,0,-1,-1,0,0,0,-1,-1,0,-1,0,-1,-1,-1,-1,-1,0,0,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,0,-1,0,-1,-1227133514,-1227133514,-1227133514,-1],
-		[-1227133514,-1,1840700269,1227133513,-1,1227133513,1227133513,1227133513,1227133513,-1,-1227133514,0,-1,-1,-1,-1,-1,0,0,-1,-1,-1,0,-1,0,-1,-1,-1,0,-1,0,-1,0,0,-1,-1,0,-1,-1,-1,-1,-1,-1,-1,-1,-1,0,-1,0,-1,0,0,-1,0,-1,-1,0,-1,-1,0,0,0,-1,-1,-1,0,-1,0,-1,-1,0,-1,-1,-1,-1,-1,0,-1,-1,-1,0,-1,-1,0,-1,-1,0,0,-1,-1,-1,0,-1,0,-1,-1,0,0,-1,-1,-1,-1,-1,-1,0,0,0,-1,0,-1,0,0,0,-1,-1,-1,0,-1,-1,0,0,-1,0,-1,-1,-1,-1,-1,0,0,-1,-1,0,-1,-1,0,-1,0,0,-1,-1,-1,-1,-1,-1,-1,0,-1,-1,-1,-1,-1,0,0,-1,-1,0,0,0,-1,0,0,-1,-1,-1,-1,0,0,0,-1,0,-1,0,-1,0,-1,-1,-1,-1,-1,0,-1,0,-1,-1,-1,-1,-1,-1,-1,-1,0,-1,-1,-1,-1,-1,-1,0,-1,-1,-1,0,-1,-1,-1,-1,-1,-1,-1,0,-1,-1,0,-1,-1,0,-1,-1,0,-1,-1,-1,-1,0,0,0,0,-1,-1,-1,-1,-1,0,0,-1,-1,0,-1,-1,0,-1,-1,-1,0,0,-1,-1,-1,-1,0,0,-1,0,-1,0,-1,0,-1,-1,-1,-1,-1,-1,0,0,-1,-1,-1,-1,-1,-1,-1,-1,0,0,-1,-1,0,-1,-1,0,0,-1,-1,-1,-1,-1,-1,0,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,0,-1,-1,-1,0,0,0,-1,-1,-1,0,0,-1,-1,-1,0,-1,-1,-1,-1,-1,0,-1,-1,-1,0,-1,-1,0,-1,-1,-1,-1,-1,0,-1,0,0,-1,-1,-1,-1,-1,0,-1,0,-1,0,-1,-1,-1,-1,0,-1,0,-1,-1,-1,-1,-1,-1,0,-1,0,-1,-1,-1,0,-1,0,-1,-1,-1,0,0,0,-1,-1,-1,-1,-1,-1,-1,0,-1,-1,-1,-1,0,-1,-1,0,-1,0,-1,-1,0,0,-1,0,0,-1,-1,0,0,-1,-1,-1,-1,-1,0,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,0,0,-1,0,-1,-1,-1,0,0,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,0,0,-1,0,0,-1,-1,0,-1,-1,-1,-1,-1,-1,-1,-1,0,0,-1,-1,-1,-1,0,-1,-1,-1,-1,0,0,-1,0,-1,-1,-1,0,-1,-1,0,-1,0,-1,0,-1,-1,-1,-1,0,-1,0,-1,0,-1,0,-1,0,0,0,0,-1,0,0,0,-1,0,-1,0,-1,-1,0,0,-1,-1,-1,0,-1,-1,0,-1,-1,-1,0,0,-1,-1,-1,-1,0,0,0,0,-1,-1,-1,-1,-1,0,0,-1,-1,-1,-1,-1,-1,-1,0,-1,-1,-1,-1,-1,-1,-1,0,-1,-1,-1,-1,-1,-1,-1,0,-1,-1,-1,-1,-1,0,0,-1,-1,0,-1,-1,0,-1,0,-1,-1,-1,-1,0,-1,0,-1,-1,0,0,-1,-1,-1,0,-1,-1,-1,0,0,-1,-1,-1,-1,-1,-1,-1,-1,-1,0,0,-1,-1,-1,-4194304,4194303,-4194304,-1,4194303,-1,-1,-1,4194303,0,-4194304,-4194304,-1,-1,4194303,4194303,-1,-1,-4194304,-1,-4194304,-1,-4194304,0,0,-1,-1,0,0,0,-1,4186112,4186112,-1,4186112,-1,2147475471,-1,2147475471,-2143289360,-1,0,2143289359,0,2143289359,-1,-1,-1,-2143289360,2143289359,-1,-1,2147475471,-2147475472,2147475471,2147475471,-1,-1,-1,2147475471,-2147475472,2147475471,-1,235340673,-1,-1,-235340674,-1,-1,0,-235340674,-1,-1,-1,235340673,-235340674,-1,-1,-1,-1,235340673,-1,-1,-235340674,2118066063,-1882725391,-1882725391,-2118066064,-2118066064,2118066063,-1,235340673,-1882725391,-1,-235340674,-1,-1227133514,-1,-1,-1,-1227133514,-1,-1227133514,-1,-1,0,0,0,-1,-1,0,1840700269,-1227133514,-613566757,-1227133514,-1840700270],
-		[-613566757,1840700269,-1,-1,-613566757,613566756,-1227133514,-1,-613566757,1840700269,-613566757,-1,-1,0,0,0,0,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1227133514,1227133513,-1,1227133513,1227133513,-1,-1,0,-1,-1,0,-1,-1,1227133513,0,1227133513,-1,2118066063,0,2118066063,-1882725391,2118066063,-235340674,-235340674,-1,-2118066064,2118066063,-1,-1,-1,-1,2118066063,1882725390,-1,-1,1882725390,-2118066064,-235340674,235340673,-1882725391,0,-235340674,-1,-1,-235340674,-1,-1,235340673,0,0,0,-1,-1,-235340674,-235340674,235340673,-235340674,-1,0,-235340674,-1,235340673,-1,235340673,235340673,235340673,0,-1,0,235340673,0,0,-1,235340673,0,235340673,-1,235340673,235340673,-1,-1,2147475471,-1,-2147475472,2147475471,2147475471,2147475471,-1,-1,0,-1,-1,-1,2147475471,-1,-2147475472,2147475471,-2147475472,-2147475472,-1,-2147475472,-4186113,-1,2147475471,-1,2143289359,2143289359,-2143289360,2143289359,-1,0,2143289359,-1,0,0,0,2143289359,-2143289360,0,-1,-2143289360,-1,0,-1,-1,-2143289360,2143289359,2143289359,2147475471,-1,-1,2147475471,2147475471,-2143289360,0,-2143289360,4186112,-1,-1,-4186113,-1,-1,-4186113,-4186113,0,-1,-1,-1,0,-1,0,-1,-1,0,-1,0,-1,-1,-1,-1,-4194304,-1,-1,4194303,-4194304,4194303,0,-4194304,-4194304,4194303,0,-4194304,-1,-4194304,-1,-4194304,-4194304,-1,-1,4194303,-1,4194303,4194303,-1,4194303,4194303,-1,-1,-1,-4194304,-1,-4194304,-1,-1,0,-1,-1,-4194304,-1,-4194304,-1,4194303,-1,-4194304,-4194304,4194303,0,4194303,0,0,0,0,-1,-1,-1,-1,-1,0,-1,0,-1,-1,-1,-1,-1,-1,-1,0,-1,-1,0,0,0,0,-1,-1,-1,-1,-1,0,0,0,0,0,0,-1,0,-1,-1,-1,-1,-1,-1,-1,-1,-1,0,-1,-1,0,0,-1,0,-1,-1,-1,-1,0,0,0,-1,-1,-1,0,0,-1,-1,0,-1,-1,0,-1,-1,-1,-1,-1,0,-1,0,0,0,-1,0,-1,0,0,0,-1,-1,0,0,-1,-1,0,-1,-1,-1,0,0,-1,-1,0,-1,0,-1,-1,0,-1,-1,0,-1,-1,0,0,0,-1,0,-1,0,0,-1,-1,-1,0,-1,0,-1,-1,-1,0,0,0,-1,-1,-1,0,0,-1,0,-1,0,0,-1,-1,-1,-1,0,0,-1,0,-1,0,-1,-1,0,-1,-1,-1,-1,-1,-1,-1,0,-1,0,-1,0,0,0,0,-1,0,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,0,-1,-1,-1,-1,-1,0,0,0,0,-1,0,-1,-1,0,-1,-1,-1,0,-1,0,-1,-1,0,-1,-1,0,0,0,-1,0,-1,-1,-1,-1,-1,0,-1,-1,-1,0,-1,0,-1,-1,0,0,-1,-1,0,-1,-1,-1,0,-1,-1,-1,0,-1,-1,0,-1,-1,-1,-1,-1,-1,0,-1,-1,-1,-1,-1,-1,0,0,-1,-1,-1,0,-1,-1,-1,-1,-1,0,-1,-1,-1,0,0,-1,-1,-1,-1,-1,-1,-1,-1,0,-1,0,0,-1,0,-1,-1,-1,0,-1,0,0,0,0,-1,0,-1,-1,0,-1,-1,0,-1,-1,-1,0,-1,0,0,0,0,-1,0,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,0,-1,-1,0,-1,-1,-1,-1,-1,-1,-1,0,-1,-1,0,0,-1,0,0,-1,-1,0,0,-1,-1,0,0,-1,-1,0,-1,0,0,-1,-1,-1,-1,-1,-1,0,0,-1,-1,-1,-1,0,-1,-1,-1,-1,-1,-1,-1,-1,0,-1,-1,-1,-1,0,0,0,-1,0,-1,-1,0,0,-1,0,-1,0,-1,-1,-1,0,0,-1,0,0,-1,-1,-1,-1,0,-1,0,-1,0,0,-1,-1,-1,-1,0,-1,0,-1,-1,-1,-1,-1,-1,-1,-1,0,-1,-1,-1,-1,0,-1,0,-1,0,0,-1,-1,0,-1,-1,0,-1,-1,0,0,0,-1,-1,-1,-1,0,0,-1,-1,-1,0,-1,-1,0,-1,0,-1,-1,-1,-1,1227133513,-1,-1,-1227133514,-1227133514,1227133513,-1227133514,1227133513,-1,-1,-1227133514,1227133513,-1,-1,-1,0,1840700269,-1227133514],
-		[-1227133514,-1227133514,1840700269,-1840700270,-1,-1,-1,-1,-1840700270,-1,-1,-1,-1227133514,-1227133514,-1,1227133513,-1,-1,1227133513,-1,-1227133514,1227133513,-1227133514,-1,-1,1227133513,-1227133514,-1,1227133513,-1,-1,-1,1227133513,0,-1,1227133513,-1,-1,-1227133514,0,1227133513,0,-1227133514,-1,-1,0,-1,0,-1,-1,-1,-1,0,-1,0,0,-1,0,0,-1,-1,-1,0,-1,-1,0,-1,0,-1,-1,0,-1,0,0,0,-1,-1,-1,-1,-1,-1,0,0,-1,-1,-1,-1,-1,0,0,-1,-1,-1,-1,-1,-1,0,-1,-1,-1,-1,-1,0,-1,-1,0,0,-1,-1,-1,0,-1,-1,0,0,-1,0,-1,-1,0,-1,-1,0,0,-1,-1,0,-1,0,-1,-1,-1,-1,0,0,0,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,0,-1,-1,-1,-1,-1,-1,0,-1,-1,-1,-1,-1,0,-1,0,-1,-1,0,0,0,-1,0,0,-1,0,-1,0,-1,-1,-1,-1,-1,0,-1,0,0,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,0,-1,-1,-1,0,-1,0,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,0,0,-1,-1,-1,0,-1,0,0,0,-1,0,-1,-1,0,0,0,-1,0,-1,-1,-1,0,-1,-1,-1,-1,-1,0,0,-1,0,-1,-1,0,-1,-1,-1,-1,-1,0,-1,4194303,-1,0,-1,-1,-1,-1,-1,4194303,4194303,-1,-4194304,-4194304,-1,0,0,-1,-1,-1,-4194304,4194303,-1,0,4194303,-1,-1,-1,-1,-1,-1,0,0,-1,-1,-1,-4194304,-4194304,4194303,-1,4194303,-1,-1,-1,0,4194303,-1,-1,-4194304,0,-4194304,-4194304,-4194304,-4194304,-1,-1,-4194304,-4194304,-4194304,0,-4194304,-1,4194303,-1,4194303,4194303,4194303,-1,4194303,-1,4194303,-1,4194303,-4194304,4194303,-1,-1,4194303,-4194304,4194303,-1,4194303,4194303,0,-1,4194303,-4194304,4194303,-1,-4194304,-4194304,0,4194303,-1,4194303,-4194304,-1,-1,0,-1,-1,-1,-1,-1,0,-1,-1,-1,0,-1,-1,0,-1,-1,0,-1,-1,0,-1,0,-1,-1,-1,-1,-1,-1,-1,-1,-1,-4186113,-1,-1,-1,-1,-4186113,-1,0,-1,4186112,-1,-4186113,-1,-4186113,-4186113,-1,2143289359,2143289359,0,-1,-1,2147475471,2147475471,-2143289360,-1,-1,-1,-2147475472,0,2147475471,-1,-2147475472,2143289359,-1,2143289359,-2143289360,-2143289360,-1,-1,-2143289360,0,-1,0,-1,-1,-2143289360,-2143289360,-1,-1,-1,-1,-1,2143289359,-1,-2143289360,-1,-1,-1,-1,-1,0,-1,-1,-1,-2143289360,-2143289360,0,-1,0,2143289359,-1,-2143289360,-2143289360,2143289359,2143289359,-1,2143289359,-1,-1,2143289359,-1,-2147475472,0,2147475471,4186112,-2143289360,-2147475472,2147475471,-1,-2147475472,-2147475472,2147475471,-2147475472,2147475471,-1,-1,2147475471,-1,0,-2147475472,2147475471,-2147475472,0,-2147475472,-1,0,0,-1,-1,-1,0,0,0,2147475471,-1,-2147475472,2147475471,-1,2147475471,-2147475472,2147475471,-1,-1,-2147475472,-2147475472,-2147475472,-1,-2147475472,-1,235340673,-1,-1,-235340674,-235340674,-1,235340673,-1,-1,-1,-1,-235340674,-235340674,-1,-235340674,-1,-1,0,-1,-1,-1,0,-1,-1,0,0,-1,-1,-1,-235340674,-235340674,235340673,-1,-1,-235340674,235340673,-235340674,-1,-235340674,-1,235340673,0,-1,-1,-235340674,-1,-1,-1,-235340674,235340673,-1,235340673,-1,-1,-235340674,-1,0,0,-1,-1,-1,-1,-1,-1,-1,-1,235340673,-1,-235340674,-235340674,-1,-1,-1,-1,0,235340673,-1,-1,-1,-1882725391,2118066063,-1,-1,-1,-1882725391,-1,-1,-1,-1882725391,0,1882725390,-1,1882725390,-1882725391,-1882725391,-2118066064,-2118066064,0,-2118066064,-1,2118066063,0,2118066063,0,-2118066064,2118066063,-1,-2118066064,2118066063,0,2118066063,235340673,-1,235340673,235340673,-2118066064,-1882725391,-1882725391,-235340674,-2118066064,-2118066064,-1,-1,2118066063,-1882725391,-1,2118066063,1227133513,-1,-1,-1,-1,-1,-1,-1227133514,-1,0,0,-1,-1,-1,0,0,-1,-1,-1,0,-1,1227133513,1227133513,-1227133514,1227133513,-1,-1,-1227133514,1227133513,-1,1227133513,1227133513,-1,0,-1,-1,-1,-1,-1,-1,0,0,-1,-1,0,0,-1,-1,-1,0,0,0,0,-1,0,-1,-1,-1,-1,-1,-1,0,-1,-1,-1,1840700269,1840700269,-613566757,-1,-1227133514,-1,-613566757,-1227133514,1840700269,-613566757,-613566757,-613566757,1840700269,-1,613566756,-1,-1840700270,-1,-1840700270,613566756],
-		[1840700269,-613566757,-1,-1,1840700269,0,-1840700270,1840700269,1840700269,0,1840700269,-613566757,-613566757,0,613566756,-1,-1227133514,-613566757,1840700269,615364315,-2995932,-612368385,-2995932,-3595118,-1227133514,1226534326,1840101083,-2995932,-1,-1,-615364316,1228931071,1840101083,4194303,1840700269,-615364316,1838902710,-3595118,-614765130,-1837105153,-614765130,-1,1840700269,-1,-1,-1,4194303,-4194304,-1,-4194304,0,-4194304,4194303,-4194304,-4194304,4194303,4194303,-1,0,-1,0,-1,-1,-4194304,-4194304,4194303,-1,-1,-1,4194303,-1,-4194304,-1,4194303,-4194304,-4194304,0,4194303,-4194304,-4194304,-1,-1,-1,-4194304,-1,4194303,4194303,-1,-1,-4194304,-1,4194303,0,-1,-4194304,4194303,4194303,4194303,-4194304,-1,0,4194303,0,4194303,-1,4194303,4194303,-1224736769,-2396746,-1224736769,-1224736769,-1,-1227133514,-1,-1224736769,1228931071,1227133513,-4194304,1227133513,-4194304,1226534326,-1797559,-1,-1,-1,1227133513,-1,-1227133514,-1227133514,-1,1226534326,-1,-1,-4194304,4194303,-1,-4194304,4194303,-4194304,-1,-1,4194303,-1,4194303,-1,4194303,0,-1,0,-1,-1,0,-1,0,0,-1227133514,1227133513,-1,1227133513,0,-1227133514,0,1226534326,-4194304,0,-1,-1,-4194304,1226534326,-2396746,-2396746,2118123519,-2118066064,-4194304,2118066063,1879565297,-235340674,-2113986673,-235340674,-1,0,-1,0,2113986672,2118123519,2118123519,2118123519,-234881025,-1882725391,-1879565298,-234938481,-3677199,2117663870,-57457,2118066063,238615678,-235340674,235340673,-235340674,-4194304,-235340674,238615678,-235340674,2118066063,-2118066064,-1,-1,2113986672,-4136848,-57457,-1,-1,2118066063,-1,-2118066064,-2118066064,2118066063,-1,-1,-2118066064,-1,-1,-1,-2118066064,-2118066064,-1,2118066063,2118066063,2118066063,0,-1,-1,2118066063,-2118066064,-1,-1,-1,-1,1882725390,1882725390,-1,0,-1882725391,-1882725391,-1882725391,-1,-1,-1882725391,-1,-1,-1882725391,0,2118066063,-1,2118066063,-1882725391,-235340674,-1,2118066063,0,235340673,-1,235340673,-1,-1882725391,-1,2118066063,-1,-1,0,0,-1,235340673,235340673,-1,-1,-1,-1,0,-1,0,-1,0,-1,235340673,-235340674,-1,-235340674,235340673,-1,-235340674,-1,-1,-1,-235340674,-1,-235340674,0,-1,-1,-1,0,0,0,0,0,-1,0,-1,0,-1,0,-1,-1,0,-1,235340673,-1,235340673,-1,235340673,-235340674,-235340674,0,-235340674,235340673,-1,-1,235340673,0,-235340674,0,235340673,0,235340673,-235340674,235340673,-1,235340673,-1,-1,0,-1,235340673,-1,0,235340673,235340673,235340673,-1,-235340674,-235340674,-1,-235340674,-1,235340673,235340673,-1,235340673,0,-235340674,235340673,-235340674,-235340674,235340673,-1,-1,-1,235340673,-1,0,-1,-1,-1,-1,-1,-1,0,-1,0,-1,0,-1,-1,-235340674,0,0,-1,-1,0,-1,-1,0,-1,-1,-235340674,-1,-1,235340673,-1,235340673,-1,235340673,0,0,0,0,-1,-235340674,0,235340673,-1,-235340674,0,235340673,-235340674,-1,-235340674,235340673,-1,235340673,0,0,-1,-235340674,0,235340673,-2147475472,2147475471,-1,2147475471,2147475471,-1,-2147475472,2147475471,2147475471,2147475471,-1,-2147475472,-1,2147475471,-2147475472,-2147475472,2147475471,2147475471,-2147475472,-1,-1,-2147475472,2147475471,-2147475472,-2147475472,-1,0,-2147475472,-2147475472,-2147475472,0,-1,-1,-1,-1,-1,-1,0,-1,-1,0,0,-1,-1,0,-1,-1,-1,-2147475472,-1,-1,-1,2147475471,2147475471,-2147475472,-2147475472,-1,-1,-1,-1,0,-2147475472,-2147475472,-1,-1,2147475471,0,-1,-2147475472,-2147475472,-1,2147475471,-2147475472,-2147475472,-1,2147475471,-1,2147475471,-1,2147475471,-2147475472,-2147475472,-1,-1,-2143289360,2147475471,4186112,-4186113,-2147475472,-2147475472,-1,-1,-1,2147475471,0,-2147475472,2143289359,-1,-1,-2143289360,0,-2143289360,2143289359,-1,-1,-2143289360,2143289359,-2143289360,2143289359,-2143289360,-1,-1,-2143289360,-1,0,-2143289360,2143289359,-2143289360,0,-1,-1,0,-1,-1,-1,2143289359,-1,2143289359,0,-1,-1,0,-1,-1,-1,-1,0,0,-1,0,-1,-1,0,0,0,2143289359,2143289359,2143289359,-1,-2143289360,2143289359,-2143289360,-1,-1,-1,0,-1,-1,-1,0,-1,-2143289360,-1,-1,2143289359,2143289359,0,-2143289360,-1,-1,-1,-1,-1,0,0,-1,2143289359,2143289359,-1,2143289359,0,2143289359,2143289359,2143289359,2143289359,2143289359,-1,-2143289360,-1,-2143289360,-2143289360,-1,-1,2147475471,-1,-2147475472,-2147475472,-1,-1,-1,2147475471,-1,-1,-2147475472,-1,-2147475472,0,2147475471,-1,-4186113,-2147475472,-4186113,-1,-4186113,0,2147475471,-1,-1227133514,-1,-1,-1,-918552583,-2396737,-1,-1,1226535497,-1794049,4186112,-1226535498,-1,-1227133514,1226535497,-2392065,4186112,0,4186112,-1,-1794049,-1,4186112,-1227133514,-1,1227133513,-1,1226535497,-1224741450,-1,4186112,-1,1226535497,-1,-1226535498,-1227133514,4186112,1226535497,-1,-1,-1227133514,-1,1227133513,-1227133514,1227133513,1227133513,-1,-1227133514,-1227133514,-1227133514,0,-1227133514,0,1227133513,1227133513,-1227133514,-1,-1227133514,-1,1227133513,1227133513,1227133513,-1,1227133513,0,0,-1,0,-1,1840700269,0,-1840700270,0,0,-1,-1,0,-1840700270,1840700269,-1,-1,-1227133514,1227133513,-613566757,-1227133514],
-		[-680806950,-134480386,-579946661,781667237,1739839980,1327993801,-134480386,-134480386,1807080173,235340673,1227133513,-1,-1672599789,-33620097,-1840700270,-1672599789,1874320365,-235340674,-1,-235340674,-1,235340673,-235340674,-1,-1,235340673,-1,1874320365,-33620097,-1,-1,-1672599789,-201720578,-1,-1672599789,-1,-1,-1,-1,-1,-1,-235340674,-1,-134480386,-1,-1,-1227133514,-1,-1,-1227133514,-1227133514,-1227133514,-134480386,-235340674,1227133513,-1193513417,-100860289,-1,-1193513417,-1092653129,1327993801,-134480386,-1227133514,-1,-134480386,-235340674,-1,-1,-100860289,235340673,-1,-1,-100860289,-100860289,1227133513,1193513416,-1,-235340674,-1193513417,-100860289,-134480386,-1092653129,-1,-1,1327993801,-1092653129,-134480386,235340673,-1193513417,-134480386,1227133513,235340673,-235340674,-235340674,4186112,-1193046473,-134414850,-234881922,-2392065,1193513416,-1794049,-1,1228927561,0,-1226535498,1226535497,-1,-1,-4186113,4186112,-1,-1,1226535497,1224741449,-4186113,-1,-1794049,-1,1227133513,-1,-1,-1,-1227133514,-1,-1,-1,-1794049,-1,1226535497,-1794049,-2392065,0,-1,-4186113,-1,-1,-1,239068033,-4186113,237208449,1195176392,-1092120649,-1226535498,-1193513417,1228927561,-235340674,1226535497,-1193046473,1327461321,-3727361,-2392065,-236741506,1193243080,-100925825,-1092120649,-1,4186112,235340673,-1908867983,237474753,954700743,-134677378,-1224736778,235340673,0,-1,-1,0,-1227133514,-1,-1,0,-1,-1,-2147475472,0,-1,-4186113,2147475471,-1,-4186113,-238609282,1912595342,-235339778,-897,-234881922,-2143289360,235340673,4186112,-1,-1,-1,-1,-2147475472,-1,0,-29425537,-1882719119,2147475471,2118066063,264764417,-235340674,2147475471,1912136590,-235339778,-235340674,0,-235340674,235340673,-235340674,2147475471,-1,-8065,-2118058000,-1,2118066063,-1912136591,-29424641,-1882718223,-235340674,2118066063,-2118066064,-2143289360,-33496961,-4136833,-4136833,2143289359,-1,-267918338,-264241154,-1,-1882725391,-1,-267918338,-1879048207,2146966543,-2143289360,2143289359,-2143289360,-2143289360,-264241154,-264241154,-1879048207,-1,-1,2146966543,-2143289360,-1,2143289359,-267918338,-264241154,-1879048207,-1879048207,-1,0,-1882725391,-1,-1882725391,-1,1882725390,0,2118066063,-2118066064,-1,-2118066064,-1,-1,2118066063,-1,-2143289360,-1,-1,-1,-4136833,-2113929232,-2113929232,-2113929232,-1,-2143289360,2118066063,-2118066064,-29360129,2147426191,2118066063,-2143289360,-1,-1,0,0,2118066063,2118066063,2118066063,0,-1,-1,0,2118066063,-2118066064,2118066063,-1,-2118066064,-29360129,-1,33496960,-29360129,-4136833,2143289359,2118066063,2143289359,-33496961,-1,2147483632,-2143289345,33554431,-29360144,-2113929217,-1,4194303,-1,0,-1,-1,2118066063,-2118066064,-235340674,235340673,-1,-3734655,-235340674,-3734655,-1,-234881025,-1,-3734655,-235340674,-1,235340673,-1,238615678,-3734655,-2118066064,-2113929217,-1,2113986672,2117663870,-238558223,3677198,238615678,-1,-33037314,-264241167,-238558223,267918337,-29417586,-234881025,2114446321,2118123519,-57457,2118123519,-2113986673,-2118123520,-4136848,-1,-2113929217,0,-1,-1,-1,-1,-1,-1,2143289359,-234881026,-33037313,2143346815,-3677199,-1,-1879105649,264758270,-57458,-4136833,-4136848,2147483632,-2113929217,-2118066064,-29417585,2147483647,-57472,-2143289351,-918552583,-2396737,2147483641,-918552586,-7,2147483647,1228931071,2143289359,4194303,-2143289360,2143289359,-1,-1,2147483647,4194303,-1,-7,2143289359,-1,-920949319,-1,2143289359,-1,-918552583,-1227128842,2147475471,1227133513,-1,-4673,-920346631,-1,-2147475472,-1,0,-1,-1,2147475471,2147475471,-1,2147475471,-4186113,4186112,4186112,-4186113,-1,-4186113,-2143289360,2147475471,-1,-2143297521,-4194304,-1,-2143297521,2147483647,-8177,2147475471,-2143297521,-2143289345,-4194304,-2147475472,2143297520,-1,4194303,2143297520,2147483647,-2143289345,-1,-4194304,-2143289345,-1,-2143297521,-8177,-4186128,-2143289345,4194303,2147475471,-2143297521,-1,-2143297521,-3505,-1224736769,-1,-1227133514,-1,2147480143,-1227128842,1227133513,2147475471,1227133513,920351302,-4673,2147475471,920351302,-1,0,-1,-1227133514,1797558,-1224736769,1226534326,-1226534327,2143297520,1228931071,-920351303,-2392074,-8177,-918552577,-4673,-2392074,-1228922896,-1226535495,-922738704,1226534326,0,-1227133514,-1227133514,1227133513,2147475471,-920948154,-1,-2396746,1224736768,1226534326,-2396746,1228931071,-1224736769,-4194304,-1,-4194304,4194303,-4194304,0,-4194304,-1,-1,-1,-1,4194303,-1,-1,-1,-1,-4194304,4194303,4194303,4194303,-1,4194303,-4194304,4194303,-4194304,-4194304,-1,-1,0,-1,-1,2147483647,-4186128,2147475471,2143297520,-8177,-1,-1,4194303,-2147475472,-2143297521,-8177,-4186128,2147483647,-4186128,-1,-2143297521,2147483647,-1,2147475471,-4194304,-2147475472,-2143297521,2147475471,2143297520,-2143289345,2147483647,-2143289345,2147483647,-8177,-4186128,-1,-1,-4186128,4194303,-2143297521,2147483647,-8177,-1,-2147475472,-2143297521,-4186128,2147483647,-2147475472,2147475471,-4186128,-8177,-1,-4194304,2147475471,2143297520,-2147475472,2143297520,-2143289345,-1,-2143297521,-2143289345,-4186128,-2143289345,-1,238615678,-235340674,-1,-1,-1,235340673,-1,235340673,-1,-1,-4194304,-1,-459650,-234881025,-1,-238615679,239075327,-459650,-234881025,239075327,-1,238615678,239075327,-1,-1,-234881025,-1,-1,-238615679,239075327,-1,238615678,-459650,-235340674,0,-1,235340673,-1,-1,1672599788,-1,-1,-1837704339,-1,1840700269,-1,-1797559,-1,-1124738323,-234946689,-612967571,-1671196216,-680142099,-67437186,1840101083,1840700269,-204527725,-459650,-599187,1672131930,-1225335955,-1194916991,-1091446548,-547927131,-1198373,0,-613566757,-1,-613566757,-1,615364315,1840101083,-614765130,1226534326,-1198373,-1,1840101083,3595117,-1225335955,-2995932,-612368385,-1,615364315,-1,717094655,783267803,-615364316,1672599788,-614765130,-613566757,1840700269,-1,-613566757,0,1227133513,-1,-613566757,0,-1,-1,613566756,613566756,613566756,-1840700270,-1,-1,-1,-1840700270,-1840700270,1840700269,-1840700270,-1840700270,1840700269,-1,-1,-1840700270,-201720578,-235340674,1840700269,235340673,-613566757,714427045,613566756,-1672599789],
-		[1840700269,-1,2008406765,2123540820,-546326565,1419070117,-545604258,-1518379089,-1344844891,-1,-67437186,-25756803,-226886402,-238558223,1874279325,-1916345487,495855379,-302630115,-302063761,-1879565298,234881024,264758270,-29417586,-470681347,-1611055981,-59245570,1874780015,-1614880914,-37748739,533645166,-28317698,503201555,-1,328083219,-34087955,-234938481,-3677199,-3677199,238558222,2117663870,-328083220,-472922141,-301998097,-604020817,-613566757,-1519561045,-9586981,2125289131,-8405025,2118066063,615167322,238615678,238615678,-3734655,-1,714427045,781667237,-235340674,-613566757,-1193513417,-1092653129,-1,-1,-713360002,-1227133514,-3734655,-714230053,-1636377747,-1227133514,716027611,1839099703,716963326,-1125206181,1871652755,-1126273225,-713360002,-167903489,781667237,-615364316,-67240193,-546654886,235340673,1518395504,2126495711,-11343500,-1,-2995932,1522474895,615364315,-2995932,615364315,-1,1228931071,3595117,-3595118,-2995932,-1227133514,-4194304,330151643,-1222997,-1221141658,937135567,-1209157925,-23360058,-306783379,2146260644,-1225335962,-1533892257,-612368385,-2147426177,626174299,637493156,-1530920972,-2113929232,1521309355,-1511131404,-1530920972,-2113929232,-1,33513380,-627947228,-538422235,-256000294,-15800449,-137484525,1187505150,1341783294,938883869,1570614999,-25206865,-581079369,2117483293,1880090611,-2117606415,-1882208285,2118115311,499048303,-302580867,-1837105153,-1,-1224736769,67437185,1739569115,680142098,1838902710,-1,-337711241,-1513495310,1955588083,-1883119376,1588672668,-1107952899,-681677001,-1,-1840101084,-1,1838902710,0,1838902710,-1840700270,-613566762,-1,925718187,-302580878,-309730062,-33496961,1816732818,-4136833,-327795939,-25215073,-58835169,-2118066064,1908867982,-4136833,235340673,267918337,-459649,-29360129,2113929231,2147426191,-1,-33037313,-3677185,-29360130,-238615679,238615678,-234881025,-1,238615678,-1912602609,-238615666,-3734641,-459650,0,2113986672,-2118066064,4194303,-1883185040,238615678,-2118066064,-1882725391,2118066063,235340673,1879105648,-517106,-57457,-3677199,-2118066064,-238615679,-1,-239017872,-1,239075327,1883185039,-2117606415,-235398130,-234881025,2113986672,2117663870,1882725390,238615678,-1,-459650,1883185039,-1882782833,-264241154,-1883242482,-264758257,268435455,-3677185,264758270,267918337,2143289359,-29360129,-235340674,2143289359,1908867982,2143749007,-459649,-264241154,235340673,-1,-1,1882725390,1879048206,1912545166,264700801,-1879105649,-238615679,267918337,8191,2143289359,2147483632,2147475471,4186127,-2143289345,2147483632,-8177,2143813616,-1879048207,-264249343,-2,-523279,-1879049201,-2143297521,-264764418,-1009,-264242161,-8177,-1879571471,-1879048194,2143806449,-1879048193,267918337,2143297520,-2143297536,2147483632,-16,-2143289457,33554304,-4128769,2118123407,-8065,33554416,-2147475457,2143354767,-29368193,-113,-4128784,-2118058000,-7281,2118066063,-267919359,-29424753,2118066175,-1882718335,-238615679,-458866,2143354864,-2113937281,-4186128,-33496976,33554431,-8065,-8065,2147483535,2147475471,2143289344,-8177,-2147475472,1912143870,235347953,-238609295,-234881025,2143755377,-1912142975,-2143297521,-1,-3728271,-1908415601,-234889201,235340673,-238608399,-1908408321,2143297520,-235339778,2147482751,-33030271,-1879048193,-29417473,-2113986561,-2118058000,4194303,-29417473,-1882725503,2117657598,-4128784,264765313,2143297520,-1,-8177,8176,-2143289345,2147475471,-8177,-8177,2147483647,-2143297521,-4186128,-1,2147483647,2143297520,-16,-4186113,-238615679,-234881922,2147483632,-3734641,-1908408322,235347967,-2143296625,-235339778,-1908867074,2143748993,-234888305,-1908409231,-238615666,2147017615,-4186113,239068033,-1,-4186113,-1,-2143289360,4186112,-2143289360,4194303,-4194304,-1,-1,-2143289345,2147483647,-1,-2143297521,-1,-4194304,-4194304,-1,-4194304,1226534326,-4194304,4194303,-2396746,-1797559,-1,-1,-1,-1,-4194304,-1,0,-1,-235340674,-235340674,1228931062,2046360697,-3734655,-1191182345,954700743,-818151943,237474753,-102797377,-1912142962,-234881026,238615678,-2,-1,-3727361,-1,-235347071,1094713343,1228733950,-262529,-1092849208,1195376639,-1191449033,-237214847,1329590199,1228931071,-1192780674,-466944,-1191451775,-1228661121,-1192780674,-235080759,1193050111,-1794049,-1090716162,-1090519041,-1094246401,4190793,-136077826,955835278,-953708081,2147483641,2046555087,-2013061122,-237011842,-134221890,-1191449033,1327993801,-196609,1228927561,-234881922,-1,-102990338,-1224736769,-922738704,922742207,-1226535495,-918556081,-1224736784,-918556081,4186121,-920351303,-4673,-2145083399,-1,-1227137009,-1802225,920952831,1794057,4186121,-1224736778,2145088070,-2395575,-1226538993,2147483632,-1224744945,-1228926391,2147478975,-4194304,-1,2396745,1228931071,-1797559,-1797559,0,-1,-1227133514,-1227133514,1227133513,-1,-1227133514,2147475471,-2147475472,-918557255,922743366,-920346631,-1798721,920949318,-2143289360,-1,-1224741450,-2396737,-4673,-1227128842,-4186113,-4673,-1226530826,-918557255,-1,-920944647,4186112,-1,-1226535498,1227133513,-1226535498,4186112,-1224736769,-8192,1226539001,-922738704,-1,2143297526,-1226535498,2143297520,922742201,2145091568,-918557255,-1193041928,-102993993,-2400241,1912398721,2143296625,-1859599,-2009079281,-1092119993,953974776,2011135481,2143297535,-458753,1908874353,2147476353,2147023998,239075327,-234881025,-1,1226534326,-136351817,-3924495,-819489847,-954438089,-952110641,-234881921,-134414721,-3924495,1327993801,1054904248,-818351154,-1325405121,-134484418,2147483647,1226534326,-2143289345,-920949322,-2133945,2044960376,2143559665,-819491271,-2042822657,-134487154,954441720,-953970232,-238609295,-820022329,2143297520,1193513416,-1193045560,-1191182337,1228931071,-136348169,2147287039,1193513416,2046615951,-953970760,-4186113,920949318,2147475471,-4186113,1912595342,-1,-104653185,1327199177,-1,4186112,-4186113,4186112,-458753,-3727361,-134218242,1054831175,-1,-1227133514,920351302,2147480143,-134479874,2046616463,-1327988746,-134480386,-1092653129,1327993801,-1193513417,-1,-1,-1193513417,-100860289,920351686,2147480015,-134480386,2147475471,-1092653129,-1092649865,1327989641,-1054826504,-1193513417,-1227133514,-1,-1,-920346631,-1,1227133513,2147480143,2145686095,-1224736778,-1,-1227133514,-918552583,-2143289360,-1227133514,920949318,-2143289360,1055167047,-100860289,-2143289360,1908867982,-234881026,235340679,-1226534321,235340673,-1224736769,4194303,-1,-2147483648,-4194289,235340673,2143289359,-1,-2143289360,1908867982,-1669333005,1945623534,2143289359,-1670260500,-3800306,511105891,-1669332996,-67642387,-1228931057,-1912142975,-922746865,235340679,886730047,-1433458790,2146950123,-1638979689,-2145086897,237212097,-1910008889,-134677378,920949327,-1908867976,-10,235398142,-267918337,2147483647,-1912085506,-235340674,2147483647,-1882725391,2118123519,-1879048193,2118123519,1912142961,-267975793,-3677199,-1849737313,-37100289,478142444,1672607997,1651636467,1844902783,-231080719,-296761373,1920400242,1672599788,-1879105649,1838073699,-923034162,-1224769610,-1879048208,-134480386,469425507,-1226936522,-33299969,-343005411,-2009071631,-1092653129,-820483015,-134480386,1739839980,-1126273225,781667237,-579946661,579946660,-1159893321,-134480386,-579946661,-680806950],
-		[-1227000105,1600896500,-308383945,-1396769537,2077444772,2043085185,748047013,1055363785,1567872476,-2728075,-33622181,-201722667,2144816036,-235951195,-202787618,1127873783,1533916892,-1737177697,-580012197,238878719,-1092316599,1056963711,956298183,1094709175,-918815234,-2009079793,2011205567,-1803557089,-134882578,-2114183680,-1677836525,-920547127,-575809059,-465568401,-48385092,2147221502,167772145,-268427265,-130350465,-922705927,-14970241,926654462,-14879817,-309403661,-25740387,2143354752,2109138173,-6369,-1610668161,-230088707,-305149969,-296757229,498531486,-1912610033,1844889375,-63314690,-229647124,310319247,-1613824004,-305136509,-8302,-37798786,-5113709,-2080965888,-272703635,-4186016,2118246240,-29367409,-1908465665,-3677297,-1879507841,33546352,-2,-264305665,-234945537,-64513,2143748097,-57457,-235397234,2118059007,1912601615,-1908408321,-33037327,-234881025,2143756273,-466817,2147483647,-459650,-2146960370,-29417487,2147023887,-58256,-234882033,-1908867073,-1908867975,-1912136585,2143756287,2145091568,-918552583,2013265023,-1193315391,-461874,1912339961,-1910803889,-200770,-266306,-2392586,-234888161,1639242219,-533525,-1467007874,2147477407,1467080567,-1260019728,-340805675,-234881025,-100860424,-1224736784,-134217730,-235340674,2013265022,-918552704,1329594366,2050441117,-728881378,-4194420,1938343023,1644117919,-306583651,-29367425,-528875523,-26132500,-477167617,-113,-473956355,-36755471,-306249869,-272630657,1674566556,-264757249,-1879507855,-1879507071,2143813631,-235340674,1882726385,-1879105538,-2113929232,-2113937281,-234881026,-33488897,268435441,-29425537,2147483647,2147426303,4194191,29425648,-4186240,1241513871,-918618000,-234881935,-134218626,2147478976,1341914615,-29883393,-134414855,268435329,-33094657,-1882725391,-1908409231,33554304,-3734529,-29614465,235340798,-1798664,-234881031,-2143297521,2145086911,-920948145,-4673,2145091583,1227137023,-2143297521,2145160783,-116916226,1994691142,-115569080,2145907705,-264758257,1879048193,2030566409,1338506249,-2097730,-920944647,-264757263,-264245249,-807178248,-577,-264764418,958172743,1071644663,2145086902,-3670031,1879049200,1882719217,523278,-957874753,-1,924611017,1206361599,1071611902,1073734535,2132770942,-1191411714,-822567809,-953969720,1054834680,2147483128,-4186113,-1092575240,2147220977,-1908605313,-1,-134479874,-920346634,-819487105,-100860289,921146310,-1191182345,-3640,-16,1222348233,-7170,1226534326,1325137407,-805330993,-14618503,-1311216586,-13042561,2146999375,-1879245199,1059558343,-805306375,1340603913,-7169,-134480258,-2017205264,-1912136079,-2147480144,-1193513033,-100860417,-807105095,1071353287,-920150407,-29360129,1226600905,-1090978313,-952303624,-1908867471,-920612935,-234881026,3727360,2143749007,-33087489,1912136590,-1,-264241154,268434433,-264248322,-3670017,-264241153,523264,1090296887,-152597066,-163904897,-1086328890,-1882725378,-2114395249,-234881039,-1879049103,2143297521,268434559,-912,-268369921,-807144969,-920683015,-2113993743,922743366,-953906119,-3513,-137945103,-1222926346,-807407625,-1076068361,-2113986673,-958173624,-100687985,956035006,-825760783,-2585530,1061158911,-941329863,-16998833,-805339201,-1076040073,-1092850121,-235404290,-952110641,1323790281,-1326973895,-113,-818351154,-954503754,2046619583,-818086785,-939561393,2118123407,-1912275471,-163898354,-104333951,1323299447,2010903030,-264248321,-1056969649,-1879072769,-1224741441,-897,1190957055,-1176243720,1996264447,-29819778,2145677881,-16801841,-1092119607,-234938481,-1189375434,-100720689,-820216329,-62385,-14450744,2130706383,-1075864521,-1073774641,-817892401,-1,-1191383617,-234888305,1329328718,2013068926,-7808,-458767,-3989505,2046820350,-102461511,-102797383,956301311,-1054904393,2044489271,-204401,1056960447,-135818295,-1191382577,-1090781698,-152792503,264758271,-234938495,33496961,264758271,-1879565297,-238615679,2114446334,-1908465777,-33489807,-1879507841,-1879048194,-29818882,-2114387984,2147482736,1879515134,-1879049202,-165157889,-1090785345,-134414337,-134247353,-817913911,1193050111,-399,2144903160,2118090639,1918762143,1674566642,-25173010,-205667471,-1617707777,-1813119006,238854142,1786732453,795589354,-675807249,-941821889,637485019,2146270031,-25773283,-617769265,-625005089,-907191457,-16783020,-1224741441,2143894370,1535112921,922144187,-612370723,1840700269,1839500729,-614162439,-3505,-1,-1224740282,-1794049,-1794049,614165942,-615360805,-1196033,-4186113,615957650,1772057289,-34481940,-150999626,1586486558,1604319965,-1353725981,1879179023,2126438303,-3145729,-277693665,938903878,-604620003,-583372943,1063482366,2145686095,579145534,1535115257,-1158760613,1636915035,-679478547,-1736443292,-1159824009,-851513543,-67436674,-1228922881,1837704347,-1224736778,-4682,-612368385,-3595118,-1842504846,-29409390,-1728353,-412371499,-14974977,-1645908641,-1189087259,-147014657,-17079298,2130238943,-114287173,-805947041,938302294,-238027549,2147476239,-1078479507,-1218715658,364494046,-574770321,-353643030,2127735124,-347084813,-1226870857,-1105293497,-113467831,-101016247,1228266221,-1086332948,15800448,167763948,-1,793692013,1765506487,-257953794,1817931190,-1209141505,-1210331667,-11934465,-1,33496960,2118123504,-33554417,1535115263,-11379857,-1512957109,1536342815,1837105078,-1219313665,-14974992,2125313759,-1839497217,1539913435,1845460987,-625564417,-1817326090,-22183962,-1774087,-41043,1530920975,2146257234,-2965514,1533927417,-598017,1535063741,-2365100,-1227687045,-612368398,2143940054,2143264761,-1217562949,-17979093,-922789057,-40842,-906593413,-923944096,1221197723,922712927,-925123724,-598026,-19730972,-1228927553,-612368385,2145689599,-307381397,2143890877,-2337,2147481311,922147693,-919152931,-3588110,2147480143,-1228331886,-1797559,1327267326,1739839323,-612368385,-2799324,716027867,-546460763,-615363146,-1075548169,-11409036,1194043155,826130143,-237011842,-957790365,-9561604,-738238481,968604759,-244159755,-295174265,-134217841,-4679,-201588737,-546390795,-1530920963,-1260221121,-407708822,-1159889035,922744543,-545792389,-579210022,1023337231,-615037165,-131073,1804017663,1161753417,-747653157,1737971675,-1669998228,-34093293,1737969335,582737920,-171100325,-35156810,-613296678,-545792642,-715960521,-136350775,1807144795,-645925414,1638510263,1739971838,-69771265,-1191182337,1673199359,1329590199,-613501093,-1226662145,-69370113,-1193511991,-2400256,-1092379522,1224938057,-3640,-1863224,-1193771009,1327996414,1672990391,-167838210,-1327533111,1873786843,-1225538029,-713163009,-2733916,-546325122,-3728258,235340673,-234889216,-234881025,239074430,-3734655,-234881922,-234881025,-714230053,-615167323,-238558223,-1412241035,2138618961,735439610,-328331,-571278988,-143271195,714788206,-33915586,1591143286,-204593378,-1113928340,-673386882,-429000973,1594303385,328083219,-1527186306,-1648804707,-558375961,-674306181,-1212153841,-1812513891,-57457,-3677199,239017871,3677198,3677198,-1,-1,238615678,34087954,1669981426,-327155713,-25559810,4194303,-328083220,-499237101,-470164243,-28317698,503201555,535902353,-268730951,1264283062,535450919,696180591,-235324257,-4120432,-61872142,-58720257,-676702939,-151142693,-61946003,1430257663,500098931,-33496975,-1879507841,268435455,-29360129,-234881025,-234881025,-1882725391,1908932593,2143371248,531708060,-1841945363,-1819679502,-1651623709,-268837650,2143740785,1941519518,1874279325,-260514706,235398129,532668399,-226492417,1840659229,-3603326,231729150,-335217,-1879186559,-4186113,200998257,2144818271,-176489090,-624992352,1518730590,549446305,-1410377813,1420470949,-75827878,-235946882,-1512939777,1891625685,-536914177,-2038964493,-305404417,239074417,443930979,2110334879],
-		[1567346154,2117531774,-956375322,-92610845,-307316770,-1088426505,-67707922,1427606473,1052588716,-949239948,-40127697,-1730351945,833482725,-817685670,-1350672598,1693171441,-88081555,-1091316791,-1624059977,-1092183187,1726988031,-544327444,-76163105,1727511511,1991375607,-275025050,-546349729,-1751544681,-1518444625,-1518541858,-1522343169,1971315342,804104875,-2010333185,-43078737,-100196879,-188411185,-1329328193,2051006734,-826204546,1933311231,-1597576,-200860289,1254880831,-210165949,-236351686,867629949,1570167029,-539369482,2103963593,-306824403,-546877447,-286999495,1073516543,-234931105,-1062928455,1609525243,-1192249381,-1714436125,1062928118,-1126173849,1325268479,-37797985,-551881994,862959579,-610403017,-285751491,324167988,-25796643,-1814315034,-302392418,299322277,-456206493,-1176535609,-641804641,-956330369,-818187266,-1188823112,-235397234,-952373314,-1094708672,-920882233,2118098943,-153095753,838340529,1060893630,-29884415,1338309696,-102240264,-1090552241,-1040244850,-80577076,-96863601,1459297918,-339082110,234219507,-1480721934,1912323758,1034868540,-1477855371,2061352955,2103243191,-1345469203,1038838731,401750237,-819481895,-994123921,-1092234391,1589629823,-121159823,-677154343,-554845834,-73552932,-815988793,-172566695,-249851337,2115348206,-104251521,-32506001,789053149,-491915522,-143131656,-1548560027,-739814486,-569941058,987935060,499499758,-713589827,-673206313,-1193308217,-3266797,-807243464,328641365,614953913,864149744,-1149329585,-1090633889,-147892757,-525315,-1220776833,-102264903,-1088425481,1882725390,-14160391,1075551238,2033970622,-1575943,824217992,-1580039,2147475983,-101130114,-1191182833,4136511,920940025,954526841,-1109656610,-49295750,-922280455,-352321547,-13173058,1819761125,-1226057809,996816154,-75583716,1666186103,-1342234640,-75546735,-1669937249,-310904579,-255377431,-1409352474,-1769998593,-33923156,1982009342,-1146391117,1335545582,1391982231,-29893377,-843982611,-251781895,-1283900243,1958838137,2071300085,1843854808,-705873155,-1080101011,-1651579621,-13669430,1992279711,-334849,1511385850,-539879685,-725657681,1891589201,-1533544450,-1343554187,-168823158,-33030146,-1371923185,-1344976213,-168839205,-627943718,-1430263506,-1879049103,267968513,2147016719,-1908867087,-1908408321,-234881922,-3727361,-4186113,2147476367,1054903352,1054834175,2143756287,-102265409,-134479874,-1193316937,-1090985079,-680014145,1672600301,-1225801985,-1782710273,501117126,-1363945507,939060135,2079160829,-378570181,-1233585035,908063743,1874317238,-819753415,714299227,1466941305,-1126596615,-3505,-1092918769,-235077634,1999106105,2046599103,1228729911,-29421623,-2592769,826220486,-1224937481,-1103371266,-807443577,-1090321906,-2115723265,-4551,-806904201,-50073153,-308024057,-130026506,1803416520,-621824037,1721461413,1342177279,-170511125,-488640521,-1663230642,-1339755912,1866866558,-1156292994,-546391187,-17976631,-613697627,-695346047,-25725459,-1213958602,-135681361,747985263,-546456851,1325604855,-1670730049,1464803599,1055832555,-275026119,548192255,-273958429,2012199661,-67305488,-1497565349,-1090783525,-545592923,717222351,-3260417,1873592185,-1157958182,2077178861,1533917679,-1706214673,-201331021,-645924897,-67179338,-1837111252,-273028441,-69311729,-1094112073,-713490953,-1090915640,-100733001,461860406,-1118835804,1840167916,1205829767,-610943494,-1780088925,-219168777,-563126662,767423956,-958362961,-313172619,-96537790,-12852850,-386412874,1565341141,-1912995977,-610436381,2069888588,-613270890,-441349,-225184962,-1879049201,2114388478,-1879048625,2110059087,-2009079681,-1386511362,-1484787079,-1785175047,-932256582,-1196562,-151030850,-949027861,-270008743,1873393491,-1181911052,1402595231,920566639,1073674396,266826288,-819091475,-808981346,-18090628,-1209215658,-17474579,1181579971,-1381097541,-1187277834,1889507531,-688238594,901773019,-1342177282,570054394,-1344803851,-168300889,-958804122,-574898469,1457746359,1993221277,-1453592579,-1134581084,-1879110585,-234909135,-1868360,-2145681418,-1196046,1228333055,-8183,-1225932814,-919150604,-1169,2144491227,-1196046,2146282939,-920951649,-2143290513,-1227732700,2145689599,-1911996428,1328390158,-2349,1671132903,987188949,-134676977,-329520981,1409023374,-1531383361,926940751,-310378545,-13963649,-708879681,2146557838,-1671613061,-147103873,938833375,-1174864722,-817912655,-150019074,-1073861777,954972830,-1208746753,-1043333875,939234494,-115019191,-908918787,-822388993,-147396644,2028648357,266065884,-827683905,2125896518,-1228865649,-1208584506,-1224084874,-1837705493,629137535,934737599,-12584177,1228909988,-2418956,4183727,-612374195,-307365013,-638977,-624957132,1542324105,-1342212441,-267206769,-30542161,268413660,-1433225555,1585963023,-318767118,-4169821,-2118066049,-31658369,-264249369,1902837737,737851069,-1613237746,1869854527,-1363715199,-303309998,-469793910,-1636718290,-170928770,-1351103570,-39878770,1877923547,-1880102192,-707334804,-21775261,1492571487,-623010833,-906008104,2135500764,-610114727,-1237590913,-184594976,1550509898,2137913076,-1663916804,1202711988,-545269613,-1516467908,-473964417,-1610612739,-1610670850,-2118116209,-1611063310,-7931777,-229688082,-1613017687,-1623597842,1928829899,-570950671,-1214761,978988415,1312488819,-1082261781,-538593571,-539371678,796376027,-17064503,-1074283540,1995515775,-239895882,-1912135695,558950526,-33086493,-17745773,-1106319615,-15800449,1237336047,-238213865,1053768551,-1209092262,-1908826930,532118463,-1233716354,1191146430,1606117983,-269254977,-1125144406,2134545774,-1010704517,-688022226,-680135745,-1552984145,-1281610338,1964896117,-537151748,1920135100,267558906,-337219607,964677177,1793490585,-831210661,1808616338,1801983,-833405443,2146934781,-1879192740,-1074386145,-824984081,808922937,-1312033093,2133847550,1204808415,-1075334162,-909644184,2128146870,-1208164455,-249050177,-1175495659,-1221006446,2146147802,-33886248,-421547088,1910465414,-620909223,2045749794,-50992278,-1910160400,1851719651,2113994735,332789999,-501769229,1842504943,-923873,1840774143,-465921,1840368293,-2146962505,-2147478967,-1191239799,1227066825,-1672487020,-1837171865,-1224744946,-578815132,-3866495,-29556239,-1275191473,1054970574,-169443703,-680810772,-294118411,-403038309,-1446969630,783431475,-296228613,-1091633242,-988681356,1599557054,-713490434,-1400187729,-591725375,-33620281,2145123760,-1384415298,-1075512193,-725680839,583167745,1054708174,-3588113,1021172971,-134768696,-623018307,515555184,2058343932,1908997134,-633747603,-234996609,-238993424,-503389953,-1879113729,1056956319,1593833502,-1301732897,-438153302,2100946687,-479157521,-1795215379,502988511,1920532353,1672456191,-1947309571,-545276106,985538429,-537313283,-205920397,-2143293013,-304218243,1613919688,-810258473,-54658267,668908331,-613566625,-1642365713,-528486074,-233456385,1883238696,-411489419,-279982083,1945561070,-1529882593,-1615249699,994643389,1069541350,1073707151,-299172390,-1839498522,230038781,-3595022,-691404894,-1228288538,-1149294209,-316948659,-1211181577,-1227962073,2122268431,-2131243809,2129099897,-908708134,-907577133,-2127112513,-2131333655,-19749664,939524070,659420861,804126128,1742465405,-1879054702,-743606541,-135698754,-1218455881,-742072357,-616964482,1592159583,-1191202827,1003860924,-1452479201,1719138302,-591536259,-1210126466,-1245777570,-272514562,-35160547,-32251393,-295992833,199910685,-369270893,1878522352,-59031634,-1986036641,1975442687,-137314317,-1116340579,-20541735,-1343064219,-1074040194,-1210281865,-822018055,-1222115833,-920611896,1195111375,1094447031,-100930122,-102501761,888039287,1671126447,-647431175,-275022997,1568135525,-34415109,2078708047,-1107907748,-841098009,1589641003,-2642949,-29427877,-1879050283,-1398161541,-2084500561,1586978687,-3733329,-760234764,-237550837,762967263,1937036755,-1213285525,2132767687,-675516994,603961262,-67256481,2009129599,-1484914818,-167985829,-1497498113,-1350632194,2002103252,2108158966,-1150429715,-21250565,-488777057,-19204738],
-		[-421659785,-1077776433,733301637,1977285567,-13421071,1537957723,-436286698,-558034189,-560829513,-382083217,-2535509,-1350590641,-104899025,-67667018,-541885635,-33603937,-538532350,-295855656,-1971109201,1567487531,1589294213,-23724587,-121635343,1522382331,388956144,-890373315,-596051105,-273812187,1752954435,-805601476,-1628962952,-135434327,-109744387,-621551386,-573838803,-1174934147,-542765130,267056574,-277168643,1520398271,760569691,268291763,1656354676,-1077945493,-1108562744,-1292468846,-8388907,-40112673,1935101900,-114311173,1972367291,-12722224,-738764562,-318841233,-1233422497,2138737403,-1283980329,-656937044,-90030150,-577182738,-226789465,2060790974,-1893859662,-33755159,-1637914994,-1525448751,-23384165,934657917,469028319,1853333055,1861746350,-24662625,-1116816653,-626131633,-613024002,-135987322,-1090544200,-67166209,-822084017,-214375937,-1075342850,-973537343,-18571722,-82078729,939487095,493702595,-16862726,1294974774,-83899601,938334189,2109468618,-39237125,-729415723,-34406401,-22180133,2002777274,-147099201,-469835841,1135306082,2004759897,527385327,-924336525,-1906347091,1207827965,1995764650,-1695553586,1136510591,2074048468,-1168402721,-690487365,-675880724,670543285,-56762377,1060828460,-1812070749,1847473629,1273888506,-1385177233,1243578875,-71827588,-92356893,-690367017,1590537889,-353093797,-119033605,-445071875,1073145238,1982436991,-2348619,-974196801,1607687148,1854918476,-830476965,-1019281412,-551493765,623894526,-142763103,922156541,786327511,-37562369,-1405403474,-1781948453,-659984936,1872734767,-1162430133,-46645281,1820302607,-38547653,1899970429,-1130967041,-655968593,-2055747617,-774216849,-941142411,-575432229,-52999490,-319257379,-730598098,729775039,-142844361,2141844183,1962931834,2145332729,1632041342,-4542483,2102868972,2117848045,-599528785,-1088494594,-577417389,2138714919,534235134,-1276388356,-546080672,-42745895,1484625911,-604389714,-910334404,-1611870278,-1414406403,-170993745,-476065441,-637648965,-1107296470,1811311709,749983667,1287117695,-738482532,-93720641,202855772,-5980769,-285233719,503177007,395173874,-832051333,-369447042,-379404359,-270730369,-436208933,-214544585,-68685854,1989275606,2143119331,-268702261,-871769955,-1745096317,-1695401937,1341874111,-1712326754,2071913201,-1879200033,1597496847,501996370,778686461,-1075163207,-1242180533,1743634419,-1712876194,-1080927747,800344423,-67143690,-539005441,594437627,667745583,1072674616,-679536642,-1611800961,-218333372,1610035166,1044344234,2144337748,-863627786,-1938916945,2132536893,-1207960046,1275002357,-412614778,828693880,-72594947,-974315637,2106848998,1802492862,-1090528198,-825955265,-1615393793,-1707106664,-414333133,1675882935,2143868438,-1076634129,1577050034,-969181881,-355050715,-604422169,-1075986825,-1901122457,1610080234,1509752305,-1123618835,-1090793487,-480741126,1869593854,-152092650,-105656881,1054039860,-168842324,-1342255409,-1193477650,-1626738775,-763693717,1742143378,-270053378,-1636317191,-607657057,-1923421481,1593673727,-511749889,-285296786,-168893458,-85467653,207219383,2045718054,-953518123,-421294147,125239195,1802131283,-336920577,1230790600,-578956033,-556845794,1072545783,733592044,1704977328,1537678967,1472003193,-1597826,-614142385,-1125806961,-614628994,1453159152,-1552084162,-612500833,-604001287,-307785145,987192169,-408977409,-339806507,-986833029,989795819,822083496,-507947781,-528946949,1371453437,-548546637,-640820514,2009136493,2004866495,2009588719,-1715602844,-511180812,-137894806,-124297218,1840675095,1977705471,-1910648842,-540264485,-643891395,-1116869835,-805832261,714599285,737148847,1441445727,1610612734,1576925947,1794646015,-369230081,-274137103,1048317815,-270154081,-696961301,2012623739,2137733529,-1210289218,-1622083586,1604052470,-1353533197,2147483574,2070801692,2147155834,-403775531,-612420262,1440030458,1693916234,-1186643989,-1830201512,-335685827,-1908408582,-92586244,-1078984802,-1042198797,1039728108,2076114179,-269760677,-1170638081,-369123563,-711001365,-1682186338,-1082218050,-201327553,1872675674,1734303159,-9442370,-963414049,-1216385061,-1355874569,-1276533989,-1896948615,-4350595,1568669647,-1116785359,1062024367,1551366764,-817890755,-69206623,2079839743,-562173193,1867838463,-303923297,517846268,-587464935,1210526839,792325854,-311480467,-678832373,1992137119,1671003132,-864493953,1825568495,-1074757766,-609226129,-336552425,-1124979288,-21500090,1860033255,-808527922,-692584467,2146361023,-344146349,1823342332,-1817205121,2128770743,2076817749,-1796540133,2104179771,971824890,939323071,-7226,-822251059,-38078609,-895217793,1940908015,-683674729,-176308865,-111427970,-285346833,-327714,929026839,1207263739,-62950273,1062207427,-1081249433,-421576715,1071544627,1144880895,-605071893,-16777609,1291583225,-1570779609,1587461338,-287993273,-1131533188,-1075416071,-1108380943,-1107741945,2109700805,-1881146377,-1580204327,-873738389,-279742593,-1489240881,454993261,-389572704,-212535827,-33830993,-1078538509,-1932150058,-305593149,1293811445,-372332717,1191182335,1944515447,-545374433,1548871927,-10896282,-344203624,-278047313,255320060,-1161826617,-146816993,-834937362,250707911,2147188220,1643206911,-1168060305,1803542525,717940034,-436470276,-75858627,-17844315,-404914449,2046356430,-67191715,-1889667639,-320996003,-305137491,2112747231,-679778721,1558828271,-845254858,1070827983,-661064323,-218509889,-466178849,-235106414,-1080136841,2008545205,-825521676,-402925291,-1208653860,-537428001,-575800648,-1168147217,-181212758,1994600312,-1437604098,-310747393,-810779211,1807644365,-136893453,2147417831,1735195359,389792238,1303562197,-280548114,-1292169224,-914653481,2059330461,995717561,2134896507,-170205193,-374015023,-77734106,1672186058,-1029841034,1073545089,-654836742,2054389483,-84701618,-490801156,-1792608547,-117712001,2145340127,1736276987,-70014225,-1212458129,-1627425927,-409657355,-554201347,-1075858689,-99951492,1304802286,2125987837,-161620306,-1145052207,-1275143809,-705204358,2119605879,-1781666117,790787418,-30742663,-1351929937,1198042612,762142684,519934247,988388255,-235202625,2014919983,-1578106881,-579749522,-202933121,779270901,-544212898,-220268707,-1959284865,-75596123,-403801959,1810890749,-1226576323,-594608267,1845489087,329103293,2046253121,-747837569,-562567937,-628175120,-671310265,1839869439,2094910725,-59773077,-112266352,-1898586057,1054849535,-805523400,-291382536,468973962,-612500772,-57464,-1525925,1816110204,1928310631,-609959953,2137519839,-403062786,973070454,2105395574,-55732229,-73763405,-210207798,1440721247,2130654956,1073561978,-829068371,-608719490,671083789,-1483557130,-1747270547,-1253361925,-72380455,-408304173,-282075653,1987521759,-475670629,-186458438,1736179582,-759425,1475836719,540342575,-27263604,-151585957,-1074008242,-251685388,258856873,-313820803,-731906137,-606945285,-1107597362,2142223807,-117834146,-906031639,391765503,1243213753,-1679902850,1778341631,-1136742291,-1546785089,-1283721349,1968534910,-864308074,-18351565,1067348205,-4732322,-52436034,217823535,-1116320308,-1154094179,2080173747,910029518,1245249015,-1895887426,-1126294648,-907214989,-1140850813,-29914118,-1915805953,2015213481,-12846511,-806070921,-79761538,1120663442,-1879507057,-1158570307,-1013623554,-1311450129,2144815487,-684723371,-176784011,-72623573,-623772171,-354962514,-553681621,2011881133,-628368015,-1193314947,-84296737,-347282109,-59345927,-2066409573,-1529024858,1520663929,-26222625,-302581321,-813750371,2135904216,-1184250051,-30000143,1978347062,1005500278,-268570665,2056203253,-644257338,-546328066,-107119406,-596194436,1875719386,1327397884,1742499391,-1143242755,1741098942,-67713633,-1688310994,-792756440,1473635327,2041561055,-16843019,2145068753,-296037377,-135349398,-411353240,-2108499388,1056833497,1196841967,-35723265,-1048724562,-747962537,473576637,-68356225,-1225936181,67841973,-12877889,1307989592,-306380802,-210542915,-17045449,-269517538,1188265709,-36340743,-595393,-941799519,580899823],
-		[-38878002,2011130346,-570921451,-33577063,-29377665,-1206403100,-1013202963,-1367507969,2066658638,-424679437,-68225769,-1098991250,-705133569,-251700743,-1555089429,1877867335,1872625650,-138061321,-1102596489,-1251328301,-1074922577,-275338009,-1109023672,1606078908,-666130,-1615857739,1863802236,-1098975403,-948174921,-6037571,-339218573,-1107298518,754941626,-881899283,1051394047,-1499729177,-948610880,-1676386483,-88099275,2062544703,-483397009,-1090527377,2108522495,2146400561,1489893047,-563872013,787864981,-1946226706,535673859,-811741803,-143655954,-371208428,563156459,-235160083,-956348567,-474480993,-29442849,-766022866,1063250911,1998493146,1978906077,-40902645,-1218479106,-8194123,-1149632521,-467894669,-160585957,-1143116619,-600064538,-550078249,-33955929,-545614197,-1321541827,-1154844937,1875738579,1402418378,645587903,-276989974,2141028139,1899481134,-294125639,2145842297,-445120533,-826360530,-1235560967,-20213782,-858132579,-1082463661,-352912385,532601749,734905583,1056515931,-23085569,-145771045,-1308772121,-1354768777,838262631,-863044763,847238984,-650179854,634498040,-30183963,-8483842,1941127165,2079061501,-1082302622,2138832877,-1074317435,797679150,1609022437,-151307331,-547885193,-820623632,-137429481,-541488169,2062939374,2004303759,1019444030,-557880998,-58788855,797019893,-948966186,532598137,-226590726,-306243026,1878947647,-839504418,783529964,-1966240870,2094198245,-302739714,-68498161,-340134225,-458227815,-16897,1474230257,1065020401,-1531244043,1171258831,465363849,-88620806,-269489234,-1115739795,-318898695,769121181,-555164465,2006690811,-143755561,-1211320322,-118075893,-268514946,804585467,2008280551,-20566050,-1276152123,-1074532481,-450655747,-708878732,915653103,-1242042881,1936058239,-57936660,2112287683,-583008937,-170976532,1006607983,-872685581,-811860483,-1248879748,-67715112,-992305558,763147047,-207621325,-617112143,-79345967,1428675291,1463164141,-542326785,629028281,-312247338,-193331206,-156297236,-1786988593,1044307055,-1278252051,-436355817,-1427132535,-1250116471,2095492580,-288145,-42059799,-1350206538,2143514527,-252282666,-560681661,-412934311,1689905295,-218401416,-27153,1595661911,-323525031,-382731825,-31656298,-1344375944,260521387,-383975431,-1300666755,-1867323559,-805527887,1572841317,-1255284817,1965374191,-1343463336,1945735487,1610516118,-145772626,-54641289,-459426763,1330380671,-8391683,2122184949,1676868041,-615807258,-312554435,-1247069569,2133848027,-3463205,1006554812,-1161036035,-1484490643,399228571,1315910484,-23279663,-1472217089,-1143129729,-1652621701,-1652934420,-1649148724,-808920283,1841658879,2060790453,1918500373,972487660,-1141114913,-1351321737,-405232697,-4338772,-1229542297,2059353087,-1359168677,-613561061,985327606,-23383310,249296446,-422970405,-824204625,-235954465,-606585735,-628686859,-45798677,-311362121,-81357329,-1413752911,-85665301,-325062977,-70023344,-814892465,872152987,1811053804,-1031939329,-2013268309,-135725187,-813441141,-1540563805,1042270611,259503867,-1799779910,-426284130,1046216412,-1118303,-1619428162,-962925585,-673383525,-571273478,-1283491332,-262563,1579148663,1748081425,-1078729285,525195369,-181618817,-187695207,-111202473,1560262447,2034728959,2134474749,-814684533,-631832867,2101313530,-22316099,788529007,1476376271,-2018019907,-107054678,-77415686,-71457361,-1747460167,-167838215,-1135647745,2058575144,1031748047,-1500922465,733928216,2119490915,2094997438,-51928983,-10637378,2111819736,-42833581,1761558268,-1234464033,-134338585,-1351254901,1052080053,-570452506,-681132578,-1107548289,-1116742679,532283255,-1231457475,2129103870,2111551995,-152446282,-1088762068,1064519922,535250923,-50612317,-673322025,-174205110,392687867,-320953407,899389925,1994346137,-277695843,-543187458,-29104680,-1342781479,-1252664449,1044769247,-1468009621,-673749904,-67306581,-141922341,-1214809653,-151259465,1853749739,-1092659429,-1099844,-593571077,-749765172,-1143129449,-1098138693,-134218009,1313286141,-365733385,-1482828971,-503714357,-545984607,-1365517123,519306489,-856555613,-706810930,-1000931778,1796960485,1609507193,-1101334061,-1311114626,-22020231,-8499431,-1214775800,1476259783,-426870274,1740471683,-325060673,-1518569714,-44649025,2046812590,-482607463,1873407887,-689573927,-361910891,787462382,-990063342,1291427087,-591421873,-268501634,-1054802564,-1815249033,-1611303687,-1292044897,-352486090,-1881286081,-34522489,-1614862713,-1611860483,-739968289,-168841257,-1884698115,-144130265,-363413523,-108265730,1760798681,-1075367783,-811548,-16877569,803171065,-437616679,-137396229,2009822911,-1956053011,-1896404210,861859518,-1933389577,-1145130037,-706302266,-147723,-1621648021,-73598550,-24643268,-4465160,1570043559,-680838079,868153107,-1614757433,-1082311698,-2022933033,-285884715,-1215437067,-1699262602,1592260276,-224424646,1743472637,951048077,-202844166,-1770680906,-405274697,872132218,2146852974,1474803121,1541271889,-4670181,-1216529076,-1185099337,-47333384,1255104511,-842065358,2108152765,1858270590,2060973535,-16941609,640278573,1960831660,1812961455,2110609879,-29049090,-2147603,-101222453,-1820607489,-1639036197,-475070821,-756820804,-12723903,1335358207,-1360797717,-1611039886,-873006604,267287147,-730113047,-412147973,-85987366,-42106955,1742597964,921409743,1475344807,-1482690671,-41089,-339951685,-1176031681,-69533918,1739151839,527734642,751467735,-542379077,-135463082,-17104897,2142970299,-983957530,-142890169,-299926345,-202782098,1060816791,-323755140,934837342,-808596517,-438460437,1064479802,-604762601,-19932294,-541953,-1362108585,-52189813,-1087620243,-677429352,1739483439,-1626423551,-1090675077,-848515397,-18154547,-180450692,-415785925,1072969401,-1681920907,1073348579,-117296166,1300051385,-240415149,-87496113,2004776398,-591466498,-508379657,1605709679,-9289,1609000863,-36266275,-138510356,1063755870,-186793243,-638722322,-570430276,-371191867,-34471662,-805831819,2117328877,-1210190542,-85198434,729447999,1534590975,-86575444,-72694467,-897813323,1132066303,-396366529,1200962522,-1117504298,-1380225158,-1224841749,1467613151,789667749,-438083693,-155189252,2063510524,-864954437,532462061,-2047545556,1969223639,-438342183,-537523346,-1161324482,-235959074,2124414845,-1025919042,2145869640,1918885373,-149450305,1395847087,-2027757130,-63071050,1164537423,1677448127,-1344597889,2004611066,1728043989,-505164931,1063144440,2108439038,-1011653266,-12643953,1509513059,-305437778,-1751425165,1046145750,-1716685353,-1075909817,-446729288,-647000371,-386141713,1501421406,737475700,-294150404,-553678034,-403372548,-106324623,-1141441555,-759440585,-1171611971,2096591183,-642466306,-981543047,2045068465,1605232986,-168083984,602922107,-425402413,1574955469,-1164446732,-1074031497,-371409042,1822425007,-1914782280,-1346503281,1979447187,1810672100,-2082001415,-1816938503,2143002166,1053357849,1512488829,-2117147985,-56893442,-1099886236,-159665337,1205129335,-603982852,-1900298625,1800763366,2061991477,-25278801,-58642630,-1187145362,-586445953,2086632869,1931466687,-1621378729,1987739132,-729594898,-579904881,-709499056,-182832257,-8494253,-1331290,1928947893,-1342195755,-1256049159,-1614446833,-448022855,-225100209,-13764897,956168301,-673186824,-23339273,2119933373,-356730885,2042598829,-145916001,-1825228298,-690294489,1938809303,-835004066,-1107911413,1773763231,-1689618586,-1341214777,1845455102,-1613280324,-1615594254,-109588486,-51964633,2077202403,-1076298804,-2703902,791941358,-319598785,-507617337,-439432338,-536886412,-1076618450,-742449696,-1443906006,1005322025,-35684485,-1882248423,-203751938,2124990303,-167772301,-170588174,-1644757065,-3323917,-1470497281,-473956363,1537733618,1457258191,-1337080621,-4717826,795645949,-1445987430,-1351421451,-2884658,-779193608,1576533494,-2102411818,-235686997,-19433006,1270518646,1769838315,-67176335,-8721041,481459661,-1162528775,-1073826071,-1082856011,-655655753,-1241592306,1821807599,-85114900,1609660013,-160514310,-438437275,-1069564488,-67904139],
-		[1342127029,-1208172820,-1618172209,-37880678,2146693999,1534412605,984754174,-671295505,-1946950243,-461644419,-441541130,787787319,-1659715329,-268619787,-605323887,1500567675,1520516091,-551032332,1744210888,-1075329187,-892601618,-1685656215,-1067226313,1845268094,2003561855,-1736975619,-478291405,-7390121,-257436007,-314655761,1861847023,-543604996,-153687554,-369115523,-1125253735,-605297509,-739248513,-582141338,508382582,-536913153,2103246521,-127862225,-216378705,1458434043,2147090383,-1389379610,-1723935489,-362420226,1439547375,-1821638658,-545888785,-322994243,-556045877,996818111,1517813657,2104448894,1874313214,1668702205,-272741898,-503662441,-268664995,-777144419,2095054090,-571624197,2012184222,960427485,-1229259873,-269035795,-224791986,-39878829,-1146095395,-814818094,-1640880132,-253250866,-847823130,1670719167,482301685,-157024332,-189465362,-375401935,-739773227,-1283359937,-420201002,-446776850,1471269887,-672404611,-293936417,995351291,1677717467,-74121349,1973271995,-1621625873,-1698695491,-139476995,2138042365,-156698713,-209784590,1554215115,-218370127,-744816674,-1373025377,-341515045,1954527039,-240180484,-1208784834,-399385,-1680282915,1501449197,-154455369,-3934606,1910264685,-680611881,-197477125,2076114017,-766517325,-24510514,1610446524,-51758889,-289944387,1022556795,-1216950792,-708977313,-972299274,-2030315529,1942978277,-68580781,-1099538888,-377594370,2079268860,-772045866,-288430325,-738418874,-809580169,-221252201,-10196009,632924475,-1241810452,881802228,-234985517,-806890797,1131733386,-1610756158,-1103428313,-405845520,-807716874,-2035352134,-1677992669,-1682954778,-546378759,-18919726,-4427825,-21301257,-92627984,1861091015,1876941759,-1209111314,1591233262,-1075175688,-46671893,1564853239,-1770229251,1284683739,734117375,1658780503,182963616,-1146880403,-1159988753,-1141506669,-1483167879,-321917991,1065352659,-2121010850,-1411542105,-1040316545,2008831721,1040102627,-1913557170,2136964943,-2080851234,-189010946,805205228,-269191428,-555030635,2144843646,998168047,-1266899467,264486909,1556083659,-8543250,2017442782,-675287363,-38343381,-42996807,-1671432425,2093059303,-545307011,-1100267636,-235419724,-74121734,901611007,-773918785,535549399,-873072433,-1614856298,2142453108,1457896527,-1208386885,-570463721,-608186881,-1212180790,-221287447,-2097488452,-25401002,-1074171762,2117918518,-1107830810,-1096851809,1599856956,-1708262401,-1183228418,1534292860,1598021094,2054017895,-172581903,-1083763082,527425327,-21539663,-10093831,1369373614,1020428269,1466789263,2010107327,-495116737,-42148070,-10541393,1338627036,-29906002,1641969491,-608832977,2100788602,-176236801,-790896680,-161760274,1580578863,-88478522,-27716549,2144028581,-1914273813,-678834249,-977281590,1711143254,-1078016795,1061516495,2143254010,-188144005,1793023999,1199532782,-409610377,-1104151625,-4608634,667186685,2088167398,-1213531301,-673258774,-552891409,1676638254,-1096080973,-134677009,-1749270733,-272667652,-1090584714,1492897790,-1612345769,-176324230,1701035256,-593284641,-136929412,1517813502,-369773255,-570600322,-344608580,1540803974,1609940179,-553829426,-1477001219,1056766845,-1982861708,-52599234,-1048937,-673982550,-763631799,819015071,-710560713,-1313058134,1042202018,-68231185,-1111099461,1870003099,-1979753497,-33652905,1413410708,-158204949,-1636352189,-59769145,-1080045064,-193749001,1944480623,-1411652110,2094915417,1792718164,-52570149,-1075971681,-1443987730,2112873455,-273467886,2137890302,-33971017,-574645353,2143284595,-671668242,-2115895658,-134442,791771959,-79843857,1979480055,1206619113,1845097687,-554483788,444440294,1407123374,-621769924,-486572563,-11575301,367493116,-135051694,1839185911,-1211126298,-607742093,-411314221,-574758666,-805821429,-1242827778,-84083460,-547756119,-637534593,-226561541,-70361858,-721513235,1507303167,516802718,1388764549,-105486890,1721759663,-1275463109,-1157630419,1586230263,937999041,1966857079,-1082524,453759994,-1863809,-269371950,-948568325,-303451209,-303178253,2144824693,-269526187,-371987205,-726301201,-809541211,1711276023,-759240983,-593310001,-604381490,-806472837,1968925132,945534638,-1442843873,939393007,-1142948419,-1637876294,2113863167,-1345413121,754511471,639021979,-1670926353,1529593275,-739017223,-455442449,-5412449,1342163007,-809797633,-496140433,1853549995,-1319518727,-1075538158,-1597116722,2112880559,-253037130,-728990730,-554745317,-1682638949,995883671,-1275226134,-20854095,-230822019,2143231985,2147254263,-10620434,-2105437,-1277764905,-293962231,-134349257,-8552749,-515947650,1006478779,-405570613,-1629651393,-716448769,-458312207,2043181005,1326815614,-84183585,-641862085,1004488250,-924961160,848779085,-117460099,-609028114,1964588973,-118393346,1868297212,-44564492,1753185387,1870001775,481659816,-908077929,-1150911364,-1277166988,-1485361168,-780836375,-671222669,59727123,-10293465,-1275534113,-893583470,773585898,974550865,-491913235,1867087709,980414326,-43811466,1941404927,-1153564939,-295201803,-537399414,-1219650317,-420537041,-488636929,-1355199905,1026995607,-880122025,-1713134626,1064713087,-1462479425,-2637825,985084379,502739706,-898637875,1070300635,2010088778,-943907202,-875091223,-38601473,-1434497426,-674242945,-202005526,1067177953,-491819361,-1092769561,-78188831,-1285623849,99953119,1333890477,-824452625,-372392298,-731948467,1069094713,1680728728,-771819046,-126674706,-1482171402,-333767047,1673584613,-77760656,-8417283,-609308429,-320944523,-270008997,-1485161487,1064550245,-2197827,-86052993,-30188562,-67852682,-1050145,-245528612,736630223,-1261754753,1562365659,-337737731,-1386328293,1902066582,-746785291,-1258316509,-746587269,182435838,-1766423713,759135175,1989410557,1064666486,-1354846586,2041970084,-328274354,-52144761,1956249505,1451010173,670940794,1843842871,-626035265,1073443839,-155060803,-746913863,-273040201,-67272839,2004293566,-874023073,-835721286,-67223590,-537674328,1305729647,1876582397,-1880600725,-1175105994,-713093377,1307502208,-1210354570,-848364801,-1217266209,-176439430,-46336286,-521575189,-1294304209,1027547105,-945311879,1523547566,1872486067,1072127595,-1790477340,120944693,-1137995841,-37787018,-1411448961,-235147305,-1712324673,2059257718,-322512931,-111330345,-1191271480,1542269942,2074803947,-286459913,1587526989,-277952513,-671286146,1052162101,-606274730,-610975748,-1115819289,-1100035273,-69370513,-17052610,-44124933,-320889046,-92587077,-823315514,-1725281425,-474481804,2060686206,-755572133,1117255483,-202797128,-306217346,2010044959,-857125137,-445129889,687793086,-1142957068,-286557828,468169678,-1008901625,-964167924,2147033059,-1023472805,913045239,2095359231,-402725153,2146910207,-1106520786,-825439618,-283249349,-907448521,-52302793,-188918485,2143285037,771527935,-324420793,-607395855,2068641017,-1127697420,2147458834,990182550,-318910618,1856893818,1283303735,-545330370,520066110,939441903,-97724986,-1617664545,2033580113,-1182811394,371156863,-503593986,1094447094,2139299129,-551692122,-78675985,1207392597,765226335,1035592947,-1212167713,-88620321,-134692913,-237504905,-268454965,-1697459282,1328527289,-1124366529,-1109913706,-375689405,-1923507731,2076893055,-1074431542,535741945,-135286785,-126143019,2095824767,1604299771,2038415357,1610292970,-975320497,1993251798,-1108166721,-25755958,-614928019,-1092665345,1527819263,413075405,-1237732689,736061431,-1677805325,-81826055,-57867411,2141962223,1673489659,-649037869,-1711751177,-40698117,-523044517,-1086344017,1273687665,-1335913741,-92410153,-1092275202,-538969153,-1748549901,2080373596,1789881923,-130845090,2078259959,1257760254,2011384551,1673382908,-629501060,758900302,997312201,-1091102849,-323568582,-1126251268,1501098487,-574633002,-173541145,523284214,189836735,2050457326,1604800405,1442697137,1509414911,1555897324,-262144009,1576058861,-151601158,-1644470533,-1146152987,-71901587,2075219121,2096059806,869166814,1341946271,-373461106,803724999,-580587673,1721358718,1874572382,-79175682,521041910,-320610625,2121723883],
-		[-129966866,-1835881523,-268881740,-1168151300,859747668,-90507324,-302523909,-1316028881,1190849891,-1248006010,-590913553,-268686083,-1720305923,1601748895,1658794459,-1881360963,-111384262,-340267713,-700015243,-201864162,-1268015240,-988853794,-281046531,1334769662,-1344947359,-1141773865,-1400983681,510516841,1828285878,-40193,-185340133,-306299015,1722804143,-142672399,-272022451,2145251163,-1233126469,-2059930833,-1883629686,2142723903,-1208750213,2011531230,-431411601,-1547714857,-291867298,1325869827,-1369654806,-290070609,-1076243041,-1108344842,-151003297,-360759313,-370151061,-1126208449,-830500869,-1119884323,869842749,-47096101,-19968196,-1636436819,1797880607,1542258111,1532750847,-300371013,-745803828,-1258377756,-301990473,2140205599,-67240975,-315623701,-135790689,-1369310985,-1647873005,-138451073,-1099104388,1730082027,-1256243563,-1626870788,-956305490,-1157759530,1340967421,-898105605,-1619617284,1591958036,-1780136193,-134313987,1757077407,1354223483,1685446362,-25234965,-771826313,1030740499,663206783,-607191076,-273388260,2069036782,-37932066,-759242802,-842433575,-1074626997,1073587942,-1622152705,-2081596930,-1508508641,-422090082,2012739965,-156247747,1479692257,-52126940,1731426404,-1569262509,-576192785,2071133438,-818416154,-403456050,2147457711,-467310628,-2359403,-1074164,534569599,-907612207,1442475975,-8912901,731388920,-453020585,-307430148,-1015033705,-311495209,2112315386,-272105749,1741360437,1132304571,-769827974,2008088348,-71309397,-539289740,1055817213,1664801118,-213977641,-577117924,-1093276078,-68327681,-102597122,-10228395,-69929501,-1350635526,1067253738,-536946838,1232334865,1842192118,-231490961,-324224545,-995437001,-1844161683,-1518902596,1866202567,1924622171,2012643215,-1208569410,-844129025,-675283473,2085180094,1866300410,-70586697,-1294050322,-232784083,2013212535,-16778373,-1292513354,-691024017,1316784543,-1102054441,439829936,-1630835770,-715244055,-5331513,1783887597,-935463505,-159974514,1065344868,-612904209,-71305441,-88297538,1603993559,749794676,1893462443,854519679,-15911427,-704841777,2110872063,-1248883,-10813585,513064745,-395840819,-1582968,2137459218,2126804291,1868562255,-539029521,-1117997095,-17834011,-335741198,-1644447173,-1099604683,-617128073,1492587772,-241336341,-339747075,-470466593,-658793725,-1142499618,1875793787,-1151697273,1272900858,322846169,-1653610050,-721723539,-1095329797,-1518804225,-42270920,-545305457,1539235740,-688186385,-1116610955,1841095198,-84219019,-606150980,-789479622,-138412064,-1164808469,-612403201,-1103036617,1867951871,-84426931,253714428,-134352844,-1082142209,-960692517,1508585189,-5313562,1065057755,1607973819,-235001133,-1513431080,-1070793964,234145081,-544744519,-1087064129,-612631743,-890032692,-306397876,-1078379,-1121715018,-1394954285,-1503658786,-1619808788,-42091265,1977611485,-1133339140,-1124106325,-402928323,753531864,1442105231,-2084345105,1554300715,2143022074,-672277267,-612797500,985343419,-50421895,1064762874,-683686446,-1095670278,-27171247,919967675,1136295295,1008205287,-1411797025,-42292010,-319724545,-4202563,-522195559,-1619034237,-1342607922,1048513983,-1644183847,-136972880,-1393365219,-1122239299,-1215320337,804058718,-809584893,858155485,-269228215,-554704995,-68804621,-748260407,-235805852,-2021743982,1860075295,2042363336,-305929249,1340591422,-1057493135,1055837891,1374394363,-449102609,-12245377,-111583254,-1120478177,-296798321,1305770879,-1665975984,-1403555104,-101257217,587388920,-136349441,-894511652,-707448331,-1185262669,-1210853815,-1308638497,-143032777,-470695687,-156052305,1044371258,867942383,-432592147,-1479853315,-273681166,1864201433,-26481785,1337872568,-697349587,2084437933,1050278523,1330773997,-34082821,1744728813,1976221679,-119810097,-549486822,2143274879,2142198267,-633425474,-1567591322,-1695140993,510342100,-273303565,-285238325,2086238895,-1077879511,-271058625,-875045325,1844807669,-272379977,1599601596,-180454093,-470341345,-875332997,1521835255,-973915137,-643928087,-1645515523,2124406270,-1342734982,-621375607,1543487287,-95184001,-562180333,-1427049539,-1345882701,1585430477,-24684689,-224412708,-135472398,-726685893,2127470075,-231840362,-361931874,615508575,-1436029545,-958798043,2139093169,-1074086675,-138478341,-178610955,-16801817,-1947668769,-661212609,-2075489546,-1284385169,-1310264451,1585438086,-326131634,1815758589,201325518,-302762061,-856699663,1568574902,-311239084,-460961,-1143375110,-1087011141,-572001833,-50412794,-305670249,-1275204228,-882393121,-102519125,1130101370,-1229046757,-1616940033,-1410401964,-687914021,-201465017,2146827643,-37750793,-441811287,1584780157,-1720912554,-1876101777,-67916231,-1120442500,2135653561,261973455,-168396457,-906535655,1039080655,1358672882,-552927491,-953027281,-1581189,-1293961802,-1074201489,2145386141,-620891154,1962306303,519274026,-76168361,-1342622337,-181019297,-268445267,1605615311,-144704592,2079653881,1588033979,1684943933,-2106011791,-34644225,-543171701,-1155940498,333380949,-1078067474,-173460088,-144357,1868819283,-543183209,-941666625,-538552625,2111831983,1558181438,1461673644,-1108136209,1073348559,903069662,-1521266802,-1076370691,-153848965,1576615679,-5047377,2132704317,-1073774854,-1138756225,-67181097,-1228296389,1492975387,-572916233,-1346443309,936803772,2118327989,-1372464294,-539516945,2065657803,-892408881,972939247,-1593422,796638948,-1074071685,-810962137,-1222248579,-13803644,-1107364013,-536936609,-10553753,932697964,1307130554,-321220649,-710035649,-880804805,-80019621,-269060500,1040096107,-1503291426,-184768481,2078796749,-11870313,-1380329349,2071849365,-8388938,1609152497,-872494595,-87077969,469036507,-1128529985,-1367229285,-1102911509,1864832378,-1183581251,674675673,-801141137,-88170513,2142544243,2147417531,-368115961,-1884332076,-839000516,-1654924499,-2065088012,1407121259,1600896720,1030197047,-708970584,-1617448098,1992154213,-1142960331,1962777002,-1361977459,-488804759,-539505957,-83970588,-1682294110,-1654261157,461206651,-1714028582,1741897679,1790376751,-219390440,2032000895,-614241723,-37949477,343928814,1693061918,-66553858,-61107285,-542283883,-207715633,2111307635,-442837345,-105113866,-1661499410,-279038642,1576719734,-1092094251,-706086162,424395631,-34226665,1333699797,-1083187715,-295886934,1310669678,914112462,383768829,-1689527471,-33950722,1776276063,-31838322,-71059142,-270876810,-1082466861,-6424449,2142824315,1286391755,-539013323,-979653745,1944968951,937652076,1001857983,-204475017,-328083522,-1718978146,334296317,-674314308,1843858931,1069071419,-1627440201,-1662583393,2138906487,-1345402145,-823021701,-1695036173,-1122330249,-1387287875,-403186819,1734827235,-415702019,-538329930,1048202478,-1221592177,1382809002,-559744035,-1629632646,2136133093,-1234829703,-1257317769,-4218670,-118560837,-43232707,-1367877257,-1547698441,1072913134,-268739866,801056588,-419566725,1902826998,758120054,-567435396,400475075,-864298001,1807473332,1053883609,-4830205,1474224942,-96776302,468145612,2013003721,-36729858,1876376344,-47454292,2010241081,146615772,-66074818,-1780052045,-1181024962,-1883529126,-1376780458,-1897144114,937400305,-1879377929,-67406849,-1585496193,-281151628,-866719369,-134562115,-973851140,-151857671,-430969993,748600192,2136210924,1037867887,-1351352567,-547946513,-1888149669,1578259862,-164094636,-335718609,2134783419,2146362715,-899716113,-70779493,-18427177,2079063515,-1678827531,266337142,1071032510,326627261,2102258170,1606116830,-38977583,2007967694,133928689,-620789882,1538111570,-154699885,-316671466,2145171161,-404750609,2105105334,1872321369,-1120006721,-537297675,2126330807,-11577391,-12714057,-738363959,-335880497,-149061850,-1384126793,1907480575,-422598779,-269701557,-1208616322,-903152449,1365767115,-17375774,1660763992,-140030625,-140814388,-1597829,1515681503,821419181,2143659763,-1097892431,-179545497,1736389628,2082204504,1867365662,-437388157,-553800327,-1126449209,860004063,-1574007657,-337675617,668649455,1601166134,2021646128,1043527679,1277687219,-1007955253],
-		[-541088903,-1746064591,-7000643,-208340058,-201409923,-1598410,-638715794,-11741624,-805850129,-1117871252,-1248839987,-152178810,-559040132,1053548540,-39660052,-85608500,2111176702,-72780313,-553652358,-1325647795,2067128696,-233866722,-45189146,1056582607,-620808195,998180595,-33654027,1413442263,-67158105,-1545101475,-1124475737,936736191,-17319939,-178623313,-848237602,-67248385,-268468466,2144063156,-816955534,-578827265,-4281032,768306881,-1889765993,-858390820,1231481079,-488097365,-302188620,-1073970177,-40176193,-1213748943,-985666017,-1975652970,-1342223307,-1132479142,936842176,-1281371802,-110144073,-1296047045,358568951,-412736043,494599806,-1681151786,-1620119633,-5130282,1608417227,-8933401,-1711829511,1878859386,-1652033040,-538800346,-26752449,911416111,-856216248,-575217750,-1967264530,466353970,-2048461573,-269128005,711894011,-607404051,1592786889,1568500569,-167840330,-544243866,-356534535,1301276649,2079497021,1856618285,-1346638338,-1546,2062165481,-286790420,-1183207698,1789591294,-1353521155,2138549175,-1511742467,1480052533,-91291890,-16945249,-1722688323,1860022939,1811607726,-743592577,-1168852113,1676666799,-2101635419,-202392104,-509869,1538312304,-1006633101,-286706693,-105407077,-944265291,1973284815,-205530650,-430092756,-1077086311,-440713366,-293634657,1165976180,-604029633,900463633,1066382031,-1260683290,-113526831,-1916272803,939343033,1072348912,-1636907145,1508371451,-169639706,-269489555,-57016451,1894710511,1599003807,799616886,2141039934,1406037246,243267149,-676627050,-1491818373,-117684261,934247678,-144580609,-587352081,-1610894197,-138323203,-351405258,-553709585,-73662949,2130245489,-1152386213,1274006503,867122909,2003778034,-1158468657,2053810911,2147415286,1807542782,1034403718,-309788699,2027764411,-763373967,-583275282,-255925892,2132057973,2137619318,2050922491,-951487385,1901944751,-147849638,-680053799,-872940627,2063203580,1420816566,-2080933151,-738528548,-8923585,-27939888,-323026978,-369136809,-939672745,-134234626,-1213469209,-606086193,-1552580843,753351229,-1184414665,-136378641,-417666369,-1342742370,-1611220881,-403899401,-1444823821,-307245637,1189085182,674412542,-537945226,-1444620899,-1225916949,-1208268577,868150942,2088738269,-362333122,-1428455558,-405018965,-67189521,-604846657,-1198557474,-285267427,-208626827,428580350,-336599281,-655033417,-629818518,-1286603363,1341550440,-504513259,-521150793,-782375,-637888523,-778322024,1937636287,-2086670909,-335781943,-1218494625,-1071712261,-1982070787,-1099469061,1526425023,1998387891,-272648257,-139542593,-1430291910,1657502812,-269221955,-100913594,2002761506,-210797881,-1113799081,1070852095,1290852080,1581118527,-377111777,1071111615,-1107894421,-1280647337,1508351962,-910689731,-606146657,-578855077,-1665253385,-166273122,1387691383,2109992940,-621281361,2057285220,-447958259,-1950883893,1933564044,-1328661,2067758951,756936951,-912334614,-17040506,2006672994,-335566851,2144462682,616225566,-429435946,1476386708,-679938178,-1084239329,675281239,-1356942097,-300441713,-332455073,-293759127,-1223311942,-1282523953,-1102610459,787513228,1456464286,-704782635,722861589,-578841701,-1369326337,-1915552787,2039724922,-26290828,-1618660609,1858075851,2076646677,-2392073,2003400853,-1446569585,-1345016685,-159391800,1811627623,1389182974,-303569098,-1699745863,-1251632209,666660863,1991638904,-1310329863,-1837891859,-1612677703,-188874775,1863773237,-2693124,1962930111,-579877897,-310790246,-1636125942,-68957703,1860340374,-1275992941,-149094401,-1141707779,-1361452072,1711043314,-37821733,-151782610,-430453281,1181704191,-1670964915,-155238501,1534964767,-1424032483,-558904585,1235089723,640154929,-1415074897,-102406980,-1712203083,-1279541570,2006974461,-1900281875,1733296109,899491711,1592622578,-201417,-1028689764,-444640041,1281490748,-1312846605,-102769810,-1277993217,799309815,-1024481544,-134752187,797161527,-543167255,670493227,-336593711,-281329735,939488101,-1883771178,2141052415,-538989921,-1209292307,-1258970007,1836826615,-1923648957,-706093314,91327475,-55710542,-1266073635,1865310327,-339072581,-407646252,-1174458637,-383580,2013229023,2012732927,-1108513417,770927327,276135868,-1494239881,2063319742,-545792041,-315245381,-41999114,-403062913,-274465065,-705548310,-713180978,-2130326,-1620086789,-311167085,1751960745,1566209147,-335820585,791770559,-1011091208,-591792642,-119578625,233540058,-1138271745,-819724577,-15204385,247289818,-285362726,-169622515,913975287,-1681920675,-237897157,-10197232,-168720134,1935304660,-634809474,-296001703,-572785010,-92554248,1330638072,-580149274,1877078462,-570992902,-3152594,2063195627,-156363673,-303304705,-1711736155,-75678725,1945976757,1398228913,-12,-36896810,-278481158,-562073765,-25169178,-7360513,-142623241,1747446719,-60852350,-941380939,-1286087266,1922259669,715319245,1455148765,-38914198,821861295,2119135189,660448839,1526719353,234356031,-287607045,-373580817,1796440062,2145804259,-644186761,-823927960,-634652684,-555241473,-1096167977,-1019743578,1708107684,623373807,1038302149,-1385693706,-92910095,-101724957,-1209699025,-35131961,-138422435,-7606298,-914523657,-38340629,1862004671,301753245,-1293575177,-157374913,1793965310,318053931,-1126259845,1072361320,1541406650,2001970971,-111215367,-6692945,-1020793434,-579342537,-221882049,-1745348484,1159296477,-1367367877,-810058118,2012944750,-176720652,-109584491,-168436746,1417276411,-1082364417,-697322514,-674333275,-1074020825,-306316289,-1882196483,-574629937,-55776901,-714244754,-104043873,-1216894386,-1795332241,-14458395,-86107814,-306284337,-35655748,-1907934471,-1729189957,-940859663,-941885989,1811323653,-402654881,-403334161,494907366,-1669526051,-1118869869,-1571357211,-542451869,-1065235,-605415454,1210802122,2101278365,1592652925,1727150045,-353501327,-38119011,-585715397,1121821685,-548611714,-1221136585,-302718121,1061418687,-1489372717,1542913535,-282538406,-2034414791,1749024119,2004606715,-64620093,-1271148185,-1333889537,-16814937,-220348946,-223171075,1072429054,-1369496902,251129463,970800884,1439676277,-544545281,-270035026,-1351057996,1174362075,-680135172,1768482175,1301903755,-957241563,-71385813,-6327345,-1516263941,-610795545,1973286579,-1577339163,2139860237,-1265694074,-1078526114,465366143,1699431213,-1470303239,-1378093,1602741971,-101179953,2060331966,-873020950,2078264208,-35569739,2094677231,1710433755,-78647841,239647961,-1888895072,-79904770,-1122533633,2143277007,-311306894,744381561,1324871675,-86556738,1308314569,-997311441,1878830542,-2064650770,1005490165,-1227870595,889008407,-273318309,-1343857619,746016955,-613161167,-235438378,2008739822,-1381501527,1270605483,-1260668961,-75497857,-187171611,-2001342522,1845157847,1457175219,-2093940769,-24129057,922724718,2145074900,-322972698,-423649367,1139037950,-1349849602,1905129037,-1151482401,1332067300,-1485587509,935186397,-2020879379,902882577,1815838651,161640180,-107430218,1867968767,2147180543,1664495480,-219628287,1937686385,-613294083,-637577537,-5784107,-1191650049,1727944054,-1619026181,-138776454,2093007741,-20818025,2079781336,2008628602,-1139016513,-1620607239,-917718370,1926942270,-2125470012,-875890187,2043538943,-331381411,2130333136,-174080066,-909480603,-1208110097,-1003307330,-1225786403,976747455,2064938767,-1902822418,-79696676,-1613447298,-42692935,2144327351,-3687509,1332606957,-1217101833,-1500012666,-1180977281,-1692992935,2128432359,-235723874,1318025650,1609162435,2146827647,-275835079,-9474085,-147894529,2140618018,-1250088621,1539011519,-168460789,-1880227877,2146956849,1601661908,1272439799,1434839826,795374579,-1818247346,1006599082,-1186996738,357186516,-572645399,266581725,-604545028,-235709846,498253305,-1619931225,804641787,-1262007045,-646452266,1992027943,-1418232636,-1702898926,-355242531,-1220674032,795016931,1324310397,1241168830,-14715137,-1449924545,-8944643,-788939169,-12605537,2145378047,527392251,1770781183,-538587141,-37753200,1475729199,-442375992,-265363605,660209535,1836021690,1403571087],
-		[-511836441,-740008773,-1084770827,-348398601,-1634099387,1482481235,-541118506,-1263567515,267905503,-1704787987,-18981001,1951342509,-1342525443,-657224230,526573055,-1192238514,-1164182547,2138232703,-553830596,-127172612,-5179602,679271222,-983434934,-629323828,-1698662169,1838772154,804242409,-1078002289,-1280508806,-111433062,-9693188,1554949498,-1476876019,-536920641,-10660097,-2821217,-134350345,-1275160629,-134353123,447998842,-856283649,-1110143001,1572794923,772538094,-409636867,-270616033,-592802881,-269026896,-1141450247,-117690958,1995011627,1934377119,1811939149,-865774585,-610308564,-314212533,-340869254,-22102529,-203464905,2142935907,1310635925,-731979816,756856539,-572723234,-184648862,-1092163585,2141581174,1062502361,-356518725,1560220658,1992406841,-170725397,-428767489,223408868,2138996217,-270733313,-254296100,-1176719577,-1409614339,2076048361,-185967526,162528729,-1096182469,-1966358731,-1007693874,-1698435281,-73762841,-168870721,512749557,-455248334,-1379078940,-1913706203,1258018735,1842290556,-1360473347,-574786181,1560197444,-308381697,1790554421,1741553655,377155319,-743511705,-364975481,-1160069767,-57187972,-1619740183,-168445969,-1167361825,-552665999,-738347046,-948175875,-1621401673,-134641753,-115350666,-419665201,-304133190,1534618111,1861901391,1792777534,-1417695761,-1143568130,-598748649,2042486357,-287608971,-13517448,2105020767,-407929354,1606064621,1957390737,-277092017,-2031853705,-1769155621,-33823461,428371703,-2063864589,-704653332,-806562446,-549523718,-840960657,-1883312273,-68330115,-203037746,-805573543,-1396772870,922082302,-1620561035,1315272401,1753218524,-1178733953,1778351067,1405647791,-1477521411,-1351442923,1982801815,-1217056946,-639701190,-17278017,1052110718,-840597507,-570495489,2147214831,-1172148945,-23402025,-1417920527,1944419152,-116422017,-1687687185,-973360449,-712274393,1601946427,-2080506135,1256914653,-545293114,902982646,-1401569403,1475607803,1619551580,905902971,-1679442999,2044849140,-278924420,-2539781,-1117828497,-22138893,2133130694,-24653905,1875864399,-382994069,-1124079378,-2560135,2142727167,-170753,-589442833,-97641491,-56841682,1438146415,-620814380,-645996609,-1360006657,-281649793,-10485761,2004058069,-1677877505,-68198583,-877659410,-18499910,2054647510,-1182902721,-10029593,-305698825,2127322174,-185013068,-1247814245,2130673150,1877982974,-580683877,-1619371530,-1208470434,-134891919,-1568886665,-1073742417,-1352023153,2070670175,973045743,-1217987425,-479870584,-56111825,780465084,-269519171,-1486807686,-1753928155,-1100249113,-52470042,-1078297163,-8389297,1551214317,-1684574486,1137667038,-136638017,-953617690,-1611867333,930862832,1588105708,-1444189237,1596747311,-2118145,-177905668,-861069387,-153534722,-402703364,-355614927,-1397520387,-1615953495,-1930044963,-325117490,-1353451077,-610553857,909986510,1941545978,-571093017,1584063934,-134492203,-538259734,2029607161,-881862661,368266188,-570949858,-84177570,-75827277,-1081607311,-1845234767,-105195244,2071190493,-8626409,-72567489,2147466431,-310445627,-910636063,-1123274322,-572545929,-677270562,2105491325,-1618038851,-271879063,-445874702,-1082896785,1139791177,1999076203,1959262659,182189559,-169938849,-1617894534,-134753281,-536946693,666589028,-1239574049,-1680876827,-273394961,-1692015779,-410071137,2145307727,1743719512,-970836225,1589599230,-1268924907,-25730427,-1480853761,-302029074,-1225787654,998299631,-75661329,1332865401,-576903364,1688190695,-134521074,1705753591,-1682587654,-646686241,-411075704,1969661349,1154196707,-138556477,1332934649,-338008477,-2097294338,-297715867,-13669474,-381353991,-343986867,-373377417,-419927880,-163631389,-546504833,-1554717505,-353637189,-540059313,1464070868,-1079532797,965663198,399653452,-1082130444,2126464989,-1634254968,-1084811415,1775906675,-268476464,-1102775833,-208666755,-558911541,-1192370838,-197625359,-80151087,-294135112,1318894558,-12657445,-67385361,-290199717,-55664389,-115458099,1688530358,-671285923,2122096350,-638292777,230802357,500136783,-1076470614,-1879850473,-148717896,-2061781000,-746903171,1875873331,-1511328705,-68898821,-174543370,1610336191,1877282687,-343133722,-62920865,-1074944259,-1149997079,-9718049,-682727581,-50715,934041178,-1097256285,1790700799,-539029956,-67499537,-67155091,-1359028103,-758432021,-38298662,-386010902,-1076757995,582182529,-1483241813,-43524197,-1812204033,1928158973,-280011053,-1033394326,-471368941,-167784465,-271586818,-74462478,-138950834,-69583894,-696800161,-1077285894,-738564993,-136924619,-707813629,1983769309,-824763493,-1464111685,-1912635908,754265023,-545333455,1609115423,-593895425,1123670527,-595989586,-135529946,-1137658341,-421342497,-709491226,-76808363,-685287010,-1162610710,795125749,-47614721,-1390152847,1009598198,1056144859,1006527918,-1682000430,-405284354,1765695486,1941620606,-1346450063,-1100776581,-73487401,-1824008266,-646197485,666105773,-1891370139,-1275068645,1576755190,-278946062,2063394557,-609298433,-68222985,1459346920,-306340027,-21569635,-2014643205,-1017964868,-757073239,734703547,-1560712330,2104823362,-270940204,-1161597495,-1141505090,1502399993,-1699701325,-1141638274,-88211585,-1382199753,-409193025,2044965388,2080062207,-1417235306,-165718561,-816945281,-555221146,-1635157270,1987650299,1002307403,1591527678,-1078462787,-8611045,-1084321883,-705372878,1904239357,-814261266,-1140925845,1833621471,-353372257,-1147011089,-1388579087,-140650881,-476409122,1047396341,1056838523,-286037001,-1434522511,-608193025,2130004887,-1888159809,-1648567307,-898182030,-63290123,2088499578,-855810337,-279354771,-743736082,-1088435132,-38077964,-66200371,-1625915483,-1503676513,268271592,-975710228,1836808175,1293899647,-293618589,2137423226,-742941366,-1135067279,1590406686,-806398727,-151847137,2105532312,242564189,-1778012291,1802025131,-776057476,1335664509,2009006059,-902054928,-1152731015,-269074593,-33580047,-1649689826,-477701,-606079365,-86287225,-828451521,-1111853191,-86163789,-17835034,-537372421,-681577645,1037004635,1609563121,2036316029,322567007,-1180771169,-1355424041,-405423217,-1205829829,1973237399,-188294161,1073081080,-682102785,-2074084675,-1648499257,-184643605,-16794306,-152636231,905084239,-1188184065,-210307623,-31646784,-1617446151,-180896066,-135858265,-558932039,-237931906,-1343185298,-1493219917,-542671147,-705438323,198823679,1901827575,-41944785,2116013876,-578994954,-423789587,-33704646,2104400859,1337716345,778827357,-10553910,866898687,-408563713,-306385672,-136604291,-306430482,-1359624501,-1587544065,-59419796,2079874799,1976115129,-762385963,-406940676,-25658529,383638493,-705767080,796388777,-89404937,285412143,-2057409617,-1377553,1888729211,-854591873,-428517379,1200811559,-141614199,-294656513,-1145184721,-631434104,-68436868,-1982206053,1600953294,-47286761,133955448,448486139,-562806821,1927081367,-374739393,-1721770377,1941263285,-2726113,2021457886,-256412809,-903425,-1879744969,2061975009,-352899723,-293806082,1911350694,-219809051,1979117295,-332956002,-844243989,-547882513,240025543,-1084232075,-1226860033,-1210187893,1585331663,-206851281,-176475393,-223102212,-38274435,-549520897,-625034246,1724347391,2033863399,-1364379846,-171525147,2113132511,1760035831,-181029393,2046567840,-640356875,-335956290,1719910271,-4753673,1929320043,-14373794,-1415079075,-1555154341,2104163215,1030027183,-659624334,-154228005,1327494139,1946990589,2132635066,-22033475,1416613720,-539002017,-917574,-1280524564,-635185170,2140577503,-695927987,-735854601,1971518818,-1345886219,-82969221,-99754699,-84308545,-1191246997,-1644849669,971957704,-304210217,246020953,1597495227,-151074820,-516499825,1002207400,352238559,-317194409,-827665703,-46900258,-950317377,-6293261,-1770050733,-610279553,-675910409,1722997401,-1317306477,-1353963611,-1720281601,-334628512,1810882555,-398467493,-5669197,1858394478,-33731402,-663784773,1056701943,-134752428,-163602447,-690634836,1897788397,2080298914,2099214282,-1621944465,-1661258017,-67177233,-1947479454,-1073781501,-105709569,2143137278],
-		[-558215398,-365517321,-1932332313,-25166245,-832616987,2003812285,-9380868,-1815351379,2088615903,2082717687,1070951385,-8650773,1877986617,1610609206,-240124419,-537133091,-202851913,-243351562,-826185121,1413413268,-1896125134,-2080558108,-314681381,-1074561629,536371918,834549711,-1980027665,-1628995649,-511801717,467658559,-1344352266,1595061143,914078717,-1476985218,-784369923,2105432987,-1568673458,-507539688,399735732,-173345041,-1657127778,-1216490337,1610239743,-390136938,2146434947,-412295041,-788358,-235749569,2130153199,1793721275,-33556617,-22157825,-1085671601,-680181767,-1132468004,1423564028,-273549377,-318812673,1987733375,-17434885,952104811,933207925,-673192727,-536929302,-1250054729,886295004,-152466437,-1109796001,-275554825,-108030121,-590635049,516939721,-2232464,-1244988673,456227355,-1073786938,-405111979,1004262903,-209829961,-541140326,1472066490,2111827013,-569115841,-726811202,-5313577,-1196586841,1863516052,684965821,-148374704,-1078101121,-1241580561,-304545,-587761019,1051448823,1929116570,-1073758277,-629177699,2011846783,1004519481,-613122389,-344238601,-550588931,-756288193,-1678392904,635338225,-1494819869,-278960129,-241270809,-484519571,886279446,1676580839,-41947820,-269420041,1800470383,-1980353162,-694289639,-2115067907,1590071138,-1074005229,127852543,-68194762,2142462955,-578848923,-1143493481,-2672643,-197,-740561539,-1672219729,-2086797326,-960497666,-570950220,-588040217,1994288886,-35963717,1660646323,-1846804527,-538838562,1832533822,870242713,1240987503,-570605421,-615186437,481810913,1886746860,774895290,1595780477,-1220281345,1946115989,249458293,1869213684,-538969064,-148441426,-1366414350,-1954937218,-67156805,-50358571,-715466011,-63456577,-673333411,1875623899,-1077944878,-1080834387,-1219267515,-1728787554,-1211775108,1626592491,1674327498,-1612856385,1305337855,2038421157,-894640855,1047133183,1878970106,-201601411,-235016262,1987559103,2092684149,-607680901,-910164193,-1522507948,-421200482,-939524145,-71966980,1606411514,1924083217,-1918044358,-1302925356,-1763887377,536860027,1878929246,-86279851,-998631778,-839252194,-1510126337,-1103381539,501610465,-142383561,-1492270351,-586156053,-4276229,-686227458,934870135,2008981966,-1349603578,530438996,578254325,-1073893717,-91769605,-1080089237,-1275757753,665704677,-135581191,-1611678262,-1660944825,-140363862,-805634633,-545333805,-142804083,-35143959,1407993823,999022334,1408038907,-167903785,1942974330,-3844713,2078244599,788515816,1610547083,-122454341,-1086883586,2011593726,1607860175,-709595221,-268507853,631503743,-1082398341,-549241386,-1897995957,837910154,-236103814,1266605175,-134505480,-17983012,-23816387,-1618237085,-1420296681,2147404227,-645435585,-546410571,-482660667,-1760217127,-18882755,2062530425,-1436160278,-285346564,-8790339,805289945,-268454025,-185761922,-475078754,2056382191,1905081084,-100733461,-481311009,-1919948891,804186937,2034461983,-610403588,-1343144322,-354420770,-536876586,-21044243,-329273441,-1823162425,1468988627,-646238777,1592256661,-883051811,1541045631,1557806719,978845525,-807699473,1270841035,-147656682,1465891703,-711009026,1602205450,-172572807,-279007897,-64159918,-1444200513,1609261033,-101258757,-1984172657,-43010578,1839198185,-2016156498,-192430369,667642831,-4195356,526818586,-557942419,1725430539,-1962672477,-1275331857,2138764014,-1494808880,-1146129413,-811303335,-1489001483,-139510857,331325128,992769662,-1430295780,-1579225221,1603140049,-646645155,-1481058967,1387227123,259972981,-277893721,-1236148225,-209918777,-293887241,-824615426,-183275179,2147343853,-89279429,1659621166,-21531298,669775359,500015819,2089511599,-147505461,-269033182,922300765,-556799052,-1120738193,2134626166,-1169691915,-1190506553,-1544849451,-1623625002,-169873570,-293717000,-474300457,-71719577,-789865519,1857721682,-908092161,-823445753,1085225150,1266512981,1136910263,-1963798533,-1076420705,1608501717,-1091455246,1517280377,1635630448,1575926655,-438592901,1449370477,628882428,-35778571,-475398220,-1883832434,1979112438,-526511089,-1784070493,-608780300,-29401185,2046397887,2082373046,1881145135,-33339953,-342350659,-1074282777,-34146980,-7417867,-700064258,-206911617,828766949,-557981763,1333716987,1710729023,1722624495,-33698962,-135303586,2104217919,-144783633,-88736531,-284068594,1073544155,1050536188,2131590996,1962362807,1738528255,1037033423,2130638317,-562957354,-46148369,-152339790,-538711938,1858321039,-479389828,-170434563,-805438479,2046711763,-1032020457,-204475461,-26272873,-71090585,-1237653726,2104342263,-246786101,1978776502,-1812992550,-1485016577,926897103,1940905358,1493105905,2105174711,-67838098,-1919294827,1526721534,-1215594499,1545452031,-1293685517,-42478352,-1677861636,-1750956307,-18257893,-146280516,-103558153,-7830941,-34473734,-39913153,-437415941,1672998058,-1208044677,-540066725,1055889387,-539757451,1945631675,890691510,-271068193,-1396785929,1052696567,1289204418,-204234443,2137109493,-1079601249,1006497721,-940196006,-237546203,-1241646147,-1101317171,-211321025,713276307,2010061135,-81456241,-1845758230,-419972289,-1226015930,1190649591,1607817214,-268472458,1965932267,-35915894,-48890635,-1782597957,-776818085,-279003522,2145873337,-1132501193,234073565,-1736843685,-137070226,-1412432138,-60893646,367980175,1028939129,-1113219333,-1208485086,-679478806,1206235133,666582523,-16992837,1567489965,-1887461433,-5646501,2071966685,-46208301,-622366594,-604132428,1874030327,-1398284669,-353419274,2117778517,-1112541355,-1721262792,-448067914,-537731095,-1145111186,220200186,-411117761,1659828959,-505920807,-1113682318,834264221,-1345323537,435534695,2012577727,-33988755,1975377703,-524294,2130626814,-1373741581,-101986794,444592124,-1544560913,1987573727,1874850511,-805844051,367351291,1689162947,1826166745,-2130082,-69214248,-1216629802,-557866595,1682701389,-571543846,1161737023,-226915274,2016345919,-1145352513,-761332291,1800842205,-47186225,2145369375,-42511236,-1287854437,1039577055,-176275459,2129653115,-201394210,2130537620,-683688097,1274882942,536866799,-563416133,-109268370,-1108920801,-95233177,-185491012,1417669308,802683177,2138902740,-2062050470,-1114343373,-608307763,-143242785,-1420952019,1509359311,2139665637,-17341521,2025676539,1001760759,-112203661,-1115707970,1721741821,-159041862,2147462121,-25701126,-1484825493,-546681665,1170200543,368292895,-87982434,-528358850,-545785135,-21017221,-137518363,1538260958,-93491629,-634570499,-43136259,-906072873,-598368389,-1076761107,-979528708,-1654142098,-740824791,-223367130,1994272223,-1426375781,-1610679472,-402957489,-54796677,-11944515,2147297941,-27052297,987557627,-218367572,133686526,2141447591,2060058363,-1108260613,1503637420,-1536823651,-404750860,1849206751,1805434335,1308314258,1736938615,1557747421,-49397879,2147198198,-1074384420,1718410715,-1351100190,-1277166594,-242489105,-6100806,-674302600,-1615347743,-92152169,-11039779,1465367492,2009858045,938078063,1703798686,-1914282202,-1259626881,-1091275521,1878916326,-1376162305,874608109,-31614031,2038430778,1253455863,-235449477,-1181197326,-1484947497,1312154975,-679740706,-1435763719,1069407994,1417281437,-303170131,938483493,2140354030,-617693541,-310018307,475958509,-2004183233,1073512181,1417665641,318504751,-1968248415,-1621119542,583464669,2097057657,-629178378,1392452335,-346817865,-629375618,682884155,-547365890,-1629037083,-721520641,1599812535,2140544983,-480657665,2062349503,-755253546,-277022308,-1687184275,-14843954,-171982913,-1539918113,-1110525573,1011742610,-1080133899,1802305015,-538483477,-1644742731,-159421487,1744822111,401974707,-587207185,-1317868837,-670501222,-709924187,-319037987,1734825398,112609647,1195073529,-807405698,-438474788,-461940417,1073575487,1204787711,-652880501,-1753751681,-67635725,-580913240,-1175851011,-153223446,-272772965,2129092267,-196378699,230422583,1264568986,2147344322,743953259,-126616580,-403495427,-289548294,-8920578,2145709015,2129996921,1054726087,1487925211,2028781546,-48241707,1152249845],
-		[695073279,-807142165,-69561573,-608924353,586678270,-229418,-1342487465,-1624768633,-268837697,-268612140,-1054134724,-1091377201,-1319176258,-139381761,-1292134421,1971247423,1172143793,-1078347858,2011572455,2130247404,-1701913705,-36848305,1066163591,-46322339,-135430220,-1075318865,-1495277059,-11141303,-22626441,-1223230978,-153219095,-1091698894,-604112417,-539493377,-1411582083,794132207,281804798,1971223289,-1352465131,2104885162,1579120010,363320839,2124998775,460857273,-855675092,1810882365,2095938751,1570766510,-134550788,-269107969,-51318841,1878760826,-807408146,-689046835,-610408494,1609431953,-13911377,-1309040913,-1213358187,-612893018,-1201681,-1097220385,-111559020,-104145932,-119948469,-1812611728,1401890531,-352321602,786011642,-667436472,671042045,-2022137947,1609424375,-545900071,-1551043747,-44897289,-308348337,-134582477,-71094082,1877794750,-1342595457,2067263225,-295116950,1907162539,-1189347458,-1395654850,-573350945,2109650835,1769987055,-1783474182,-134762753,-202670709,1945857783,-42754881,1610342304,2077191095,1006617595,2007725563,-67901061,-1603088641,-325327878,-542534673,1575436263,1971189751,-1090067523,-555762189,-281309317,-287540809,498521450,-1201940883,-781322694,2078244811,-1130041749,-1183076442,-1076115018,1599520054,2046633567,-69664769,2096805631,-843592299,-2125865,1818673115,664267441,-134446723,535649787,-71421953,-279060482,-1376139266,-1243630705,-107351313,-883032426,-1620269381,-12616922,-72049415,2012738357,-277027481,2105521790,2129624913,-6969354,-68193921,-272797769,1656221343,-254817921,2012001150,-8389146,-163742012,-2000961,-503400298,-67184282,1450503013,804353399,-872723027,-125842629,397006810,-668992290,1517121262,-1747730854,-1244300811,-1346955650,-1359776991,-781992081,-33794936,-709413034,-889335815,-1191299285,-1251938820,-1244700763,1920960124,-592730114,-441677138,-697305587,-1079520327,-1073762356,-450366788,-136450293,1039138294,-1610919947,-1141229718,-839422338,-1092226053,-337728878,-101060689,-748773408,-54802113,-218908903,-1146265863,-212074627,293330743,2069360614,2104451005,-2114309,-1754211021,-25731725,1920987039,-822150986,-43614901,-1917038017,-1090601991,-1075190085,-726575161,1408103870,-17171106,1031732585,1560148791,1725613534,-568257739,1865505759,-1896102485,905920325,-947922506,2146949050,2119525491,-285459367,-1621970033,-54544513,-705987912,-1074530316,-706874649,2054025063,-1935790245,-944382129,-36749329,1274924158,-6660779,-1288771101,-711196961,-1191248967,266844914,-141118097,-65610,-1804371969,-1107493381,-1177061214,-1415735825,-60520929,-134440729,1608446435,-1249787969,-135957285,-208916,-138428611,-84606978,-449445409,392953571,-1099079814,1234595387,-57884691,-355109497,-1118873605,-33838083,-268756675,532593029,-71448228,-539038723,-1121190421,-182239308,-1983986690,666879721,-672848436,1106646740,-1694122575,-612369204,-95324673,-1808851987,-551693194,-1140870819,-1384655878,-1237607490,1305595871,-29784829,-1136781345,-134498467,2124266737,-1082286249,-77103413,-1592792066,1546567515,-236323521,1672178671,-696473237,-1078986962,2084487021,-1695285357,-558647907,-1345067010,-287136088,-172248146,-1355026914,-25306695,-436292289,-1632477413,-1342456848,-238290281,582709078,-1580322945,1770982643,-8471073,-1293306177,2004784615,862953213,-3351564,-84030603,-1160713102,1875686329,-438322314,-176525313,2096410622,1268174639,-839254187,-595611718,-1585784874,1589630142,-1619075755,1390533311,-1686184218,804842431,1335974723,-888942251,1743847421,-36074901,-305144690,-168857937,-1211128515,1451065337,-212148618,-1471807941,-2047353922,-572558453,-6347779,2045198733,1560207839,-85759466,1252898792,922475188,-421836238,-117981426,2004592589,788380648,2147153917,-268825410,-427622545,-303672141,-17137665,-3687539,-1343981569,-167380247,-1089544597,2144729587,100644793,-4802901,2059001830,-1104423941,1207664575,-531796332,1979174510,1801705841,-164719182,-1328977,-1218782075,-135037241,2042879993,-302260754,-236978197,-1512678651,-42238039,-187977813,-2118585386,983206705,1872752175,1528546275,-138414852,-608193873,-16803481,2012708322,-1815166989,620330494,-755283224,-1477559094,-757151875,-1146401796,-1096157453,1330286935,1868792406,519535578,2120613884,-1233027657,-1744880401,-40460475,266797007,2111284182,-16909106,-1615037094,-1342883025,-1292179733,-541703246,1910685376,804511591,-1078723841,1555496761,-1148076,-180656721,-33572681,3593389,-283426959,-8393561,-100753414,-547749955,-394446477,-291770763,-866326739,1063845491,-1841505116,-573690050,778826174,-911672904,-1947709500,-45026881,-1073795682,-135154893,191495039,-2045249029,232456061,-34754593,-16909315,-350666987,-135146065,-1983316297,805304785,-269046566,-1075806225,-827343942,-759194097,-631589162,-111018101,-1390853,-849368321,-223016141,-822272185,-809014414,-486842481,1802600333,-537055243,-18350169,-271937677,-1654702222,-412092770,-640309024,-303059580,-219820099,-6897795,-329297003,-1074964015,-1366476746,-51806601,-4392593,-1745893413,1738110679,2111811583,1879006702,-4920129,-705301573,-2105522,2008730285,-1163468858,2078176125,-818809361,-1327524406,-1210258118,-1488455941,1051576063,2105143253,-1175970665,1608220653,2145879135,-1880658005,-408097570,1767271604,-13208854,-190646849,-1351931657,-426533509,-461841569,1567063917,1705539439,-598278602,-208482342,-1139016809,-504430732,972615197,2092274121,1247612669,-52351907,1601892661,1096666860,-152325153,-210109955,-28241415,-2495721,-1816314385,1768420862,-138457237,-151257351,-567570793,-557880391,-403184204,-578857377,-1289301,-672088069,-540483863,-1657176838,-1661231418,-814877703,-1213371991,-571523142,-276828746,-823135912,1852661108,1358486080,871007198,-646218178,-1284228099,-98054641,-856258,2003550047,-1230579719,1842733031,-176232179,-1478894813,-1076349491,-724255991,-313266177,-1210086443,2034794195,-998703105,2115444029,-1128941645,1459613307,-1513324569,-1132730700,-227156114,-176366098,-1647027081,788056814,-201889272,1470621494,-274942466,2120961631,-1078069847,788052959,2074068827,2024800253,-8480806,-420323617,165658351,-780144661,-12814905,-1864401781,-1712480129,2137972479,-547719304,-80806403,-1554391572,-69356594,-864513,-68723719,-1074981162,-38175199,-2044486726,-890425365,-1101337371,-1374112058,-187439243,-471015537,512223487,-402822089,-319979619,1631574975,-571486482,1770479269,-758729735,-1671949313,-136325634,-1074334401,-290105966,-1409439771,1609765887,-90279993,-147524033,385351068,737667897,-1104282261,654045100,1606062563,946846588,-1044996121,1677180385,-94405802,-539003089,2111827763,-349323585,-2034238657,-275817879,-805535842,-1253445731,-1275429304,1810628374,-654905962,-864786533,-1213670017,-949393259,2012716542,-412381771,641678585,-1678796160,-1143998487,-1222068715,-3024897,-153202983,-271523081,-1141967985,-369508938,-3858525,1492910045,-1443102881,2134061813,-352378889,-9505077,1633573654,1324873780,449670382,-1564475490,-9044785,-104351521,-941033004,-864159334,266169817,398193789,470718359,1815682157,-575168582,1067935720,-392641028,-1229718355,-136614661,-203454994,-1465929228,-834820038,1790602095,1020456395,1862126439,-1120118924,1526574423,750720925,-121786885,-1384139411,-1084786193,-1630621857,2111364791,972025705,-1814452619,-1918534126,-3363957,-419598408,-608570443,-42176606,-911447037,-277021207,-1342312856,-1040722501,1446999919,-244687025,-1599246,-544008465,534509949,-1275139333,1977613390,-279283721,-155855877,-1409286183,-270542157,-1766539277,-409486025,-925125639,1131342556,-741982795,-84558795,-613220582,2085347323,-607395898,2068814559,-1328677545,-3952645,1800550894,2076098478,-541430277,782493069,1600110497,769642092,1879030437,1044330982,-1635945601,-833527262,-417628497,-337076755,-445980194,-1216399377,1442840411,248945341,-2113930382,-345254154,-1342734859,-203537268,-1543928241,-137472283,-948535305,1878430547,1171104378,2096096827,1741256747,-841918602,1800926173,-239289474,536346610,1771385333,-20238483,-135307300,-109497607,-87,536773707],
-		[-23119961,1722069335,-221381697,-89145601,-622227106,-542311753,1860169591,-1155236899,-31999437,-1073827802,-1955211789,-283248705,-1082130594,1554462718,-484555089,1226735572,-553921313,-838860817,-18288973,-537120769,-1212223493,-1354777185,-71827467,-1078275093,-1422050345,-1164540545,-8987219,1507606093,-749769235,-1111004931,1066399640,408641532,691775870,-67478146,1610514367,-844194178,-1624638593,-537526307,-635781378,1610306910,-726270163,-713319718,2123421599,-348260622,-5511713,-610941373,-30429204,308002686,1810758390,427776146,520077161,-1103758371,-1097372866,-225576553,-1611338250,-104381409,1069473789,-1997571202,2011098503,531299317,-672989313,-1184383523,-1324354213,-112935601,-77497417,-176170729,-85852929,1195352061,-578921857,-77349313,-1177576546,-1879130506,-1789142025,-521339773,-575672707,-486699009,1434218911,-76683785,-281580564,2147164116,-1108084825,2125839791,935310847,-6057129,-206162707,1266024423,-1224771611,-8782502,1165935611,-143414409,-1405233441,1395654143,-50602701,-1140883938,-1216356689,1040039870,1064925151,-1176520961,1810593996,-184033312,-1280819876,2145872701,-183820290,-1379928115,-1915798084,1841390255,-1243239329,-403980429,-549653519,2125802303,-49422569,737794397,-508137766,845595390,-1267926537,-159793669,1507852173,-469806347,-805444422,1255134886,-371925776,515435661,1045130111,-114074409,-607431241,2138036157,-560259114,1059290558,-2902017,1606414111,1655435519,-592319714,-1211121714,-1700430948,990044031,-583639224,796614388,630019629,-142963201,-688521220,-332497525,2142633909,-671293538,-548556625,-1750615221,-1615497235,-235053313,-258507185,1136428562,1811935719,1404914335,511409089,-1351689667,1704197855,-137691806,1492606443,-1082949655,534080759,1206646927,962427895,-76840965,-226615835,-1207683,1667063422,-1645756712,2011066351,973010887,-1074305894,-680136833,-76497671,1290370707,-303073635,255122284,-42682510,-377668610,-566813748,2146402251,-1344550683,-228720641,-17296239,2118870775,-1208317514,-604049453,1621518572,-470024967,-209816642,-296366110,1020059631,-173624601,2013259937,-137719953,-1278215169,985089855,-1095864077,1044053823,-607651999,1102051185,-755106355,-539491229,-33853454,-273180875,569307053,-1311924225,-270932546,-1074415637,1341042601,1050364855,-17664005,898496970,-50631954,-841097393,-1104413333,-687939665,710671039,-1210604105,-664993962,-1291850914,-967527,-567851409,1811803887,-51020675,-621327745,-1196441797,-184861485,791000047,634632099,-1031799197,-322578437,1542946282,1893662434,1819773374,-897229642,-32145409,-671962793,-930850,-139747335,1522365018,1669067066,-660930955,-1090791971,1071046526,-940920675,1299691421,-84658916,-71320322,-620757399,2147294831,-269762409,-1753404931,-1908246542,2141583359,909062611,1858074610,-1669406939,-29900433,-889235098,-84886090,-286974019,2145905861,445382552,-34120893,-879002115,-1506668065,1970469818,1714618357,-567429155,1878979821,-419981769,-1123520993,1806608444,-1085281095,-864840193,-402854781,-1905665096,-1100517827,-121669873,-109793029,-541606530,-36294724,1065179492,-26266426,2096340861,-115685389,368439023,-1610933772,-1244431417,234066813,1379321235,-63050699,-53533705,-687702658,2078669757,-1678770195,-135045410,1631313905,2134269951,-1393645009,-803254977,-403706861,1060490695,1836702751,1735879076,-1610747906,1794113008,-1084328065,-205010001,-540400259,-1644495955,-738483609,-671335595,-627925029,-56678787,-67135694,-287385602,2102603514,-75743369,-1422034433,-1730154578,-134743105,1912532206,1105551230,-575308189,1029109663,-79006260,2113793405,-957094154,578808799,-379065371,1304100333,-1397031963,740142655,794734591,1342045123,-947988138,-730964937,-887227017,1554429851,-135811107,752340890,1922937524,1928850071,-217335955,-1850609912,-1342333845,1991704567,-238371859,-282857482,2096096221,-1462793605,-1100692667,-292004330,-743392161,-134986320,-2034708421,-1649427970,-269716785,1572039087,1861208778,-2516542,-545850026,-2072119052,1297903103,-116923333,591754984,-35418145,1828157055,-2080459330,-1105299110,-1225012713,933713903,-119563521,1604198286,2097985535,2058212825,-717800047,-683671593,-9720834,-387663537,2145347453,-97203301,1336636081,-1090530817,1268703051,-18997425,2117579870,-325519237,-1082010690,1079705599,1808789244,-1411428353,-1275068678,2130673113,-7396475,-411707311,-2017477250,-537331269,1814888238,1853158235,1387050512,-536874101,1873508350,1740099540,-298329221,-1213565961,1581186781,1784670971,1804015346,445362167,-1678102856,-23272718,1532869710,-171053057,-660605536,2060320249,1483544567,-8669297,-673645063,2143109629,-289411706,-437003945,-285744651,-1208348741,-836895057,-94995417,-1356335617,933688573,-207880277,902216278,2136817343,-796939281,-690610467,-1349804386,709941855,-681367813,1325307871,-561786903,-1078833425,-1931551678,-338854021,1841798527,-151524459,-1079989249,-1451236066,-556577047,-1180736211,1328693215,106671958,-1653811359,585022397,-20727297,-1155637521,-1485644135,-1111173922,-513090375,-335547523,-483208450,1239282534,-75403291,1305853591,399449017,-1615103431,-395797269,-312294445,1598811926,989279977,-119748865,-1951465705,1821596628,-1934098446,935292395,-1956647692,1862229910,-1096649289,-899424271,-1395274018,-93946971,1264507811,1587950619,-34605189,247584833,-473174273,-218447954,-272165897,201326567,-1640822277,-26411665,1609963244,-604151513,1835888879,1191178408,2147193291,855307259,-564683845,-415376067,-1480604974,1170185631,2074059960,-554508866,436168447,-1427153975,1509220105,-503515429,-815796357,1994817279,1693158807,2077956062,-692070401,-1749819587,-550915,-17335859,-79833483,-87105,-1927283780,-78648651,-1283545866,1500491466,-141034498,383057318,-1028768789,2099634750,-109625565,1442651235,2063264199,-1614079252,-609807419,-12720802,-12846145,-1443402212,-152064199,-748315289,-1401446674,-1098986,1333389691,-1750140937,-1991934114,-1963475043,-748312833,-327254507,802144086,-1128819592,-128208015,-302170241,-782845125,-1879118907,-1841400516,-1906901556,-1158760135,-1110076418,1790786797,-841098355,2096320510,1808771671,-608729618,-1090720013,-201394181,-1880229778,1232531263,658501412,1474735226,2012930011,-982557817,-286872681,1033597250,-806154274,976551675,-67158123,2117828543,-1376935945,-1176626715,-35230225,1874060493,-84165145,-1686111539,205512701,387865950,-1361726497,1663295166,2128695282,-17858049,180346398,-40896910,-33583425,-1308953411,-188754129,-113843523,-513923122,-1610615332,2062548346,1944516516,-639568065,-1554546881,667582455,782131965,-536937819,-1111503507,-548381503,1835201012,-1082851460,2069121005,-282340401,-958426337,-138420747,-563120137,-557992257,-27910594,536869371,2004838347,1600615906,797944405,68605916,-1108353814,-325090439,2034474255,1853520382,-1611148994,1581773713,-219678793,1037891575,2112501517,-1214292370,-23134841,-1342800275,1008619004,-1376524297,1351340783,-256436493,1270845429,-1857034323,1960836923,-1620075974,-1210589361,-772866229,-709615897,-1395261513,1648220897,-68735109,2113368959,-1416133815,-912311329,-730612013,870579158,-1149985833,494294986,-286792737,-677380099,-471861281,-323025330,-18895878,-1882754609,-1343313161,-2063795334,-1094005033,2147475151,1996487418,-1495169,-895506666,1070360536,-13160137,-556859809,-725634788,-1980652049,-155505726,-337785625,-1423038715,-864227385,-7723102,-258048065,1300123391,-1354928641,-1683227882,2085207791,2046380027,-14288491,1868296380,1983565813,-536949258,2144064429,-651550818,-1148465359,1356472017,-1167336257,1807457515,-134368910,-1832397901,-1007819010,-285230211,-343223533,1177711356,-578891139,-42615558,-1946209573,-1588108997,-175184195,-1730289731,-587235329,-655994885,-1262616662,-1157965985,-805429428,-1613136150,-807619553,1866431614,2146139539,-103821918,2044583741,-286540039,-1112752681,1874721023,-5328037,-2690617,1841158605,441448443,-296766531,-143200273,-1516538214,1861963644,1039891803,1524628462,-364212449,1163656879,-186270212,-578253353,-663822370,1879012060,-41975825,-448823569,2129636733,2109524927,-1614840994],
-		[-1376466056,840890103,-152398355,-4547649,-2107130255,-822108185,-979895637,-1275596046,-17843335,-4051009,-2014382210,-1495438355,-1225941521,7241279,2010492393,-543305,2130703992,-1142972589,-1610951977,-94650502,-1159363206,2130335445,2146925027,-44073019,-306580161,1550843743,-412120346,1063254951,-205540961,1039121385,-536891419,-639664449,2106318801,2138636244,330419959,-616620438,-436678876,2130587997,-1610731970,536337463,905793530,-1889577394,-883036251,-721160204,-1681778783,-18980865,1811118764,1342131930,-772319449,-1343498753,-412910866,-1646006325,1609889535,-72616601,-216212500,1002167986,-20453475,-480412681,-260608024,-1258356033,-130578435,1073197019,2108030643,2139338411,-173708383,-135374851,1866931805,-51948476,-134367098,-134580794,2003955487,1943369723,2134799706,-1150332297,-1971873314,1476107327,1592778733,-36784664,-164179986,1032806299,-1812005921,-501755002,-271880193,519937019,-2109581,-5300498,-1090555921,-355535969,2033086135,-5229,-1116045527,-285385929,-365073,-1613953035,-97886531,1874491383,-699073708,2096883330,-403243067,514062810,-1355291777,-689122883,2104932095,528346075,2145252961,1686010847,-973816201,400686519,-1126993205,2013227876,-1177702278,1027860395,-1262735489,802020794,-268528001,-341485993,-295781787,1836579810,-1342740195,1551414865,-422166545,-257180799,1607719824,-1751206049,-218909748,1847267861,2137638641,-1682252033,-1191707171,-1836160001,1946148861,-288653571,2147081726,-1166624788,-1673626628,2145976029,460652159,1937602513,2128542418,-1828735455,1866661871,-677521,700019129,968686439,2063585019,-42165957,-969695305,-1499751217,-1082400791,-117594479,-912261637,-1141440616,-615989851,1744722938,-1687355401,-36832658,-239110225,-206573869,2080335559,1580191291,2029878782,1941891067,-63258827,-11848779,1030051253,518912910,-1211367951,-679614790,1576996024,-537067715,-337614729,-222300485,1604009255,-1208362753,-205657673,735048589,2005661614,931880815,-65211093,-547434511,1000109356,-637946914,-1083245321,2107090663,-42369794,-1676944516,-894779935,939518881,-421994857,1471642566,-297570,-61349937,1467654381,1995937277,-1144002791,1567194072,-1370376004,-662073381,-131571024,-168105042,-218991161,-1679834978,-1418012676,-274762427,1878105203,-571609501,-1134005206,897357534,526970670,267054846,-1342750886,-8461457,1985886984,-887244802,-872810809,1899757550,-744625705,1565476717,-1712101509,-1223643239,-386162245,-1747026020,-2036477953,-477233666,-743207475,-1787938627,-1912681238,-101449860,-474207122,-1350866630,1648295617,2042494671,-1646411847,1871904547,-2034241541,1992163311,2070347479,-451153924,1670691837,1696610199,-1360073,-402927855,-226992646,1158672099,-1549890819,-153098121,-290259337,-1147146241,-279986178,-914408503,29223881,1460075170,660589889,-412104821,2111807116,-1632108841,-1916770533,-1250246673,2032336381,-1411421185,1389657850,1571782449,-1916813601,773838590,1706812295,1206583006,-608317829,905707389,-1263541441,-677708047,-496256326,-1347829283,661643226,-156246027,-781861630,-1619072409,-633490648,2141882844,1589309904,-4425251,2009022177,-537199311,-1614831621,2074992446,-2192697,2004579815,-277717269,1059962851,-1078068293,2012986298,-1226179233,-1806534371,-1882691340,-1445397061,1853186019,1565521911,1610051939,1329757439,452938141,937627903,-37327889,-911343925,-1451589981,-34861726,-1184890899,-1183243023,-52501551,-268673060,-172303187,2080337343,-67781653,-1692534819,-1377453,-84738241,-403120718,-1962029732,905928314,-1040908726,-673342593,-220987809,307991286,1753897877,1848633151,2028396111,-555747428,-17040093,-1275220124,-32408589,81512575,2145971919,-134261482,1719992013,2050816105,1010302940,-1032544513,1874251635,-1181100818,-1101484642,1053994110,767125758,-286940036,-875094069,-88255547,-469790850,-538091521,-807403801,-135546113,2125788589,2127769533,1270316786,-311199484,-1967917186,1406892332,1853488444,-463701155,-675334126,-549464135,1509636511,-1678967055,-1360289983,-13734286,506867647,-714420265,-1610637381,-8422325,-285263313,-145100826,1496708939,2109296307,-1778450611,-269632513,-1411908753,2129362107,-805414919,1051579893,-218777794,506717679,-2130716236,665189887,1843326455,2002419679,-1076371509,-72192785,-1613768193,1942735358,-134240805,800815481,974176119,-136819090,-71371778,-354649,-1096825033,2055015593,-271877751,-171469178,-347365521,689896953,-161240136,1607006547,-1917322530,2113564603,1723745255,234314191,-34161717,2122110451,-883265561,82124748,2109597562,-2032248067,1071560844,-411308421,-139217847,-185145509,-1183441037,-1652064262,-378602519,-1351422980,1135583734,-690423332,-1075928457,-1342267953,-331362561,2138306015,-92522717,-47197345,1469562823,-856005339,-310075777,-1478920259,-20466207,-593363169,-708964706,129507195,-1127481347,-68915365,-815019530,1493687492,1848900490,-1798619218,1903865431,1168953211,-344140292,1776904787,-235732993,-353378093,-305957145,1878829250,-346384149,-34669579,-134632450,-1151862916,1068200059,-1625308751,200801256,2030026646,-629675050,-24911873,-560661089,-418416897,-355733761,-679483650,-570442300,-13443119,1837082107,-1447091495,-595705031,-2065138873,-1217448477,-1152127129,-572538919,-78053441,-369788693,-540131601,-1254129908,1677323461,-924854579,-7995641,1668214079,-433066177,-656676741,-1118793865,1991556831,-1833263205,-1746076674,-457203716,-637667858,-1429341741,-76048389,-2908202,527132591,-369164442,-4260001,-39483745,987725685,943717625,-1106460897,1594843263,-1363641353,-1644397183,1777747740,-201431106,1811670685,169856399,-841089657,-1375994908,-1619134481,-136187906,-27891790,-167928578,201008365,2003746291,-882115378,2145345886,1794591807,1946147675,1545007022,-584992519,2117852095,-156609765,-742400673,-61811915,-96928933,-2437075,2126727602,-1150305237,2062934265,2128599277,-1067917862,1610055634,1299708607,-841224201,1526588959,-172507653,-182517965,-1694515395,-403087909,1604320212,-141570514,1733291974,-27334919,-126353722,1334586751,-1301621382,-735472671,662679615,1805121247,-1075171431,-1108247975,-16913001,2080042475,1032916451,1575836811,1056959311,-71959443,1593819109,-89215321,1603102198,1843456474,-705713293,-1096940072,1576959690,-824447109,-1454773,-1889763864,-84201474,-537362888,-101810501,-34968577,-1176633487,1670299098,1592130267,-1130366933,-105908778,436116563,2005244653,1050148638,1960828441,-1432887570,531404395,2107339989,-1296374069,-1076791401,-1673821499,2132010877,2003546111,1874261036,1358937883,-35555345,964440670,-415957011,-48107610,1337839607,-1449329900,581697430,-4199003,-35097334,2105505251,-450901140,1768813991,-1697599762,2142437232,1556010619,-1733572282,654044862,-92668481,-822141010,476555765,-694561174,-805490878,2141048489,1473882812,-220334481,900703197,-118153222,-1175486478,2142830335,-33164681,-151673898,-1314079121,-2021671937,-537018438,1866183433,1509029499,-1384192402,-1722352114,1777006573,-386461738,-773117966,-340931329,-102763873,-405017044,-3168403,-1376091506,-1611497717,-412673409,-35324219,804683667,-1412017257,-273464389,904888318,-1687183907,458661693,-422584993,-169963681,-268610856,-1879052819,-713430049,1861023031,-1778868873,-1143256621,-305951445,-1427198995,-612385780,-169515737,1910269636,-189901398,-1688209153,-319310374,-1856725121,-1300402179,1053466230,-284825363,-1076102161,1336402607,-104874569,-538607877,-1076639523,-407963177,-63441051,-276968729,-51882301,1610405766,2125395903,-285689898,33423226,-282853965,-1678060674,-1063649318,1812274021,-311854446,2137127374,1425797098,-680052901,-858391209,-4899587,474844943,1165967341,-622369164,459753396,-78563572,-1042325575,1979701034,-1691883587,603944415,-1259898377,2077390749,-987807904,960491517,-1174433793,-1209273733,1875850871,-588531339,-1581580570,-480519221,2009034417,-709995857,1876295529,954968257,954181306,-33653058,-1142105106,2146267101,-1652883801,-873826322,-1896390685,1979418538,-521538307,-580010866,661286147,1734057949,-10618152,1842872309,2100805625,905858558,1207023285,-1653443343,2102361069,-1612722180,-674567210],
-		[-28576025,982514622,-1421869717,-67603252,-71615757,1070956396,-479470,-3277727,1854775167,-352846355,-1166053538,-615125761,-547374331,1727246735,-404175058,-283836551,-1821489429,-1328398609,984776639,2061357037,-1628118418,-608174422,-559955971,-237008181,-405649409,-273174817,-4206717,-176378423,-632981161,1540269819,2107633949,-1091770067,-1244284495,-1543512126,-224419459,-982827406,-1438982549,1272413458,1606121331,-1217562215,1384883149,1028909190,-6194737,-1084231501,-945891521,-1108099205,-609557577,-1078006346,-69992460,494885851,-279397905,-96012311,-1095422209,-4238399,-546969217,1508333411,664721371,-965841935,-1229534232,-1652998995,1266593207,1035991929,2122511533,-135361857,-421110633,-1880672322,-1091830785,-1678043785,-1527911578,799076095,-77883410,-966796998,-13915420,-9633892,1828707950,-1073764465,-1966102059,-32869189,1072488798,1502604030,2012145907,929716991,-170345105,-1623217223,1838742507,-1174896656,-72035274,-1216365457,-469827599,-5800069,290447123,800946163,2070907836,-1075678026,783932094,2119926205,-15757027,-342233169,-1946253387,2113616559,-541679619,-1617972385,-604248902,-316691239,-1083197211,-1235415571,-159010907,2147458534,-688026753,-2033463788,-5353737,-1540637731,2146957421,-1627424909,-1713402932,-644634321,-461685891,-1612481410,-2024294018,2141641726,-1744882341,1454138010,199134171,-1554517057,2061051886,1471515646,-33605717,-28091532,-251669067,-964962659,2128503544,-570490911,-1319559686,-16189441,2084364187,-1107445775,1358675934,-235364643,-549618743,-1565206183,-269504490,-1851853858,-268732749,-143158323,1598806703,2012725215,1961719789,-268582930,1870375767,-384357637,-1895826497,-670801937,-878780708,-152647737,-1231719429,-237771172,-172401321,-1194331682,910031860,-1628050820,2104999871,1202976062,-11017527,-596181795,-1684312069,-421692353,-1160841941,-476053522,2033056638,-1705384066,-1859670729,-1999650830,-1347423515,-506646801,527804388,-92440721,-319980321,2111462623,-438637774,-102107682,-294347965,1874589439,1995440127,-571380594,-579994498,-122046401,-93355538,-772358617,603702715,1067380087,530382827,1802875903,1868545277,1541398483,1583349111,-283200550,-2131231085,1685733356,514175483,-1090576987,660623224,-19055367,-707332327,-570606627,-287506803,1925275961,-1217400131,-295715841,-1097340449,-38820579,1861820919,-134939501,-22077989,-538968234,-1353778178,-74464425,-1194536961,2063588095,-87921849,1937334188,-57028905,-1897141765,1868523510,953891445,1502281453,-122095938,-100745288,-1073799841,-1997562563,-12438025,-818891729,2145771190,-1617168651,502292305,-11436289,1710110322,-1083553392,-1244668421,-109652021,464666622,-58772001,-570688194,-1307840003,-84093154,2071935391,764392101,-136358353,1036827646,-1769899,912248559,1709600704,-85992577,-610918409,-42207769,-279513157,-1695769449,-481576467,-807833809,-2047918961,-102568367,-138414277,-78890946,-267521,-1687162915,-68255785,1735814263,-419971332,-3533892,-1566577675,-3458113,-951567621,-771842081,-539614103,-547357807,1995169263,1738485175,-1291881794,2078489557,1138628479,2078015342,2097088474,1852666555,-691535963,1559485181,-1946224551,-110644713,-1357985291,-379345557,-219579405,1069875156,-1411450600,-336593281,-101720070,1714339492,2010021183,-1212285258,1728051688,-1073859028,1307933693,-1092731111,-966395090,-42476738,-34079275,-168920237,-679779565,-1347430913,-1614906897,766369747,-176524821,1972358335,-1342542061,-199944237,-29891588,-184550411,1342094203,-543950409,-681351210,-673710121,-1489330770,-1507871268,-8564995,-1097958450,2126420671,435920891,-738331003,-35308757,-1357513748,-303906833,-484981777,261982841,2100690337,-818588585,-84941833,134201311,-394807561,-1485373697,2108528333,-202451129,-475015177,-1143480401,1610349311,-990103065,-1346603075,-1915158557,-73427027,-86122661,-302268805,1006168121,-1510876195,2084884349,-1221861606,-319828867,107634674,-113810051,-739952709,-937820322,-19153034,-1116210594,-1006634433,-67688969,-1616438731,-1231632419,530739951,-1149404744,-296773658,-252373010,-1036519297,-151519315,-1612022919,-412381441,-594153499,-1461321985,-1074270657,-1947770882,1576517629,-25182320,-109849723,-881340935,1071217597,-4367118,-907452634,-288713122,-1712119809,1314895706,2054175697,-1174480897,-1293459857,2143358820,979366567,1430247340,2131754206,-1049882,536180222,-1106510873,-1973521556,-1382131331,-41484313,1576467166,859624559,-44613775,-1231462936,-742736918,-1272972993,301399423,2141697703,-204643857,-1045153829,-322324281,-1914773738,-1879406871,-944256043,-304729097,-1686325537,-823665733,2133667831,2146922798,2111270391,1338608859,-805322769,1072946591,-277679119,-271349427,-1980383250,-614617649,-437897906,-1207974274,1660929863,-961700625,-403089825,1988097862,-323590218,-1883799697,-179307528,-221974730,-1192427841,-101193867,1022699321,971956730,-642156834,-1345060916,-14164174,-1922403540,1859546065,-285799174,195689293,-1122313,-1130399015,537810649,-420574314,-654369320,1585426043,-2962434,851324542,-974164099,-34573706,-822613569,-1753933317,-305410094,-315768965,-1888023709,-19923225,-135007109,2140128894,1172298551,1441261065,-1248972935,-168632357,423613407,-739028246,-43615490,-33857,2147464552,-34956309,1331636007,1465875962,1833416569,939487986,-36242977,-1234625569,-1846743058,-107544836,757922995,-391223703,-50893127,1475313215,-281252369,-8927251,1340073727,-286569763,2032923284,-1680935817,-5801489,-332180023,2100887283,1833956348,-134333155,-572526658,-710685285,-155192625,1962655214,-147941434,-1722585210,1607621856,-760011059,-576789,1266155420,326764796,-84116012,-1208734409,1207664549,-1617963013,-1986134021,1607652607,1136459708,-1145191107,-79785010,-1101078660,2123320795,-839967653,-1611940637,-1141815746,-101811745,2095978207,-107743441,-385889667,-71472846,2146950585,1001860802,1675754735,-722583936,2010177510,-413166035,-207398065,522790015,-137496449,-271835651,-289324417,1073694379,-304742651,-1686271494,-1619715724,368553411,-1627533828,1807155002,-369411462,209043163,159231864,-1148275224,2106572079,-877800962,-1758899987,-1969867140,-150073737,-135137304,2027995135,-537575567,-931209251,1441594397,-542532097,-1443500687,2136473209,-1342669668,-415912914,-216859185,-1782581484,-84108357,-1272193155,-1644253701,1447033662,1861442928,-1118822466,-613372433,1047355468,737000185,-696346689,-872980929,2146434651,-405131949,-1075841033,-1207081024,-809115717,-169249225,-268971045,1184493022,-1778421771,-1912881227,-1073744965,-8406362,1067202411,-1885357856,1316748471,1209913145,-9031806,-10788871,-25167115,588820475,1862137851,-591598678,1870599144,-1187774590,360607710,1585093887,2138544246,-1862344833,872147440,-964805763,-15728898,-919866188,-1376237764,-1208025694,50053024,2134441438,-2019043107,-1079263849,-973108551,-1074532997,468529663,-847675781,1881140156,-295928339,-202375253,-34134381,1798278703,1876368125,-51314883,-218124338,1926233451,-352563075,-21559778,-1663137801,-336013121,-631422980,-691016019,1848376312,-1682064202,-1229406283,-907060289,-2103329,497383327,937949765,-89161729,-327161762,-76038685,-84046917,-967385145,1979166942,-75567625,-101929989,1018159038,-82120411,-2082264449,-276909077,2134603388,-531891741,-99029454,-441843726,-1611268387,2125951195,-1709195307,-2885163,-160811653,-34620545,-1510013249,-1327038229,1845493503,-562087297,-18517265,-1502897728,1962729977,2122084702,1849327611,-489142497,2130624254,-1663676441,-1379664010,-508808072,-646447671,-1682866241,-65013287,-1148091657,-1521373709,-556286016,1063779801,-1494884358,-571506709,-1015417251,-579012954,-2095071937,2140494582,-415564129,-1640630709,1978521525,1861253045,-531120642,-445802633,-698321156,-781420874,-404694105,-872578701,37350122,-1209683443,-575767945,-128002366,-6453260,2131781215,-371296525,-204718509,2109202038,1307274751,2062662539,-505528625,331347301,-1109000810,-1678158202,-1357514763,-828777187,-134648322,-9822241,2029780571,1966276154,-298189382,-39983714,-135334578,55963625,-406176577,1811410618,532043609,-806100996,2007549435],
-		[-1342288418,-705171491,-587727937,-406350001,-135483877,-438321857,-506729234,-33849754,-598917249,-195428898,1048433648,-41100706,-34615082,2121676273,-1694519553,1789877693,1995333623,863392921,-574117257,2012091833,-1681003563,2147155669,-686366747,1072208834,1711138021,-74072179,1337956147,2073966526,939441903,-285443851,1944820158,-1480607778,1744760319,2147362731,1377736697,-169371665,-42737729,2114280887,506587442,1025997558,-411077129,628845786,-76169565,2007460575,-882069543,1448315227,1556084204,-1151600710,-14177346,-259136105,-1913675849,1189673468,1669232763,-1124468136,-676758083,1866067705,710514649,-679543302,-121905166,1491684160,-2744395,2141015480,1061355421,-14618561,-1205918373,-1489994290,1859891181,836681658,1778379579,-536964225,-541429875,-93388873,1073699613,1575631898,-103825845,-2137837,-833163778,1978354807,1710161343,2124077013,433535983,-1367900677,-17842359,-167319289,-34321438,-2114880045,-405333125,2113504639,-1081204753,-1179453694,-220363873,2078599987,2038424687,-54038703,-52767938,947566891,-102051585,-1174480391,1605303151,-1494350263,1602190824,-1262662505,1273634431,-10767073,-43061284,-842432581,784947958,-268763396,-1090528361,-676462593,1392325529,-1687030960,2012869885,1257234358,-1580341419,-680269204,1060633575,1776256959,-672138755,2029189718,-641478923,-1396720385,-1625333762,-814483780,-541069441,-1437056198,-1652839048,61796051,-1141576273,-71315713,-1078176991,-848954394,1878916885,-162615366,-29950609,1593434107,803131100,-641759681,968871628,1608429551,695115771,-1904623887,1844499263,-2796037,-578830491,2130132421,-333728838,918155754,-1082209875,1566789359,904370127,-587505925,2146766494,1563854463,-1712472177,1992187647,1354583444,1442444906,-39845948,1876938235,-1493733381,1977067519,-740445751,-1223924317,1457257903,1798992635,1822374779,-1751204019,-1148452873,-210503441,2147465020,-860407405,-1613302547,-1103496508,2092779516,-1245851842,1072250795,2110848363,1473349530,-646587539,-377901122,-1658884389,-226566301,-1241780753,1467907749,-1444020289,1056810866,1072299689,1810322538,-527073,1853282173,1681310968,-1208066401,2138267055,-8840706,1668676723,1840181763,-1615868965,-572573886,-1664484371,2075097209,129094890,1934163449,-436371662,2136997846,1974877677,-574850346,1958467569,258691957,-331244338,-405861417,260046334,-43455190,2126377968,-1241942032,-173564940,-2099258515,-1615331433,239003132,2004612971,-708867234,1018128470,-806897925,-1494954024,1056435684,2096097256,-193462764,-950554,1654510293,-71314753,979402230,-912196681,1777689339,-2055473685,2147018679,-620791827,-1309187151,-38814763,-35364001,-671978114,2135764861,-813052314,-142606617,220692215,-16785425,2113396702,267752265,-1562067261,2040778443,1340141349,-21561441,-34735831,-269387869,1319762621,-274927665,-9344460,-341874772,1169555455,-360056653,-629805153,-285463169,-907022040,-1766461889,-142608898,1801436413,-26780273,-1151415235,-1698972295,-17827603,-154152993,232533962,-21103654,2108599182,-143393794,-203558275,1443619823,-1209041410,-159781785,-141644951,130498035,-142909505,-277507586,-1008468836,-270147669,1424219513,-654451617,-403157393,1897174868,-73039921,803730350,1802950014,-571563181,-1142017093,-1522688838,-702611675,-1118001732,775413709,1744779699,-1157667620,746552495,-525889,1925133763,-618612678,-1494707009,-127542274,-738987204,-1627460633,-315631634,-1237338645,-710475956,-1110610226,-92283157,-183286426,1467998131,-574906437,-4372997,1858032276,1485987194,-223349765,-1883308305,-34740257,535032691,-1176308487,-226295945,-234785154,2075725806,-67868737,1978596319,-1885901755,-102763121,-85610753,-182197282,1861071611,-404820452,-1713448967,-1081082669,-689470601,2134481276,-889610297,-672825755,2004538015,-475919412,-1093243359,-320651988,-71871518,-206969909,-1709849,-1751130494,-472122162,-345544332,-941658164,-704815160,-924849345,-561546498,-1157631010,-289420808,2147077055,-1780514881,1777856315,-280604681,1020476993,-208262223,-1225462800,-416285703,1551331039,-1711672366,-138185801,-581694509,-141663257,1570763964,-678433307,-941671619,-60835369,-1682579763,1876376199,-672153666,-3181953,1461094249,-492902481,-142110829,1955002079,-356845029,1462415358,532347763,-74139305,1566391738,-1791049758,-143366339,-176426313,1878579946,1568317359,1739545303,-167781398,-1879406973,1862182321,1048030865,-1195841029,-268882180,-848335043,526625761,1067446271,2129501756,-1121992771,1740571646,-437390979,-659595786,-2053378106,-1193279805,-2075420723,-2081502225,1473183210,1835767707,-1100141315,-37241253,-1879966351,1476308791,-1351104974,1583330234,1855979493,-1372841553,1995963261,1308601846,2140863462,1226521208,-1485129937,1986257395,-173689373,-21165,-184586701,-8348303,1683685363,-1762011161,-1580713,-1145709078,-423777375,-581971201,-160220531,-1617768686,1627159422,-936616031,-1091123622,-138412579,1605097448,-1682998217,-542931073,-1683047460,-23199789,1868771191,-170233745,1574592511,1969945981,1609547503,-1073808521,1909716022,-1237337249,-823659533,-5326597,-14844119,182190011,1992247742,-1890726923,-1947765380,-344445507,-50909342,1784624111,-558897403,-482847909,-340038177,690606051,-2018386169,-676081977,-104354437,-537414210,252865114,-671224906,1001373183,-1571045570,666490997,-9199717,-330599305,-553836689,-1524106951,-12665029,2109325016,-211264136,-139887380,1825290330,-1420583577,-756836709,2022304693,-384770603,-335567633,469605326,-54561093,2012339198,-304182303,-69526545,-538652693,376171975,1727821558,-21504775,1856763643,2039934004,2135882222,-890423298,-1189132358,-809644225,-537592282,-1095078769,477216651,-369720659,-33570833,-1297089281,-1174440014,-136878305,-907906499,-468260097,-42215202,-1906131127,1270281175,-424020135,2128115515,1045347799,-1276346506,-134463632,-1250582531,1854914532,1920462318,-76025177,2068804779,-1143333378,-425798279,-1146081298,-185630738,-596427793,-697306407,-288515099,-1090752529,531877721,-1879605685,1906963967,-67180701,987652053,-44041219,-218155923,-1615464509,2130378502,-252276817,2067488326,-290069265,-692077059,880336327,-1086484314,1054974975,-42732613,-807542788,623883923,927706079,-1483934795,-135269041,-1352991129,-764432897,2102656635,2060442822,-537794193,-281021520,2147128615,-392695989,490175228,-50995379,-2703881,-1082263595,893975437,-841744405,-1686137041,-668811,-136445989,-46303430,-858034691,1626397311,-796964929,-1684086317,779017071,-1349284741,-547512649,466968543,-885065225,-1108381723,-1752515665,1840212893,2012714858,-773860607,234575071,1608162813,2125977383,-31064134,2087377407,1237138974,525983485,-61027761,-1078826753,-1644186646,-1166087768,-359532593,-109911151,-679661574,-13879112,-142714890,1228356043,-154283013,-1155802185,2011681782,-270378311,-840337441,2007657115,1071373758,1526496958,-629588442,-1816306578,-1078396484,-222412937,-354418785,-6045516,1348725887,-12682379,-1043408102,-1276126436,-675439785,-1797271563,296710004,-470101585,-492584970,1035894103,608092187,-188485726,1936162557,-147853924,-428392451,1020886395,1998481860,1987051519,2138128343,-146855703,2008399741,-12681220,2061475635,2147219573,-151085063,-1665409320,-1342841289,-75539109,1541406702,-286699564,-446386946,-806738203,-1271035139,-61880995,2111812061,-847273217,-21565491,-407973138,-1646100505,-20448325,-15752705,-1345359967,1257238382,1987648027,1832597459,-219170893,-940622097,2109324723,-176562940,-738335683,-1346400260,-371269917,1257201598,1616465341,526727805,-518153,1995235197,-1363071749,-59916791,-572080295,-1884160002,-306187873,-1174583446,1561917397,2109795831,-948469859,-109068395,-807408146,-1427877,-1176442393,-143788681,1716382045,-537017473,1476225533,-273547786,-1226837799,-83888961,-160189633,1514142937,-546833086,-41103367,-1127549265,-135670533,-1478623490,1957590821,-2080615297,-173172503,-350323974,1968800667,-184964537,-1611403106,1407778752,670278623,2006938595,-1544561411,1743688283,-747650665,-604209205,1149236492,-260065641,1742712627,1540901106,1440385693,1036900341,-545649291,1609558417],
-		[1349761599,-135425549,-136626561,953414879,752875434,1575696095,-616634018,-302254129,-1074790557,-545468994,-1094715587,1593571130,-85426434,-671121673,-1617076886,1865983206,-1077888009,-754025227,-163066506,-1823506457,-1376141590,-708825559,-1105723401,-478374601,2134289162,-1209543486,-1973383617,-78649281,-125870212,958724159,-6825089,-54550577,1635741681,1017922493,1065279485,-362204291,-1611236234,1796872156,1424590167,519470908,1249640383,-45106049,-20467800,1000195835,-539903522,1635496926,-660467203,2097151103,-667452513,-1984581651,1404942808,1002426843,-276859930,-14225349,1721492221,-3182893,-1377890314,2046688682,769552325,-654985217,1733525456,2008644799,1798774647,473036267,-1269892295,-1252074236,-1180436670,467591163,-1363689850,-170175672,-641794317,-701035649,766906056,2004677630,1463760413,-1412231749,-2088763474,-168846605,830256959,-1396247591,1593177723,-1610642441,-101144338,-1375774775,-67133528,1180639079,-268460226,-214008113,-405276451,-879575838,-93982994,-12482665,1887336273,234327550,-1298305033,-277872969,-1544133121,-772542469,2109659133,1105460911,-81825054,-404946961,199748540,-1147735058,-88344593,1810824859,-4263218,-1111198946,-145235615,-298648195,2138255358,-1417859653,1893399959,1540545269,-278503477,-623166053,-809568513,1703403385,-1829771521,2085502439,-1078134341,-1078216797,1341639887,-46141500,-86189761,-1392551955,-1885345812,-310453252,-1212763207,-1702179243,-721905922,-678829064,32669454,-452935833,-147147044,-59836449,-274859303,-70294797,-86687787,1386446659,124646006,1811784703,851828066,2080308037,-1813708934,-199888954,-2013856070,-1487020627,2077488974,-571658339,-125830245,-637899527,983257995,-547685326,836747115,-554700881,-437841967,-663767151,-1510741267,-272942095,1516927998,-1344380002,1869529063,-182763682,1035040031,2097016699,-1412018261,416247735,-847612177,682484185,1357884927,-105059441,-933602341,964515038,-386140201,-1983913992,-121638149,-831479209,-1677744889,-1443578213,-805314561,-574922833,-886661242,717131203,1068419643,-2106493319,452672111,2036751807,-1627471521,-67112587,-79743905,422608417,-540983315,395765654,-278949202,-44142714,2103203149,-190367786,-1659945359,-878859398,-28139537,-587610325,-274156453,516525583,1064726012,-44378140,1551531767,937082591,1073295349,-249563465,703448491,632199661,-4210696,-1610629250,-1778630693,-86247217,904273398,-1125205029,-840697893,1586478587,-605065497,1575162099,-304435201,-111215141,666887161,-77665442,-1221085705,922707710,-708952213,1296803197,-1906469506,-1331719430,2128055401,-673221323,2109603694,2136692220,-875024723,-259068681,1424881401,-1765249935,-1712095402,51702879,-1074078148,-134568456,-930502165,-1697055330,-1376264337,-1309263907,1809766549,-1392519253,668802430,-349318199,-1645281796,-1701060948,2147219455,1784378479,1921088821,2145295694,-825239746,-1481061313,-1544817159,-1119904678,-470898857,-16934378,-135595545,-686826116,-476184720,2122296795,601508549,-840983617,-1620432422,1044610404,-1556218909,-1620378194,-4916236,746562943,2146680831,798504427,-284459531,-1080992489,-537356545,-637938725,-294257185,-570427337,1668930458,-231150612,1958987647,-941621255,-419454541,2136044785,1773633405,-1615369059,-780371018,-169492925,-772545319,-396418691,2137079805,2111858066,1070856820,-822216741,-1081119297,-274280705,1069390659,-1812836417,113479373,-1616913411,-2009127178,-24103039,349126122,524287547,-190976042,939454453,1996476863,-141115398,-67149965,-1075483729,-828925702,794820011,800943758,-1412433958,196764862,-256710097,-270563897,-77688577,1072313781,-1895842186,-1404567841,-50417801,-1763209265,-1099317298,-313575715,-18665,-403117061,-1229075713,-471020588,-26181834,-270902678,-1823350785,-542129190,-1627460721,-1907135562,-925892971,1484758862,-1780487558,-625060120,-342233237,-87012498,-1073774801,935012791,-36733809,-3223705,1746681830,-887095857,-572674113,-86343843,-1052873249,-811616388,-553451553,-1378156557,-26027925,2137783216,-320365093,-288508935,-282773774,-1614841387,769368733,-1519447559,-419455042,-480563969,-675883155,1040113559,-1345674085,1672707326,-1360037521,-981600706,-80756748,-1803690242,-541073547,-1101029819,2051161547,2113850804,-1085867525,-520211025,-637509,1091812116,1971265403,2071428346,2013056621,-1380688401,-1469096253,-1343423437,-1935919242,268418067,-25782216,1567618555,-303605762,426486727,-1276250466,1606287101,-1073754162,-839069865,-1140865448,212524410,-285491363,-304463905,1055059510,-1745028549,-1163035504,-829495334,1570682388,1046732431,-207389115,1553432799,803776447,-273069527,-80298083,-1113068033,-1345339665,1073155302,-840435747,-150995714,2115726453,1973286143,-2491959,-1278232641,-1503726721,-1142334030,1999568268,2124386275,-77631939,-1631790880,-1161955400,-408168321,-186796201,2113101716,-604602571,780294846,-1209565587,-1087767169,-479531330,-405865090,596631518,-637684749,-1653954019,-1461761061,-678462262,971503070,-170961122,2012200821,-134615281,-1107628423,-1233913956,-1829538243,-269369473,-277897253,-1370062871,-1854120105,-173148185,1474280215,-1322814482,-1076913299,298569724,-2303362,-1678117291,1646243988,-1701891349,-1002441136,1939518489,1065320415,1724890846,-1677722427,1073279405,-673055787,1204524017,-11865031,-19300868,2122276662,-358318185,-16779397,-167772953,-537469873,545078926,792588157,2097020919,-592478532,2046800627,-1745592385,1398964054,-394360833,2139061550,1073199023,-550547479,-1100034585,-1080408118,-1510744635,-692454188,1541379791,-70704068,-443352164,1777253823,-671293525,-29967170,-997332633,-5000871,-624494661,-1969251437,-576327939,-290554641,-145071512,-4465545,-6357393,-1376060817,-152068121,1429317103,1928028127,483889067,-944046356,-604588657,754843647,-1275661826,-83968516,-1497419217,-151013122,485874410,-706290009,-2235998,2112356333,1866907646,-202318598,-1175682473,-138555528,2079718902,1940705437,-210041608,1992809423,-560646530,-167845977,1203584823,1891350366,-269662211,-1185483039,-214094623,-5419833,1933237998,-1629299721,-554958951,-489328769,-289464337,-1149493682,-269222788,-3115081,-563245827,-1350734865,1492903379,-515141638,-1558773839,-1510947906,515833053,-23603270,2118035436,-403833900,709621629,528334751,-33867783,-170074993,-1861094977,-1115970067,2144325615,-176208525,-277225483,754907297,-1520063682,-422469634,-633402647,-1554225001,-893389895,969487222,1907342275,-110100547,1152384766,-2900609,-2042692611,-541863963,-88748035,2088630525,-180951810,2080298487,395559036,-1292124925,701349631,-570523938,648408511,1338930743,-1751738385,370761659,-810691219,-279515362,-215557869,-542273364,-590377419,-1073775634,1744043748,186891261,-1157636120,-1656327170,-21160086,-874787235,-1762920633,-1713627283,86767950,-16820009,-278987521,-151136793,-814749363,-1401575877,798432055,-1214186965,-537464849,-104175693,-1729699842,-136579213,743948131,-1209329606,-541661265,-1381536427,-1664235564,1521417711,-37751009,-303073300,2075452782,1953440654,-1619124484,-12823067,1492835527,1140062995,1307573885,-622925571,-406402309,-1434639593,-1111558411,2117154418,2096332471,-4232577,-639640220,1207591871,-991266852,2079208571,-6448321,2136518543,-7549989,-1143808257,-941949009,-654966929,-2140433,1139068230,-201673138,-2019823129,-1543554846,-127804484,1041984156,-253755426,-787801,-88867602,-470951691,-22619177,-1152288928,-537929987,-663060497,-743804034,-37750947,-581477538,936873818,-557058,-95434249,515829405,-1879973889,-638126727,-1094419716,-1227005478,-823473034,-1277217385,-1636057318,1250275178,-313033345,1000340951,770663870,-67910017,-1499614917,-1163666587,-306751505,-268474369,-1617692089,-394934594,-1153696306,-17311915,-587613337,-580944394,-581846421,-9885,-671762609,1811675374,-448594973,-1652868356,-1276515490,-1444152097,2131918761,540733312,779859361,2056635375,-204591217,-1162025601,-876678401,1876603191,-1480861748,-1817852308,850385403,967233490,1581071217,-1401072805,1596573634,-1275449501,710926333,1341636574,-1099174209,-1145047572,-1419772164,-1648362259,977204279,-143704211],
-		[-1076929079,1979314155,2061025778,-2105473,-404017154,-210284570,-85726277,-1222902145,-134474945,-1026140209,2146589363,1066383253,1339913199,-609403217,-74753675,821917923,-571262490,-1082132345,1298527741,2128722797,-69764481,656798711,1539768292,-1293156353,-338231857,1978104373,-296095617,-275111,-814172611,942911966,-50675809,-872631303,-538136837,-394788865,-63975431,-431626017,-1117962760,-4196419,-716702722,1417082774,-1372070407,-278283841,-546603820,2120807901,-241213602,-1482261071,-540698881,-1216895490,-1358984322,1911395577,1597337392,-508070027,-35653981,1936784639,2008931683,1824405441,-1614940260,-1782581867,-1624244231,-876620547,2018498808,-1632373253,-501668164,2110766370,-1480690218,2120193332,933462969,2050481386,2005877981,-1212752389,-619087702,-778679113,1323607669,1876587839,1070563189,1999363485,-293632065,-742611142,1967365649,-183637061,1048571446,-483008641,-178362401,266563393,-647278618,-147981381,508361164,1608512988,-625697799,-26231991,-856995157,-63644673,-151136260,-261899399,-176269857,-639248641,-7910145,-43325673,-46268440,-377815597,2002727837,-2029146449,1274986237,-1125129971,1966857611,-1620087065,938425015,-366882309,-116475350,1877113277,-471911941,-438513172,1202814511,-277676129,1994040079,-873465601,-286523737,2146425209,2079192765,-403184257,1472347890,-36490,-1093272189,2146434941,-1077443043,-33820740,-250364070,1593407777,-454039041,-8408344,603813308,926570862,-579338241,-823281665,-1109446729,-1084229009,-1472272551,-337716190,-283107547,1068695451,-403316873,-275850377,-274795309,-1415591940,2106191711,780377337,2105421756,1808523119,1834970611,-1093028973,-27528193,-1263607889,-974128848,1529642750,-1486219883,-130704074,1063632853,-268550258,2013094849,2110247955,15941220,-541591947,-2017995401,-134791218,1399582197,-571602396,-44958763,-23090113,991815935,606562630,-29376554,-998247427,639888315,-654910561,-892103263,-613950785,-1214279186,-7518349,-402665523,-741614702,1793850126,1517225391,-537469442,-442556803,-1145194681,-539495075,1730273279,-134436865,-1303327757,2094923437,-4270356,909080148,-405409798,-8521729,1711258554,-805383555,1939072447,-1670928161,-1141383941,989439955,-17843503,-1925755015,1062645047,1910210025,-414777608,-591929414,-1252457904,-404458497,-269239877,2141188031,-622985233,-648039476,1013922908,1941798629,-641927169,-1984823631,-4529412,-276912150,-1895879220,-171966533,2012950921,-1240199778,-831128628,-446241317,-1588726915,2144677603,-824217601,-1661473538,-42275330,-591928403,-745547315,1392553588,-117481801,-189923473,1230913311,2127551859,-37904577,1465542873,1038331833,-176220191,-69359623,1073143485,1743514298,-1396711753,-277505028,-1024612909,-1237484114,-50334738,335012287,1739388329,1656745899,-382233713,1305062039,-79691834,-167905537,1333622666,-1614003713,-139461798,1356319967,-1162958061,-12714021,-1293202281,-1328021549,1843781627,2094919406,-1681983521,817853437,-814907953,1062890167,2113141381,-335616284,-1162061077,-68191151,1794499743,-180626964,-113189597,-604324928,-823198872,-1561079381,-1931110941,-29923883,-647848313,-282760452,-1745670425,-84033593,-93889756,-438486275,-1147142273,1858674683,769252828,-1009111046,-1309382939,-1748272957,-29574986,-1182994241,1597228342,-34483005,-551688627,-767727635,-815300805,-75538967,1858861181,-1348839305,1797768149,-1084632004,165865359,-562573911,-621183361,-1076273834,-566788492,-541606021,-805630486,771028926,-381691474,-1174437911,-307275460,-69214749,-6962243,-546076706,-353518881,1955033455,-452996577,-1615478819,-382207502,-914885828,-120094748,-964337957,-341885705,2130706164,-1375864009,-898389131,921918951,1647246390,1727392702,-201371649,-75629841,-168407385,-1766610245,2098607995,648516605,-134742181,-642744355,1669098237,-33736337,-311062529,-1409361034,2131754479,-134221314,-588699153,633794530,1056734045,1441214363,-1077485891,-251183120,-4999363,-1208239118,-1281688721,-402719239,-1644184212,2145361894,-614609997,-397033481,-191365709,-540149828,-110893708,1054728315,528457668,862448092,800373758,951890023,2008917464,1463506405,-1931086497,-3638529,1224671165,834625535,250598239,2003344798,1206603102,1992598687,-1809842334,2009890744,-1352212567,-424427521,-1175207463,2012081407,-1162121607,-288899075,-511706657,-1342639423,-608195212,-679846033,1575813047,-18490981,-59019278,-273549992,-1522900343,-25184291,-1729954153,-1079069026,1721728510,-967909393,2089806839,2130045429,-671264143,-222699106,-27297706,2075524339,-793084193,-268984879,788096351,-624048962,-206946497,1607218675,252497767,1071212670,1508503036,-1707476866,-715530393,-503498924,-7364911,-819666689,-542499764,1623180253,-177484935,-679481868,1014864851,-166341817,-1117827,-1752711812,-283199245,-69209866,-83894277,-990110477,731506686,2025028446,1515144597,2058959207,-1587440171,536853181,-1049825285,-287313489,1219886527,-409377602,-1722564653,-570726753,2063588819,-570721170,2011463663,-84795905,1979419615,-1075996228,-1789923673,-1891632337,2146875100,-1749418721,-1040495779,1451177319,962830589,-1884051160,-698782010,2061322551,-1090687901,-603993125,741793739,-31460355,-138778190,1778359727,-1186497685,725613411,1580103419,-301992450,-327812993,-909258881,-772554787,-137284561,-1090733642,2071959863,-15861035,-746586337,-526394637,1600634590,-1751778436,-688727301,-16926535,-182748357,2147200835,1641380221,-1897378035,1397179975,-480662677,-336073058,-544337655,-315270714,-255443982,-270098,-540817039,854682505,1892515771,1840576109,1329020889,280622076,999427254,-16814245,-1107573522,-537926038,-1045255715,-35988788,-1610620061,-1835083617,-184700954,1407135647,-59277810,-1729454337,2036272554,-817910551,-401133989,164478457,-15162501,-1751255043,-1116211211,625822868,-940263121,1877826995,-832218114,-4440969,-1144791057,464501079,-34026155,217313277,-827541,-17006601,-434749458,-430833833,-2085146745,-8544914,-590505,2107244388,-810291974,905714141,-287515105,-545532804,-1380363009,-15960555,1607925117,-808609106,-17564680,-1543537172,1576663134,133830461,1924788452,2012171197,-16875851,522942250,-44053013,-1108769795,-605190029,-98473779,-944818177,-1079771205,414650869,393081463,-1108031350,1464723774,1873854197,781678453,-568697239,1727823149,-1124470913,-1582659164,2132983534,1606364662,1805484002,-811902255,801369679,-1698794067,-59235101,2123672697,985072625,-30476428,1039997395,-85856262,-522231082,-1733042197,1216069082,-679741441,-134372015,-1446445377,798992095,1055212493,-478285953,779992311,467917821,-69537411,-1135089695,-1100769,-44586881,1333705149,1674297119,-608960773,1191115771,-71317778,-780929316,-69248426,-16952333,-16849409,1333246717,-1152395313,2109046663,-654391441,1878945761,-1076105933,-1124698178,-203057222,-46092,-1095795115,-1108044805,2053567035,756027145,-386469076,2067791731,-572228178,-544214151,2094882796,-299653,1357557599,1197211508,-849924099,1860005883,-1118640473,-1302857762,-19205217,-536905640,-1077115802,-131223,-1219954807,511550398,1073606391,-673403395,-557003227,-1426338305,1018093311,1039882343,-160809059,-277642369,1497096031,-32266762,-940445744,-71478277,-138090634,1055769394,-304098081,-773427982,-1346404631,-143284259,-134222809,1980447231,2024384491,-206605319,1576525279,1329839551,-53091270,-414814942,-203964841,2104646125,-557206625,-155814402,-2023690572,513670155,1844144107,-1550032940,-270841868,-909250577,1874901123,-31300625,2129747950,-8913475,2111829456,1542957037,2134765375,1608171006,-825526537,1397744338,513015534,-1460980297,1576707691,-1080127651,-1354760706,-1628443674,1918290901,-1091709713,1862258015,-692323745,2075849373,-541729800,-50874418,-570476730,1904181134,-1032061361,-1075348611,-1342186948,1869085693,1407445434,-1106603045,-455003986,-684213779,-343458861,-1216972345,1828449055,-136184113,-865303,-471367699,-776800324,1928983231,-1661105239,1598295866,1983765367,-1259128845,-16680450,-337707144,762150142,-551171841,-1632636132,-536974537,2045047781,497926592,-627183954,1767534542,-629317762],
-		[1807134719,-18948193,-135858755,1643378927,-1611277199,1274772207,771440477,-1079288901,-336134465,-1514160148,-1342497669,-1183023777,588502249,-480254285,2113154543,-631262329,1038040838,-277872931,-560755809,1524459215,-1209335129,964579294,-788543885,-1359422417,1327429870,1891467259,-68982836,-1365509139,-20972321,-173292365,-402688801,-620929827,-1895387158,-1279263211,1851723475,-641655171,-1640056802,-890108330,596590046,1544644269,1541308404,2127527690,-1219586326,-1302925615,-276068353,-1634484741,368895805,1861201515,-1175761057,-1109911580,-1151406117,-612565573,-168464388,-285212946,-312639907,-75499676,-16808566,-18219684,1344776143,1204445139,-325194021,220263490,2142184380,-81101321,-22245519,-1109583905,-1489641793,-883966211,997941041,-168985105,776709807,-702124547,-1941495825,-1616245429,-1351762434,912780143,-180472139,2071901486,1776283412,-2018574721,-307366146,-1232552765,-48638375,-1745275233,2078508797,2087439395,1340734694,-68096027,-1563951108,-537202754,-15405461,-1111853637,-335554119,2058190707,-1073897603,-1292141470,905816844,-1359663396,-1493733641,-2951433,-1359561237,-1275369,-174457505,-689471661,-1951134801,-89392827,-1043385601,-1108872427,-1578179497,1819954687,601734783,1710077359,-754987138,1782561533,1991438935,1833422752,-681198193,-1143212170,1575742075,-6432649,-621364548,1606417395,-144308225,-587629014,-88092209,1433778141,527420145,-76255281,-1287660423,-1392925993,-1079138313,2107614446,-21212180,-1075849062,-1732650243,-504136194,-1691402265,-339771524,-205761923,-506478599,-483875936,-158101545,-572985866,-1311023267,-2013307723,-314734723,-1610625399,-10756130,155025406,-47319491,1778080506,-98065873,1984596926,-176755667,2138952509,1165665456,2132864996,464701254,-711245825,666430687,-1732281089,1840668028,-1488208772,1772933947,984513199,-587532945,-1208123685,1862265341,-1477643532,-437281238,-124781357,2008972031,-1981809694,1561145333,-890896453,-214608417,-1015175802,-1085282817,2011066931,-826300676,-512000006,-1896349865,2112313858,-1114636306,-714181713,-1932417025,-375423785,-44829218,1572284670,-185647649,1703932566,-1215539205,-411075649,-624968373,546758639,-2968004,-1075322889,-540959646,-37766665,1740999656,1189945183,-348405831,-713369362,-1701168433,-1293354997,1329425404,-1108528725,1868559357,-1342705486,1324498788,-131278884,-950575379,-1351114117,-1355169869,997859263,972050085,-67132591,1958723379,-456531809,-51709441,396332531,247414162,-84936705,803741435,-1210462721,-147366550,1539308862,919301756,1264164782,-42896257,1384494463,-28054666,1044085349,-22446398,1199565155,936058645,-1781028,-270020615,-1323067746,-14689490,1059945650,-219224369,-335880840,1602107263,1454685810,-50358939,-1316034625,-428907690,-1457258868,2147384191,-1018757588,1787293662,888715263,-2338105,805063667,-117973082,99712821,38238154,-1181614521,-1421890562,777839391,-1230517260,-562169877,-547724806,-137760777,929622522,2059370380,-50367553,-923411474,-1132648717,2079258751,-209715597,-194687006,381016029,-467699797,-270565569,-539862036,-134359395,-559516937,-2035558625,-312754758,2126487034,-1646657,1073299126,-562583081,900713719,-895328594,-101976223,2060402031,-1622684033,-1257545,-85430794,-11604938,2136864755,2133065531,2095176701,1069438413,24117155,-271325650,-1051265,301849803,-1361372535,-579076577,-1636043905,-1082138690,-409417741,2073034524,-274309453,1836432978,-1511078087,-679487553,-1049256328,921157438,1971093214,-18155299,-23331348,-1577059427,1588068271,-184723527,1802154749,2138535899,-604030153,-282602536,-1632782794,-72892718,-4590629,635943894,2061416886,-671428693,1977378622,-565218961,-851018388,-1426939927,-862016708,1259026431,-208117855,-117969442,-269640103,-3277190,1909681582,-1057215788,-90968353,987613602,-1023721879,-1914773506,-1686153547,-409503748,-100885009,2056716079,-226902661,-273350955,-2094917645,-134837793,-306203298,-1646421537,1283972336,-1382877833,-1694605031,-1108418721,-55750937,-117965333,-9016387,-1546719245,1047914978,-559690021,1993289321,-954311425,-1242170534,1001911641,-1753469447,-159421530,1853749247,-806106625,-165894,-218628202,-352650819,634154907,-603979847,-68687886,1989945191,-4194562,935127775,-209830971,2006558459,-1484809061,-637790761,1475041960,-1142662081,1070529173,-640860119,2135552506,-69492807,502267902,-437887029,2112610131,-471475220,656139627,-1628521221,-16993642,-499124229,-76791917,-1393198529,125272047,2068631931,-419668293,-50726183,-558060647,-422720529,1330378712,-7550529,903786031,-25329683,-1644453265,-30475202,-123738117,-800078857,-71388165,-289421013,-959965013,399143663,-362091754,2025324271,-1577591108,-1344800985,-21539010,-1089339409,535625646,-612390693,1962365814,1911475372,-537068713,-243307284,2141058822,-155666915,1804316159,-592154212,-1133719072,-338168253,-539241283,-780147954,2112993272,-1082215753,2014748350,-51741585,-1502628361,972279663,-1090652301,377318571,-42975273,-1710510405,-142197991,-521181970,2110979301,-606080257,-981468195,2141713373,-279023917,-3801434,-1259421137,-945336357,-681608258,-889269283,-1888296066,-49030695,-572531338,721181588,-92407890,-605571147,1174904667,-558371841,-1610758147,-1095041217,1942744889,2136350550,-1602359553,-1088968961,-717458069,-536972297,1040110837,1918868963,-1124271686,1958640099,-104860169,-228775026,-1103429998,1877729023,-265945601,-218443829,-1695158563,-1354773011,-143466513,409779099,-1076007633,-723862718,2062286723,-1826965522,-56794335,1467242463,535346684,1060849614,-320408082,-34603270,2096582459,-134238093,-539197717,-58789391,2138832589,-1163677741,531558367,-595669153,-69698,-1308758421,1991671218,402253269,1939576745,-673288201,-1226326239,-367473202,-1445086922,2110848924,-169617411,-874597763,-289497089,-541245709,-556862511,-144321638,1272438621,-100991524,2013187803,-609428242,-738476803,-20976066,905802927,719485015,2046473212,-1231561238,-852508689,117437276,799529465,-1224808061,2021555935,2130308662,-941654201,1071642511,-34800993,252328483,-1081623125,-134269001,2034670332,2000616927,-747807825,-1483643146,-186096421,-1118048377,-352561187,-295625033,1669298143,-4468813,-271747999,-2122145,-581965048,-4719847,-1923377514,-264373058,-1104228431,1805885164,-34295169,-52715382,1069473144,-303628819,2146393918,1794350415,-1132486667,-1342412802,-1076175021,-1385725993,1609662205,1941765497,-1080585556,-371888150,-112015886,-507710277,2128539578,-1133510934,-1656297764,-136610854,-33883398,2129264474,-134438999,-1603593,-730408131,1306259310,-1147409444,-277348558,487979525,-1645709323,-277966967,-589300154,-1862676482,-1180472597,1836768131,-637555715,-2122981750,-2097626147,1073376511,2147408379,-741644289,1635774055,-1344897189,-940410969,182220542,1030664167,-1614430357,-388921450,576716795,1810184827,-1092894839,-187732613,-17696449,-1813001857,1057151999,-915264579,-72898631,-838911176,2010042299,640404988,-1478536485,-1622168917,-478740595,-1258570753,-573794,-564724770,-41952403,1438284977,251000793,2055887523,624918477,-423724598,-176243738,-2364705,2059369006,-16909825,-1074407639,-945840469,-26271890,-1260520469,401000427,1922529240,-1646362633,-170971202,1014955919,-269677057,-1990295746,1874577025,-1619247795,105027798,972930143,-274477241,-75895907,1543487271,-637572637,-606898810,2141132270,-1617154436,-1851267014,624688581,-70279170,-1234219361,-18874449,60673531,1541405165,-942743811,1673455351,-188257810,-1545877009,-1882965001,2129113598,-1146126460,660730603,-285489698,-650175041,-143952007,1053943471,2132217922,872038475,-29640257,-405817897,-637612039,-90458366,1823123351,-11616474,-1482762318,1040055073,-1080123833,-19304617,2022295550,-1389114009,-1295124485,-1835043,-276113444,-1610883208,377419678,-69813329,-25967113,-1078496267,-1259422575,-174260817,-302101526,-808257468,-214042753,-8701602,1026271671,-356515846,1590947267,-144830799,2079269298,-389173382,-413245962,-841523649,-822216705,-145231910,2069298369,-143954057,-731616343,-273088897,-1388855571,-89852229,-1106264386,-147873884]
-	],
-	[
-		[-1227133514,-1,-1,0,-1,0,0,-1,-1,-1,0,-1,0,0,-1,0,-1,-1,-1,-1,0,-1,-1,-1,0,-1,-1,-1,-1,-1,-1,0,0,0,-1,0,-1,0,0,0,-1,-1,-1,-1,-1,0,-1,0,0,0,0,-1,-1,0,0,0,-1,-1,-1,-1,-1,0,-1,0,-1,-1,0,-1,-1,-1,-1,0,0,0,0,0,-1,0,-1,-1,-1,-1,-1,0,-1,-1,0,-1,0,-1,-1,-1,-1,-1,-1,0,-1,-1,0,0,-1,0,-1,-1,-1,0,-1,-1,-1,0,-1,-1,-1,-1,-1,0,-1,0,-1,0,-1,-1,-1,-1,-1,-1,-1,0,0,-1,0,-1,0,-1,0,-1,-1,-1,-1,-1,0,-1,-1,-1,0,0,-1,-1,-1,-1,-1,-1,0,-1,0,0,-1,-1,0,-1,-1,-1,0,0,-1,-1,0,0,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,0,-1,-1,-1,-1,-1,-1,-1,0,-1,-1,-1,0,-1,-1,-1,0,-1,-1,-1,-1,-1,-1,0,0,-1,0,0,-1,0,-1,-1,-1,0,-1,0,0,0,-1,-1,-1,-1,-1,-1,-1,-1,0,-1,-1,-1,0,-1,-1,-1,0,-1,-1,0,-1,-1,0,0,0,0,0,0,-1,-1,-1,-1,-1,0,-1,-1,0,-1,0,-1,-1,-1,-1,0,-1,-1,0,0,-1,-1,0,-1,-1,-1,-1,-1,-1,-1,0,-1,-1,0,-1,0,0,-1,0,0,0,-1,-1,-1,-1,-1,0,-1,-1,-1,0,-1,-1,-1,-1,-1,0,-1,0,-1,-1,-1,-1,0,-1,0,-1,-1,-1,-1,0,-1,-1,-1,-1,-1,-1,-1,0,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,0,0,0,0,-1,0,0,0,0,-1,0,0,-1,-1,0,-1,-1,0,-1,-1,-1,0,-1,-1,-1,0,0,0,0,-1,-1,-1,-1,0,0,0,-1,-1,0,-1,0,-1,-1,-1,0,-1,-1,-1,-1,-1,-1,-1,0,0,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,0,-1,-1,-1,-1,0,-1,-1,-1,-1,-1,0,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,0,-1,-1,-1,0,-1,-1,-1,-1,-1,-1,-1,-1,0,-1,0,0,-1,-1,-1,-1,-1,0,0,-1,0,-1,-1,0,0,-1,-1,-1,0,-1,-1,0,0,-1,-1,0,-1,-1,0,-1,-1,-1,-1,0,-1,0,-1,-1,-1,-1,0,-1,-1,-1,-1,0,0,-1,0,0,-1,-1,-1,-1,-1,0,-1,0,0,-1,-1,-1,0,-1,0,0,-1,-1,0,0,-1,-1,-1,0,-1,-1,0,-1,-1,-1,-1,0,-1,-1,-1,-1,0,-1,-1,-1,0,-1,-1,-1,-1,0,0,0,-1,0,-1,0,0,0,-1,0,0,0,-1,0,-1,0,-1,-1,-1,-1,-1,-1,0,-1,0,-1,-1,0,-1,-1,0,-1,-1,-1,-1,0,-1,-1,-1,-1,-1,-1,-1,0,-1,-1,-1,-1,0,-1,-1,-1,-1,-1,0,-1,-1,-1,-1,0,0,0,-1,-1,0,-1,-1,-1,-1,-1,-1,-1,-1,0,-1,0,-1,-1,-1,-1,0,-1,-1,-1,-1,-1,-1,-1,-1,-1,0,0,0,-1,-1,-1,-1,-1,0,0,-1,0,0,0,-1,0,-1,-1,0,0,-1,0,-1,-1,-1,-1,-1,-1,-1,0,-1,0,0,-1,0,-1,-1,-1,0,-1,-1,-1,0,0,0,0,-1,0,-1,-1,-1,-1,-1,-1,0,-1,-1,0,-1,-1,-1,-1,0,-1,-1,-1,-1,0,-1,0,-1,-1,0,0,-1,-1,0,-1,-1,0,-1,-1,-1,0,0,-1,0,-1,-1,0,0,0,-1,-1,-1,0,-1,-1,4194303,4194303,2147475471,-4186113,235340673,2118066063,-1],
-		[-613566757,0,-1227133514,2118066063,-235340674,-1,-1,-1,-2143289360,-1,-4186113,-1,-1,-4194304,-4194304,-1,0,-1,0,-1,-1,-1,-1,-1,0,0,-1,-1,-1,0,-1,0,-1,-1,0,0,0,-1,-1,-1,0,-1,-1,-1,-1,-1,0,-1,-1,-1,-1,-1,-1,0,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,0,-1,0,0,0,-1,-1,-1,0,0,-1,-1,-1,-1,-1,-1,0,0,0,0,-1,-1,-1,0,0,0,0,-1,0,0,0,-1,-1,-1,0,-1,-1,-1,-1,0,0,-1,0,0,0,0,-1,0,-1,0,-1,-1,0,-1,0,-1,-1,-1,-1,-1,-1,-1,0,-1,-1,-1,0,-1,-1,-1,0,-1,0,-1,0,-1,-1,-1,0,-1,-1,-1,0,-1,-1,0,0,-1,-1,0,-1,-1,-1,0,-1,-1,-1,0,-1,0,-1,-1,-1,-1,-1,-1,-1,-1,0,-1,-1,-1,0,-1,0,-1,-1,-1,0,-1,0,-1,-1,-1,0,0,-1,-1,0,-1,-1,0,-1,-1,0,-1,0,0,-1,0,-1,-1,0,-1,-1,-1,-1,0,-1,0,0,0,-1,0,0,0,-1,-1,-1,-1,-1,-1,0,0,0,-1,-1,-1,0,0,-1,-1,-1,-1,-1,-1,0,-1,-1,-1,-1,-1,-1,0,-1,0,-1,0,-1,0,0,0,0,-1,-1,0,0,-1,0,0,-1,0,-1,-1,-1,-1,0,-1,-1,0,0,-1,-1,-1,-1,-1,-1,-1,0,0,-1,0,-1,0,0,0,0,0,0,-1,-1,0,-1,-1,-1,-1,-1,0,-1,-1,-1,0,-1,-1,0,0,-1,-1,0,0,-1,-1,-1,0,-1,-1,-1,-1,-1,-1,-1,0,-1,-1,0,-1,0,0,-1,-1,0,-1,0,0,-1,-1,-1,-1,0,-1,-1,-1,0,-1,-1,-1,-1,-1,-1,0,-1,0,-1,0,-1,-1,-1,-1,0,-1,-1,0,0,-1,-1,-1,0,-1,-1,0,-1,-1,0,-1,-1,-1,-1,0,-1,-1,0,0,-1,-1,0,-1,-1,-1,-1,0,-1,0,-1,0,-1,-1,-1,-1,-1,0,-1,0,0,-1,0,-1,-1,-1,-1,-1,-1,0,0,-1,-1,-1,-1,0,-1,-1,-1,-1,-1,0,-1,-1,0,0,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,0,-1,-1,-1,0,0,-1,-1,-1,-1,0,-1,-1,-1,-1,-1,-1,0,-1,-1,0,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,0,-1,0,-1,-1,-1,0,-1,0,-1,-1,0,-1,0,-1,-1,-1,-1,-1,0,0,-1,-1,0,-1,-1,-1,-1,-1,0,0,-1,-1,0,0,0,-1,-1,-1,-1,0,-1,-1,-1,-1,-1,0,0,-1,-1,-1,-1,-1,-1,-1,0,-1,-1,-1,-1,0,0,-1,0,-1,-1,-1,-1,-1,0,-1,0,-1,-1,0,-1,0,-1,-1,-1,-1,-1,-1,-1,0,-1,-1,0,-1,-1,0,0,-1,0,-1,-1,-1,-1,0,0,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,0,-1,-1,-1,0,-1,-1,0,0,-1,-1,0,0,-1,-1,0,-1,0,0,-1,-1,0,-1,-1,0,0,0,-1,-1,-1,0,-1,0,-1,0,-1,-1,-1,0,-1,0,-1,-1,0,0,-1,0,-1,0,-1,-1,-1,0,-1,0,-1,0,-1,-1,-1,-1,-1,-1,-1,-1,0,-1,-1,0,-1,-1,-1,-1,0,-1,-1,0,0,-1,-1,-1,-1,-1,-1,0,0,0,0,-1,0,-1,-1,-1,-1,0,0,-1,-1,-1,0,-1,0,-1,0,-1,-1,-1,0,0,0,0,-1,-1,0,-1,-1,-1,-1,-1,-1,0,-1,-1,0,0,-1,-1,0,-1,-1,-1,0,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,0,1227133513],
-		[-1227133514,1227133513,-1,-1,-1,0,-1,0,0,-1,-1,0,0,-1,-1,0,-1,-1,0,0,0,-1,0,-1,0,0,0,-1,-1,0,0,-1,0,0,0,-1,-1,0,0,0,-1,0,-1,-1,-1,0,-1,-1,-1,-1,-1,-1,0,-1,0,-1,0,0,0,-1,-1,0,0,-1,-1,-1,0,-1,0,-1,-1,-1,-1,-1,-1,0,-1,-1,-1,-1,-1,0,0,0,-1,0,0,0,-1,-1,-1,0,-1,-1,0,0,-1,0,0,0,-1,-1,-1,-1,-1,-1,0,0,-1,-1,-1,-1,-1,-1,-1,0,-1,-1,-1,-1,-1,0,-1,-1,-1,-1,0,-1,0,0,0,0,0,0,0,0,-1,-1,-1,-1,0,-1,-1,-1,-1,-1,-1,-1,0,-1,-1,-1,-1,-1,-1,-1,0,-1,0,0,-1,-1,0,-1,0,0,-1,-1,0,0,-1,-1,0,0,0,0,0,-1,-1,-1,-1,-1,-1,0,0,0,-1,-1,0,-1,-1,0,-1,-1,-1,0,0,-1,-1,0,0,-1,-1,-1,-1,0,0,-1,-1,-1,-1,0,0,-1,-1,-1,0,-1,-1,-1,-1,-1,-1,0,-1,-1,-1,-1,0,-1,-1,-1,0,-1,-1,-1,0,-1,-1,-1,-1,0,-1,-1,0,0,-1,-1,-1,0,0,0,-1,0,-1,-1,-1,0,-1,-1,-1,-1,0,-1,-1,0,-1,0,-1,-1,0,-1,-1,-1,-1,-1,-1,-1,-1,-1,0,-1,-1,-1,0,0,-1,-1,0,-1,-1,-1,0,-1,-1,-1,0,-1,0,-1,0,-1,0,-1,0,-1,0,0,-1,-1,0,-1,-1,-1,-1,-1,-1,-1,0,-1,-1,-1,-1,0,-1,-1,0,-1,-1,0,-1,-1,0,-1,-1,0,-1,-1,0,0,0,0,-1,-1,0,-1,-1,0,-1,-1,-1,-1,-1,0,0,0,-1,-1,-1,-1,-1,0,-1,-1,-1,0,-1,0,-1,-1,-1,-1,-1,0,0,-1,0,-1,-1,-1,-1,-1,0,-1,-1,-1,0,0,-1,-1,-1,0,0,-1,-1,0,-1,-1,0,0,-1,-1,-1,-1,0,-1,-1,-1,0,0,-1,-1,-1,-1,0,0,-1,-1,-1,0,-1,-1,0,-1,-1,-1,0,-1,-1,-1,-1,0,-1,-1,0,-1,-1,-1,0,-1,-1,0,-1,-1,-1,0,-1,-1,0,-1,-1,-1,-1,-1,-1,-1,0,-1,0,-1,-1,0,0,-1,-1,-1,0,-1,-1,0,-1,0,-1,0,-1,-1,-1,0,-1,-1,0,-1,-1,-1,-1,0,0,-1,-1,0,0,-1,0,-1,-1,-1,-1,-1,-1,-1,0,-1,0,0,-1,-1,-1,0,-1,-1,-1,-1,0,-1,0,-1,0,-1,0,-1,0,-1,-1,-1,0,-1,0,-1,-1,0,-1,-1,-1,-1,-1,0,-1,0,-1,-1,-1,-1,0,-1,0,-1,0,-1,-1,-1,0,-1,-1,-1,-1,-1,0,-1,-1,-1,0,-1,-1,-1,-1,-1,-1,-1,0,-1,0,-1,-1,0,0,0,-1,-1,0,0,-1,-1,-1,-1,-1,0,-1,0,0,-1,0,-1,-1,-1,-1,-1,-1,0,-1,-1,0,0,-1,0,-1,-1,0,-1,-1,0,0,0,-1,0,-1,-1,0,0,-1,0,-1,0,-1,-1,-1,-1,-1,-1,-1,0,0,0,-1,0,0,0,0,-1,-1,0,-1,-1,0,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,0,-1,-1,0,-1,-1,-1,-1,0,0,0,-1,-1,-1,0,-1,-1,0,-1,-1,-1,0,0,-1,-1,-1,-1,-1,-1,0,-1,-1,-1,-1,-1,0,-1,-1,-1,-1,0,0,-1,-1,-1,4194303,-1,4194303,-1,-1,4194303,-1,-1,4186112,2147475471,-1,-2143289360,2143289359,2147475471,-1,-2147475472,-1,-235340674,-1,-235340674,235340673,-1882725391,-2118066064,-1882725391,-1,-1,-1,-1,-613566757],
-		[1840700269,-613566757,1840700269,-1,-1,-1,0,-1,1227133513,0,1227133513,-1882725391,-235340674,-2118066064,2118066063,1882725390,-235340674,235340673,-1,-1,235340673,-1,-235340674,-1,235340673,-1,-235340674,-2147475472,2147475471,0,-2147475472,-2147475472,-4186113,-2143289360,-1,-1,-1,-1,-2143289360,-2147475472,-4186113,-1,-4186113,0,-1,-1,-1,-1,-1,-1,-4194304,-1,-4194304,-1,-4194304,0,-4194304,4194303,-1,-1,-1,0,0,0,0,-1,-1,-1,-1,-1,-1,-1,-1,-1,0,-1,0,-1,-1,-1,-1,-1,0,-1,0,0,0,-1,0,-1,-1,-1,0,-1,0,-1,-1,0,-1,-1,-1,-1,0,-1,-1,-1,-1,0,0,-1,0,-1,0,-1,-1,-1,-1,-1,-1,-1,-1,-1,0,0,0,-1,-1,-1,-1,-1,-1,0,-1,-1,0,-1,0,-1,0,0,0,-1,-1,-1,0,-1,0,0,-1,-1,-1,-1,-1,-1,0,0,0,0,0,-1,0,-1,0,-1,-1,-1,-1,0,-1,-1,0,-1,0,-1,-1,0,-1,-1,-1,0,-1,-1,0,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,0,-1,-1,0,-1,-1,-1,0,-1,-1,-1,-1,-1,-1,0,-1,0,0,-1,-1,-1,-1,-1,0,-1,-1,-1,0,0,-1,-1,-1,-1,-1,0,-1,0,-1,-1,0,0,0,-1,-1,-1,-1,-1,-1,-1,-1,0,0,-1,-1,-1,-1,0,-1,-1,0,0,-1,-1,-1,-1,-1,0,0,0,0,-1,-1,0,0,-1,-1,-1,-1,0,-1,0,0,0,0,-1,-1,0,-1,-1,-1,-1,0,-1,0,-1,0,-1,0,-1,0,0,0,-1,-1,0,0,-1,0,-1,0,0,-1,-1,-1,-1,-1,-1,0,-1,0,0,0,-1,-1,-1,-1,-1,-1,-1,-1,0,-1,0,-1,-1,-1,0,-1,-1,0,-1,-1,0,-1,-1,-1,0,0,0,0,0,0,-1,-1,-1,-1,0,0,0,-1,-1,0,-1,0,-1,-1,0,-1,-1,-1,0,-1,-1,-1,0,-1,-1,0,0,0,-1,0,0,-1,-1,-1,-1,-1,0,-1,-1,-1,0,0,-1,0,0,-1,-1,-1,-1,-1,-1,0,-1,-1,-1,-1,-1,0,-1,-1,-1,-1,-1,0,-1,-1,0,-1,0,0,0,-1,-1,0,-1,0,-1,-1,0,-1,-1,0,0,-1,-1,0,0,-1,-1,-1,-1,0,-1,0,-1,-1,0,0,-1,0,0,-1,0,-1,0,0,-1,0,-1,-1,0,0,-1,-1,-1,-1,0,-1,-1,-1,0,-1,0,-1,0,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,0,0,-1,0,0,-1,-1,-1,-1,0,0,0,0,-1,-1,0,0,-1,-1,-1,0,0,-1,-1,-1,-1,-1,-1,0,0,-1,-1,0,-1,-1,-1,-1,-1,0,-1,0,0,-1,-1,0,-1,-1,0,-1,0,-1,-1,-1,-1,-1,-1,-1,0,-1,0,-1,-1,0,0,0,-1,-1,0,0,0,0,-1,-1,0,-1,0,-1,0,-1,0,-1,-1,-1,0,-1,-1,-1,-1,-1,0,0,-1,0,-1,0,-1,-1,0,-1,0,0,-1,0,-1,0,0,-1,-1,0,-1,0,-1,-1,0,-1,-1,-1,0,-1,-1,-1,0,0,-1,-1,-1,-1,0,-1,0,-1,-1,-1,-1,-1,-1,-1,0,-1,-1,-1,0,-1,-1,-1,0,-1,-1,0,-1,0,0,-1,0,0,-1,0,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,0,-1,0,-1,-1,-1,-1,0,-1,0,0,-1,-1,0,-1,0,-1,-1,-1,-1,-1,-1,-1,-1,-1,0,0,-1,-1,-1,0,0,-1,0,-1,0,-1,-1,0,0,0,-1,-1,0,-1,0,-1,-1,-1,-1,-1,0,0,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,0,-1,0,-1,-1227133514,-1227133514,-1227133514,-1],
-		[-1227133514,-1,1840700269,1227133513,-1,1227133513,1227133513,1227133513,1227133513,-1,-1227133514,0,-1,-1,-1,-1,-1,0,0,-1,-1,-1,0,-1,0,-1,-1,-1,0,-1,0,-1,0,0,-1,-1,0,-1,-1,-1,-1,-1,-1,-1,-1,-1,0,-1,0,-1,0,0,-1,0,-1,-1,0,-1,-1,0,0,0,-1,-1,-1,0,-1,0,-1,-1,0,-1,-1,-1,-1,-1,0,-1,-1,-1,0,-1,-1,0,-1,-1,0,0,-1,-1,-1,0,-1,0,-1,-1,0,0,-1,-1,-1,-1,-1,-1,0,0,0,-1,0,-1,0,0,0,-1,-1,-1,0,-1,-1,0,0,-1,0,-1,-1,-1,-1,-1,0,0,-1,-1,0,-1,-1,0,-1,0,0,-1,-1,-1,-1,-1,-1,-1,0,-1,-1,-1,-1,-1,0,0,-1,-1,0,0,0,-1,0,0,-1,-1,-1,-1,0,0,0,-1,0,-1,0,-1,0,-1,-1,-1,-1,-1,0,-1,0,-1,-1,-1,-1,-1,-1,-1,-1,0,-1,-1,-1,-1,-1,-1,0,-1,-1,-1,0,-1,-1,-1,-1,-1,-1,-1,0,-1,-1,0,-1,-1,0,-1,-1,0,-1,-1,-1,-1,0,0,0,0,-1,-1,-1,-1,-1,0,0,-1,-1,0,-1,-1,0,-1,-1,-1,0,0,-1,-1,-1,-1,0,0,-1,0,-1,0,-1,0,-1,-1,-1,-1,-1,-1,0,0,-1,-1,-1,-1,-1,-1,-1,-1,0,0,-1,-1,0,-1,-1,0,0,-1,-1,-1,-1,-1,-1,0,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,0,-1,-1,-1,0,0,0,-1,-1,-1,0,0,-1,-1,-1,0,-1,-1,-1,-1,-1,0,-1,-1,-1,0,-1,-1,0,-1,-1,-1,-1,-1,0,-1,0,0,-1,-1,-1,-1,-1,0,-1,0,-1,0,-1,-1,-1,-1,0,-1,0,-1,-1,-1,-1,-1,-1,0,-1,0,-1,-1,-1,0,-1,0,-1,-1,-1,0,0,0,-1,-1,-1,-1,-1,-1,-1,0,-1,-1,-1,-1,0,-1,-1,0,-1,0,-1,-1,0,0,-1,0,0,-1,-1,0,0,-1,-1,-1,-1,-1,0,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,0,0,-1,0,-1,-1,-1,0,0,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,0,0,-1,0,0,-1,-1,0,-1,-1,-1,-1,-1,-1,-1,-1,0,0,-1,-1,-1,-1,0,-1,-1,-1,-1,0,0,-1,0,-1,-1,-1,0,-1,-1,0,-1,0,-1,0,-1,-1,-1,-1,0,-1,0,-1,0,-1,0,-1,0,0,0,0,-1,0,0,0,-1,0,-1,0,-1,-1,0,0,-1,-1,-1,0,-1,-1,0,-1,-1,-1,0,0,-1,-1,-1,-1,0,0,0,0,-1,-1,-1,-1,-1,0,0,-1,-1,-1,-1,-1,-1,-1,0,-1,-1,-1,-1,-1,-1,-1,0,-1,-1,-1,-1,-1,-1,-1,0,-1,-1,-1,-1,-1,0,0,-1,-1,0,-1,-1,0,-1,0,-1,-1,-1,-1,0,-1,0,-1,-1,0,0,-1,-1,-1,0,-1,-1,-1,0,0,-1,-1,-1,-1,-1,-1,-1,-1,-1,0,0,-1,-1,-1,-4194304,4194303,-4194304,-1,4194303,-1,-1,-1,4194303,0,-4194304,-4194304,-1,-1,-4194304,-1,-1,-1,4194303,0,4194303,-1,-4194304,-1,0,-1,0,-1,0,-1,-1,4186112,4186112,-1,-4186113,-4186113,-2143289360,2147475471,-1,-1,-1,-1,2143289359,-1,2143289359,-1,-1,-1,-2143289360,2143289359,-1,-1,2147475471,-2147475472,2147475471,2147475471,-1,-1,-1,2147475471,-2147475472,2147475471,-1,235340673,-1,-1,-235340674,-1,-1,0,-235340674,-1,-1,-1,235340673,-235340674,-1,-1,-1,-1,235340673,-1,-1,-235340674,2118066063,-1882725391,-1882725391,-2118066064,-2118066064,2118066063,-1,235340673,-1882725391,-1,-235340674,-1,-1227133514,-1,-1,-1,-1227133514,-1,-1227133514,-1,-1,0,0,0,-1,-1,0,1840700269,-1227133514,-613566757,-1227133514,-1840700270],
-		[-613566757,1840700269,-1,-1,-613566757,613566756,-1227133514,-1,-613566757,1840700269,-613566757,-1,-1,0,0,0,0,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1227133514,1227133513,-1,1227133513,1227133513,-1,-1,0,-1,-1,0,-1,-1,1227133513,0,1227133513,-1,2118066063,0,2118066063,-1882725391,2118066063,-235340674,-235340674,-1,-2118066064,2118066063,-1,-1,-1,-1,2118066063,1882725390,-1,-1,1882725390,-2118066064,-235340674,235340673,-1882725391,0,-235340674,-1,-1,-235340674,-1,-1,235340673,0,0,0,-1,-1,-235340674,-235340674,235340673,-235340674,-1,0,-235340674,-1,235340673,-1,235340673,235340673,235340673,0,-1,0,235340673,0,0,-1,235340673,0,235340673,-1,235340673,235340673,-1,-1,2147475471,-1,-2147475472,2147475471,2147475471,2147475471,-1,-1,0,-1,-1,-1,2147475471,-1,-2147475472,2147475471,-2147475472,-2147475472,-1,-2147475472,-4186113,-1,2147475471,-1,2143289359,2143289359,-2143289360,2143289359,-1,0,2143289359,-1,0,0,0,2143289359,-2143289360,-1,-2143289360,-2143289360,-2143289360,-1,2143289359,-1,2143289359,0,-1,-1,2147475471,2147475471,-2147475472,-2143289360,-4186113,4186112,2147475471,-4186113,-1,-1,-4186113,-1,-1,-4186113,-4186113,0,-1,-1,0,0,-1,-1,-1,-1,-1,-1,0,-1,-1,0,-1,-4194304,4194303,-1,-4194304,4194303,-4194304,-1,-1,4194303,-1,0,-4194304,-1,-4194304,-1,4194303,4194303,4194303,-1,4194303,-1,4194303,4194303,-1,4194303,4194303,-1,-1,-1,-4194304,-1,-4194304,-1,-1,0,-1,-1,-4194304,-1,-4194304,-1,4194303,-1,-4194304,-4194304,4194303,0,4194303,0,0,0,0,-1,-1,-1,-1,-1,0,-1,0,-1,-1,-1,-1,-1,-1,-1,0,-1,-1,0,0,0,0,-1,-1,-1,-1,-1,0,0,0,0,0,0,-1,0,-1,-1,-1,-1,-1,-1,-1,-1,-1,0,-1,-1,0,0,-1,0,-1,-1,-1,-1,0,0,0,-1,-1,-1,0,0,-1,-1,0,-1,-1,0,-1,-1,-1,-1,-1,0,-1,0,0,0,-1,0,-1,0,0,0,-1,-1,0,0,-1,-1,0,-1,-1,-1,0,0,-1,-1,0,-1,0,-1,-1,0,-1,-1,0,-1,-1,0,0,0,-1,0,-1,0,0,-1,-1,-1,0,-1,0,-1,-1,-1,0,0,0,-1,-1,-1,0,0,-1,0,-1,0,0,-1,-1,-1,-1,0,0,-1,0,-1,0,-1,-1,0,-1,-1,-1,-1,-1,-1,-1,0,-1,0,-1,0,0,0,0,-1,0,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,0,-1,-1,-1,-1,-1,0,0,0,0,-1,0,-1,-1,0,-1,-1,-1,0,-1,0,-1,-1,0,-1,-1,0,0,0,-1,0,-1,-1,-1,-1,-1,0,-1,-1,-1,0,-1,0,-1,-1,0,0,-1,-1,0,-1,-1,-1,0,-1,-1,-1,0,-1,-1,0,-1,-1,-1,-1,-1,-1,0,-1,-1,-1,-1,-1,-1,0,0,-1,-1,-1,0,-1,-1,-1,-1,-1,0,-1,-1,-1,0,0,-1,-1,-1,-1,-1,-1,-1,-1,0,-1,0,0,-1,0,-1,-1,-1,0,-1,0,0,0,0,-1,0,-1,-1,0,-1,-1,0,-1,-1,-1,0,-1,0,0,0,0,-1,0,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,0,-1,-1,0,-1,-1,-1,-1,-1,-1,-1,0,-1,-1,0,0,-1,0,0,-1,-1,0,0,-1,-1,0,0,-1,-1,0,-1,0,0,-1,-1,-1,-1,-1,-1,0,0,-1,-1,-1,-1,0,-1,-1,-1,-1,-1,-1,-1,-1,0,-1,-1,-1,-1,0,0,0,-1,0,-1,-1,0,0,-1,0,-1,0,-1,-1,-1,0,0,-1,0,0,-1,-1,-1,-1,0,-1,0,-1,0,0,-1,-1,-1,-1,0,-1,0,-1,-1,-1,-1,-1,-1,-1,-1,0,-1,-1,-1,-1,0,-1,0,-1,0,0,-1,-1,0,-1,-1,0,-1,-1,0,0,0,-1,-1,-1,-1,0,0,-1,-1,-1,0,-1,-1,0,-1,0,-1,-1,-1,-1,1227133513,-1,-1,-1227133514,-1227133514,1227133513,-1227133514,1227133513,-1,-1,-1227133514,1227133513,-1,-1,-1,0,1840700269,-1227133514],
-		[-1227133514,-1227133514,1840700269,-1840700270,-1,-1,-1,-1,-1840700270,-1,-1,-1,-1227133514,-1227133514,-1,1227133513,-1,-1,1227133513,-1,-1227133514,1227133513,-1227133514,-1,-1,1227133513,-1227133514,-1,1227133513,-1,-1,-1,1227133513,0,-1,1227133513,-1,-1,-1227133514,0,1227133513,0,-1227133514,-1,-1,0,-1,0,-1,-1,-1,-1,0,-1,0,0,-1,0,0,-1,-1,-1,0,-1,-1,0,-1,0,-1,-1,0,-1,0,0,0,-1,-1,-1,-1,-1,-1,0,0,-1,-1,-1,-1,-1,0,0,-1,-1,-1,-1,-1,-1,0,-1,-1,-1,-1,-1,0,-1,-1,0,0,-1,-1,-1,0,-1,-1,0,0,-1,0,-1,-1,0,-1,-1,0,0,-1,-1,0,-1,0,-1,-1,-1,-1,0,0,0,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,0,-1,-1,-1,-1,-1,-1,0,-1,-1,-1,-1,-1,0,-1,0,-1,-1,0,0,0,-1,0,0,-1,0,-1,0,-1,-1,-1,-1,-1,0,-1,0,0,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,0,-1,-1,-1,0,-1,0,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,0,0,-1,-1,-1,0,-1,0,0,0,-1,0,-1,-1,0,0,0,-1,0,-1,-1,-1,0,-1,-1,-1,-1,-1,0,0,-1,0,-1,-1,0,-1,-1,-1,-1,-1,0,-1,4194303,-1,0,-1,-1,-1,-1,-1,4194303,4194303,-1,-4194304,-4194304,-1,0,0,-1,-1,-1,-4194304,4194303,-1,0,4194303,-1,-1,-1,-1,-1,-1,0,0,-1,-1,-1,-4194304,-4194304,4194303,-1,4194303,-1,-1,-1,0,4194303,-1,-1,-4194304,0,-4194304,-4194304,-4194304,-4194304,-1,-1,-4194304,-4194304,-4194304,0,-4194304,4194303,-4194304,-1,-4194304,-4194304,-1,-1,-4194304,-1,4194303,-1,4194303,-4194304,4194303,-1,-1,-1,4194303,-4194304,-4194304,0,-1,-1,0,-4194304,4194303,-4194304,-1,4194303,4194303,0,-4194304,-1,-1,-4194304,4194303,-1,-1,-1,-1,-1,-1,-1,0,-1,-1,-1,0,0,-1,0,-1,-1,-1,-1,-1,0,-1,0,-1,-1,-1,-1,0,-1,-1,-1,-1,-4186113,-1,-1,-1,-1,-4186113,-1,0,-1,4186112,-1,-4186113,-1,4186112,4186112,-1,-2147475472,-2143289360,4186112,-1,-1,-2143289360,-2143289360,-4186113,-1,2147475471,-1,-1,-2147475472,-1,-1,-2147475472,-1,0,0,-1,2143289359,-1,-1,2143289359,-2143289360,-2143289360,-1,2143289359,-2143289360,-1,-2143289360,-1,-1,2143289359,-1,2143289359,2143289359,-1,-2143289360,-1,-1,-1,-1,-1,0,-1,-1,-1,-2143289360,-2143289360,0,-1,0,2143289359,-1,-2143289360,-2143289360,2143289359,2143289359,-1,2143289359,-1,-1,2143289359,-1,-2147475472,0,2147475471,4186112,-2143289360,-2147475472,2147475471,-1,-2147475472,-2147475472,2147475471,-2147475472,2147475471,-1,-1,2147475471,-1,0,-2147475472,2147475471,-2147475472,0,-2147475472,-1,0,0,-1,-1,-1,0,0,0,2147475471,-1,-2147475472,2147475471,-1,2147475471,-2147475472,2147475471,-1,-1,-2147475472,-2147475472,-2147475472,-1,-2147475472,-1,235340673,-1,-1,-235340674,-235340674,-1,235340673,-1,-1,-1,-1,-235340674,-235340674,-1,-235340674,-1,-1,0,-1,-1,-1,0,-1,-1,0,0,-1,-1,-1,-235340674,-235340674,235340673,-1,-1,-235340674,235340673,-235340674,-1,-235340674,-1,235340673,0,-1,-1,-235340674,-1,-1,-1,-235340674,235340673,-1,235340673,-1,-1,-235340674,-1,0,0,-1,-1,-1,-1,-1,-1,-1,-1,235340673,-1,-235340674,-235340674,-1,-1,-1,-1,0,235340673,-1,-1,-1,-1882725391,2118066063,-1,-1,-1,-1882725391,-1,-1,-1,-1882725391,0,1882725390,-1,1882725390,-1882725391,-1882725391,-2118066064,-2118066064,0,-2118066064,-1,2118066063,0,2118066063,0,-2118066064,2118066063,-1,-2118066064,2118066063,0,2118066063,235340673,-1,235340673,235340673,-2118066064,-1882725391,-1882725391,-235340674,-2118066064,-2118066064,-1,-1,2118066063,-1882725391,-1,2118066063,1227133513,-1,-1,-1,-1,-1,-1,-1227133514,-1,0,0,-1,-1,-1,0,0,-1,-1,-1,0,-1,1227133513,1227133513,-1227133514,1227133513,-1,-1,-1227133514,1227133513,-1,1227133513,1227133513,-1,0,-1,-1,-1,-1,-1,-1,0,0,-1,-1,0,0,-1,-1,-1,0,0,0,0,-1,0,-1,-1,-1,-1,-1,-1,0,-1,-1,-1,1840700269,1840700269,-613566757,-1,-1227133514,-1,-613566757,-1227133514,1840700269,-613566757,-613566757,-613566757,1840700269,-1,613566756,-1,-1840700270,-1,-1840700270,613566756],
-		[1840700269,-613566757,-1,-1,1840700269,0,-1840700270,1840700269,1840700269,0,1840700269,613566756,-613566757,-1,-613566757,-1840700270,-1227133514,613566756,-1225335955,-1198373,-2995932,616562687,613566756,1837704338,-1227133514,-1227133514,-612967571,616562687,-1,-1,615364315,-1797559,-612967571,-4194304,1840700269,615364315,-1225335955,-1837105153,-1225935141,-3595118,-1225935141,-1,-1840700270,-1,0,-1,-1,-1,-1,-1,4194303,4194303,-4194304,4194303,4194303,-1,-1,-4194304,0,-1,0,-1,-1,-4194304,-4194304,4194303,-1,-1,-1,4194303,-1,-4194304,-1,4194303,4194303,4194303,0,-4194304,0,4194303,-1,-1,-1,4194303,-1,-4194304,-4194304,4194303,-1,-1,-1,4194303,0,-1,-4194304,4194303,4194303,4194303,-4194304,-1,0,4194303,0,4194303,-1,4194303,4194303,-1224736769,-2396746,-1224736769,-1224736769,-1,-1227133514,-1,-1224736769,1228931071,1227133513,-4194304,1227133513,-4194304,1226534326,-1797559,-1,-1,-1,1227133513,-1,-1227133514,-1227133514,-1,1226534326,-1,-1,-4194304,4194303,-1,-4194304,4194303,-4194304,-1,-1,4194303,-1,4194303,-1,4194303,0,-1,0,-1,-1,0,-1,0,0,-1227133514,1227133513,-1,1227133513,0,-1227133514,0,1226534326,-4194304,0,-1,-1,-4194304,1226534326,-2396746,-2396746,2118123519,-2118066064,-4194304,2118066063,1879565297,-235340674,-2113986673,-235340674,-1,0,-1,0,2113986672,2118123519,2118123519,2118123519,-234881025,-1882725391,-1879565298,-234938481,-3677199,2117663870,-57457,2118066063,238615678,-235340674,235340673,-235340674,-4194304,-235340674,238615678,-235340674,2118066063,-2118066064,-1,-1,2113986672,-4136848,-57457,-1,-1,2118066063,-1,-2118066064,-2118066064,2118066063,-1,-1,-2118066064,-1,-1,-1,-2118066064,-2118066064,-1,2118066063,2118066063,2118066063,0,-1,-1,2118066063,-2118066064,-1,-1,-1,-1,1882725390,1882725390,-1,0,-1882725391,-1882725391,-1882725391,-1,-1,-1882725391,-1,-1,-1882725391,0,2118066063,-1,2118066063,-1882725391,-235340674,-1,2118066063,0,235340673,-1,235340673,-1,-1882725391,-1,2118066063,-1,-1,0,0,-1,235340673,235340673,-1,-1,-1,-1,0,-1,0,-1,0,-1,235340673,-235340674,-1,-235340674,235340673,-1,-235340674,-1,-1,-1,-235340674,-1,-235340674,0,-1,-1,-1,0,0,0,0,0,-1,0,-1,0,-1,0,-1,-1,0,-1,235340673,-1,235340673,-1,235340673,-235340674,-235340674,0,-235340674,235340673,-1,-1,235340673,0,-235340674,0,235340673,0,235340673,-235340674,235340673,-1,235340673,-1,-1,0,-1,235340673,-1,0,235340673,235340673,235340673,-1,-235340674,-235340674,-1,-235340674,-1,235340673,235340673,-1,235340673,0,-235340674,235340673,-235340674,-235340674,235340673,-1,-1,-1,235340673,-1,0,-1,-1,-1,-1,-1,-1,0,-1,0,-1,0,-1,-1,-235340674,0,0,-1,-1,0,-1,-1,0,-1,-1,-235340674,-1,-1,235340673,-1,235340673,-1,235340673,0,0,0,0,-1,-235340674,0,235340673,-1,-235340674,0,235340673,-235340674,-1,-235340674,235340673,-1,235340673,0,0,-1,-235340674,0,235340673,-2147475472,2147475471,-1,2147475471,2147475471,-1,-2147475472,2147475471,2147475471,2147475471,-1,-2147475472,-1,2147475471,-2147475472,-2147475472,2147475471,2147475471,-2147475472,-1,-1,-2147475472,2147475471,-2147475472,-2147475472,-1,0,-2147475472,-2147475472,-2147475472,0,-1,-1,-1,-1,-1,-1,0,-1,-1,0,0,-1,-1,0,-1,-1,-1,-2147475472,-1,-1,-1,2147475471,2147475471,-2147475472,-2147475472,-1,-1,-1,-1,0,-2147475472,-2147475472,-1,-1,2147475471,0,-1,-2147475472,-2147475472,-1,2147475471,-2147475472,-2147475472,-1,2147475471,-1,2147475471,-1,2147475471,-2147475472,-2147475472,-1,-1,-2143289360,2147475471,4186112,-4186113,-2147475472,-2147475472,-1,-1,-1,2147475471,0,-2147475472,2143289359,-1,-1,-2143289360,0,-2143289360,2143289359,-1,-1,-2143289360,2143289359,-2143289360,2143289359,-2143289360,-1,-1,-2143289360,-1,0,-2143289360,2143289359,-2143289360,0,-1,-1,0,-1,-1,-1,2143289359,-1,2143289359,0,-1,-1,0,-1,-1,-1,-1,0,0,-1,0,-1,-1,0,0,0,2143289359,2143289359,2143289359,-1,-2143289360,2143289359,-2143289360,2143289359,-2143289360,-1,-2143289360,-1,-2143289360,-1,-1,-1,-2143289360,-1,-1,0,2143289359,2143289359,-1,-1,-2143289360,0,2143289359,-1,-1,-2143289360,2143289359,-2143289360,-2143289360,-1,-2143289360,0,-2143289360,-2143289360,-2143289360,0,0,-1,-1,-1,-1,0,0,-1,-1,-1,-2147475472,0,2147475471,-1,2147475471,0,2147475471,-1,-1,2147475471,-1,0,0,-1,2147475471,2143289359,2147475471,-1,2147475471,0,-2143289360,-1,-1226535498,-1794049,-4186113,-1,2145088079,-1227128842,-1,-1,-1226535498,1228927561,-4186113,1226535497,-1,-1227133514,-1226535498,-2392065,4186112,0,4186112,-1,-1794049,-1,4186112,-1227133514,-1,1227133513,-1,1226535497,-1224741450,-1,4186112,-1,1226535497,-1,-1226535498,-1227133514,4186112,1226535497,-1,-1,-1227133514,-1,1227133513,-1227133514,1227133513,1227133513,-1,1227133513,-1,-1227133514,-1,-1227133514,-1,1227133513,1227133513,-1227133514,-1,-1227133514,-1,1227133513,1227133513,1227133513,-1,1227133513,-1,0,-1,-1,1840700269,1840700269,-1,-1840700270,0,0,-1,-1,-1,1840700269,-1,-1,-1,-1227133514,1227133513,-613566757,-1227133514],
-		[-680806950,-134480386,-579946661,781667237,1739839980,1327993801,-134480386,-134480386,1807080173,235340673,1227133513,-1,-1672599789,-1,1840700269,1672599788,-201720578,-1,-1672599789,-235340674,0,235340673,-235340674,-1,-1,235340673,-1,1874320365,-33620097,235340673,1840700269,-1672599789,-1638979693,-33620097,-201720578,-235340674,235340673,-1,-1,-1,-1,235340673,235340673,-134480386,-1227133514,-1,-1227133514,-1,-1,-1227133514,-1227133514,-1227133514,-134480386,-235340674,1227133513,-1193513417,-100860289,-1,-1193513417,-1092653129,1327993801,-134480386,-1227133514,-1,-134480386,-1,1227133513,-1,1327993801,-1,-1227133514,-1,1327993801,-235340674,-1227133514,-1193513417,-1227133514,-235340674,-1193513417,-100860289,-134480386,-1092653129,-1,-1,1327993801,-1092653129,-134480386,235340673,-1193513417,-134480386,1227133513,235340673,-235340674,-235340674,4186112,-1193046473,-134414850,-234881922,-2392065,1193513416,-1794049,-1,1228927561,0,-1226535498,1226535497,-1,-1,-4186113,4186112,-1,-1,1226535497,1224741449,-4186113,-1,-1794049,-1,1227133513,-1,-1,-1,-1227133514,-1,-1,-1,-1794049,-1,1226535497,-1794049,-2392065,0,-1,-4186113,-1,-1,-1,239068033,-4186113,237208449,1195176392,1327461321,1226535497,-1193513417,-1794049,-235340674,-1226535498,1193046472,-1092120649,239068033,-1224741450,1194914248,-1092382793,-134414850,1327461321,-1,-4186113,235340673,1912136590,1327989641,-819491271,-235078146,-1798721,235340673,4186112,-1,-1,1794048,-1227133514,-1,-2392065,0,-4186113,-1,2143289359,0,-1,2147475471,-2143289360,-1,2147475471,1912136590,-1908867087,-459649,-234881026,-235339778,-4186113,235340673,-2147475472,-1,-1,-1,-1,-1,2147475471,-2147475472,-2118058000,-235340674,0,2118066063,1882725390,1912143758,-2147475472,235340673,-1,-235339778,-2147475472,-235340674,-1912136591,-235339778,-1,-1,-1,-2118058000,-1,2118066063,-1912136591,2118066063,264764417,-235340674,-29425537,-1,-1,2118066063,-1,-2118066064,-1,-1,-1882725391,-1,-1,-1882725391,-1,-1882725391,-1,1882725390,-1,-2143289360,2143289359,2143289359,2146966543,2146966543,-3677185,-1,-1,-264241154,2143289359,-1,-2143289360,267918337,2146966543,-3677185,-3677185,-3677185,-2143289360,267918337,-1,1882725390,2143289359,1882725390,-1,-33496961,-29360129,2118066063,-1,-2113929232,2143289359,2118066063,-2143289360,-1,-2143289360,-1,-2113929232,-4136833,-29360129,-1,-1,-1,-2143289360,2118066063,-2118066064,-29360129,2147426191,2118066063,-2143289360,0,0,0,-1,33496960,-4136833,2118066063,-1,2143289359,2143289359,0,-33496961,-4136833,-4136833,2147426191,-29360129,-29360129,-1,33496960,-29360129,-4136833,2143289359,2118066063,2143289359,-33496961,-1,2147483632,-2143289345,33554431,-29360144,-2113929217,-1,4194303,-1,0,-1,-1,2118066063,-2118066064,-235340674,235340673,-1,-3734655,-235340674,-3734655,-1,-234881025,-1,-3734655,-235340674,-1,235340673,-1,238615678,-3734655,-2118066064,-2113929217,-1,2113986672,2117663870,-238558223,3677198,238615678,-1,-33037314,-264241167,-238558223,267918337,-29417586,-234881025,2114446321,2118123519,-57457,2118123519,-2113986673,-2118123520,-4136848,-1,-2113929217,0,-1,-1,-1,-1,-1,-1,2143289359,-234881026,-33037313,2143346815,-3677199,-1,-1879105649,264758270,-57458,-4136833,-4136848,2147483632,-2113929217,-2118066064,-29417585,2147483647,-57472,-2143289351,-918552583,-2396737,2147483641,-918552586,-7,2147483647,1228931071,2143289359,4194303,-2143289360,2143289359,-1,-1,2147483647,4194303,-1,-7,2143289359,-1,-920949319,-1,2143289359,-1,-918552583,-1227128842,2147475471,1227133513,-1,-4673,-920346631,-1,-2147475472,-1,0,-1,-1,2147475471,2147475471,-1,2147475471,-4186113,4186112,4186112,-4186113,-1,-4186113,-2143289360,2147475471,-1,-2143297521,-4194304,-1,-2143297521,2147483647,-8177,2147475471,-2143297521,-2143289345,-4194304,-2147475472,2143297520,-1,4194303,2143297520,2147483647,-2143289345,-1,-4194304,-2143289345,-1,-2143297521,-8177,-4186128,-2143289345,4194303,2147475471,-2143297521,-1,-2143297521,-3505,-1224736769,-1,-1227133514,-1,2147480143,-1227128842,1227133513,2147475471,1227133513,920351302,-4673,2147475471,920351302,-1,0,-1,-1227133514,1797558,-1224736769,1226534326,-1226534327,2143297520,1228931071,-920351303,-2392074,-8177,-918552577,-4673,-2392074,-1228922896,-1226535495,-922738704,1226534326,0,-1227133514,-1227133514,1227133513,2147475471,-920948154,-1,-2396746,1224736768,1226534326,-2396746,1228931071,-1224736769,-4194304,-1,-4194304,4194303,-4194304,0,-4194304,-1,-1,-1,-1,4194303,-1,-1,-1,-1,-4194304,4194303,4194303,4194303,-1,4194303,-4194304,4194303,-4194304,-4194304,-1,-1,0,-1,-1,2147483647,-4186128,2147475471,2143297520,-1,4194303,-1,-4194304,2143297520,-4186128,2147483647,2147483647,-8177,-2143289345,-1,2143297520,-8177,-1,2147475471,4194303,-2147475472,2143297520,2147475471,-2143297521,-4186128,-4186128,-1,-1,2147483647,-2143289345,-1,-1,-2143289345,-4194304,2143297520,-8177,-8177,-1,-2147475472,-2143297521,-4186128,2147483647,-2147475472,2147475471,-4186128,-8177,-1,-4194304,2147475471,2143297520,-2147475472,2143297520,-2143289345,-1,-2143297521,-2143289345,-4186128,-2143289345,-1,238615678,-235340674,-1,-1,-1,235340673,-1,235340673,-1,-1,4194303,-1,-238615679,-1,-4194304,238615678,-459650,-234881025,-459650,-3734655,-1,-238615679,-459650,-3734655,-459650,-235340674,-1,-1,-235340674,235340673,0,-235340674,235340673,-235340674,-1,-1,-1,-1,0,-1672599789,-1,-1,1837704338,-1,1840700269,-1,1228931071,-1,1805545271,-34014082,1840101083,1192109843,1806415322,-167903489,-612967571,1840700269,-1636172546,-234881025,1841299455,-713959188,1838902710,236744247,1671393207,783267803,-612368385,0,-613566757,-1,-613566757,-1,-615364316,-612967571,-1225935141,-1227133514,-613566757,-1,-613566757,-1837704339,-612967571,613566756,-612368385,-1,-1198373,-1,-713163009,-1067045,-2995932,-1192109844,1838902710,613566756,-1,-1,1840700269,1840700269,-1,-1,-1,-1,-613566757,-1,-1,-613566757,-1,-1840700270,-1840700270,-1,-1,-1840700270,-1840700270,1840700269,-1840700270,-1840700270,1840700269,-1,-1,-1840700270,-201720578,-235340674,1840700269,235340673,-613566757,714427045,613566756,-1672599789],
-		[1840700269,-1,2008406765,2123540820,-546326565,1419070117,-545604258,-1518379089,-1344844891,-1,-67437186,-25756803,-226886402,-238558223,1874279325,-1916345487,495855379,-302630115,-302063761,-1879565298,234881024,264758270,-29417586,-470681347,-1611055981,-59245570,1874780015,-1614880914,-37748739,533645166,-28317698,503201555,-1,328083219,-34087955,-234938481,-3677199,-3677199,238558222,532561695,-302523411,-472922141,-301998097,-2113929217,613566756,1519561044,2123516196,-4136848,-1521309356,2118066063,-3734655,716897662,-170439771,-234881025,-613566757,235340673,-714427046,-235340674,-1,235340673,1672599788,-1638979693,-1837499138,1673133310,-1227133514,-3734655,1672928109,239075327,-613566757,-716027612,-1067045,-1669595650,-680609957,-1093720173,-681677001,-235209346,238943999,-2667611,-2995932,-167903489,782939482,235340673,-1522474896,-16417,1519577460,-1,-615364316,1522474895,-1198373,613566756,-1198373,-1,-1224736769,-1837704339,-3595118,-2995932,-1227133514,-4194304,-1519536421,-9611605,1828117350,-908358193,1835906779,-13805909,1532718518,-640238,-308580933,-613591377,-1198373,-2147426177,-626174300,637493156,-1530920972,-2113929232,1521309355,-1511131404,-1530920972,-2113929232,-1,629088128,636335835,802203739,-707269634,-27201409,-1669595650,-266564024,-202927159,-304378426,-688530707,-1410213661,988706619,-1816133648,-1639455746,1912545167,1871176702,-528482319,-38323309,-302039139,-3595118,-1,2146285270,-1975517200,1598957340,-1803550731,-581079369,1192109843,-1107952899,-303171845,-1638405122,-62914562,-1615898715,-27135745,-680142099,-2118066064,612967570,-1,-3595118,-1,1225335954,-1,-1227133509,-1,1520143814,-1815485187,-1837696131,-33496961,-1816732819,-4136833,327795938,2143281135,1941560558,-1,-1,-4136833,-235340674,-33496961,-1,-1,-1,-2118066064,29360128,-238558210,239017857,1883185038,-234881026,-29819777,235340673,2143289359,-1,-459649,-235340674,-1908867983,-1908408335,2143289359,2118066063,-4136833,-2143289360,-264241154,2114388879,2146966543,-238615679,-3734655,-1879507842,-459650,268435455,-1882782834,268435454,2147426191,-238615679,-2114446322,-2114388866,-1882725391,2147483633,-2113929231,2118066063,-1879048207,-3734655,29417599,-234938495,-264241153,1912602623,2146966543,-234938495,-29417600,238615678,2146966543,517119,-264758271,-517105,-1879048207,267918337,-267918338,-2143289360,2147426191,-235340674,-2143289360,-1908867983,-1908408335,-234881026,2146966543,235340673,-1,-1,1882725390,3677184,-29819777,-1912085519,-2113929217,-3677199,-267918338,-8192,-16,-4194289,-2143289345,-1,-1,2147483647,-2143297521,2143813617,-1,-2143297521,-2143289345,-523266,-264242161,-2143297521,-264764418,-1009,-264242161,-8177,-264764418,-2143289345,1879565297,-264241153,1882725390,2143297520,-8177,-4186128,4186127,-29360241,-29417473,-1,-29425537,2147483535,-2118066064,-4186128,-2113994625,-2143289345,-33489024,2143354864,29425536,-7281,-29424641,1879565297,-29417473,-29417473,-1882725391,238616574,1912595455,-2113929329,-4128784,-2143289345,-2113994625,-4136960,-8065,-1,-1,2147475471,-2143289345,-8177,8176,-235347954,-1912136591,235340673,2147476367,-1912136591,-1908408321,-8177,-897,-1908416497,-3734655,2147483647,1912136590,-238608399,-234881921,-1,-897,2147475583,-33031055,-523266,-1882718223,-33488912,-1,-2143297521,-1,1912143870,-33030271,-29360129,-29418369,4194303,2147475471,-8177,-1,-2143289345,8176,-2143289345,2147483647,-2143297521,-4186128,2147483647,2143297520,-2143297521,-4194304,4194303,2147475471,238615678,-458754,-8192,1912595342,1908875249,2147482737,-1908416497,238609281,-2143748097,-1912135682,239075327,1908875262,-235340674,-1908409231,2147475471,-1912135695,-1,2147475471,-1,-4186113,-2147475472,-4186113,4194303,-4194304,-1,-1,2147483647,-2143289345,-1,2147483632,-1,-8192,-4194304,-1,-1802240,1227132342,-4194304,1802239,-2396746,-3511,-1794049,-1,-1,-4186113,-4194304,-1,4186112,-4186113,-235340674,-234881922,2395583,-1910541362,-3997183,-1864129,-819491271,2013000654,1327989641,-1092649353,1908874353,-897,238615678,-234881921,4194303,239068033,-1,235347070,-1325400065,-1224999297,-197122,1327730231,-1090519041,-136544770,1191644727,-1193312257,4194303,1329132489,1325407358,-238608898,-134422015,1194914761,237477503,-4553,-1224740279,-100925825,-1090519041,1329587073,-1226530817,1195110856,955835278,2013265401,-1226535489,-134418498,-100925456,-1090981961,-134221890,-1191449033,1327993801,-196609,1228927561,-234881922,-1,-102990338,-1224736769,-922738704,922742207,-1226535495,-918556081,-1224736784,-918556081,4186121,-920351303,-4673,-2145083399,-1,-1227137009,-1802225,920952831,1794057,4186121,-1224736778,2145088070,-2395575,-1226538993,2147483632,-1224744945,-1228926391,2147478975,-4194304,-1,2396745,1228931071,-1797559,-1797559,0,-1,-1227133514,-1227133514,1227133513,-1,-1227133514,2147475471,-2147475472,-918557255,922743366,-920346631,-1798721,920949318,-2143289360,-1,-1224741450,-2396737,-4673,-1227128842,-4186113,-4673,-1226530826,-918557255,-1,-920944647,4186112,-1,-1226535498,1227133513,-1226535498,4186112,-1224736769,-8192,1226539001,-922738704,-1,2143297526,-1226535498,2143297520,922742201,2145091568,-918557255,-1193041928,-102993993,-2400241,1912398721,2143296625,-1859599,-2009079281,-1092119993,953974776,2011135481,2143297535,-458753,1908874353,2147476353,2147023998,239075327,-234881025,-1,1226534326,-136351817,-3924495,-819489847,-954438089,-952110641,-234881921,-134414721,-3924495,1327993801,1054904248,-818351154,-1325405121,-134484418,2147483647,1226534326,-2143289345,-920949322,-2133945,2044960376,2143559665,-819491271,-2042822657,-134487154,954441720,-953970232,-238609295,-820022329,2145689593,-100860289,1193045559,1329331710,-1224736769,-818089009,2147024894,-1092653129,-819491783,-134484418,-4186113,-2143289360,2145088079,1228927561,-2143289360,-1193513417,1327002569,-103055745,-1,-1226535498,-2392065,-2392065,-1,-1193046473,-134218242,1054831175,-1,-1227133514,920351302,2147480143,-134479874,2046616463,-1327988746,-134480386,-1092653129,1327993801,-1193513417,-1,-1,-1193513417,-100860289,920351686,2147480015,-134480386,2147475471,-1092653129,-1092649865,1327989641,-1054826504,-1193513417,-1227133514,-1,-1,-920346631,-1,1227133513,2147480143,2145686095,-1224736778,-1,-1227133514,-918552583,-2143289360,-1227133514,2143289359,-918552583,1055167047,-1193513417,1908867982,-459649,-1908408335,-235340674,-1226534321,235340673,-1224736769,4194303,-1,-2147483648,-4194289,235340673,-1908867983,-235340674,-1908867983,2143749007,-201326594,-275896547,2143355023,1670326163,-993181,1945557870,-1669332996,-135145236,-1228331875,-201720578,-1191641999,-33620097,886730047,-1433458790,2146950123,-1638979689,-2145086897,237212097,-1910008889,-134677378,920949327,-1908867976,-10,235398142,-267918337,2147483647,-1912085506,-235340674,-2118066064,-1,2118123519,2114388865,2140022547,-273163282,-33554417,1670301539,234823567,-1636418797,-474005613,-33669345,-1651579011,-25215086,-264700815,330381468,1920400242,1672599788,-1879105649,1838073699,-923034162,-1224769610,-1879048208,-134480386,469425507,-1226936522,-33299969,-343005411,-2009071631,-1092653129,-820483015,-134480386,1739839980,-1126273225,781667237,-579946661,579946660,-1159893321,-134480386,-579946661,-680806950],
-		[-614633801,-137147861,1021213549,-201397467,2010335909,-3727873,-615098661,-169370241,1467209564,2111201274,-579946529,-749316129,-1564480593,-545720354,-33950497,-819880713,-613565972,-376684545,-613501221,-104398463,1226796983,-1600569,-1192779785,-1092575240,2013003257,-8191,-136277937,-1803550721,477486590,-3932161,469622556,-1224741057,1572059611,-465911149,-421040417,-2009079281,-29367807,-4186113,2017140111,-1228890183,-908358193,-3907023,-1209569671,-1889,-25740387,2147426191,-230038286,1670265059,-1610668161,-230088707,-305144673,-1817100292,-33628317,-3728258,1878514590,-4135955,-1647384707,-473031923,1652111587,1946076031,-495846414,-1618526227,511631212,511064851,-204473106,-4128784,-208672781,-1651623709,-1793,-2113937393,1845493746,-4708369,1917452061,261094627,-269352964,268428287,-1879049089,-1912078463,-1882726287,2147483647,2118123519,-33489920,2147483647,-1908408321,2143756273,-8066,1912086513,-3728271,-1912085506,-29425551,267918337,2147482750,-234939279,1908875256,-1908867079,1912143864,1908875263,-920346625,-920686665,2042634224,1226798017,1912139710,-266823,-2392577,-235078018,-262658,2145091574,-6241,-508240261,2146951163,-1467007874,-6241,1467080567,-887455745,-545924374,2143362928,-1159892530,-955834370,-955841650,-1945756431,-1397302429,1228331874,-67706909,1982939036,717176819,478093311,1676748643,-32899103,-1636172557,2146445283,-201381986,1652504607,-58785940,2113995007,-1870725261,-1639366804,2113928434,-205061135,-1615447277,-234882033,-1879048207,33554416,1908925326,2147475583,264758271,-114,2143813519,-2113930225,-234881026,2113994751,-1879048207,-29417473,-1,-57457,29425536,2147475583,-4194304,2143354752,2130771840,-29818882,-100664193,-1220545985,-807407625,-57345,1055364031,-268369922,-523265,29884288,-234881026,1879571470,-1879507855,970780231,1999082494,1105489399,-820019714,-4186122,-1228922881,2147483647,-2143289360,1226538998,1227137023,-8177,2145160783,-808976393,1071944262,-807629241,2032661497,-517106,2143289344,2143812616,2030566415,-2097743,-920944647,-1879564290,-264245249,-807178248,-577,-264764418,958172743,1071644663,2145086902,-3670031,264242160,267912190,523265,-1188561479,-918552577,2145153479,2011667961,2145353719,2147476367,1088389239,-1191411713,-1091003266,-102526008,2145353719,2147483135,-4186128,-1054826504,-138149362,2046616463,-2143293889,-2009338306,2145088457,-818090567,104857599,-1227133511,1193045567,1193517055,2143297535,-1222348234,267919359,1192988102,-1187209650,-939524097,2129097656,-1327960970,1912577998,-247463938,-104340481,1327468489,-1580033,1340603913,-7169,-134480258,-2017205264,-1912136079,-2147480144,-1193513033,-100860417,-920129543,-1222643722,-819683335,2147426191,-1192981065,1910469198,-102527553,2043085711,1054831175,-1798721,1327461321,-1908408335,954428999,-236723202,-918557255,2145164303,-1187212873,-1073963593,-113544193,-1,-958100040,-165453186,-1224765498,1059581944,-1175982466,1340076030,-1573432,-15,-2387393,-918555658,-136077833,-2124865,-922776074,-1879048193,-4128881,268427391,4193280,-100667408,1055366647,-1092587575,937194439,1340568502,-1054344113,2117664767,-1059090375,-16834640,-102797769,1191182334,2134703729,2134580622,-152603719,-941362698,-17039880,-29622785,921146311,-266574961,-1980013954,-12607553,-251663297,2118059007,-1086327873,-14684169,2046619583,-1086586882,-2297417,-2017198081,2117926399,-134538225,-100671104,-251658633,-1210289209,-29564032,-817893297,-165706761,-1222872121,1881109447,-953901063,-952369666,1056764879,2147483647,-2322874,-958402568,-16974215,-1879048207,-958100040,2143490639,1327267321,2143294415,-819757113,1325342654,-1075351559,-1042223113,-234884153,-817889287,-100929994,-3989889,1056960446,2043085710,-3989520,-1908408335,-1908677745,1225199550,954437176,1193513415,-1193045553,-1325400074,-1224933761,-235078146,-822087746,-2015400007,-1192976904,1323068983,-18550151,-1879048207,517119,-3677186,235398142,267975807,-459650,-29360129,-267918351,2147483535,-1912135681,-267912191,-458754,458752,2147483632,-2143297536,-29368208,-1979736186,-920350138,-824205433,-134247353,822050742,-918753346,-2013201009,2011667959,-2116255618,-39944338,-1644746755,-4201570,-59105040,-294690946,-302612579,238854142,-1782595622,-692595363,1869315991,-941821889,-603988060,919762630,2142061794,-1523152561,1837096870,-312765834,-8392135,-1224741441,310370319,2145091581,1534514911,-1224741447,-614765130,1228326045,-1224737937,922742207,1203051,-1225933978,-1225334785,-1794049,-599187,1837710189,-598017,-57345,2143248351,1774188379,2089805811,-1208225866,-1616920796,-543426579,1677245531,67100673,1676626011,-2621441,-436282077,2146868047,-229118987,-1961788677,-834941014,-137012505,-1158554113,-340001543,-1225336467,-580608004,-341668166,-1598035293,-1124607513,1260687945,-168571985,2144491222,-612368387,-1224736778,-1227128833,-3590437,-1837105153,-1843109998,331341807,2146190447,719179500,-379879431,529904942,1499460600,-688161783,1054605190,-545721864,-545776193,1706622519,-18571265,-25689089,-1393,-710952724,-489332742,1577058301,1609946973,-311511210,-1526784113,-1615496515,-1092915786,1021308919,1340601343,1342177279,1805947465,1675915831,27201408,1677664129,-1,-88326637,-1726222849,-704651284,-1824522241,-629186652,-1210348087,-21562497,-1,33496960,2118123504,-33554417,1535115263,-11379857,-1512957109,1536342815,1837105078,-1219313665,-14974992,2125313759,-1839497217,1539913435,1845460987,-625564417,1819721663,2138500575,-1774087,2147467218,2144487723,-1533870163,-1220618,2146893789,-10,-1189034,-2366238,-306188803,1838907385,-1531529493,2133712884,1828109149,611166507,-1212196033,2147442806,930511739,321764192,-11927653,922712927,328973172,-598026,2127752676,-1228927553,1841299455,2145689599,1533918059,2143890877,-2337,2147481311,-1531520147,-612968739,-309772302,-1224740273,1837704338,-1797559,-1327727106,-1091315877,-612368385,613763364,-68306981,237874085,-1840099914,-1075548169,-1521358476,1672193811,901627615,-237011842,-496416925,1605245436,-1384161297,-1191461801,462813500,-346058247,1589583759,920949327,-201523202,1498896219,-747253825,2113928952,-1364543652,-1911204009,-952109345,-1065347,1227472126,-100798785,-103928923,-257,1329587073,-1663105,682604507,1194306469,1639111405,781073389,-69375067,1837112830,1739499155,-1192707091,-613032065,-35822190,-1158693121,-1126736642,-36479490,-1859585,-169170139,784004955,-578820205,-234881025,582941950,-1327529985,1840634605,-69432,-1635848393,1738640694,1840699355,-1090979348,1806883565,-579884379,-1124139137,-102465152,-201458306,-1193315383,-235143553,1093111735,-1090519041,-613038189,1805077413,-3997183,-67240578,-67441755,-168100482,3728257,781539035,-69772162,-545587842,-327937,-715830363,234881920,3734654,-3677199,-897,-728385627,1610438655,-1908408433,465559471,1907013082,-1786921249,-186777737,-143786261,-340406467,-76546341,-153321993,2123290772,471823313,327820947,-469959170,2122234686,1944890813,-227709601,-295215,-227410947,-57457,-3677199,239017871,4193294,4193294,-516097,-57345,238615678,34087954,1669981426,-327155713,-25699074,4194303,-328083220,-503193837,-470164243,-28014594,502619923,536574097,-269402695,1264283062,535450919,694533999,-237126497,-1728368,-61872142,-58720257,-677530331,-153092389,-59848851,1428127743,500098931,-33496975,-1879507841,268435455,-29876225,-239067137,-234881025,-1879571471,-33037313,-1845419873,531708060,1844853522,2118771954,2144216860,-226894610,-234429583,1941519518,1874279325,-260514706,235398129,532668399,-226492417,1840659229,-3603326,1941506035,2080374129,264707825,2147475471,1946492286,-801117270,-176489090,-1522483281,-625302786,-1601871887,-1410377813,-728410795,-76890758,-235341858,637397295,-258621451,-1350738005,2009938927,-302104801,-3734641,-443729172,-272696467],
-		[1839978347,-451437,-1151633961,-74786530,461098391,-941658185,729743227,-1254375969,-546702363,-941656097,-1644767815,2021577174,-254302466,2136060491,-1351877461,524286831,-6922251,2044683670,1862265127,1334389949,-420094993,1868398461,-268702021,-1143148932,2006559958,-409112113,2147149807,-151060665,1979669807,-70795279,-557890817,1975590270,1521465045,1447032287,-1346232587,-29491343,-1074072881,-821882417,-200925186,826077633,-231084610,920616441,-1086463090,2080183871,-1882195779,-819359942,1304427389,1069528858,-538317574,1130339390,-575218852,-1318588457,-18564042,1073541054,-235388961,-1061163082,-537958411,-1191990697,-1610887581,-994185736,-815337495,1037824767,2113921007,-89805521,-1648952453,1205291942,-881329161,967480655,-1816215715,-500435461,-25174035,-296572588,-212930577,-337949383,2113855341,-119569799,939228167,2130670518,2116025406,-807100423,-1088656328,1340313023,268428279,971013561,1073738673,-3932240,-264568706,-1312612353,2045210569,2030018105,2147426190,522873293,-97386866,-555508097,-353467261,-38929411,1471248894,-2115764571,-848332948,-105011030,1794099931,-1264091570,-540358804,1307535947,401750237,-1090047656,-56991889,-253349735,783962990,-656752816,-688198215,-1344930456,-73401382,263984127,-185215121,569503305,819984223,1981783961,-606338,1877982587,-359732242,-916918791,-470811923,735989546,-14769157,-1512131929,1040068095,-44118165,-135303337,1329307078,-711326,-103485575,-1812268189,-624696834,-1145438977,-1114773773,-81953,1060427740,-34079761,165739983,-2133047,-1075351554,-1883177858,-807440328,1055391678,-1074228296,-2060295,-254219137,-2387975,-115109953,-236982154,1226805302,-1052803697,-103058871,104394736,-829319681,-271661281,-302081,206957551,-61702191,531871724,67063293,-307011885,-260160322,-1515488043,-1222115399,-60861601,728952043,-25175683,1564238824,-67764881,2129505435,-1879709697,1994911688,-1179146389,2068550430,-1663682588,1400480575,477002962,-1078592291,-1395210954,1339923903,2075289085,-472002401,-286603805,-112592963,-1627431123,-572538985,-142855682,2134605431,2079867902,421411255,-1381044518,1727381011,1870358750,-113544577,-1228631095,1189084750,1912372351,-918588425,-3730489,2010947535,-1,-272736669,-339822726,-818045511,-1364216492,2044621006,-1160954899,-1292575113,616492991,375093084,-781537475,-136274433,-2663042,-818217735,-615700581,1327331283,1977449723,-818282503,-1226403127,1329594367,-885064194,-1397496195,-1091911813,-33592745,-191971841,-224257,2143576526,1996291575,-1093111735,-819490888,-7808,-819490225,-953906122,2147480271,-819817529,-235347569,926646143,2063465743,1672062099,-592489051,716365787,328130402,-142758446,-1103591426,1706020571,1566490077,-546734812,-1327334146,-612403475,662544000,1675588296,1339655039,1398931455,-273319235,-1773228554,2147426255,1601545571,-1386312505,-8786121,2028171705,-1615167183,1727751667,716004093,-119145731,-45229588,-46668052,-98576897,-119829047,-1944229083,-985666506,1874783013,-1124273463,1261021997,-34619400,1706149788,-68838417,-546653459,-36614855,1021212372,1911806575,1467473150,-680674050,-615030786,2013265407,-1090521389,2145818075,510375134,-1497899533,-612504261,411041267,-1638976161,1056961631,-101264161,1774184495,-682533122,-1056567497,-1497762606,-819953266,-1327390978,-952308033,136053633,-2627693,-1225402661,-136344265,265213413,-655479,-134284437,-1321779589,1228264927,-954168664,-1194508939,-1090585745,-135549506,-1636847939,1565875699,2145684670,-560579673,1860030023,-1364231844,-480258,-22221596,-62915569,2117853809,-113246785,-885035599,561971327,-478610314,-1570267569,887619126,-492210309,-319072549,-134253634,-407962645,-1612186023,-219564205,424507380,-1378228321,845069167,536803484,266826288,-1636980765,-95949679,-563350148,-18033313,-562734098,1672313550,-605151297,-1879337985,1818204362,-1837477889,-306186533,-234881025,-168143109,1947724789,-1401426265,-1764110489,-75776289,996372920,-460446563,-941887493,1402972845,973016137,-817917391,-1868361,1802230,309182450,-598017,-8183,1533919227,-1225334789,310372653,2146883295,1837710178,-1531525998,1531524831,-2346,-613564419,-1224739111,-1326858765,1911742453,-1040197922,-1082403155,-436740395,-1174901175,-1652852589,-541352519,-230224335,1996951503,1803550273,-388107,-1648403653,265214985,-1533204213,-4408449,-319916577,1191181521,-352279314,-1879112811,-1682487955,-570790063,-319363841,-139528389,-102791490,-113569223,-7143541,1325092808,1559685085,799714266,-2173225,-338708317,-617744645,-906026881,-623707345,-1837123841,2145086909,-306184193,-1837718314,-1820262526,-621357409,-29424385,-613565569,2146297078,307375549,2147426182,925711199,-597923,-727996041,1882013823,626174331,-704701405,1602177406,-32134869,-1817331715,-4145156,-466899,-257418,4186094,-1890979201,-1431586947,-38183888,-1819687690,-716587122,-403972317,-268918789,-1837717153,-3222177,-1351169746,1939570687,1873655083,-1901438992,-1442886173,-487725074,1156491999,-1388577553,1505716825,-2100746323,-1163625523,1778375679,-1484829202,-271246403,-20976204,-1610678420,1493167543,1001511167,-1114468657,-37617668,-1912602626,-1945110413,-2117085153,-1912604643,-28908673,1838144285,2985431,1845491742,1204763875,-1101006146,2037840694,-407766856,804699003,-170997985,-321581682,-6285,1793488827,-1373951714,-18459862,-1211374083,-807144963,-201720578,-1990806357,-399378,-1117260801,-17235969,-17239937,1023143497,-581173297,-307785927,-715391042,-1963003441,-1179502316,-74809377,-1109550079,-135288083,-747818744,-341939540,1979709407,926636661,1819977715,-545732635,1603922160,-1273400106,2130042623,-537136690,-71308638,-1011556871,-164975961,-487207491,1060023735,1536998251,-1435087009,-2397745,1438332796,-1885015119,-305659911,20971126,-1409319417,-272633791,-1639452162,-1226310210,-473960840,-1770003905,-1193308545,-1226334641,-509148167,1073704014,2137325566,-1363153024,-1480876617,-296241881,-118535457,-102268913,-1208568529,-549589657,-579946509,-2889825,-470565988,-495435887,-295928708,-202899457,-304349185,499646350,1872078691,-115029,616495591,-1225006082,-819470322,-1094189170,1224744841,716709875,1839900974,-333522,-714233557,-171165744,-2977,-1155304904,-153096714,-1006688257,-479545509,-1581250260,-546056356,-1784843579,-193105025,1937473423,1738269563,-101112450,1538914003,-175833210,-195667041,-185188978,-551589207,-127951481,-893470885,-260714659,2145876053,-1648975129,1978662661,-261628682,-1073832689,-1812895923,-984748036,-114704,-72122627,-497426532,-1845724615,-2113,-1615806547,-59762674,-63979521,-272677793,-1616382028,-839766031,888536339,-208020226,-528748581,-171577161,-20537985,-318834530,-9159842,-33567681,-696574235,-1674461998,-713427675,-34093154,1655831405,-1644697614,-1183243281,-268961649,-590702,1673518994,-1878442082,-1611956239,1228667116,653912033,1261288433,-1486658563,-609000289,1852759035,-148332,1454979053,-79677729,1605821298,1609962941,-356975362,-134611986,-587868355,766376779,1035934525,1838912118,-687869315,-1515360928,-1814168322,-307381246,2121744381,-308764688,2115829639,1839141496,-4481586,-624943153,2143253055,-1834680375,1535079023,-541920037,-1342769424,-606374222,-125835619,-623706147,-1042907137,-914375233,2013257210,-1128062117,-1140935764,2130411510,469237103,-153750786,-1074927681,-34903321,-919341362,-276991625,-197165341,-35946595,167534462,-30541441,-403998109,-140855043,-340785695,-71311569,140476415,-1913045135,-1779933185,1716190566,-884666577,594778261,-906009993,-12845175,-2592311,-1979748785,1094217209,2147480176,-1090519426,2011143239,-1908604930,-546453779,-1091180043,1599140862,401864355,866740730,-2662541,2130669049,-52529364,1998944505,-1554560139,1610096607,-33555215,2143297535,-547990310,-749643809,-1363548117,-393217,-204382342,-202392459,-559287121,-1342493698,-1159890457,2079915662,-342132067,-1418832133,770571695,-1879048817,-1417808038,131287717,-1431388290,-66644005,1553290971,-215521299,-680669478,-134496683,1807597015,616201527],
-		[-1770700804,-1260195889,-545928213,-692387842,-540286162,-328507909,1569605550,2096961990,-120406097,460970866,752467187,-859836857,-1484789250,479121848,1006500157,-13730657,-715302659,-541425793,-8456358,-1880228338,2126465957,238021412,1958352635,1580070666,-688176528,1459362669,1960292142,1871704583,-35355145,-675481739,-1784739469,-262214358,1203425050,1052638914,-671296297,-1653084865,1945631572,2006626007,764568459,460566192,-1925209683,1675049963,-160543723,905543550,1568602108,-302614017,-22090261,-17002497,-1108362441,-309241869,-1828720571,-570567430,-224464049,-763921813,-1170742412,-21840209,-541542514,-1648407618,-872757282,1643659198,-1552684069,-27405284,-67437730,-564531871,-1229258819,859619239,-27789347,1040177612,-38839841,-820253239,-1128317139,-541137486,-808790354,-590358563,2023390043,-2556017,1945836551,1056674814,-1998556231,-2015399986,-12844607,-2084102146,-1254355569,1073254391,-1076890970,-545571841,2000664575,-1366303021,1617952575,2103037623,1743176319,-201704453,2007678377,-403839011,-555845638,-405047351,-1212221450,-352601476,-1107821671,1872682358,2075651362,993819623,2069821695,-139610409,-5509590,1332996042,-856016167,2078277599,-715130274,-1091929353,1266108182,-285281340,-1711348321,-102906473,467109768,1093322751,-161501541,-184829089,535147114,233994190,-1345287747,1604008318,1894490077,2146130940,1055869941,-135267202,119011326,-118752312,-279528835,-817929921,-377761829,1040187385,-303464492,-649688145,-473498215,1073739262,-143143849,-1431394871,-679543364,-1227786277,-1410499755,-816239638,-1230126377,-942642270,2117698014,-68093208,-1241532110,-275157442,126943078,-1533845547,-1536827441,-70631842,1380690207,-304644239,987586550,-1193410699,-1380536834,2132741421,-301997514,-620460307,-1228087359,-1241646755,-1375736130,-337752185,-1142132993,-306398243,-672596505,-1618166373,-1713145604,-677585,2138042127,-1422984209,-1276452229,919949270,-92897283,-61082539,-306791523,-1688289305,1524628159,60388967,1870983167,-1226907937,-1745977349,2005650934,-808196370,-799195942,-562666060,-1142644775,937131766,1945826779,-1167624745,-1350698442,1977037709,-1375949138,-397410334,401366943,-1353843489,-541921697,788516735,924482874,-277876910,-170133629,112191467,-1687905618,-1650852154,-213024786,-327230609,-954298497,1520418614,-393225783,-1091121345,2059007929,-546898529,-141623361,-1227055297,-1885215364,1868428030,2109593419,2146886143,920552926,-547881857,-1398895590,-1105609729,-1133084687,-31879176,-1625243777,1590156831,-19732097,2114959743,-1131175938,-1117814882,-547500219,1047494119,-2186241,720370427,-471740442,1675576175,-629951185,-1114770502,1775729427,914994170,-1668882497,-271857679,-824899,-1154406409,-539150731,-75634690,-143413505,1589567341,-581435412,-2335441,-452302861,-1083279543,-875102493,-345032708,-15220880,913307951,1960799995,1851694811,-1365586439,1489828796,-50596913,1138281038,1962442681,262137843,1586708095,2135910386,2012804215,1132005616,-383283201,2133573614,-346308645,-1645216789,-998244493,130871231,-33325519,1899878911,528172655,-1162145509,532987837,-5252102,-746759186,1927541754,-10560517,-274744339,-37096819,-836852993,-1311560367,1862237502,2113891775,-1826885649,-1141989265,2147104690,-1952260100,-1211666953,-1308714307,-139595009,-1095312411,2062284662,-1109420249,-875959307,-860963943,469607799,-882905280,-1082135685,1878381262,-1361662098,-374539077,1056955027,-3555646,-101204594,-281055794,-57751692,1731644442,536583813,-823013583,-84750917,-134353221,834513895,-1175228466,2134898423,385736381,1610050735,-213275204,-1112876197,-935530576,1274606479,-62914639,-855835137,-1225003976,1288395689,-1092034647,2146682927,-273692717,1001543415,-102327815,-1839866115,1803513078,-1302997261,-1627615302,-672137603,-1680492755,-1207967233,-1619297796,1807290109,1929099111,-1073913863,-929940,2078555249,-328729567,-604809539,-1109553254,-1573652522,-1129316491,-537962340,-92519938,-486925327,2068949727,-143284130,666258795,-12679584,-69731587,-240939021,-580190356,502258158,-1521501525,1576959471,-1297882140,1093508453,-752141351,936868429,-976540685,1710947258,-140772563,-873734178,-306708757,-687865860,-76687329,1744240096,-155325185,-537202691,-682893931,1585217144,-1623444161,-9332097,-235160235,-639644437,-2015436932,-193817679,-40172866,-1090782161,-204497601,1610312700,-336150817,2134277522,-55740650,-35668505,1002159550,1029697991,2111353070,2145132253,-84489298,-33956215,261472221,-648855554,1449110846,-48505348,-55332137,-20132883,-269583689,-451161,1673517999,-234906530,-134259277,-69408321,-61956257,-1267271694,1480046818,1605091315,1904111494,-1402214722,-101196737,-706904451,670424574,2134602355,1070188272,502627805,-1225490487,1071375038,1331732461,1021149183,-421724730,976182975,1240888055,-1648722630,-1042,1341643567,526349518,-495558721,1583345639,-1232601893,-744631221,-68330769,-1226592257,-53565970,-96499737,-1414386976,-625346283,921681407,-1653744265,1828642815,-148940995,988905335,-1156727373,-68983809,-1059111429,1719502571,-483917856,1717557231,-831332387,-138473871,2137391937,-18448953,2147319369,335510247,2098544060,984976446,-1347422249,-613008961,-101785953,161976248,-1193937221,829907801,521329066,1539235044,985349876,-744605409,-361366978,-156278915,-14426785,1593758659,1793154212,-42172870,-600705833,-922936417,-594459373,427687925,-1229996081,-772997403,1433878207,-1659261322,-1819091203,416749742,-2093000809,-685049093,-67109386,-1971979522,-806884118,-650130035,-5344330,-120598697,-13685328,-1498833429,-169083375,-35168911,484574701,-118234244,1487597271,-1631252487,964113143,-101394,-669969,-554172949,-1703969,1677271035,-276955265,-212506642,2138436761,-13045783,-151060481,-67967041,-111288410,1944022471,-1091250188,-1010827405,-251790181,-268568142,-279145989,-838949017,-675611353,-408824115,-1888585737,2003789275,1055952627,1165096943,-8556581,1468002246,-1385197827,-21368175,-1235597325,-140038670,-8815746,-438319685,1866465185,-4727074,-842043909,-405889114,-52455494,-84112428,-1173691143,-1078593543,-550406690,1566211579,-1495542786,-690441410,-21043074,2011032239,1609498581,-1646302508,-1296077745,-1367947932,-781000994,-202801165,-2107562,247016188,-173097746,1872451556,-1187185028,-547882723,-520637473,492174302,-1252002890,-1141475075,-1377130559,2104950649,-1247949703,1448000511,-575879173,-679555925,-1082786305,-134779423,-1775780105,-4145327,1462105070,1934359945,-592332489,-1850736677,-1649451349,2143255651,-1883751113,2111241438,501026687,1912299477,-286563346,-1124090141,-9044999,1006105564,1021308895,-624953378,-1149649317,-807665778,2130386341,737632251,391863294,-659710026,-805852673,-889792855,-885014738,-108268289,-1212215105,-141602,1005508271,-1917117521,-471944772,-407218763,-685804118,-67240043,1838848827,768207723,2138109001,1870657277,-1363684113,1842993977,-325078935,2129524895,1633681373,-1779049602,-811631960,1877231989,645739755,801064922,-51254299,-1209409570,666877735,1708915655,-899064587,-1749856290,-229155118,1836920303,1927183583,-162169361,2143250414,-410165901,-134808637,-1123042633,-857006221,-127070724,-102571394,2010116095,-1073866161,-2035353091,-806359042,1039755724,-1342444863,536854511,-117641474,-1307190530,-548422221,1878768347,-134275601,-92078085,1933569354,352287647,-369497557,-1084756162,-1076568069,391884708,-553652309,-865349665,-1578696796,1601631991,-2046967154,-244090889,-888734757,2052371933,1836957435,1673396015,-1006650344,-706903554,2137247670,-352718664,-638691413,1606152703,-671932457,-1143147778,1506655557,1656588268,-1851020590,-143138825,-1084266569,1525163882,-168365189,2111811067,-1847281954,2115120621,-1758544005,-1098044630,360310430,-270565380,-420761923,1868376950,1539052380,1841185938,-507552323,1656363307,2146991563,-50410574,-1883115989,1607023471,-11587850,-1627390211,-46401241,-852891224,1872223770,-38627585,-538519567,-345131427,-51396876,-1128530436,2092536700,1076469751,-419448981,-552865846,2088730351,-306784803,-442044565,-160037,-616106132,-308294803],
-		[-102835017,-279447591,2078488466,2060917747,-1359245321,1056797246,-546308676,221804518,1585312247,450625471,-1765021138,-587466027,-1617768110,-358877279,-285887427,-188811355,-587085642,-546966177,-1075910293,-1109131673,-1016692821,917891055,1740907451,1272129675,-751182865,-537945879,1018911210,-139198481,-287593906,-1087373461,-548569850,872396798,-1259540491,-1222915170,-49449,-465587297,1793474299,1490204263,-37770273,-574166854,-1073753233,-590941269,-1421873922,1874310719,2143286783,-582328791,-336874129,1974160095,-1488991497,710094814,-573046788,1897363813,1600778862,-37814787,-1101792403,-1261259879,-1353450321,-1350766,-537208951,-2034401321,-1111524541,-51400807,1601348167,838351596,-91808919,-74614830,1811603453,1528821982,2094492398,-67256873,-1744971150,-338192441,-270574173,-1074016380,-2088805604,-1278669106,525967294,-13676683,-413671947,2011690063,1589549673,1249902318,144276127,1022808663,1870551039,-1586906124,-1076364745,-80004745,-1078566145,-70591560,-419741723,-565275789,-28378694,-822381029,-1376358682,-1455163929,1866329980,1475326845,-1619013777,867106815,-302058675,-118692993,779058707,-1646266369,-1120611329,-2080392241,-1136182030,-1647665611,-1815391475,-1886697737,-168080641,-1396738243,-134492762,-912769156,-1092638566,-1611164674,2074033655,1438601713,-504631988,-133190213,-547774241,1772557508,870069936,2109421311,2108174717,-170918147,-375006629,-1218978193,-235966978,2004326078,-374679053,1391815384,2116968444,-2015577697,-72290319,2104222810,-1628472577,1272336244,-1216376436,-559677574,1287320466,-327378278,-132790811,1985868619,1907802717,-2088895155,1877884913,2012307422,1475598129,1877325978,-200554273,-952108452,1510856702,-1286611997,-964700553,1592586990,1050424589,-1712818,2102879986,2012725235,-332007425,535723667,-864791182,1978754069,-954229373,-809591041,800519719,-321669857,-1089148675,41713567,1660311037,-344197673,-880980420,-269091073,2062727101,-285885987,-490538465,-537923700,-605366310,167562222,-311434249,-13413677,-159121740,2141406172,2142020463,1782486492,1834860413,1589410598,-1080116385,-185081415,-1683169986,-1627652745,-554376997,-53543121,1788329201,-831053844,1998512122,-1195836417,-716178065,-370213405,-605294689,-218638081,-1170491766,-69211929,1894627319,-33589321,-95783259,726953978,-1631952093,-1216891661,-575428762,1575827118,-739541026,-1275335377,-139483649,-80432324,-704910474,-373469477,-589858325,1206419375,-1176560642,1037885179,-743032834,2117939551,-1158187803,2144839871,-1940183318,1073191071,1811808093,359900147,1046474446,587117887,-1641157009,-111678565,-227679142,-624102468,-309202990,94314401,-1987167311,-1141907521,1851256698,1978654571,125590331,1480564290,-184717607,918453471,-1083478690,-402703233,2086362691,-814284979,-268736017,-20982548,-352403763,1811336823,-162679425,1038995450,-1350567334,-538772074,1005338619,-31473889,2107109804,-678021137,-1109899103,-1230512161,-210223512,-843980198,-944792093,-1363953729,1426056950,-18035339,-882486306,2003662435,-101408993,-1208033626,-1621628306,-1073916179,-18088980,-407178767,-290607117,-169914384,-204483592,-1275595665,-538719267,1606002142,-203705993,1845412919,2138361191,2138144749,1945065437,1673286547,-285286666,-106431757,-1124352335,-1427513661,-713078337,-1838536521,2079916027,-54870075,486069790,2044706739,-997544118,-818975030,1840872406,-872684936,-494944308,1855675261,-1179394161,-1646917538,-335873089,2065396936,-1280635457,1476095798,1998446590,-1074431495,1818089415,2030039694,-286012323,-1215823919,-156586763,-268529031,-13836547,-68738049,1940418543,-33561219,1875378167,-42157515,-1318191269,-1082691969,-353253377,1387550637,-11600425,-872441145,377060333,-116902919,2094442364,-25218128,1052716901,-709689418,1543191515,-40129551,1329969079,-1615078499,1668390207,-226264201,-403593477,-277885070,-1711882760,-588282419,-1924710407,1710804536,-349704201,-327202059,2133278207,-844193301,-223480258,-958665473,-1494287441,-141959457,-328266697,-4208641,-168305754,-1157669895,-1520992449,-394363269,266857085,-361636067,-26083877,-269129756,-13503684,-226806646,292058985,-25196135,-1795406626,-1254871613,-13074842,2146040250,1040114907,1389058026,-579272915,-809587778,1337441661,2107229754,1297063806,-72437701,-136317973,-25853953,-286543495,2130525814,-20278346,-38257409,-1142015809,-2101444905,-786466916,2134880139,-927089737,-84679684,-187314327,-339769671,-1109393422,-1227397133,-379598117,1987556109,-97787912,-726942339,988544588,1935177589,-1664466977,2066397095,2132189029,-13112322,2144287709,-1527022794,-44320661,-675250211,-1075455237,-706545922,-1225346305,1585411857,612187998,-546097013,165394124,-953500119,1457501814,908886222,-1361170585,-269068300,-1901077931,1174403053,1506727895,-6381061,-1101529938,2137960702,-1153527958,-34655523,-1547917545,1072150977,220904823,2135574237,1804070527,801064831,-286265863,-539525325,-75497826,-1104167331,-145236033,2145025982,-1345380413,2112288191,-142737761,-566323970,1402713839,-814769228,-560050705,2076514003,-646021548,-704677249,718721903,1576565737,1598028776,-561058263,-967340055,1073591919,-152978690,1874820830,-168248353,926938091,-39161865,-7342722,804149067,939473063,771716924,902799357,-51009431,-1745130029,-268507149,1327437047,-268571177,-79705,-202968113,-1655687169,-1097862318,-774127958,-9225218,-916472805,-604933314,-808162382,-269488497,-444799089,-1082131460,-600869001,-1084300166,-70275149,939466552,-243642985,314821581,-1892288772,-536887429,1604475261,-1210990980,-1076397656,536700855,-461472009,2102522271,-5538948,-1243976213,2005236596,1013309183,-41083137,-582103649,-1111884549,-422616293,-449333992,-419584297,-1141397581,-271925513,-134907470,1199287125,1933434815,-134578441,-88245249,1188559734,1677033461,-322175234,425156591,-161932305,-1916829849,-142868495,-807430374,-1410402489,-1260506145,-294415044,-168054039,-416048137,1120536174,-1376016710,356470679,-2002780225,2028686413,-815989761,2079840630,2078106023,1772736477,-604768037,479419491,501101439,-346693666,-605619212,792720830,-9376073,-1092747790,-6513486,-1631585809,1577046749,-543473694,-574696036,-680002130,-1174803718,-1127130454,-1611139273,-343474798,1671341039,1634967039,-612972782,-408454822,-405852417,-355468546,-681574497,-46140009,1709832377,1053556670,1060462483,-282511,2101173137,-148906146,-681843435,-377651715,1835985918,-39849988,1602178038,-148398466,-274531426,-1619026521,536854381,-1485641026,-308110473,-231265285,-1150337572,1902731133,-59340907,-471219501,-210780209,1944403895,-31600789,1071625770,-707287734,-952816065,-609768643,771504494,-207431411,-800292,-1226838053,-908203050,-315204129,1005459359,2045508534,1183131319,-343933269,-706592773,-160022529,1607449790,-1024548980,-19926073,-667960789,1071611573,1806956490,2136733677,-141658166,-302522631,-206144187,1232034167,-21495815,1574886911,-220864035,-6369283,-1075151011,-849252681,-873498067,-1286128667,935195639,-316743685,-1078258081,-621826547,-1113984292,2130298056,2132572065,-280233707,-1621301797,760945244,-332927394,-1295266161,1398777704,-288289305,938947670,-823681331,-135365144,-1716275206,-874239153,-1673659178,-337051649,-152502574,1072483553,-424837153,-110213322,-151011338,200752119,-1215172,-547670025,-1391919877,-1892423181,-218268982,-222343305,-114562119,-559587398,-985233091,-35794307,-28970499,-1104847269,-274483491,-1073770519,1852700637,-973471889,-1110721381,400547315,-1707115605,-1421476757,-260124675,-1921024313,-1615990049,-210297868,-1156844362,-823263777,1024311183,2079325775,-549635,2113009327,-1343882299,1869264860,-237758725,-563163478,-706936866,-1345357915,1887357937,-371195950,2109685165,-377247289,2021505943,1828023918,-4720893,567138714,1222612962,804485925,-814534694,-64255361,-277643553,664188900,2061713253,1037728907,-1143940614,935960532,-68236369,-1126208391,-457720067,2087647198,1472025085,-1146241817,-1755386309,-1153434081,-469910703,-340152386,334871333,-6994,959749855,-1073742117,-1420958161,2105482847,1623930069,-1170091400,1342143928,-721067],
-		[-1619267731,-157467178,-921976406,-1208113716,-1330343329,1589416172,1203760535,2110059775,-1087049750,-1887551781,-760615503,1095827133,-587777027,1407187944,-435271820,-612963422,-18122332,2046538836,1308221277,-115002394,-17896225,-1353495611,1048042335,-1082130860,1031490750,-1912802319,2113806826,-134414870,2080104388,1071488730,1423406938,-16304185,-1194872970,1928915622,-1682449177,2147122935,526348203,-279187737,2007815166,-1196377427,1285274962,2025806220,-13110801,226420360,-419465377,-1645219075,-815299724,2120607351,-1117795169,-1241552149,-85197953,1526037775,-677596295,-176304987,-1478559786,1852675913,1978589117,-708452933,-263221249,127922031,-1208714241,-550675801,-554523011,1996488031,500076285,-103811790,-1610883137,2011264959,-1208035719,-742666709,-1744936961,-1077151149,-2080905505,1434336876,388569149,-1062496802,-1348389537,-828378396,-578953634,-1746706945,997932153,-167914209,666450397,1134075595,-545265202,-1132473135,-1298707405,-29819902,-623147778,-1713913934,-1642177665,2081419150,-1446118485,1844934463,-1612959760,-595675138,-553988228,-739491866,1965027872,-15176197,1727516493,-1616680391,1335818207,1853527895,-413958663,-1460209661,-152896199,-1116227843,799899283,2121726931,2012958646,-34739209,2010444916,-1272471813,-68897058,-226492509,-969051138,-89871,2140611771,-211441930,-168378705,2140661939,1716387135,891625399,-1243620882,-10504325,1677636519,-83887105,-805576909,2080243127,1140748183,-21799812,989820922,-268993545,-10814779,124221503,-1898144146,-271113308,801552527,-657460,1833612597,2130638562,2027691582,2134811516,-536919251,-547428913,1906695001,-1904302164,835893955,1598023531,-10492024,-34881811,722788447,-1114047941,2146941435,-963650825,2043922282,-1899251784,-39879713,-614776836,2146907055,-831070536,350194124,-1378886737,-68944083,878525567,1019215099,1279201202,-268513459,-1132532521,1834334007,1584905663,-69410438,1364458191,-382804049,-640168066,-1312819400,1440130075,-324011305,1276619667,2109076341,2147286425,402387897,2060820859,-693114889,935028710,-420234713,-575418977,1421661931,2147089647,-2081956897,-145541122,-369688641,-335578181,-697368773,-35719072,794350579,-44504097,1534982647,-70255171,871346053,1034407359,-664044340,1239350197,-856694949,-617627649,1056556535,-1377173742,-1147161223,-1896214665,-700475718,266816232,-1578107502,-58728681,258621038,-165843203,-1209294180,-791288331,1341047422,-541133443,-16799780,-1678509391,-3301638,-62046821,995027951,-170525250,1324464373,31453083,-621352361,-622921858,-1638954313,-1174675461,-475159061,-2081049666,802743773,949733313,-579465412,1084022437,1035821303,-69080082,-188946575,-620522585,1334825595,-1477967873,-297940217,-101208068,-1768721410,-1891683882,-1174405639,-671295331,-696254595,1964319079,-707264904,2130431311,-1226983553,-17700610,2120398073,2004875311,1806595453,1065316175,-1052934213,-1615069259,-35674377,1996029158,-219676775,1842863029,-193241366,-537793837,1073413058,-202508929,784170677,-342884819,-25494571,-540348697,1944302333,-1024557210,-822421025,-71369345,-318842635,-1241581621,511590382,1439615934,-1076232710,-1075319114,-824710426,-1117127465,-54664345,-751045362,1070388606,-688390481,-679477267,-1168641570,1524498431,1807271668,-205076554,-1638747339,1876032221,1868011487,-578871581,636480719,520092597,1853226524,-1398475651,-560612617,49796995,-268972635,-537266461,-344794795,-136322349,736942555,2130608106,-570798211,-367887,1978662140,-1747458562,-671098073,-917209137,-402678278,-1058275841,2074262746,-146827653,-276936103,-604606650,-1896679522,-1545000324,-838865417,2061885551,-1348962500,2138556273,-323642378,981194232,-1085192355,-1201743955,-1157985584,-1211105553,1474805021,-18913351,-1997968337,-587727205,-1246764553,1040045679,-1610686482,-430241539,-336520484,-344997891,-43032787,2144288543,2078137082,-72526087,1970482891,-296796545,-885268785,-303771841,-956312346,-1478774878,-1245198627,1766088312,-83954497,-1259342977,-1242468673,-609541347,-1862566153,-85983894,-1990197649,-705696825,1006619483,2126503035,-856834642,-1090781579,2145139868,-269226577,-166511077,-9453601,2008398629,2115805158,1020608507,-1638483229,1031772138,-1368129926,2000420603,-1658063191,-1113593617,-1108609529,-1210381052,-548096006,2010363383,2003416290,2143136553,-1179672,-1652966885,-691578035,-35129356,-34078797,-721725522,1070527711,1999502052,451694367,-697328644,-674272517,1047916013,-10093341,2028988031,2144671929,-1955610629,-1216774411,1068481261,534239247,-247739537,-9532514,871274700,-336135043,-281221780,-402803915,-1561542769,-1824569887,763326455,-564407538,-1049362693,-1079415045,-1409295011,895657377,2072001451,-1178603237,1828384669,-1628019617,-403277381,-1074504466,-1090994193,257927117,-1678428251,364336586,-1323634819,-692487057,-1373671474,-102539458,-1639989847,-624009233,-539502129,2079579647,2126730231,1659346863,-1083473986,787742697,-102531916,116653473,-536986914,-69295227,-1275892225,-1879130209,-202941331,-17967265,-1115882075,1725357231,-292557069,-1364222101,-1250183237,464906623,-36257934,-258230067,-487113747,-21434793,-1384129297,-651235413,2037374591,-18897178,-1510623405,944307430,-1087644169,1979415311,383624924,-268470530,-1360074377,246923246,1312537411,-278726441,1307423665,-327722545,-1226870802,2028060583,-889588862,1943317468,-480780410,-436568249,-361312827,2126995455,-320607905,2035538813,-1090983109,-1074791169,-41055,1041207025,-2013497483,1051352055,-745545,-71308095,-77877426,-268675241,-1242376486,-872944211,-239243827,1811934275,1270732653,-1665163910,-1233664361,-186304648,471105247,1026219832,2077407718,-1108153570,-98041876,1607834717,1954348360,1821176702,1820307450,-1812201505,-218268166,-92573733,-268571683,-756708897,-62263553,-565194881,1323020083,-697177867,-81580692,-780854277,-2643985,1996329225,1872756198,-926802793,-1750368518,1527209505,-547357585,-562527765,1893300951,262077909,-1382580274,-975283212,-1209296445,-1112539177,-21165543,-1208354306,1915336991,-1210831657,927914487,-335560705,-306884771,-713081131,-1409323028,-1149986916,2065154110,2011415483,536327247,-830803986,-951198228,-67240354,-1646274679,955678207,1945810427,-294728940,-525093,567373723,-1249642427,199204863,-550535173,-614899971,-528912,-205718020,-1247775785,-2002388777,-180224617,-1444103590,-585163777,1644034999,2041543246,-1477836433,-1630028034,2112723932,-83919155,-608289033,-818480641,-119948845,2129116919,-688005662,-822722565,-889717449,1558899678,-153656678,-1673724483,-1061554625,-274747409,-44335366,-1792651460,930274723,-1409567431,-403965769,-169189579,-203555035,-1660964825,1509894009,343927293,1550049245,234880503,-17728581,-282593477,-74343405,-436371973,-176626193,2006545655,-1631994273,1531303359,-101975076,2042869163,-805601295,2042965118,-375195221,1937243885,1054305246,-754978956,-29598281,-282239377,881651574,1672306586,-3940937,-1421367438,-335568897,-404783183,1842311021,-1360330789,-1147277871,1879033675,1973120789,-676474461,708832551,-624232483,1877270461,-805372495,2068833759,-1653288965,-2064521798,989309399,1241028367,-781976875,-1410629953,-61223973,-555786885,1475436453,-201342171,800681182,-1078460965,-961676034,-12584969,1916737485,-1149897046,1180563297,-1073857145,-403703332,-637608717,-1755710620,-1632796682,-1610756611,1241102638,1777786234,1941342095,-654868639,929398623,-344772277,-1145327361,-577450531,1206779419,2046806009,-3256581,-704743019,-1596336386,1567421945,-270288529,709830076,-83,-69818627,1725681506,-16810113,-364386464,786902998,-1913131943,-1233322121,-842009139,-25174284,-237084794,-100926529,902551382,1039792343,-1180238665,1600123897,-1920995347,-851486825,-1212323447,-2400258,-537952305,738080767,-255015169,-1688888710,-1225828631,289037291,1744781014,905902462,-13112385,1962786559,-962085961,859622911,-572584241,2067709905,2144844249,-281137434,1794104061,-1477005347,2077227903,-152068993,-1814726658,647193471,-287319338,-412959239,-540087433,-176686919,-1159815969,-301990772,1572535711,-1761682689,730817124,1405079231,1069457642],
-		[1708455411,-91195912,-202145929,2142280145,-1542162,-134246505,-671309889,1864816430,1830397206,-1392517905,-1295569285,1808726370,-93737989,1333697953,1206747107,1203456870,264132583,-692339817,1845493210,-49158,-106956002,-113478787,-743528515,-110482499,-68182293,-681608505,1848625987,1959160246,-1352801383,-1275998262,-306216973,-1360070793,-114429002,894893895,-571394246,-1243185385,1371929019,-894109961,-1736612242,117095922,-1662168129,899347638,842724208,-9441805,-35442040,502595494,-2535560,1694496029,-1154267265,-692961284,-286329251,-86270456,-18014253,1799913853,-1082067541,2008463350,-3162443,-1247109930,1561967543,-1430455322,1877733343,1530536223,-1342750935,-171442257,-177344957,2145741783,-1495692811,1941879551,-441014625,-577471625,1811585022,-681839589,-1159611796,-321210210,-152690818,1567067115,-173662706,-1346461702,-556179618,936408781,-422684865,-556337717,-88757271,2067763183,526820436,-402933825,254934905,-1094976090,1469064619,-71173737,939240247,-849482899,-374094966,-4197650,1941708783,-105286793,-578947909,-27344,-98369,-290376451,1895874968,2078998433,-580455201,-612511777,2121104274,7839710,1601972730,-29511521,-537264573,-1679963043,1516173244,-707283026,-1416163809,-1657299091,1597963675,-1258344786,785753962,-1297118787,-1921273572,-1172800901,-1625605353,-33865731,-1120524819,-1082405458,670823103,-673202302,-37265579,-1819566617,1054859253,-246922037,-620854026,-623652041,-269625889,1855254333,-305530379,-1699741774,-1767134561,-696300053,-22087977,-303046961,-503352358,-1937768887,-471880151,-705217557,2024738271,-1343533116,-1507363988,1739394663,-1111625821,2147270098,2085479294,2131258845,-1174667809,-264765815,-5346817,-302724105,-302907718,1734221018,63351339,-687933508,2077435845,-34031683,1895406565,1090327001,-212354203,1101494271,-1648033830,-1351368763,-1185946658,-175645799,1710203542,1795631479,-202375313,529786730,-677390490,-833491211,-2251478,860728888,-1209009672,2142342909,-575742420,-805343422,2135924567,-1162003492,1855681839,1707439867,-1375782083,-545629222,-71866786,-10224642,-4231213,-44851503,-1514165897,-887781571,-2039816,-193038916,2071625468,-1342381803,-629686278,-1283479941,-1326417730,223982839,2003441582,-76085441,-643838530,-949387288,-629287681,1970721519,-1700487201,2127032959,875165712,-1025778242,-692126307,-1647207682,-1505872040,-1090635815,-597757953,-1109570698,-269242183,1802444749,-1730234883,1596345288,1207171915,-76022085,2135864539,-1363317289,-113256473,-621084965,-1399059490,-1147005785,931081855,2115233766,-159450175,-412517778,-1240620258,2062347883,-2042704267,2071895575,1469363711,-1085443577,-33636793,1507063039,-36776981,-16844147,-1738097666,-210010220,-809010445,670859099,-1152675994,-1955914497,-177274899,-572647587,419327199,-629948561,965643766,-1074287109,-17213686,584056275,-1214485446,-672284674,-70183233,-676873354,-165421162,2089516615,1795518460,-1671767147,-331406370,1733229276,-545787141,1811742585,2079627230,624011677,1578486487,-313966614,-318900388,-22013828,-17188881,-178299042,1050994541,-1083213973,-1099435368,-273303918,-3174934,939190254,1675567998,83099069,1026220007,2124222446,-1087021269,-472024777,-180557879,-75896158,-182596242,-1130643567,736780959,-1346675921,-577045522,-881492121,-538542162,165862831,-1211271441,-222433351,-1235616402,328442094,-674118857,-556045585,-969742609,-86313057,1806491117,-371590297,2090204860,2099114735,-2032293961,-571674706,-1617508897,-559978259,-269158408,-477627430,1599849959,-1510553794,1325271919,993174454,2002421547,-1238253590,-832574276,2006241147,-598483433,-363105589,-1833052641,-60817538,-1497892077,-692580790,920358790,-1258835990,-18483452,-425232469,2139082686,-104891362,-4261276,-1683798017,959010735,2111254501,-1715603973,530356173,-1658355860,375994705,-545411490,1606343349,-169602,2086235371,-145245771,-249062510,-71371969,-1420345931,-1075455137,-404982919,-318935106,-606119942,-1897939090,-604005511,-1124270125,1071325149,1305210215,-1250230394,-600123213,-287944977,732626927,-756577425,-1679049749,-898662466,-676509624,-1141254275,-659827,1532481467,1216085993,-4587521,-839464970,1608646655,1783588785,2004875181,1533279718,-1076155568,-145256146,-125066,-659200450,1841258413,985116155,-39223561,1383837811,-85275699,-118106326,1402852335,-1343844438,340506555,-1514143915,-136708867,1857441680,-1016608986,-167926305,-877985885,2130030015,-1889606701,-4199575,1983278328,-647001223,-274366529,-343998474,2012036047,-575111206,-2045784962,784066198,922086283,-1612776485,-965745614,-1381025813,1912598263,-9612321,-1880163851,-235654488,-1107443937,-537453644,-1951666235,1071349615,-562561066,-1773404481,-588521568,1302773236,-1010902533,1872489109,-1682383489,1858915623,-1428006497,-1079453273,-542285826,-1686123009,1945598943,-336609281,-10519106,-1289002065,-285246721,-270873741,821804503,1744682997,-251928641,-1670418437,-1220810295,-136364313,-883163235,-68169290,-671789333,-67436817,-1809146181,-305627267,-29761669,2012872415,-1863713537,-177750798,-71341127,-1234403330,-54544423,-1048590,-369154629,2107480791,-1146436148,1600993663,-1385881,880275454,-2016496677,-2035919625,-1078477827,-2075207745,-2021661991,-2145686705,1961862604,-210245716,1200778694,2058594959,-1677598984,-302154385,1055357118,-1467549057,1534338867,-17213542,-267429,1333769695,-1648429316,-134469954,-1553309730,-673294666,-624177286,-1791105047,-1895899143,1472033659,-1402217659,1603271167,1542291279,1073490811,-2175799,-25428817,1873106909,923229823,1723558765,-1614807428,1543465429,1535114869,-574919697,531361396,-1361388294,-209822129,-1355434755,-6291479,-170291522,-567811,-330257,-37101807,1805372783,1910325151,798843903,-48119953,-136744963,-8684181,-536899818,-347755270,-1848152407,-771866629,-85225798,-671088805,2044723190,1517178543,-1182256201,-609190465,201309815,-151833689,-537882625,1727074524,1870647357,1939486649,-273305149,2035269103,1744424765,-4207078,1786240735,795760634,-1287664698,989578703,-159711754,-27598916,-2014170625,1536577231,1911029685,-170760195,1718349901,435541357,758611831,-1678919822,1677690879,-6292869,-848628788,1261846498,-877707410,-1080052737,766172793,1881527421,-16781385,-858916029,-570495169,233659770,1786719209,-145758813,-55780402,-75528321,2015340303,1471985499,-478008339,-677487306,1063775711,-962593855,1944050623,-1656951948,-196695,-2055357046,-352325694,-1261571461,-1074140685,750780220,2144910332,2012607899,-440545364,1710187135,2052436309,1180623085,-1135100961,2013106617,-590500930,-1376349283,-1159733636,2147341495,1927197668,-18289793,-810022163,-409998726,-420381822,-173974337,-850953290,-341871309,926801838,-648301633,-1157959841,-1111781803,-353635199,1974467518,-85501761,-1824663553,-541263170,2004402158,1799302863,-1249908866,-902321025,-624952578,-1247909193,1596946158,-625910721,-469854859,-358416550,-1077076124,947646351,1067724791,-40907020,-137637832,88077781,2146745887,-356934212,-36474125,-536921225,993525226,-2086724135,-1328317485,-222396945,501044115,1649892730,-556013761,2106580722,1156939501,-222725761,-272173155,-1696825671,-211240963,-539052297,-1059696718,-1142459473,-4200017,-814776462,2110846189,-2007548477,-2144337945,2046711711,-680222804,-352922922,146759484,-432014471,-438602918,2011971389,1043461548,-1620576929,-1081021097,1199526811,-402934802,134081657,-1914800257,-1342468443,-667087072,-822910657,805224185,-292736065,-2042743843,-1983406281,1062754363,-1881407593,725329139,1935666687,-71926283,-359138473,-1965906569,-254854229,-1068010273,-774226179,-841065475,1995439487,-1112675362,-1927676955,1789623719,2010906587,-98173953,-1295598870,2073497594,-1772651541,-966404111,589283039,-220780074,-1128286593,-336674883,-67240705,1976482759,-323549353,-21895083,-79962881,1811662717,-1056946,-403782939,-310471825,-2115803619,-2757651,-311472902,1877973919,-376050842,-1908450394,1273950204,-33718273,1542386683,-473169957,-638069781,1599860217,2108586491,1769144055,2000604923,2147319647,-1957976074,493805183,-327368707],
-		[-1636827428,2012194669,532642555,-285256579,-355356255,-1161907560,2012602027,-630784842,-1989957895,-880308564,-1630423049,886767031,-1215359062,2078669207,-1023824185,-1394032809,-155307050,1932720101,1051229693,1015000783,-1673704665,1304427939,-55485346,-4460578,-1152122961,-858788645,-1254410853,2128537325,-4001959,749229918,-8734823,-408512738,-728171155,-216621081,-1164183595,-1260556326,37369751,2077646847,-31604737,-306251778,1402203803,-19221581,-1183203905,-454857674,-323476297,-269275553,-1231716941,-817890850,-270764142,-671155969,-469764181,-25231449,-1620051284,810249727,-1909106948,-689590291,-1084044320,-693799815,-711734442,1509740021,1676930543,-1457897985,-268904469,2080332791,-571378398,-882444161,-39866737,704639415,-269818090,-112646914,-327353987,-604957757,-549585067,1876344814,1442527106,-5316620,-1699812875,-1346425368,-36386209,-1125140859,-201914821,-1180320817,-738306525,1809798815,-1688548369,-856720417,2137784278,-704782983,1051050258,-1248725351,-463064,-68585279,-1623927561,-205701220,-146940230,1551500477,-143989522,1019948943,1457378890,805017586,-1633479131,1802452598,-674234609,909865083,2112148861,-1671572,-639129643,-1090547938,-36146,-1485947949,2043669558,-1627478818,1912065114,-144985569,-44339913,1571236321,-1108426791,-1176642148,-2098945,1744174899,-188320257,-30409496,1990961360,-302227461,-128459889,-575712357,-1739194829,259219398,-1075380748,2042007030,-949003521,-144837773,-338059352,553511039,-570523777,-52523782,-1963463229,-1426763975,468577957,1557112574,2079634887,-568852626,-723481,2003162606,-1659916497,1541862069,-1097899265,750173695,-699537047,1607983039,-2506370,2083163883,-135405827,1239071735,-2105558145,-1787364075,-1074347299,-874550499,-1278601263,-67121337,1818086524,-1091634003,1072689245,-7913735,-1417842802,-1682605099,-1663603049,-1319473809,1874316150,1809653247,-612426526,-1024094794,804226951,-377497453,-1153697838,-89129739,-746821187,1140836535,1874845180,-204472325,-574703643,993774003,-147411089,536792895,490859487,-503693909,-814887443,-605102361,2128999343,-702580359,-454513027,-191114778,662236665,-1531039932,-5452451,-1694818461,1565456251,533322995,-1891893769,-134763397,-68822421,-1661217601,-873156665,-134808067,-1544111649,-268457233,-648029829,1853714171,-274925570,-1511064620,-792792673,-298657285,402459815,-9461779,-1137181813,-537601282,1877928908,267823605,-1311716493,-192020956,-9199633,-428346551,-445608569,-1348542565,2111494083,-2501711,-610648597,-1208096341,-185869323,-1097219173,-1428686635,1972759500,-335845943,1190951677,1875360611,-65821777,-1388642340,1929289213,-338641687,-1113274369,1035390846,-273060647,1868362730,1268086135,1705375974,-961171554,-1078984971,-416465987,-302354317,1995697084,-1653223444,1744014842,935059439,-328512805,249495491,-595495940,-1716591235,2111651739,-118008441,-719872076,-171444293,2080300527,-1550377308,-627581475,-308052929,799800702,-933040131,-1511122215,-440634421,-880741458,2128313835,-293736583,-344719931,1475081849,-23892154,-1088065731,-340828628,1607248669,-34078726,-205340755,-33820758,-1477194969,-16867847,501172679,2053848283,-1186091529,-1100321178,-237126473,-752395029,1739325230,-135388461,-6692033,1976538910,336420847,-818761819,-960514051,-1077658627,-273590768,-71862665,-340796250,1960289143,2146295527,2145359247,-348231691,-1111951768,802888698,2112923198,-1120132367,-744728982,-536891409,-126238721,-1443246199,1610481235,1968956778,-370705665,-1363329638,-939525644,630943494,1157299186,2092422775,-411619977,-144491779,519183324,1020263622,-542908801,-352458657,-216619291,-22021138,1868157414,1610606409,935249852,-574620795,1000305135,1841232477,-270860350,-61604098,-134365786,-1126830753,-974304233,-1679826945,931124121,1658811095,-1107701633,-152338961,-202035593,1585412012,1927833575,184443199,-1668317231,-675243537,1740070879,266248047,-671569272,-691102211,-12697613,1760540389,-4796482,-568728373,1269825314,-536884876,-560417189,-638805264,-1242563271,-849685021,-28446871,-34341207,-437837701,890814143,2146895805,1667231731,-273794250,-606143117,1338946751,1859184599,-541594939,-306196017,-1248285889,1684963293,-1234504723,-1108452449,-75075086,-369688583,-407502851,-345089089,-1720746434,800810167,-562693426,-688326932,-1679593481,-1083113617,-5242893,469755246,2103410388,-268796191,-619786758,1502374395,-139439186,2063564479,-311446703,-27537815,-142607252,-810559813,-1478511653,-5046274,-806653793,-821137762,2039612414,-1154524849,1024816509,-717432453,794286103,-1447338401,-4393634,-63219745,1023408007,-1645347931,798619887,393185735,1504549814,921681726,1333170542,-715153457,-1075953418,1070343031,-621879670,2126826236,-1616457237,-1140878365,-680216651,-739083077,-163698962,-2109778001,-1834501412,-631620484,144441127,-1117847620,-1703939075,-270701,2062255830,-40043589,-553759877,-1217202628,-590748168,-281351106,1538750319,-542223914,-511844913,-136788545,1986893373,-566182337,-285965941,-678133828,-301795794,-984261270,1795094007,1771221654,-573588266,2037741536,-889275910,1842384973,-681578762,-21069846,-1113791502,-1317022494,703969087,-96764232,-40894555,1874656488,-464257869,1842314223,-1401381043,-1883767811,-205740294,-68978258,-1200685916,-318807107,2004991079,1191115706,-337970593,-302130305,-1250083441,-26304682,-4342403,-185601031,-23747225,-220341546,-886317759,-554893465,-212529881,1962342380,-257032607,1307043822,-1871776081,-220930333,222821735,-279838869,-1261437203,-247767619,-283197,1337586043,-46608543,-1394214172,1914877591,-337741577,-67160145,1450835941,-177351174,977164501,-1175609477,-1245897005,1838783835,1874064355,-1684374733,-26330115,2146235835,2076372844,2078367103,-1346109641,-68080007,-3407959,-966231033,-2080729473,2038411149,2008514406,-88223769,1923042236,-342903581,-860538881,-2126535176,-1090552089,-338172009,-1680934001,1977580283,-2020231435,-1166381141,145424127,-1236926626,-224744604,706533055,526888447,2003725047,-54005993,-640432455,-1084809385,1181015527,392134519,247004015,-72122670,-6845537,-921280568,-1406253203,1165975277,-503439138,-1375731753,981449650,-612415877,2077749095,-85023809,894423515,-62030611,2116792661,-256287829,1178332777,-2047216879,-1259390486,1172233655,-1112347719,565813239,-537487561,465550832,1737677087,-131085509,-814354970,-143200271,-1721346370,1534717213,-6334677,-1428293129,980235918,-289603658,429783037,1029569747,-271073840,-1686644753,-2018384011,1207955355,-1749563374,2062860284,-36907918,-1259900941,-141410675,1660651150,-260637169,2052841441,1859633117,-830779996,-174198786,2103900139,-1154256135,-1166093190,2143757191,-285222402,-1619151361,2055103189,-19288356,2002449375,-1302598961,-1179658,-139055905,-1092840923,1878372704,1440175867,1843153773,-1158198054,-689470062,2145382397,1987941946,-860339346,-222692005,-1135752647,-619721229,-687904137,-7610377,-185106454,-243984658,-952699858,-443302095,-445714433,-289417764,-675431006,1590943389,-1444941321,2111619069,-957433369,-270281546,-1077432585,-812944140,-130107705,-707265543,-495233955,2079603661,786431863,2010110397,-1884013714,-568595011,-655937,-539625057,-546976817,-1132799109,1796648657,2143157398,-111157347,-101990553,-5318255,201293822,-1244926579,-85516740,-1363505577,-506462487,-562191363,2097109471,-405868585,2052258623,-621848129,-1130741763,-1761874691,-103173249,-784629765,-72367318,-547320660,2144292539,-288425731,-340267913,-777568282,-566397579,-608331953,-1100255405,-103189315,-7393321,-1095068290,-1758120901,880344885,-1112707089,1070989127,-1562403506,1053917167,-304317516,-822628641,-691819918,1047820005,-413220895,209306903,533065437,-1172505987,732819157,802643567,1119156143,-145621697,-1111582727,-101204069,-1397245226,-94385155,-168445585,-207755554,-127994867,-1782612476,-1140883843,-1761620003,-1755399570,-570894372,-1077939138,-674320590,-1262848169,-1158782997,-1259602433,-1107743791,-861459202,1828415229,945274879,1001911036,1039640385,269981751,-141207843,-623902802,1830551282,471828971,-721449155,100642599,930723102],
-		[1671103275,-1280141834,-337707025,1055239901,-272132,-808328210,994473175,2140667663,-404654083,1072127468,-1091750625,-1358332181,-1476401161,-1235810307,1308291043,-1175536649,-268470312,1918858361,-345441601,-1215014593,-825299509,803465246,-138717001,-1086757238,-676545153,715610111,930863015,-21103367,1406066622,-1653626187,1865786190,-37751272,1867771990,-75964865,1609928618,-38944799,-1355449907,-1388865379,-622998579,-136540453,1668753647,-6368771,-101584975,431279088,-72434390,-70324290,-965769853,-1115963458,1869021183,-627986825,-1611166734,-278085886,-1930256543,974846952,-621881377,-205613398,-1615071545,-1973978529,-613653189,1538104791,-818187803,1550279613,1918201759,-12363854,1856961230,-461849032,-10495916,2118137467,-63963653,-866043417,-51226758,-1372718154,-135053890,-760266759,-331704854,-572990215,-2639111,-269746342,-638591969,1431533561,-6312581,-67324481,-1174538243,-537287879,-69568083,-374342341,2130699903,182450162,-444628461,1593826206,-583505227,-950244881,-860099908,-2104003,-136651051,-8994823,-2660401,1028544335,2128542677,-741416193,1467476729,-31879415,-906037013,-1221641,522644306,-11577425,-69427219,1987998974,-50626823,1206306855,1910237935,1662049014,-25429639,-51119683,-1753335862,1274273055,-612667554,-1156613675,-1191855123,-537530699,-188777509,1249376671,1954086141,1072561850,-270148059,-64689551,935288561,-1327762979,-338946109,-343999028,-739903322,-1225131169,-1344868494,-276959812,-1880265266,1054845423,1060182207,-562045561,-1624969610,964324491,-69272045,2043630846,2134814550,-67699011,1903165371,-268533761,630907335,1157554147,-25413641,1865539374,-159346159,-67808783,-1974355986,-1358977121,-2013334849,-274497719,-559498242,1205705693,-431319057,1337160216,-1075216594,-1611149377,1670854612,-475334916,-335841797,1129938895,2083489505,-583122946,630133754,-431555339,-1243244958,792545274,1996213461,-1156827771,-96602522,-1245971201,-538158866,-554572813,-867440645,1759228879,2112142246,-1585660229,2133626063,-119591546,533331899,937352890,1048432555,1370101966,1786573694,-239746116,1417441129,1962909397,-40675521,-76034634,-88867625,1595928359,-557892085,1069538958,-440598669,-671725839,-1418794145,-562049031,-781534365,905655927,1537061421,-8716417,510578430,-1244448902,1022094321,-1158070697,1857743249,-268701329,-1623204385,128498684,-411718354,-756363666,-1074882594,-165230,1727839852,-469762257,1610053103,-43406850,1064199917,-6506715,2138570751,-352388616,-6849553,-1734477841,-1585611841,-134303049,-48693521,-316811265,-676662363,-678588468,-1099696162,483585122,-1350631995,-419502130,-631308641,-1152756763,-819495425,-623192135,-1073883283,-1115967746,-1405152771,-5243241,-306318021,2128532326,-605717081,-965084975,1441681711,1809807809,-541466689,-1111523497,-324547002,-1114414541,869105551,-756737669,-419439365,-1048738,-544522251,-1090855266,1308409823,-15221321,-1946161890,-1083925898,-536888681,-1612005385,-72502661,-676456597,-240563673,-277889946,-1913291682,1530768559,-1343507981,-31279137,-1618051203,-1099964745,-100677762,-1468060322,1597937103,-1082295042,-135057459,-253515537,-1524663789,531592184,1826340861,2113559735,-1115809298,179299444,-69297154,-1417774625,1692020469,-1081427,-654482833,-41164965,-626788420,-1812595461,-143671425,2087403363,527112427,-606197890,-705241609,-1615203937,-713842744,-374344839,-1678770436,1040004863,-4229150,795570042,-275120661,1542977327,530395630,-43053351,1049621337,-1077702324,-574002,783316892,-76679184,-214149773,-176504001,-25436429,-435327233,146778098,-27330036,1987795384,-1073578,-74159285,-97161240,-556077633,-809634121,428597141,-27333377,-104395782,-1077669409,-269058177,2137962237,-155251035,-941826114,2134307485,1538457587,-1461767178,-16978003,-10750977,-426780,-36727129,-441904483,1810673177,1861352389,994049483,-171516211,-33570929,-4878280,-1956921899,2143805239,1700787001,-1165709029,1021231026,2144598770,1062729567,2147221047,-4279264,-1174604041,-185634576,-1102060113,-17966404,-146278404,-402669842,-35068691,584538783,-103057446,1840531223,2076173755,-1747442809,-1220542489,-1103773704,-2017701140,-1094477873,-860233779,-939708421,-426183747,-269636995,-135578297,-1629853715,762828543,-126060419,1879047359,-1076954113,-161421,1710288767,-141191556,-572332060,-34705574,-78055733,-84376187,-289409099,-59285634,1933300728,-185041173,1985668906,-551960708,1991155575,2146303751,-1342311561,-1217409124,-888045569,-1108082721,1364553586,-1384187933,-1276985385,1028913374,2093412857,-276960353,-822740306,-1618939017,1038726646,2141059485,2100280823,1067419640,1805381631,977086326,-38216462,-839976546,-29928291,2025809640,-810143818,-195041606,-1814038627,-2047012365,532561917,-1544307099,-1417780107,-1984571858,1073618678,1718471933,-134520895,-243453954,-336200233,-1276152908,-1512218629,2136471999,-134524683,1992255892,-289427691,-302927953,1632341982,-72193305,2113722335,-103310882,-1620316420,-330498980,-1485047423,-1761874056,-44327462,-538968709,-718294842,2078162235,-136390082,-2081984033,-744359115,-79204369,1794894760,-1205511492,-2213989,-808666385,896431735,-272862446,-88682802,2147406549,-1158302436,-134589700,-807829667,-1921042050,1360906143,-5161515,-152707,-1084432549,2012889566,802877439,-1571307668,-1350996845,-940038790,1877736412,-613155860,-1486904931,-839622721,2094396415,567548895,2132965295,197110485,-1040195939,-1724964898,1268137439,-274016293,2034386830,2121956347,-606896643,2076171838,-287381893,-34,-1173996155,1458870008,1970619894,720829951,-1187653945,-542592067,-906008578,1865407341,1929559779,-567189543,-137111553,1609684079,-962736519,1517010417,1053515663,1944647137,-187908581,-1506222209,-743970885,-816083469,-96826339,-350277709,2143194988,-103944252,-1745660033,1869927674,1190303354,-6358065,1473702895,2080308191,-672510216,-17465859,-19825729,-1146159169,-831267059,-612567817,-907308651,-88252455,1002361645,-1480197181,-683704389,-353638585,-140515721,788453223,-1616988097,1959473054,-1150256325,205514686,-1053205,-1416637737,-1123595506,-873496713,2125115357,-77660419,-704831493,-1090586889,-1435786805,-1550026245,469499895,-11344524,2042619889,1706524528,-1076134406,-1242972303,-902300723,1826733565,-1091568645,1987990706,1031515711,-558957090,-159855393,-675886598,-1863369737,-2013268545,1060592487,1902247711,-339756053,1836447407,832042974,1365229243,1583578342,-1432363649,-1745584277,-680572459,-1407294083,-1193669202,1872542743,1703931068,-4622389,463167479,-142642727,1862196077,1038474231,-1770521857,-496133,-690860052,-33952142,-1493959817,1273749311,-140675076,746426188,-227635459,1022328314,-1409112435,-1442179,2107079957,-388010244,-689448968,-1145094873,-1130594571,2138667901,-411076929,-961290614,-1166033687,2075508639,-1453396018,804093040,-908216801,-1180963151,794810238,-619055653,1979447651,-1370033057,-536880287,-287869313,-424006181,-638089447,1742715994,1725689215,-590389386,-5682945,-1123834076,-1478492703,1034926447,1575960567,920483751,-159521362,2045426255,-18274312,399507444,1858285277,1408752473,-555037042,-608264241,1188277946,-940158381,-286392483,-168992802,-1208888963,-1073900556,-1265328147,938434505,-2013313707,671020503,2046123388,-1799618839,-1285574668,-2097173,-1667971077,2141363509,-72043585,1332510654,2075516846,1827634105,-975386791,-1352168513,1834987129,519699939,-25184724,987196031,-50728067,-604001097,855490494,914202086,-1163417217,-470991526,-458458215,1323617133,1853470366,938652540,-593535503,-440528001,-178414215,1028052969,-104874725,-2015380555,933949307,-201793953,-557531143,1877458941,-402928189,-268764738,1606139862,-103284986,518957154,-587228449,1596980279,224001069,1610202733,2138521773,1790043750,1029167548,1340063551,-140525590,-1141260614,-621675019,-813705494,-550116742,-136757254,-248549506,-404105346,-1216711201,-1813046279,-23077961,-117977256,-174474259,-1134224484,-914440355,-118098453,1742720999,-1075512754,2004803326,-1054213145,-1788297217,-83899031,-120875644,1524629397,-1986135049,-885248009],
-		[-1133641761,-1044678026,-1243876839,-548208881,-10320453,-300451032,-281570945,1137278943,1198845695,-1276331522,392395679,-190275652,-21014499,-1922067187,-2114465,-147106721,-8181981,2147184366,1318817275,-67613210,-1670629796,-1724155399,1736421111,2008501802,-20230749,2071953351,1140582135,-1562975885,-294197414,-1699138126,2143239655,-51456386,-445485606,-639737861,-327290938,-570691097,1823456049,1056963963,-67174677,-948606468,1876857811,187692959,477554567,-1548792797,-1544642665,-7923853,1036888445,2000682089,2109719374,-25361826,2086132687,-907436697,-1199616195,-1146150837,-214663713,1031471066,-876773677,-1835185217,-714137671,1352654300,-1142096037,1186725620,-1563467366,-2038694409,-68755493,1441790901,-1494014982,-293787666,924839859,1047772616,-377522177,863846373,-1490465287,1976218167,-1361154003,1996209334,-1298696461,-971770897,-83576897,-1920467241,-75508953,901613567,-1287667130,-135889299,-1391462422,-637873847,-329409379,903852396,2080296427,-880460305,2137318373,-856166418,-1023412098,2128538863,1592490991,-109077637,1581771974,1840963321,1004712623,2042593207,-586285617,-412082769,-1681457799,-476516041,-1189380426,-236781737,-78403841,-1240790818,-1663901787,-1754798315,1870583690,-21628217,-844696005,-104273089,-586683569,-19990191,2122185547,1702264247,396848343,1593273823,-14041613,1788767231,669876215,1667170153,-212448420,1784656543,-68488754,1031404223,1507587296,-338169985,2116532735,-1818770689,-1757569298,-117643129,-322278659,1813771255,-124026775,-304264085,-1512317449,1199384756,-1400935457,2006441915,-345245539,-1994605222,1185526102,-1153565846,1827798251,-172636309,1802327350,-831687639,-160497665,-2351125,-647909702,1335095263,2051653615,-1818907703,-120081729,-1826167561,1534845779,-126230937,-5833573,1609951215,1878994682,522155985,-818171143,-1723859906,-46182750,-488450109,-27415106,-1293587463,-1153665539,1468260095,-1502086531,937474723,-469762053,-805309569,-281051146,1710328059,-675357790,-503358215,-399468593,-631832834,-1214260553,-1109951502,-1099477510,2121264111,1505991289,-253856867,-12669956,-1350633479,-1489712147,-109451082,-957265935,-973644331,-102650306,-386055092,1123965179,2074428777,-1023951094,-675315731,1668786877,1967423485,-1678838093,1106200172,-1096393731,-222822468,1868134902,-1015154695,2137915383,1904003694,1907077053,-1817117443,-409190945,-1652492292,-605434901,-283121729,-1342851361,-587819271,-1477533699,-746659875,-17896819,2010837947,-429009025,-28009491,1068457459,2002237005,-122492945,-1263350145,1474671351,-1092763945,-1376860058,2139617135,-404293865,-1386484903,2111781239,-273306433,-1077823014,-689783825,-34085409,-774896131,-815010918,1539767963,-439499011,1586975129,1874537055,1577008051,-8479913,-315118097,-1552155649,1960416830,-493402642,-994595558,-201458182,2007977338,-478690531,1495365945,1033895906,-257847493,-609029637,-479371267,-1393033333,1740549867,-1244290587,712504183,-79855638,-278925605,1068441053,-1695306622,-34734089,2147443549,1706531436,-1981711031,-40368667,1605089134,-971001959,2003303126,-707434509,-429330447,1525001967,1027937807,-8522835,-1484096665,1120332029,-1049628929,-263205,-40527593,-1074312213,1721228927,1470626687,-1098122283,1609687023,535277777,2056896426,-1719942241,-309597290,-1455491585,-680058152,-935330831,-6379683,802681747,-1075611138,-628400399,-213593,-10521185,-609534640,-403987934,-565973353,-793350413,-23095579,-548939921,-521406817,-954215000,2042056550,-681705618,1878908683,1869400223,-268714409,-509627477,-1647853644,414052587,-469907843,-17412650,-339446320,-823661074,-1346130963,-1361709121,-678291594,-1409305250,-69271686,2107334378,-1241856130,-53477585,-541102898,-79471939,1273469501,-406408207,-1679165489,-1321150473,-1773823905,-85901605,-188892257,-940614662,-175827074,536657533,1559211749,-361895193,-6856899,-11508289,1434578188,-1147143937,2128608479,242171859,1425666414,-541123423,-2113972810,-1084883013,-395784,-572278313,1543028213,-922931257,-50877140,-273222804,1346222644,-1343302663,-136839475,-287870029,1570756333,-559973524,-1417805853,-1954550407,-118794257,-1750275014,-2638657,939506350,1340504973,1932443055,1039605694,2029518833,-117525938,-565291108,-562071810,1475807641,-465305601,1727789946,-942149969,-1952679957,-151650690,-201656203,-1213214833,-84250691,2140943610,436157338,-941939342,-75632883,1803550548,-1544303925,-203778061,-1409516162,1832384999,-17387600,-582355212,339717748,-1413768076,-11175157,-1058441926,-26227669,-77758995,1767879282,-710202587,-553718273,-87249409,-1311846937,-1770037249,968359739,-1241548841,-1166901785,-136064292,995980739,-1880229914,1591665918,2113337279,2056387912,-1403298082,1430052340,2146937342,2112782195,2134109180,1299181518,-849408289,964685311,-503908360,-1125190947,-1514682449,396086060,1072346567,-1142214052,-1147212878,-26869763,-71390566,1912457983,969867183,-827248138,-33476,2139852231,-102928993,-1124214275,-1096886199,-138809423,-41943445,-278274049,1609546619,-22684682,-37927183,-270558243,-609239075,-1314130178,-378802233,-3213494,1878646527,-777331044,-1760095171,2107628405,1529084310,-254110889,-1488880159,-801426739,-461518950,-843661576,-1919572162,-1377894437,-1761624963,-1208528193,-723550420,-1648624833,-194254174,2147315631,-422321949,-565775922,2138153847,1073463246,-43786242,-1143064102,-1174427170,-623757371,-558111131,-269536362,-746591018,-3526913,843054666,-874012875,1605558114,-1112900098,-1511133869,1472198527,-573148037,-302610561,-1186989321,636990177,1602663422,-1108510723,1610601439,1775364989,-654354596,-1888741433,-1643135189,2113926903,-291265617,-1108398183,2074858691,-1086325328,-706809302,-1084669960,-1682,-811625625,2009653165,-274613293,-1325476165,-462160322,-888295553,-138826913,-720408849,-405037379,-1751126545,-818495571,-1745112141,908296095,1607716077,1643016703,-285213256,-117956785,-1284016166,-269246713,2117328623,-2094702801,-941967381,-4866313,-1074414473,1298987454,637507833,-962901498,771704575,1063616318,-16798154,589634255,997584863,930962927,-444606772,-1174669321,-196739,667416415,607987629,-1146934033,-1900740939,1586167279,-345197177,-147264483,-72989058,-1248681995,1610042339,508932068,1891491787,930592735,-281219137,-1525195314,-406115150,2139081811,81649023,1055611859,-1073808075,-555296889,-78676136,-1359037106,-1110552069,-2820119,-1958234133,939508469,1796517806,1933212943,-454340764,2000134919,1608970703,2144333463,-279515445,-1082539465,897572456,237881063,1335032025,1910396121,-1611224130,1543351269,-537133066,1004535643,-520493671,-2080485515,1986686423,-33567305,-18759762,-1326438417,-561228389,1065840087,-313271766,-781079173,-19661156,-1409504019,2088708474,-839132306,1336819559,-1075913226,822081353,1892986683,-949241537,-1495832984,503115771,-442762300,846725093,2147155888,-1611173923,-423203525,-805381239,-1085707819,-134302013,-407143456,-396165121,-570962441,-1275304007,2079162334,-671944721,-138680065,-698942801,2131721903,1778331387,-34701894,-1107597961,-660748413,2043074041,1317334891,-348176762,-1074200963,-894521473,-1363469827,1566537471,1072818941,1945618556,-403996721,-1346463603,-1360113186,1994342834,1073737670,2147473616,-1629489325,1542945147,1763963751,-35991811,-277447105,-820364803,-756256782,-1921779781,1524627199,-521671797,-1418016705,-106578723,-366315009,-1693474435,-4199428,-1664761897,-625004547,-4347503,-1946981889,-537919510,-1713635911,-571998594,1855903231,2113895734,-307022911,2009060029,-414745650,-21004673,-118084102,-622171705,-135873,-1633699170,2102904718,1822162621,1775696375,1979532688,-1922111619,-1681779904,-148908059,1201905606,838153206,-609256104,-806377488,1638906718,-692109881,941091583,-1644369017,1590842782,-1188058953,1073704650,1341478686,-1166616707,-1288324644,1006103870,1877867239,1136520314,1559362666,-85721641,-100733156,-371534145,-806171919,-322511878,-396694593,2025847929,1027526218,-1899282588,-1402381067,-557073429,-323097809,1601629805,2008956307,-43391112,-381868069,1864327130,-340041942,-659331747,-584760560,-1917145333,1861058521,-1718110389],
-		[-539248537,-928228993,2003389429,-568771489,-1718370330,-3774481,-84553228,2052013467,603716863,1669200859,659341287,2137772863,1450486239,-179320114,-1125122561,-77664386,1475009829,384195207,-1948811881,-407232690,-855773444,2138499965,-475194889,-155060385,-8481541,-1350041861,1941793813,-136782467,2011097038,-1423449165,-239745313,-328548544,-1006797825,-1434484737,-1354143793,-1660944561,1470960981,107735283,1945730349,-1113907938,1898921470,-1420493446,-1624949293,-476619790,1320803063,1534783399,1197366270,-1129251203,-135372843,1879043435,-9494633,1073664420,-1806525121,1878779880,-302047277,1858070703,1593562863,2004785215,1841035915,1031654622,-1248876047,-1239547918,-707554827,1828697974,-154083477,-1344020547,-37900677,2067480279,1720423803,-83396125,-1493035,-1074399467,2080300027,-1369200731,-1636444493,1429241517,-86745169,-1417679221,-1225333761,-109422498,-817967543,1912592182,-1168281413,-1243811055,1290530267,-1774481617,-25166337,-875956353,1928822486,-52544537,-806175745,-270630948,-1229339111,2130280287,-1686226478,526470655,-335561861,1599045048,-378245123,-1682194451,74096597,-86543372,803078010,2092166907,-1077289219,-1377319,501129674,2028137311,-541151305,1602005485,-1739838113,-1178879344,912256831,1844341563,-1146960149,-1085416564,-109187414,-1212164615,-1436744301,1063026655,-164193569,-1092693329,1926341175,-923813140,-153247973,-872161409,1139988463,771663247,-168634718,1855893059,2076069993,-43270802,2012495587,-1746544205,-185081954,2145682959,-34089168,-302055497,-939632649,2128555947,1995929557,1995881577,-1621469194,1859616243,-307963121,-571063713,-20091289,-444670145,-36447987,-1550932588,1977996887,1577004986,-2125326,-2238469,1542962434,-1652918375,-78696,-399838020,-1406216545,-479234069,2040995004,142349420,-1420902417,-1091119657,-178759361,-1107347590,-268976435,2096496121,-539120013,-325660714,2101475198,-809505803,454328895,-1486950051,-542490717,2046808079,-968893093,717143998,-156509941,-1661143233,-545918990,-895641,2146564506,-1571430987,1253556646,1600651002,2009909627,-14702666,-167799593,-1679261769,-1493582250,-1211261513,2012641919,-348445635,1936128696,1966078845,666610535,-665328665,-154215492,2025150252,-134548132,1724350845,-1078341807,-710021124,1532772275,1522499290,-93365642,-1613244611,-1667518361,2067292661,-567018417,-541875773,-22105397,-187736393,-4228925,1342176237,-43144129,-873711489,-358659074,-270561538,2046801579,1574927863,-1123947402,-1202604041,1568644831,2109209557,-440618065,-327947,-389175313,-33637859,-176646,-76023309,-92856449,-3146067,-77072020,316274641,2079039383,-1078630405,2113163214,-273812226,-1087406122,-746785284,-552976386,-202658178,1982833645,2121514274,-394567,-42212364,-1082661729,1893619319,-208936971,2121267743,1421867002,-290521119,1799158127,1067271960,-16821578,-473993561,-1230032929,326105936,-629426261,-1829380199,804515700,2075432174,2135940466,528432876,-939635877,-2080522794,-762267027,-608714791,-369239365,-39358483,-1327702,-1394651297,-1622713354,1531934162,-333068562,-2115049,891247389,-1884722694,-447226214,-824969316,1967118197,234876306,-1352533201,-611173446,-1326459017,-79844155,-101515939,-684649098,-807720041,748921966,-137921062,-664831181,1069407165,1998286294,-231312531,2044357550,-1493487754,-1310033028,-1343397893,-60835977,991949499,-34329,-1377986821,-1635781012,-51413767,-310378613,-897352262,-1344276790,-134514178,-1365524993,-1037079682,-989904839,-424022283,-1129944106,1596703675,-1761696181,1605878511,1341779795,-314966401,-780534294,1404298202,-1095304201,-514332818,1679264126,-9002089,-1694767029,-1582680148,-1130629178,985587663,-84180838,1543171965,1073604599,1524326118,-1789149258,-907449872,-38537347,-1084841291,532593591,-72354437,-270641693,-1466043235,2050801263,1584362974,-1120927953,115211167,530472829,687316990,2060435259,1037520761,-5657778,-1141412961,1223283991,-33559202,1601830250,1853874037,-1233131609,-3481646,-1650540769,-104997892,1334770551,-1327251054,-747835555,1992764312,1542157785,-2315273,1975503621,-829693360,534767929,-561318285,-557012186,-536870979,1851598847,869104093,-1703968994,-639639319,1976220555,-86196341,944086200,-1245701,2132631125,-610965633,-400588836,-1280065805,-574003306,-881101922,-1400983638,-967120614,1308072515,1068461086,-547695034,2095036127,-302818305,-1476599335,-570557508,1044244991,2060369247,1325395565,1875575775,-540546891,2093280551,1027973959,-505033682,-1090535502,-1732184673,-462572553,-1540448871,-539179152,-2007297163,-346727316,-811175971,-19139842,-1534115906,-109545647,-270681111,-1616385025,-948285570,1005901717,-26214722,258440700,-306250182,-168297568,-16861536,536830407,1073642471,1871572342,-1053562121,-614545795,262103781,-343987749,2106760151,-1226805603,-589386021,-465031427,-1898006273,1790440287,-678467748,719917021,1558861557,-440245249,-1116311619,-1233125506,1995689911,-1038362521,351545335,-1679827884,-1175571476,-552285185,2074916478,-202096426,-1092102161,2044455992,1541698538,1066793903,-77414403,-537977406,133595127,-398037959,-1074463590,1731894246,930608411,-58766331,-1347444744,-633348801,2128344951,-555373089,-33571443,-1754661502,-436548000,-1371816026,1555593612,-1343795583,-1499107361,-564257794,-489236533,2136985559,-1527292357,-1411533409,-1079652878,-1212255782,770686959,-808281089,-17609769,-1455434759,-122430856,1023277865,-1746289034,-1823351697,1893684190,1868546043,-1188039001,766494975,-310453333,1733326371,-558246326,-675547537,-1080820299,865942945,-1011511378,-475708929,1794972863,-1933607493,121053148,2080239359,1004218821,2033188700,1674544445,-829704107,-695915267,2142748639,2059386606,2010043897,2076126814,-1146234409,-805439493,2133835768,666762208,-369328225,520069037,1409271732,-688001571,2137946825,1048575977,1934581183,1736014814,-1611941515,-1734347948,803204986,-1157992516,1509879487,-287604657,-550638167,-839029377,-1394771989,-1771182242,-134961750,-474212485,-303436821,1588264076,2128587763,-1221277806,-91229912,-168091685,-302391819,2133610431,-1617464903,2073932821,-10508005,-742711585,2097117247,-1224741136,-240490513,-832995468,-2025079873,-830845037,-134251367,-605585421,-258119801,1061683198,1993801529,-84472297,2056439252,-1138765354,-866740530,1824849851,-714885185,-631013823,1192394439,2095247102,-1750426632,2130312574,-2030699653,2130704855,-1514824729,-1108512845,-201404471,-222947,-537301890,-820688039,-1895612478,2099634134,768964079,-612380709,-1745127077,2029960022,1837096909,-56099333,-141169157,-7414522,-50987078,-812713985,-344475913,-101361186,2119876442,2110258525,-744927834,-10494241,523102999,-98960017,-945465524,-832571788,-206308570,-105589224,-84757317,-34021777,-1996598065,-1184479793,-1216744772,1574169070,-22030389,-1075125713,-809409,-342170161,-205606993,1869529075,-806391809,1020234750,-216045350,-558423433,-138469481,-857747715,-266941089,-44376073,-957088002,-944966974,-354668757,-1227390984,1469049656,-1279725584,-896622738,1642528251,-270542611,-1890697477,-369450006,-1680817226,1316696957,-545292669,628584972,-1518798628,-1451903010,2121754878,-271124773,-548097025,-1999425,-207619970,1342089087,-153113353,1241200597,-583565780,1065257703,-1161849189,-1142239491,2082033052,2112354267,-138936667,-272304693,-1150040594,-402763861,907816509,-1650502833,-1747290561,-1610856612,1722656614,-549982317,1610456677,2117986575,325037463,2088515389,-1845887011,-1108444741,1870658509,2109388026,1401773303,1022680831,-38158436,-407245585,-1095352856,-1754271761,-1110710595,-948071474,-1175063842,-665175299,-102761987,1309655018,-1664484408,2044653039,2130690046,2004871965,-471863813,2067715963,1680864057,-574636377,-1771596,-537028286,-961031215,-1095174289,696121564,972553657,-1344864330,1874192483,968082345,1876643833,-155881738,-413249970,-438439965,-745801487,-816322465,-1344675972,-52586777,-54813317,1815334617,-634860173,-24158932,-1384124575,-54540102,-311200779,436104954,1608513626,871022303,1989582527,-1833043227,-278113,720206072,1634651319,-898826770,-83935745,1072469501,2066325326],
-		[-1145208869,-42008849,-1778910475,804462506,-148898953,1872559725,-685060132,-109113604,1978888061,-153403421,-556533894,1935754737,-182476809,-507531649,-917225632,-874816263,-42572450,-571002150,-311462575,-258552865,-560078986,867260395,-1204299779,-623633,-435553913,1978512890,-1684218261,392167345,2136668141,-415646521,-694601826,-541890875,-2122980353,-13079613,771176396,-227810562,-1276284433,-12615835,-481954092,-1009200130,124415964,-35992583,1513581259,-1611405377,2092949410,1610267618,1853186031,1367337629,976468606,1299902382,-1075595781,-62036482,735752681,1605361018,-1724910538,736910747,-691109889,2128533263,-1705320451,-1782819,2108382957,-168840321,-694512825,-894018471,-21499973,-9456981,1531940785,-1794244737,1031716143,-270800716,1771726589,2079751483,1329286651,-2754349,-154969222,2000033781,-1610686993,451903100,-1513001065,-68552797,-33727553,1610546399,-1180524822,1174006859,-1046950045,-294839041,-1230021739,1052182293,1807205567,-1091606096,-44139,-101805073,259038930,-1286963919,670428622,2142109692,1179635647,-1529446442,2023124347,1608367401,-411771041,973044110,-120422957,-14680105,-305274953,2111610565,-28403849,1910891387,1466946295,-285235437,-884247622,-266253825,-1359495192,2147350383,-1839891577,-538458763,-174453125,1199237999,1729208316,2112965527,-13113538,-93422604,-1483998225,-857737345,-323846545,-688476217,-1075087624,599742666,-397449766,-256727913,-1090772996,-4733733,-579634979,1594357417,-37750541,-135293831,2054000076,-840061029,1675685302,-1074921857,-605562918,-1065474205,-570573738,-26913,-940737857,2143134245,1556036414,-134415393,-112207394,-68747286,-568665737,-1092682791,-2080982025,1468908214,-673237210,-806652934,2010050541,1341929628,-560236776,2087701983,587106807,-54549249,-28382549,-75025553,1496709096,-1362575990,-1112540435,2004790647,-1116704771,-1078350979,-289481141,-218130729,-139473650,-285626398,-1736468169,-995268631,-1947535527,-1277543051,1996479423,-1510105169,2061461487,-440484451,-6422541,-76809731,954642940,-1984569633,-101912586,-1487146442,-277889317,-1211371539,2144861303,1959318271,2046551972,-1074801186,-1283774469,2132757183,-1078242778,2054026405,-135173125,-322968109,2146374575,-2030473251,-302942769,-1888485419,-848641289,-48440901,2109201878,2113662974,2037870591,-1246897281,964530063,-33849409,934100575,2046153714,-268583967,1045230943,929233775,688828269,712243965,-915949321,1942814717,-537968971,-1611603977,2113612991,-747209175,-10040753,1977088790,-1745436693,-241711794,1584767535,-866559406,-1884573731,1585430326,2125012857,-317817391,-1498144837,-276840463,-1584933514,1095903540,-1111725059,-469805391,-1443168993,-324559521,2084000199,-675414407,-185208305,-3690572,-1092642189,1407694324,-369780291,621640678,-186651741,-1209303041,-1414792775,-16952069,1902828830,-169521249,-1656951960,956230327,-4236050,-1745909026,-1495300137,-2155027,1215586803,484846885,-1782978636,1837021789,1285459951,-1476785430,1491595917,-1713014089,457689078,1610608377,1463704767,-51643416,23575023,-569501,-79438473,1531867863,-1887949205,-1225528325,-208439832,-872482962,1873729960,-1747008617,-151101555,-141706766,-215483573,-942317573,-1366041642,-90261529,410449783,-1349124755,-1620331003,-679826733,-149432329,-1946190165,1455086530,-117653185,-16345166,-1191461186,-1747216452,-68176961,796704606,1698395890,939126646,-1626212681,2078048126,1511243486,902673402,-555782500,-405291026,-2059833352,2013056957,2036260565,1764441307,-1955616027,636681725,-361404457,-35148873,-1153459813,-1242572963,-118508232,1054807395,939377150,1634172607,-19013944,-1681916169,-142923938,189635491,1069546495,-548422149,2011295609,494397423,-403751957,-269926471,130611180,-86712417,-1620060946,-1074763834,-503504093,-843085961,1878957951,-1612718090,-1227041620,-72613928,-849296401,2042555214,-1182269570,-639668741,2126441826,-151015335,1069013999,-553096216,796245498,1073667547,2132639647,-1636614412,-808912950,-459395267,-1359091789,-167805093,469425903,-541137030,-1217437718,-929103889,-17826226,-1187123494,-587482677,1742634525,-210263537,-340267548,-1351094034,-1646825483,1332973527,-135412995,-590366081,1039082223,-239314946,-1356677131,-1477085193,-128464513,-1151338636,-658800666,1564999405,-452984892,-299140393,-87433829,2046458083,-1413481491,-551048579,871186171,-1222723634,854327294,-1090519465,-335828234,2007433212,-1121204422,-1497409077,-94253337,1293268467,1520953076,-203098153,1869594217,-9650961,1975339956,-332015873,-140248169,-1694571137,-1638145161,-1375865005,1572243901,-944339074,1540991970,-2034788425,1073606527,-344771790,-1995882633,-1381172501,2062404343,-472777897,-197167107,-570192912,1299868390,-590170273,-1837170995,-606438497,-83429785,2006634126,-834750033,1197211388,-9053573,-439887967,-339859726,-55214298,-448291,-437075073,-445187651,-1361449492,792722917,1811601159,-689963014,863295486,-106323441,-600114434,-370819109,-1898596693,-545700898,-35137665,-577835670,-308601017,-1519259020,-117642193,-62935814,2077210326,-687104401,-712769669,534526671,1566293641,-1583645193,1021049951,731670381,-555026849,2015329997,-991034534,-1399849505,751153151,2077552373,-1219613185,-46022666,-1162875939,-671264321,-218908683,1331691487,-201459937,-71966854,1205182328,-17875942,-789743524,939419625,-1619202835,725541366,-307347495,2006445911,2084039133,-1389470248,-945342497,-273203760,-2013501953,2043504416,-136679546,-794571652,1204680444,-172556305,-1262653797,-28639410,-97009164,-1746755593,1878757309,1237297023,-782270851,1191051230,1446964990,-1283590962,-549471755,-872547629,-952786,-145394321,1035856892,-795998,-404099445,-1580769422,2136652342,1751340895,2115878398,-1073744468,-1552417033,-140419203,2046771673,1188951398,-1641072673,1858008414,-369634305,-337882243,1460435962,-798357603,1945885681,-844190597,-1687852133,-38357,1038623692,-279554181,1423631807,380614155,-1953368065,-21697713,-876643356,-541204754,-624067329,-83887555,2123348861,-456403283,935313383,-206124294,-92285506,-276857353,-32049475,-625363277,134184639,2053996052,-453366533,-278530913,-1484800907,-40055495,-943786033,-180405616,1964698780,797987732,222281517,-101911554,-186650713,-1277224322,1827355827,-634958097,-21901517,-541069450,-9587797,-6997537,-346030725,-168823817,-1074333729,-19415123,-1618051459,-312607346,1506319043,1992785407,-941637893,2143287133,-1094730165,-85990533,-1485308549,-509923873,-574685457,1574959994,-1359224909,-138021146,-1778557777,1936706302,2101214990,-1252536598,1073731470,1845397406,1060862319,-1149473146,-543163930,-1796347907,-25301639,1399519685,1587992116,-797016193,-167838402,-1150327076,-138823185,-1822467857,1072674716,-1421934785,-1200407831,1542171551,-54536742,1543042799,-267477256,-673564803,-415507585,-1085948035,-1090879521,-56397875,-1091080705,-1252086926,-73411913,-545317031,198326089,-1828459524,-9721618,-645128425,-58330163,-269517874,-394756216,-684085413,456910710,-355143320,-18371330,-776030467,-262815,-1908691025,1639354241,1264570863,979342390,-101210199,-591679497,-33614753,-138941201,-1209581796,-365445905,518975213,-67902473,2011157292,-7151719,1426054101,1736400748,-1076633721,-614054689,1181542137,1073626101,-1074873638,-356549446,-69781540,-1140886532,-44369237,-432212005,2072063224,-1228997187,1391894196,-1187310153,2144784419,-1380751441,2139334321,-874348857,1944044991,-229140901,-1100234825,-15237121,-570516148,-543294342,-59031811,-407023774,-650644602,1677540861,-280504321,-126299143,1442609647,-217520131,2058316597,-1141014811,-612443474,-611623937,-1636571173,-1216032905,-1062390042,-159467287,-149973133,-856785507,441317877,-1074559045,-151477506,914086108,-33706499,-1864189010,-44208166,2134336468,1303844279,-138711097,-944739628,393174655,1337882487,-1635920131,-1447067913,775675767,-610374413,1977460963,-71348067,-55388177,-560230665,-1275338885,-31706199,-372724010,-302064770,993779163,-182833285,629119453,1257896213,-352938265,-159492675,-109805569,-1275076692,-1958284034,-826278082,1874025817,939245468,-312557826],
-		[1861469741,-908942595,1538245601,-42737987,-908606613,-185752201,-764428561,1057472422,2143857889,1398749805,-184963079,1072953283,1845387479,-1296089289,-1761883987,-241314305,-376984965,-1092110212,1035594526,-180880793,-134809602,-572596459,-561709129,-251675145,535821485,-775379272,-957879681,-1117851649,-700580641,-144483619,909636479,-5791890,167501022,2138637997,2114199717,-412264719,-1350730824,-1096859661,-345034979,-17533185,-2040664193,-344981677,-1879908420,-240687587,653646668,-1633683211,-708479813,2052915149,-153515076,-1642406401,-69567149,1755307449,-1167479382,1637759871,-119756328,-2022969630,2037358523,-109098505,-339413378,-1248331177,-1224872211,498321334,-1815503045,-208962446,-1495023881,-901809978,-268470305,637004309,1850996603,-629338562,-608391297,-172820612,2101304723,481164781,2062417134,2085268362,-34357366,2146364287,869779046,-558051369,-655331875,-669040727,-1347723401,-580845665,-1084565159,926603991,1811674555,-901928274,-236144538,-168427785,888517720,-474261924,882096081,1037744509,-710221317,-632346037,-1252411442,2142854559,-1660993869,-1090527260,-89147844,-268441169,1205300495,-44073230,-239390737,-3148689,-64502869,-1163047947,1070591993,2065733111,1599779069,-805319209,1066191631,-135463963,-1226070137,-2039480673,-1099600205,-352404993,-1326920676,-932750659,-776048531,-605179955,-272864340,-595108943,-1637515653,-996451078,903725435,-281216389,-578921010,1870943605,-696452621,2013096415,-1192382500,925630172,1591605979,-1445122947,-1342475409,1991411583,-573196573,-856822245,-964765745,1775099709,1575451063,2077616635,-1560287748,-763428897,1715441223,-141869127,-1799488713,-1083355811,-1781157017,-746834583,-663099465,2120917550,-798232982,802764542,-1147930138,-1074602225,-308675799,729937910,-1070798855,1794719487,-20796269,-730241065,-68522322,-1140384079,-1611666825,1567461247,-307036965,1233106662,-1132552407,-120073796,-35670729,-420102856,-1746973833,1992173497,-119614745,1032698269,-917670914,-23110026,798748671,-80455913,-280012807,-1158961325,-1612120745,-337412186,-672924689,-407291155,-50991342,-10571813,-1093796225,-1210290305,-36276781,-58327553,-1728712802,-202409442,784694658,-1632643606,167271335,-353452037,2008776631,-413273093,-188359541,-1056852,754414764,1082059443,396360959,1571322282,-1808024067,-100381,738168755,-1554138123,930604013,-11575445,-1463910507,-355502657,1828591435,-1625358983,-134217737,-287986973,1995437294,-1222923265,-2063831574,-570501122,-290335347,1599584617,1129894679,-616915073,-1283466757,-604766241,1293939242,-557160717,-332577,-1378063841,416771290,-378638929,1749011402,1006427755,-805342441,-1890935303,-280551522,970176471,1204726565,-1544622466,-140615954,-547623171,-403782369,-564302094,2095121110,2084568543,-132327061,-83913597,-306331853,-131345,989325083,2130651887,-956694621,364666286,-885260289,2110782663,1877180585,-1157655617,-1149293573,-42009601,-292086018,-1246369543,1269741206,1971318819,-884857,2142746111,-1578174791,-100925994,2009044861,-66627,-807102533,2040358710,-700350747,-1975993111,-366023181,-354861639,-352356459,2113655254,-1561414411,-1693646899,-1361988295,1724887868,1793835003,-302518309,708480797,-595612245,1470037598,2130378616,-214606443,-39983238,-1410914924,-553946370,-1610680652,-40032332,-119626836,-771813467,-1477451892,-1137025602,-649607755,-820363827,-1392825930,-92868703,-84327004,-424765585,-1355251184,-543492995,-1359517889,2138455803,2142698105,-1896302274,2038145979,-282152843,-1514321830,-947918867,-147891738,855588556,1247579327,2061335963,-1666198874,1991278779,-1737570325,-1822500161,2094525949,-535770,1868576649,-1671435021,1432352477,-671564033,1742322546,-1183924739,-1749692577,-1111535749,1852110303,1440742879,1266171224,-1695156399,527937193,-325199228,-19148582,-577340427,-722485637,-1371219361,-1628498061,-1342718592,1056694127,-169149523,-1109791046,-63630278,-1746144513,1956542436,-50797602,1492574171,1421516787,-1777123825,463861525,-137370713,-275605009,-547623009,-33961989,-1241516050,-45658769,-8944657,468429139,2071680623,-575755586,-546898190,-1751406007,1608771440,214546247,-1613795989,-270769063,1845491343,1542893555,2094881199,-210376733,2071459998,669930946,-272505093,-1376603229,-1658390529,2038065013,502576115,2102388447,-1089635922,-1845790728,2044308476,2129656680,-755016286,-379958795,-408954045,-1146791229,-108834167,-1682182204,-153627205,-161481673,-641729921,-548604452,-1362174689,-442574996,-461640196,986939293,939095540,-1676706258,1073413027,-1348993302,-8653060,-1342185963,-1076282641,-553927197,2013129455,-557961234,-77860885,-582545743,-1674366153,-204407450,1070976317,-1174991137,-1155404803,-560660938,-1359790647,411017407,2131623722,-253025,1374659702,-95392905,-503349527,-1243382933,-1365526307,-1214334744,251424419,-1761747489,-1141699267,-8407073,-1418797077,-1097361034,-395149318,-943884833,1050673112,-546876341,522742584,-859911467,2047044541,-1364276235,-1835297,-1074596481,-289669275,802735946,-528575043,-286336073,-557842437,-335859403,-1303192329,-81801565,-1749583305,-269837366,-1203766282,-27788307,-1839747089,-558498305,-1241662497,-54663569,1055770598,-137576673,761838388,163440123,-75242953,-272105491,1598939132,1264188219,934656957,-12968177,-10688705,1870625763,-172652113,1505623037,703550652,-1766654721,-1074951184,-60833826,862308515,2044719983,2088721426,-119551371,1570229547,-136337873,1776163529,-70690830,-82903810,-1451343878,-588320213,374728486,-20974433,-691593395,182974452,2146933682,-201524502,-19398850,1844442750,2142494627,-918106696,-514076699,1465631263,-121733953,536805346,-421873993,2092285421,-808825119,-1657775363,-1094826779,-36602483,-43028742,-536977460,-1110245769,-775980067,1969125094,-710,-754988067,-1677757827,2075655132,-1846961620,-852781633,-13663699,-616564785,-568606991,1400566717,-541105242,-1149501763,-445483017,-538867333,-1074151803,836750311,-72374661,-134759021,-4849959,-281203714,-1370235973,-562697655,-614755721,-1292894217,-587861265,-266089100,-137403025,1870585060,-835359471,989193657,905047918,-1615405125,1360068175,737729389,-1805779971,2130926552,-238486018,-536973329,1568517738,-1208549448,-267297,-1501848865,-1091585635,1515843227,-68773963,-456737245,-939692265,1937468382,-1347994712,513689133,535213127,736966425,-1657368580,-1906909188,1518246141,-559158501,-306209913,1073327845,1854252857,-1112049409,-573706338,-34531521,1769977735,1509299991,149847868,-623276465,-1157844241,-1685209673,-1377062961,-1763715094,-1216348702,526101503,338001791,-714605249,-740861962,2063214718,1811814561,-522246739,-1342604961,-51589253,905207293,1003972348,528315222,-118490125,-167791626,-1913016900,-370839877,-1103131265,-338502501,-305013953,-1151555777,925727149,-8653454,-158351425,-442140749,-33587809,-1275625730,-1343906058,-79982178,1023332028,-562053145,-1653361238,-151770133,1515409366,-220742001,-151586615,-38281385,-187773713,-1208492387,-1018774321,1801313021,-77791793,-1297133355,-63480067,-842019331,2070937087,-747649,-1488224281,-604067120,1610520765,-1219665921,1740562350,-537367171,-175138071,1771929063,962463597,-1211590665,1944812484,201326158,-1435503509,-105398609,-241970309,-1100513467,-575577705,1056583599,1847319850,13669530,-268493861,1873977140,-1075065268,-841015349,769555770,-225453889,-10639818,-2013817603,2010528711,-4506528,-570569201,973003263,910024559,1005513343,2100591103,-1141998658,-274769865,-1008255914,1609029821,-320266897,-2369148,-34079850,905500055,-25433257,-2527406,-101296514,-287334465,-1411412332,2008529780,-134430822,-1711940180,-841041196,-14820229,602140075,-1564811366,-1343638364,795185247,-1632526609,1471723951,-1505829067,805142228,531209207,-656700675,-1082858633,-117577877,1051455743,-286528123,-1085375046,-76292771,-354226865,-1485484037,-323043378,-295747642,935165677,-1127416851,757529660,-469920162,-22094122,2056032063,787870483,508441403,719289571,-1633034761,-537162826,-572276233,2079649196,-67176459,-37883529,1875638703,-3477665,-1510371405,-19807873,-143022216],
-		[-201441290,788133529,-318809657,1327663635,1227882423,-884557739,1610514269,-507612955,586617846,493844123,-537452676,779886584,-1905594643,-151105537,1315543150,1810689008,-139469089,-345016420,-1170210903,-135452251,-1313048581,-1614323801,-268491830,-806102437,-168517827,1756855699,-407109674,-470090785,-578155282,-20991499,-1610648073,-1954600970,-914673159,-443113284,2146696685,1489653503,-17087501,-115409427,1408714331,2147450341,-932197097,-1147843626,1941555259,2120859534,-1082698313,426573751,2073802190,1542451195,1065171956,-474129570,-1751129426,1834991359,-423914547,-38813009,903263870,-1082669186,-411791547,-71569611,-544481378,1455863214,2110726063,-27594005,1072560010,-963523906,-100668376,-1199606275,-671378517,-89687185,1508634063,-1346986066,-77115427,2080284494,1766777276,-1832911617,-759500815,966787039,1549749119,-541109661,-11538571,2092908356,-553705764,1874443444,-856985667,-1083211963,-1077950613,-69222949,-136355233,-1107897730,1483428197,-1208222877,-664181093,-538585729,-696680469,-1661360293,-4985791,-1091363849,1453193629,2077226989,-994053187,1540612015,92705964,-1081082762,-836817927,-251142419,-1506868097,-2080806961,1311733371,-577907818,-1677721867,-686908981,-6789277,-470340709,-59059207,-640712721,-1333923879,1498507678,1293014777,797360095,-942958666,264896127,-461381704,2133851753,-369974349,2080235382,-430624896,2013248222,1000337907,-202001469,618452927,-543392897,-320221381,-5639717,-70423721,1999372115,1639611901,1140850667,-1082347527,-30962709,-223625251,1699593597,2074996156,1599995883,-1028213833,-949223515,-1646434101,-1630799234,385416559,-1846177913,-5248145,-370867269,-169071907,-472918281,-648029218,-571083285,1675534314,843854132,-941622530,-10692635,1592782591,-76321321,-1610884240,1530854342,-33833027,1744279615,-30313031,2142628859,-308508933,-1146126339,2097147243,-1655705617,-940573002,-1208025274,-1761680422,-1778933972,-1896546378,-1584137,-1601710445,1069527870,-1144968109,-2022703665,-893408163,1958701438,-82708626,846940126,1037041647,-588056157,-620975624,665679739,-1179681620,-17461409,-1127007743,-565326039,-135138661,-387976840,668987843,-1347559459,1305804730,-1259118741,2129892984,-76578938,-671957722,-613621493,2147046394,-608738268,2074044275,-991372057,-608849025,1430257350,1316630776,-761462921,-15767561,-700457092,498887669,1071493456,-406931907,-28838305,1534844221,-10587138,2141092445,-671154223,836237852,-808502299,-5410325,693864919,1021804523,1056622557,1578822079,-1895534724,-18223251,-269244673,-420153492,-2080730186,2130370559,-1619001431,-1988765611,-37757617,1423863437,-1103234327,-103033226,2146374498,1190656723,-629414475,1446913235,-589488179,914341886,-1162519042,-739901794,-957528024,1981790173,331217275,1985379035,2046288670,1581997357,-61763603,1205228796,-726434306,-460148997,-1211140811,-805308005,519304178,-342100106,-412377225,333395173,1513594850,-139026498,494746544,2143157115,792697149,-150739105,-1209675779,-1354891355,1983661991,1877916637,1288550134,781990858,1960703697,-270812237,-1112068374,-318906677,-1229197633,2141975167,-1821431170,1806564855,924580869,1381986548,2075367325,1694199770,1878423047,-576226023,-919418017,1705997310,-239143105,-10893613,1796126028,-27264772,-18097449,-1094915587,-80216329,58653424,-705507742,-155569051,1607588475,-422600931,1677323766,-210011706,-287744137,-251767493,1605029499,-1286084867,-591430532,-134294299,428984702,-1246796018,1809212990,-173427545,-1166033027,-402654377,-369996843,-52429330,-403474,-518262855,-206962849,-134759458,888894299,1515912663,1799286655,-1345399517,-609554618,-90245124,-1703621810,1069508061,354115589,360316782,-158882563,-403187714,1875947479,-201459981,-675309953,2146303286,-18880531,857650793,1864343163,2071717455,805226209,-1562215019,1388301038,-1576297442,-349720358,-1354773733,1198659033,-1811946249,1841028590,-628830,-78693635,-148392297,457698590,-1343250499,-1379955553,2085333444,1981715935,-883025711,-222319960,-404849938,1562147502,-202648834,-252052561,-1214694629,-1746944535,-1132740938,-109628981,-270532949,232980383,1734286681,1991941937,-270643396,-419539081,-22086231,-1482572481,-240158794,-613295154,1509864780,-328482884,-706283733,-571516209,-1421168722,-478454807,-1727138177,2004840179,1030806487,1335191255,-542689857,-1098318531,1093559803,452926310,1702821868,-407735873,-89277602,1719367148,1705252291,2079976287,-126304519,1142290769,-5775786,-1415053639,1677195743,-1107332113,1994971198,-39856750,419331579,2147425199,-1294500143,-184563281,1867967963,921155542,-420903389,1998450683,-162091046,-192956675,-15075670,-654913695,-1867028787,798334357,-822490305,-1246265897,787848419,1739480446,1856894430,1858338518,-1481789121,2070601178,1335884255,1843114746,1132375667,2147174110,-541922,-277485218,-471875770,-35553800,-201474594,-303484496,-875858090,2130827839,455966075,-1912645786,-463997603,-1164337300,-816200739,-700469282,-390300185,1034677247,1051000699,1062418423,-21138489,2116807165,-1845612963,-840959551,-1074648667,2083364760,254493143,-144506962,-1380237545,-1243210267,-329386115,-84133893,-41947797,1729322599,-377888078,-911493330,-1285940246,-1142483139,-274039843,-578835810,-78524417,1993321457,-5515043,433573366,251341803,-357417027,-1270890833,997719989,-144444106,-268713555,1823014397,863614975,-1074105380,903075443,-278741526,-48824654,-141247505,-288358438,-2038952784,-1214405310,-21284865,2136422371,-1544525453,-1209180627,1381463167,1607448575,-2128,2077733862,-568480957,-575374083,-162088985,1790177051,258998261,-612927504,-1648892001,-1098223044,-1394082577,-1448612025,534244345,-1879048661,-615527193,-822914513,-1074268706,2111501598,-1380714017,-268927370,-67585289,2105195418,1602125769,1859058580,-9724140,1052161439,-558409765,-1112737797,-822383391,2006687699,-268452035,-1118319145,-280012397,-584650753,-902146392,1740570445,860878346,-394352798,1845253177,1325365945,1878830454,-1476665410,1743615645,2130165335,487978813,-809547802,2130425324,-345220161,-403784226,-1646809113,-34767987,-876614657,2003238858,-1366655601,-1252336770,1588854781,2043772780,2143451001,1604026310,-1676259476,-81314827,-925536965,-171165762,-657729,1193102300,-1549807620,802970549,1415443197,2012315023,-1383849,-268585521,1071636198,-1084639318,-1186745940,-33573633,-22382085,-1430339596,-2131550839,-605681753,1878809343,-1193546964,-201984899,395171055,-2022486085,-303171601,-11011394,-189174594,-1428310145,418568494,989838547,-846398782,-1479584639,-15473049,632485313,-148922883,-135409673,2096328671,-96985622,1568460579,-49558227,2128522942,-1208049731,1056760598,1574750157,-671097655,-1644395617,-1352802325,-323001369,282590457,-153306019,-1258582563,2135856506,-406519809,1248335870,2144137142,-717242415,-236237861,-1094649890,-402670561,-23633297,1198061984,-71992321,-247465022,-1485867217,-1094762757,-1619659362,-204682021,-979468329,649985359,-672269457,-1632717673,1743248381,1934218901,-202507431,-1459421226,2133324887,797892540,-136080065,-370213091,1786743636,-577311745,-100679729,1469008243,-575687237,1375463831,-276825626,-215647138,-204473237,2012512733,-1783062082,-660087025,2147384224,-142806809,794528693,-538666429,481818339,511671807,1936641597,-42340621,1375190887,699632251,-90523282,1995808655,-33816742,1844557801,-856753666,-1375981569,-14342341,-35767685,-1873249069,1736306551,-604054314,-167788740,2029895079,1073182398,1003386793,515241738,-828113721,1836377951,1031542706,-547750145,2121759063,-160991501,-296614926,-297301313,1327331084,1392494195,2105245130,-688373968,867024638,-336900356,464231081,-24773489,-354850888,2145242774,-87149321,1992289759,-744829021,872365837,-1745256476,-612384769,2139012046,-321951887,1835450206,1287298815,-11151108,-742475818,-1279148097,-42216642,-346075458,-1758737548,-342427506,-168575058,-1108029538,2100428542,-897589250,-528392,-352726151,-1514374721,764475160,-353927689,-350323827,-1577078481,-147849828,844485959,-1226002989,-1645954115,-1132626273,-137068548,-30540385,1878687058],
-		[1276761531,786185995,-1126188051,-354164771,803167100,1596562775,-399531637,1071667063,-873213071,889061358,1861960818,-906002573,-881625283,-273385857,-39125315,-143854735,-1411492241,-218203066,-559915298,-54577181,1609295346,-1948320261,1572640715,-704785054,-67151377,-513934497,-95181089,-1746471405,-12600185,2108857147,2126091935,-1195382097,-172525635,-185041435,-1292962543,536403877,-207620487,2113273829,-1082340141,-1344864451,-184559747,1595916927,1860025535,-621823201,905639233,-1098097219,852340169,-1126012011,1551823439,-52068961,-294257105,-1485702345,-239421693,-336130309,-1220054641,-270439980,-1751392321,2129058668,-1816879122,-33556675,-320868675,420048175,-169807761,-814121133,-31550177,-355999073,-555881218,-336265221,1870494175,-120193426,-1950002507,1841296883,1333251913,-806101849,-742999059,-1393189639,1610586831,-642039844,1736121213,-1073908486,-455185057,-173149187,1608496503,-339358466,-25176631,1529509309,-1413491505,1574325734,-1231162817,2146325078,-303022627,42413499,-1326712321,-2018189545,-1715898929,-680657858,-106038081,2035189227,-576722547,-134264873,-840327445,1582783451,884416510,-1250951229,-3408515,-77930765,-1678803802,-143706155,-109103302,-1947111526,-369386243,997317531,-1003382057,-587433190,-366368411,-211992620,-417871873,1863559917,-903985165,-13716993,-406197383,484225023,-1158988310,-1088965735,1877971839,1563781359,-1343619083,-297845284,-33890817,-648549377,1941951239,2049832925,-82759811,-57703057,-1218490665,2096951070,-759924227,1531920194,-1386585859,634845113,367482551,-403572929,-680560763,-134656073,-1100453381,-940622916,-741379684,-1621271046,-575276602,-1091108941,-240349423,-154288480,-146269924,2012999406,1656750078,-1307989507,-22741893,-67117635,-1950118055,-321979793,-595354145,-34545685,671067644,-857375746,-25198891,-35977219,-1389625806,-272760849,1819431901,-1057071688,-1898211399,1769370989,831004541,1608938293,-291121555,-805448417,1803218549,1374091101,-1443430419,-1343862338,-948350669,-638853121,-402705198,-1086914661,-1113391633,-1156324062,510994923,-37504802,-377722162,-724047909,-243794465,1257911815,-707244118,-1879106059,-16811409,-671354933,-1099020562,-682166356,-805884438,-87361809,-662770297,-855776811,2037731194,-134619075,974267548,-4334789,-604338186,2044558847,1072814700,-553657937,-1631849497,-413168227,-422908939,-1133611412,1163512702,-302925093,-539063561,-8397988,-1074119489,-278083963,-847315450,-1133793413,-103340033,1273953275,-1485840940,-1212219393,-1682552845,-1505862411,1430198441,-105121873,-832782482,-10352657,-1076935761,-1712464469,-474008113,-67387393,-134742783,2096846195,511432534,-671172841,-1863222337,-610021049,-807012830,933422778,1235993575,-3539012,-281114850,-655506,-614801665,-1148491906,2112075580,-201331233,2060975732,652552633,-1402030082,-571326625,-227542917,-341338254,-71098371,-623676675,-883557262,-23117893,-373426745,-136152738,670366426,-784484218,-410305027,-33030819,1707599797,-48496937,-540099038,-1757520113,-1983989538,1442200307,-907315514,-33826873,-563114177,-370835885,-1561388616,-275251290,-80233473,-138743169,-1275392579,-692259156,-584072274,-159477793,-138592297,-1334124934,-173356579,-553699357,737588219,-1417032322,-356673575,-125843485,1475530717,1089187003,-107038251,-1479263290,-286032007,1945734137,-6406217,-92669474,-412721185,-152572225,-1260407303,-360982610,1694105596,805008839,-198922,2029043710,2010597743,1072390618,-1442994539,-1246761209,1808980958,-1427262597,1337810675,-413172012,-822158089,-2083,-471112709,-328401954,1909869496,-1501765637,-1778958869,-654835715,-36723850,1021271798,-545417,1821368285,-30426021,1073718781,2146467111,-1384284177,-1851593,2132786364,1761516512,-1091582481,1261955577,-2949826,2139084789,-634196164,-436345746,937705103,-201591117,-1900110338,-1107436865,704020446,-692086634,1396492287,-1641023745,-1120144198,-336652982,-814432269,1275047107,-546075956,2080355959,762803759,-809678089,-1089523972,-1075224958,-675653863,-1568698509,-304218901,-88441345,1442761983,-1137575955,-1419323755,-21236402,-134841422,-2117117189,-67421386,-1266958565,1321196463,-352380441,-353710121,-705725016,-8618132,-556627462,2065395499,-54761765,-76838049,-236497271,-362940973,-1212885121,1833516095,-1750860817,1782028027,1070815455,-128912945,-34399235,-419980331,-1124762881,-538725409,-1210191913,1994239199,1744793860,-1830290603,-1141637124,-1881522650,1454274538,-439428282,430996963,-62330926,-1906018498,1855782603,-175857666,-1243937093,-75809931,-413799201,-1252278553,1843378858,-338040199,-1388519593,-406881025,883375422,-1217317637,-68025065,-1459757698,-38019121,-1231720473,-1100221131,-1258332373,1810602878,830461949,-1209436563,-17985888,-1611301570,-1637516291,266065406,-1069942175,-1216414377,-247676931,-598632707,-1476739270,-14715525,2035829716,-384832582,-101059353,1941222849,-151136584,-228345986,2134879186,1035436122,-9060365,-1077064245,1945901419,-272951748,652840191,-430517253,-185582373,-21833745,-1242671882,1600055403,968841914,1876621183,711967614,-1321618,-268731137,-1094136737,-818163713,-1074268526,-474878465,-1143728385,-180859445,1576465838,-306594901,1700257653,-365429290,-1753319468,2144327676,-303043351,-591738049,-76846249,1105001340,-1935737604,-100959235,-672149731,-3673249,-1171392669,2146700773,1609234175,-1627666607,-1893848117,385697149,-1621262772,379442845,-1076862529,-76759107,-1618560793,-547897485,-239092784,-808530905,968812619,-1746975902,-125830082,1937719006,-1090524242,683614173,-289407001,-571950742,-7408450,1808756085,658105755,-1116745518,-343065169,-106637506,-1659896337,-274040951,-1602249539,-1669548229,-117735690,-810035809,-10906706,-587254345,-1502368401,663090543,-1311396444,-1803692802,-49365516,1952447470,2078276533,-810680715,-1627720426,-2823425,-235809249,1591656431,1593339867,-1141919299,-75534369,-470795305,-307256103,-149434411,-1075008417,934522067,-1074127105,1530915582,-109353101,1465836526,-1052906483,-1382088459,-1493328897,-218694103,-472982026,-539598353,1871691733,1895683967,-25236599,-1091389506,-134766619,-293997569,1592098206,-469860698,-274862374,1969401276,1471938551,-83965588,-310779913,-539256651,740583165,-606333500,1699266453,1061665624,-509624389,-1411091799,1945411447,-1226326742,-1729177649,1743682523,2005989695,1005497271,-1621048069,-13980931,-1426362162,-269037868,600223708,-1627504701,-1146618229,-168448325,-151632330,2062995151,1877966511,-852787613,-3908826,251637759,-692117653,-588920901,-183766087,-1714522395,748204594,1661594859,2111664509,-982598787,-1218553993,-1074571811,-1084254315,-1082667017,1023307605,-176662926,1605552894,-1663206419,-1081854977,-579178798,-101747335,-147029,-279024001,-738840651,1908251988,-1736983889,-1758398510,-552898690,-1418069283,2070775532,-1074369939,-68173863,633731307,-680565905,-1126435074,2144065846,-978748763,2145073342,-634425425,150944637,-547422348,-1880560769,1532821370,-576064089,-760580806,-1444249699,-269771393,1040088540,1022536867,-872424967,-571292001,-1749073921,-1133472154,-606898600,-1242596417,-43009618,-424442753,-161876103,1255799421,-414991433,-508073093,-1016955155,-1939343364,-350244889,-947913093,1857851221,932230122,-550746316,-1855727974,1608463822,-282345545,-27496519,1727755830,-1075077378,-778240819,-1751160841,-146820129,-999901190,-1547881697,-84148548,-1695678513,-541598785,921501678,-144711951,-187114355,-1075990551,1781697531,-1147191593,-849182770,-314968897,-335545485,-1076764857,2129465143,1132450802,502007469,2019163860,-679954580,-109405315,189785566,-895486680,838759925,-1633224948,1940446525,-420336708,-440734631,1035331423,2128326220,1559589727,-933505158,710933816,1171502303,773585142,2089855103,1408622587,1054011299,-420906309,-25709329,-1226938883,-1143478305,1593421263,1837104203,-117473597,2073542110,-1913397249,-708844121,-114578257,1943860375,-1843021585,-206504075,-100796689,-14811349,-38870290,536452891,-8567828,-193462622,-76383915,-1262520390,1992089148,-1093437333,-201505845,-1224900625,-332705889,-216793281,1777860351,-574744082,1604265051],
-		[-68801963,1969093276,-1359553628,-26814539,786949719,-40831234,-563441917,2128472319,1560154055,1339533821,2145317745,-386271453,-917777220,-261358326,-1674641729,-135664685,-810563628,-2044302373,-1351647789,865893759,-68502179,-8267908,1844430763,2039175486,-1618105505,-1645151555,-1410089153,-204702747,-15205649,1744805829,913948467,-1344291830,-252011523,-1342242817,1594813341,-675442067,-18929217,-608506040,1845163986,-134384757,-1352159237,-1139013153,1339847632,-109325833,-279053603,-569991173,1994137669,1978621387,-671352007,-1849693381,-547651204,-1971386395,-1681377348,-1124589962,-1382091971,1402682961,-960988738,-1758763681,-282591462,-270217249,1304350549,555645932,-577264185,1404287383,-1048628484,-1710258433,-1128636274,-956539225,-1284513821,-1711813657,-102350977,-1754174866,-1163075596,1029233645,-173717537,-751655127,-1075402537,1664417499,-417604106,663469499,-1889054277,-278935682,-689988112,-578295188,2120030143,-558252457,2074992195,-1277714461,-1655228694,-1090715649,1669184255,-314397262,-27932050,-1243124823,-1108426793,-622143573,-4597059,-1199592529,-674369625,1465380006,803728031,-1800573715,-407925973,1870336765,-17565354,-142083077,-1527811,-889717203,-1312817444,-579862894,-104978241,-1207997803,1985797772,-1077417667,-1077515415,-609305669,-1157769607,-8949441,1576585706,-171120417,-3309778,-876740841,1580989207,-344115247,1329539959,1587389885,197027774,1305454591,-807946407,1860304378,-36962587,-57737826,1832894385,2105466071,669429607,-127144387,-548676360,1466791786,2140891949,-1089153075,1021955893,-1116736710,1759885722,128858815,-25690318,1433304174,-948144353,1953919971,-1677942817,-566526777,33027286,737462685,-1310761510,-548012369,-1418073421,-537526337,-285328921,-849610945,-1344274709,-640683485,-33587587,-1159757961,-1150812674,-303575041,-1128462931,1322581874,-1074629667,944049621,-73483273,1556833989,2090705718,1491986687,2120070820,1561919191,-37749400,-1629825617,-46735437,-436602164,-1717057042,1804054459,-1712456705,-1233875073,-16916493,-1358312082,1513541441,-19439709,1694342127,1825044474,1743780319,-1413492259,-87892795,200751406,921648481,-1153475237,-20682801,1039675391,-205662209,2029843019,-138303501,-1847301,1983807222,-659823307,-671247125,484413367,-1074278705,-1084297416,803080952,-1511264955,-1111792409,-1124079665,-88615808,636333854,1723854315,1597961515,-6298157,-1663177795,939114199,2145230839,-1607014533,2131609256,2069888986,2073689471,-1422528775,-547619121,1597472633,-1079429749,-123798309,-551093333,-25438435,-1161991437,2138229111,-8545035,1149883871,-1213891174,-1613005893,-38846499,1574858722,-573079577,-402721067,-780160021,2096057517,-1091223877,1875506677,-67289139,-564698133,-1218725379,-1356991237,-312639748,-1817060885,-269497269,-1193345451,1834347759,-1627739154,-208750103,2144862191,-878317619,1308359941,1301858047,-71637449,-1654668984,-1158679053,511048692,-22038051,1874048894,-416026197,1283063419,-1042794593,1845492414,1559493755,2108155581,1749940915,384659453,1302461404,-67176457,-1141508177,2146169659,-1432117553,-62260226,-553692329,-657241326,-738202163,2140698539,2013109179,-178291651,-352543249,805109663,1742355071,-671222889,795600894,-1228307975,-204645551,-1311294086,1224715481,1753611227,1056424924,-353061449,-554637862,-105251462,-1794410182,1920948015,-577782311,-68749835,1857714172,-941639205,1994387966,461346510,-384573697,1017151074,-549588549,1700560413,-1721895489,-517429255,-709382341,-743447073,1173714862,-1104939362,-406847562,-1122665255,-814383766,1088353759,-1459638575,-437256205,-541591755,-887726138,-270663942,-1361071361,1771924221,-1220370821,-573578322,-1080967554,-195060257,-549986510,1056399359,1898698461,-923083089,1761434367,-1677736470,1303248821,1743253119,-354527499,-203711205,1406138355,-1594593330,-75826182,377956511,-1409286275,-1485490177,-144844289,-373027686,1873392240,-1969269153,1865030527,2044942701,-1081616451,-1614167681,1857875763,-1281361047,-440410281,463953692,-184722691,2085221242,-696441097,1944940227,-251699345,2105619710,-115160209,-4526446,-300603845,-1099927814,985132780,599514710,2130687983,1026778815,-144096997,-819087914,-1103287810,-608016538,-151183378,-747765829,-131334171,1056322558,402598875,-324051055,-459416065,-826482947,-25171558,-1395640947,-692208650,-256380619,1941929030,-286262291,-9984409,-460065796,-1210073094,1576581950,2102079125,-699375625,-688473649,536870332,-1497957786,1205509681,637179179,-269029505,-8677635,-1082665553,-547292445,-1390195331,-1290799114,-3146250,-1552026693,762281448,-72020037,2068830971,2147352511,972457471,-279664917,-291373874,1608474501,-1947484513,-570991022,1945620267,-540316483,-71647237,-382313987,-17867137,-1145571353,-33600809,1878410111,2106580627,250835183,1068235740,-1008098210,1203694931,-1101389825,-103110022,2045748213,1468908351,-1388851473,-3966234,2136464591,1607624522,-975200595,-1308635529,-1086916661,1727684348,-90575489,1876688077,1040142298,-134287361,-339756552,1805057901,-76832954,-54059393,-738738385,-1225008181,-1212954627,-755396625,770434815,1740594879,-75263503,-681872904,-857097864,-538120085,-202245122,-949838377,-1091570270,-590945815,-436999182,2010361853,-1360957697,-92553315,1877529931,-696666124,1178988316,-1200242723,-1242643789,-1369352641,1939840875,-525379684,-1411672080,2071678939,-603980294,-1363415342,1558826095,-788553790,1472067261,-100684417,-281552994,1740366267,-1103700561,-351600913,1911942698,2095447607,2146811212,-342885122,-1677721646,-463996584,-1611809453,-1110442210,-34674241,1440168447,-539600998,243171250,1933299711,2140566502,2080271864,-1619017862,-403704596,-76828865,1607885638,-26522137,2090852343,1872688314,-2015498243,-1343392546,-83899014,1043665759,1866202105,-902632309,648394126,-67921297,1131793386,936030189,1324723069,-537199478,-67443355,-554766596,-138494483,-1505902627,978842044,-1409813635,971749870,-461637310,-1073825922,-16854181,-1090946251,1284348911,-1504559267,-50954504,2046461931,-235455578,1859043252,414616573,2071929594,-2087190633,1337321434,-2654225,-41945387,-1763190093,920637039,-2166921,-1403130391,-1566056521,-370758179,-545407111,-96880837,-1912624455,1457601527,1531966268,-436876425,-1644184107,-1852425537,-643827966,-554009743,1506208735,-1468043361,-61343810,-1896350158,1794882172,-4351237,63956190,1275854765,-556072963,-1179845234,-151685825,2012010111,1878978275,-1154845389,433470335,1527742255,-329256962,-330443469,1311207256,999128028,1593654527,-1281851977,-155205701,-672952692,368110493,2000666603,2071984589,-142640133,-1543505129,-1131481105,785115102,-152071852,2041298927,-470878355,1073157631,494682105,-232362471,434502651,1671409231,-90721361,2142238681,2132754426,1471641055,-134236207,2025254326,-1619542045,-824366373,252906987,-1233819993,2130408601,1259316439,-1579472905,-572658977,1778279656,-613687908,1585379071,-767168333,1861804027,-137661529,-726663211,744356861,1869446399,1304427389,1555011311,-67116081,2106583536,1506191147,-1885668612,1525921704,-75243585,-734552113,1671279871,-5716515,-818676745,2080262011,-1048593438,-55869663,-614990850,2124797934,-181090572,1533542268,-750791829,-1678016545,-302121590,-1744892428,-271405389,2042624711,-51673610,-1414013841,-322702353,-136324387,-1576124034,2079926013,1870182910,-1444938145,-177216262,1970135033,1842813950,1832830855,1876948638,-42080785,784969187,-1743679493,2145859056,-852074563,-1955693186,1545838267,-136317716,-1409397276,-724971651,-410260163,-342024209,1058995774,905740072,1398795743,-1632702468,532635515,1453567863,-22218844,-1082475873,-1744900115,-1488196492,1820237010,2091755422,-67124287,-1113637121,-83939880,1532414940,1606417279,1667214249,532217726,-580671493,1170109693,1555283966,-683435025,-627626322,-1528449802,-288009870,1975383671,-1821667414,1255404467,1070529019,1067041783,1054660317,892697983,-1861370913,1039531995,-553168173,-177046149,-1256264225,-337675033,-914359394,1694425037,-136945665,2142696831,-1560384772,-339093939,-261427461,-1983509825,1065332139,-50484246,-941101889,-872510465,-671621129,-7880889],
-		[521928106,-2105619489,-1518745926,-4218636,1823422780,-14771713,2084959762,-6825656,2092892019,-1547784761,1733285693,1334830862,-236530988,-453133085,-12716731,1584362926,-690857,-9926657,-1749045321,-740322580,1576523035,-331399211,-347496832,2096655827,-1720995634,-954220641,-2547333,-1644822841,1068384111,2137701882,-42807461,-1870660209,-17959173,-1229193371,-1164979564,-1913735279,-1107693761,-538191689,-742756362,-61082209,2130078952,-377686040,-68503907,1962904605,-471647814,-11804809,-965296295,-298148947,-318777690,-286785556,-5296692,1572187935,1844968541,-124005445,-9760887,-218687910,-41959457,2011003671,2080304989,-1077674234,1744697301,-15342113,750777623,-1898661066,-1219663605,-347216186,-1101820086,-420912179,-287442033,-1879773314,-327170829,-235969137,-22413442,1044364511,-268486744,-557881940,-1245709121,-1628531210,-1386163289,1907031469,-159987405,2065898335,-1401635201,1073608251,-210010371,2020110282,-209059907,-1088955412,2063031227,-143764482,2051110079,-9966679,1351613802,-221325641,1473980543,-291914,-1084361057,442834159,-280758531,-67118237,1710857335,-1086390441,-291441103,1047904053,-538469385,-1150495532,-805496327,-1611137530,-553714819,1722743502,-587230478,-1611419158,997685138,584212221,-545260737,-748684355,-933167209,-671145500,-1478591634,-404165127,2113397473,-80671181,-1074026545,2074894263,-459743393,-775033085,-135299874,-1084396017,-642319492,804973239,-1548813601,-1611093261,-248406033,-67264555,-1225572481,-822127667,-134418773,-1177568281,-6656105,524245963,-1365426339,-504398041,1991989173,-1745183387,-1110999221,-1214809241,-443549394,-26290211,1610612401,-1087195218,-356948166,1832844863,-1690366435,-172032765,-556942379,-1432928275,-548077605,-1120021116,-1487212611,-559477652,-1737531412,-1746403719,-76155909,-1078413341,1524794237,-1972115714,-316806604,-327684353,2112285910,-38183366,-275168833,1732165119,1003224011,-1448677715,-941660011,603329701,1069238271,1602120061,-1143545857,-699089169,-22518177,-55641006,-324158291,-1622718450,-83997477,-16867805,1233123959,-348483545,-1645544666,-537395236,-1141086479,-1501725233,1820126828,-1476619299,-174236161,-218628281,2037714759,1052241399,-1350828065,-611409,-1368041665,2093977430,-719053185,-570529343,-279150630,-1445300442,-1153145698,-375780021,-107494470,-705552755,-138951585,1994702519,-335549190,-1375790084,-835412645,-1150321870,-209194538,-1098612994,-1111801905,-1126452294,-172679369,-231151748,-421593217,2130566974,-1075717353,1799838901,1714886645,-156305448,-562112487,1734997482,-168797400,-1858437196,-1242784914,1837049724,1304100728,-671940613,-202376546,1609178989,-201392843,1218735735,1672108222,1840185078,-413322243,-705706949,-357967105,-571052054,-396821705,1876946774,-941362073,-350375073,-814055745,1054789461,-1477784336,-86143701,-1628283973,-285674083,-272715803,-77359907,2113797095,1123674219,-74987482,1919866455,-182718468,1490726775,1584778195,-1082707980,-824808002,-759759094,1005612991,-709935457,-617236179,1170669534,1400618743,-939737354,1581721068,635041365,-859077507,1609800678,-1260951881,1323289343,-539464962,-504509561,-359686297,-169466055,2023717747,-98370,729630556,-1770648631,1836055259,-1615437257,-45946029,-1543619081,-285221777,-2262033,-167838793,-471698068,2009055197,-25796296,-163911,1940831987,-453317125,1407950396,-411308037,-523150593,-839517204,-77520937,-1705403393,-93104393,-1692401680,-1479610497,1263267805,1939454897,-136349831,-958947393,-9470141,-86222146,-1141440942,-252577977,-38322375,-945726721,-413435980,-655556623,-1210128393,-348129684,-44227619,-1142949057,-1677766681,-1344931110,2140789730,-1611797207,-268698189,1310554079,991821231,-58729890,2138323709,-672981066,1877725130,-330351389,1875995246,-808985048,-538020118,-1849967687,-318931505,2011555574,-1119949521,-102896994,2002059215,-167063559,-2049519675,-1644200129,986957295,1978744815,1473862518,-1697126413,2144989178,-1820607353,1605104567,-1086587853,-1268050533,1839903485,-590916,-1479689730,-306499636,2046279609,-1746395858,324990108,-1631662661,1876914332,-1320265611,1849642280,-51488135,-687967932,-1376002155,1742700369,-270568153,-553703501,-71434529,-1412436754,-1356079877,-211309580,-2036368097,-604705889,1460402463,-1560304130,-571488001,-1879836161,-1082921249,-1210138741,-235420465,-1745977605,-1082839363,-402663109,-441360693,-142780770,-543270977,-805635141,-136485163,-187715917,2111552938,-240157265,-436428849,-1010041997,2113594930,912028497,-1394227437,-153464019,2100796831,1332587931,1307957946,1045008322,-1465961475,-1103389188,-67119394,1589176596,-830505221,-1262135,-1080380259,2040921854,-439484969,-1482789404,-581215375,-704653971,902691701,-1581643780,-1677969658,1605691309,-671093264,-198476034,1465040879,-9556115,1435737961,1944559423,2049178606,1828975094,-696565769,-6439235,-538150003,2072764141,-1747536998,-1147408385,-163717701,-1040526865,-2017566981,-10887226,1528060883,-539336977,1877328087,-23671595,-590104825,-882936578,-386167075,1357605822,-1127941850,-1971847974,-33586049,-1430591011,-404095499,1744087927,-361758949,2143286262,1855395569,-587638600,1140576231,-206735446,-457540946,-84187906,1543344119,1811658467,-174562641,1451042074,-1266994297,-1621109785,-2124679493,-79728642,-2082484243,-1141580006,-24444971,-160040881,-1224777423,-1451794862,-35936418,-1293814022,-588390451,-1114968657,-755007767,-238555926,-815213864,2147370732,-379107002,-1900051247,1866262909,-21784782,-335578149,-1084246219,-145850561,210607810,-465675761,-135036929,-458247873,1819398135,-159909073,1266677750,2144730731,1283096563,-857244002,-1612465665,-73467435,1287574911,1578745327,1608906620,1471826935,-1294847617,-1145064477,-834487897,2076663647,929805298,-1623249219,-1927324297,898552315,1333726796,-547423437,-875589639,-1371603002,-707465354,-1611787059,-1671921667,-40961,-276833414,-140649473,-119882269,951938684,-1112769571,524274787,-1363445668,-52445258,-720937,2077738991,-1655200266,-1196970289,1037921151,-1518440466,1849022955,-59523627,-342032546,-139125837,-1375766629,-271683431,2063562493,2021554607,-52167294,-583096836,-308307500,2106159830,-273822795,1073675168,1840214429,1609985529,-849692737,-135455586,1039127548,2012891103,-1549342130,717985761,-1099309645,-752356377,2071101359,-793812105,1609546399,-714197108,-119800107,-1179782313,-42637706,-1041253572,-1107382563,-290999298,-1124130828,2142036478,-363235524,-114435403,802019807,-1829211150,-807403604,1715269119,-624984398,1942994802,1163621311,-17596980,2025840059,-419865553,-1755591523,-680175617,1719432990,1778250661,-846039747,-129951763,-1189633573,-1135091783,1467457535,-95124229,2012674491,-242303014,1051585847,-1476644372,-62032005,-1082540315,914325112,-971065941,-16935173,-33674690,-142869585,-885922829,880801059,-1613121834,1560273655,1475696109,1878900473,544659286,-103305643,-332710921,-310445585,782232509,2139086558,-54639236,1201664959,-181470977,-436213377,-612476101,1979318174,883645812,648904339,-828802478,1850430668,-1640813472,1771894487,-269111915,-1961628033,249233260,-1414537285,-677605569,-335808340,-18944571,2074566007,1413160621,-2033192998,-537072581,-252715521,-218187813,1475606143,-119564899,-577186548,-390382756,693886839,-1181225601,-67642707,-253937709,1044871103,-72880129,-1107822621,-1158153746,-329089094,1543081405,-419989909,-238566445,1890694426,-1074176066,2135944398,1879030475,1483926517,-1291849345,-103392515,-1303388451,-587961861,1767960039,1642798556,527208375,1333255903,-469935202,-38839045,-69640618,-299946038,-1085826113,1719623415,-814096422,-268452155,1005843531,-623445156,-536880389,402537867,-1950026011,2130003942,-68732566,-805583191,469463014,1571851117,-539051393,2111692724,-96275139,-251658245,-142749701,1763638765,-1747529797,1811132248,2096295159,-1885513731,-1348208003,2126461771,629144285,398858717,-809312449,-1121714545,-1108086787,-1196861930,1876868839,-271974470,-556950679,-881412109,-246915073,-271091715,-10027095,-1820345979,-439699553,-205946947,1435498855,-1881539586,536804318,-771938597,-2073053843,-894199813,1054687226],
-		[-235994121,1866457821,-478675833,-1901078458,2110585983,-1765476685,2079739898,-646222023,-1372389460,957052659,-302424857,-673190606,-97524203,-74126602,-343980068,-572739602,-546391182,-1451864710,-1086498196,-1284225,-320053250,-236490379,285044703,-281108946,-1691755485,-1192300817,-805854973,-890255689,2125331964,-844120560,-1250460258,-429410353,-113335595,-9569862,-776353793,-906003394,1341652987,-1195410058,2067553863,-2022746769,1121721214,-272500999,-1140871975,-1371906615,-63314270,1597429488,-92213762,-8553258,-546838604,-1600915041,-125856278,-1283622409,-109921882,-650127875,-816866470,-1678246082,-137596689,-134974409,-1068812,-1443039269,-1762270662,-291512846,1537932196,129416619,-14280105,-973427876,-1744856382,-1652161,-30231317,385834830,1902237683,2074007511,-1168115789,-168941921,-297274371,1602160557,1022835711,1729821657,1451228821,-310610465,-812647779,1119607926,-180645889,364973471,-126747785,-101877607,-1111670932,1865620959,-1343375941,2130553471,-182571076,945810822,-769632769,-152372282,969287525,1332649151,-1647708887,-732207185,1610345978,1342158695,1408628159,-2117040962,-1245792257,1002236504,978709423,1249431254,-1614889247,-1151778,-259444423,-172180313,-203459477,2113526222,-1633880209,1862249375,-557903043,-1648497251,1811650358,-37817290,-843959318,-1260391425,-1371675193,1061840255,-38608919,-1083265059,-273719809,1033352380,-1331542529,-1131121713,-726633481,-1074817066,1162261573,-8824945,-1622193435,354151927,1342159997,-1315971329,1065937404,-1377970582,-1134659873,-536876065,939363327,-77785367,-6592621,400749784,1466630109,1897444240,2086244686,1606666763,-474445733,-1229103700,1878716268,-487744052,805037394,-606810332,1869462246,1776023359,-554883334,-7407429,-1746014277,638157271,-352338177,-1484922980,266469243,1528552702,-1447055910,-176163621,-840970650,-1105528962,-1623836886,-50858039,-1340129331,-1184235566,-37998425,1609358184,-1141049482,-1471155015,-1898007558,993918903,65910655,-626099281,-1790660769,-1476526189,-1426589491,-55061591,2097000054,-19444342,-105393993,-1077942937,-151061553,2113816314,-565250066,1047985534,-161264132,-675285899,-2294881,-33732355,-1749681217,-1284776983,-1216903747,1401936399,-648360978,-1278263762,-338318597,-407513976,-1640600463,1609373407,920891388,-40653910,978237396,-436488193,520071006,-1083244545,-1075597616,-752927491,1808383731,-1497366746,-1025349645,1048541854,-671359026,835669611,1592360954,2074984414,-1678253031,1563328220,1458237934,1538750399,-4330753,-4767809,-1633978546,-124283938,-1466962190,1056403370,-346144817,-58929224,1643047851,-1314458629,-276135940,-1612216628,-18891441,-299901993,-1948901930,358002138,1741551565,-1435451402,2000940406,-1665470505,-1174675587,-220655189,-608323649,-1134581945,-287946337,-142871089,-1011101820,-25982229,2044198907,-181012074,-941665543,2140702691,-1636565579,-1094874330,-548569115,-1234857777,1341882356,-157032813,-1445070364,1539305211,2079046638,2128981339,2012086205,1586479095,1740635198,-2683073,2079182586,-505305113,-556106657,737787319,-36448011,-537002737,716946990,-1795575957,1856339826,1503641325,2140601340,-1423235845,-270485947,1069226384,-1511141825,802150599,-1094856735,250278893,-1887900195,1534885351,767600109,1994184761,-36727211,-780414037,2103608411,-1124520403,-46158145,1287493631,-93114753,2080337919,-1493205506,-764680529,1605347815,-181174467,-2030354867,-1178610311,1472852655,1257847779,360537251,794688379,-551595157,376258295,2126380771,-1234306049,-1157906449,-150995969,-814498581,1466293497,1732075951,2112334143,-1411920649,587177982,-529334922,719058838,-1085096709,-1477017683,-671093034,2040199902,-19370723,-1156589736,2005302253,1945399985,-570431362,-478355719,-58886667,-1078263873,-854197458,-577236550,-493617205,-898389108,444262389,-640946970,1055251046,-67659849,-6571313,2039187255,-1167004046,-160440481,-235573377,1677439339,-1661010481,1985324795,-1409302549,-1627914244,930441135,-1229765127,1995275027,-272374818,-1151488385,-1219626076,-269635590,1723512829,-1746085041,-682178733,-214441985,123323375,1383442814,-1243636744,-1912867716,-39068267,401831083,1476125693,-1079083413,-1598029914,-741626348,-588465572,-604007919,-594678789,-944259185,1042519931,-41473,2138928603,-423929313,-587983748,-436208485,-623675823,-745540760,939442120,325643519,1850130429,-3172357,-824197267,-289161282,1403854205,2063312246,-1628555797,-817431669,2108570567,929557308,-1342308394,787965525,2004877277,-139284809,1020206630,1537323864,2023682006,2007496135,2008329203,2078011326,-1174735877,-1435797845,-17451042,-70419715,-420237313,1962343350,1740592294,-1192790562,1969813631,-1175718917,-378011729,-294441427,-1421479169,-151082008,2113273342,1979111902,-1209805848,-929145761,500669655,-1510414605,1776248951,2147346269,-109265970,-330269027,-1550373529,2146291360,1047756793,-68829512,1332404217,-331421005,2067761341,-471443593,1762131902,-136069200,-1144672421,-16830530,-1088760067,-188942383,1912265662,2142800737,-152785164,-205027668,-521300757,-424735241,-957101028,-1128752417,1594070495,-806558507,-1110655489,-137481076,1774059146,1534906061,804447466,2026750742,535753875,-540022791,2062516093,-142972993,2119553021,1588848093,-209338548,-856195874,939122659,-162205786,1107228662,393116909,1857660771,-1308934498,-298419458,-68813825,-545309458,1993979577,-1236008993,-487587853,1175846198,987709311,-805375028,792278973,-274833633,770177910,-1109427585,-1023943823,-655763983,-209980802,1876164287,-304136707,-1214263827,712474099,-1130101670,2010051503,964765397,-539377681,-541086286,-640952400,-403723779,-411215510,2062315327,-468730393,659324365,2120889819,-604932227,1575597111,-339477273,2146753351,2113829885,-273093285,-78647722,2076691704,-304136405,-374932497,1683912621,443522387,-108107778,-1635917827,461351935,-1351647457,-289336642,-1382687857,-1099438118,2071986129,1300393724,536522621,2029890818,-1164488,-268507267,-1489047569,2108619634,-1283723985,-1844644273,-422547491,-1107820546,936961509,1329441785,-1614056260,-825603,985018043,1973328982,1978603263,-1346389052,-1367573618,949485419,2147369895,-1088438969,-885057541,2011984959,-1349075333,-1929435681,938408303,-1751149635,-1128993847,-1125325941,1945033979,-653386785,1902720895,-678734958,-247578835,-671226444,750642129,1876934654,-152326915,-832742414,-1985009287,-203424391,2103685607,-338036149,-3330151,-814071989,-1141212418,1977351127,-151553594,1563021015,-331743883,-403664,-137833,-42495099,672378226,1868208755,-2130969662,1540587439,-370345330,-83945819,-671273399,925781109,-402879107,-1109983617,-306317316,1269792695,-1157700898,-1799407109,1033108947,1840420606,-55744802,-38019361,1610544563,-740328705,-336300273,-1024574744,-362038702,-650424406,-176432899,1132230137,-556926165,-176349269,-67211365,-836524819,-1105822730,-1565077533,-4214786,-1917064219,1593824191,-1663763210,-1703177218,803663093,-1082406957,-1143005189,-587772042,-1216677378,1939035071,1316151295,1974773542,-1120270529,-513296405,1998662527,-1919058119,-202020874,1541236559,-3419394,-142615042,-1951096658,-805374083,-145689,-21086347,-1090292745,-239108113,104201058,-2016578718,-154190121,192806910,-1444284577,1943797230,-1616474501,-369124114,1834866601,-1170884129,2097119055,-1115830305,-621806864,-1693490511,2130181374,-296042538,-537907717,-675495329,1499311066,1798887294,51612436,-1096376913,-2382854,-48375866,334314412,-1418597139,-4358425,686554619,-329521704,995091325,1810803580,-1120082234,-1616380243,1642781635,-16810510,1868157860,1742004081,-1374044170,-337660995,2146260350,1610050047,-392202844,1476196799,1999529887,-292001043,936492977,1853714363,2109046585,-52242614,-1342806044,-2113963790,777756407,1719585516,2079844325,-139265319,484439179,-2098573381,-539581618,-1142227972,2100369267,-805737109,2104725434,-1334576905,-1522737955,-67978006,456769278,-546875397,-46154242,-1218629234,-540198018,7110591,799006715,-1476723203,1461315030,-1746147570,-138168361,-681982013,2142500763,-2006810865,-1683165329,1542929919,2110337910,2144959783,-307926017]
-	]
-]
-//            var th = [1,1,1,1,1,0,-1,0,-1,-1,0,-1,0,-1,-1,1,0,1,0,-1,1,0,1,0,-1,0,1,-1,0,0,0,-1,1,1,1,-1,0,-1,0,1,1,1,0,-1,0,0,0,-1,0,0,1,1,-1,1,1,0,-1,0,0,1,0,-1,1,1,1,1,1,-1,-1,1,-1,1,1,1,0,0,0,1,0,1,0,1,0,0,1,-1,1,1,1,0,-1,-1,0,-1,0,-1,0,-1,-1,1,0,0,1,0,-1,0,0,-1,1,0,0,0,0,0,0,0,-1,-1,-1,0,-1,0,0,-1,1,0,1,-1,0,-1,-1,1,1,-1,1,-1,1,0,0,-1,1,0,-1,0,1,-1,0,1,-1,1,1,0,0,-1,-1,0,-1,1,1,0,-1,1,1,1,1,-1,1,0,-1,1,1,1,1,-1,1,0,0,1,1,0,-1,-1,0,1,-1,-1,1,0,1,-1,1,0,-1,1,-1,-1,0,-1,1,-1,-1,-1,1,1,1,1,0,-1,0,0,-1,-1,1,1,0,1,1,1,0,1,1,1,1,-1,1,-1,-1,1,1,-1,-1,1,0,-1,0,-1,0,-1,0,1,0,1,-1]
